% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_data.R
\name{add_dis_age}
\alias{add_dis_age}
\title{Easily add the disease-related age variable to a data frame}
\usage{
add_dis_age(data, t_init, id_var = "id", time_var = "age")
}
\arguments{
\item{data}{the original data frame}

\item{t_init}{A named vector containing the observed initiation or onset
time for each individual. The names, i.e. \code{names(t_init)}, should
specify the individual id.}

\item{id_var}{name of the id variable in \code{data}}

\item{time_var}{name of the time variable in \code{data}}
}
\value{
A data frame with one column added. The new column will
be called \code{dis_age}. For controls, its value will be \code{NaN}.
}
\description{
Creates the disease-related age covariate vector based on the
disease initiation times and adds it to the data frame
}
\seealso{
Other data utilities: 
\code{\link{add_factor}()},
\code{\link{get_teff_obs}()},
\code{\link{new_x}()},
\code{\link{pick_one_row_each}()},
\code{\link{select_factors_and}()},
\code{\link{split}()},
\code{\link{valid_data}}
}
\concept{data utilities}
