% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgp-assist.R
\name{add_disease_ages}
\alias{add_disease_ages}
\title{Create the disease-related age covariate vector based on the
disease initiation times and add it to the data frame}
\usage{
add_disease_ages(data, t_init, id_var = "id", time_var = "age")
}
\arguments{
\item{data}{the original data frame}

\item{t_init}{A named vector containing the observed initiation or onset
time for each individual. The names, i.e. \code{names(t_init)}, should
specify the individual id.}

\item{id_var}{name of the id variable in \code{data}}

\item{time_var}{name of the time variable in \code{data}}
}
\value{
A data frame with one column added. The new column will
be called \code{'diseaseAge'}. For controls, the value of diseaseAge
will be set to NaN.
}
\description{
Create the disease-related age covariate vector based on the
disease initiation times and add it to the data frame
}
