% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assist.R
\name{add_factor}
\alias{add_factor}
\title{Easily add a categorical covariate to a data frame}
\usage{
add_factor(data, x, id_var = "id")
}
\arguments{
\item{data}{the original data frame}

\item{x}{A named vector containing the category for each individual.
The names should specify the individual id.}

\item{id_var}{name of the id variable in \code{data}}
}
\value{
A data frame with one column added. The new column will
have same name as the variable passed as input \code{x}.
}
\description{
Easily add a categorical covariate to a data frame
}
\seealso{
Other user assist functions: 
\code{\link{add_dis_age}()},
\code{\link{adjusted_c_hat}()},
\code{\link{split}()}
}
\concept{user assist functions}
