% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-misc.R
\name{add_factor_crossing}
\alias{add_factor_crossing}
\title{Add a crossing of two factors to a data frame}
\usage{
add_factor_crossing(df, fac1, fac2, new_name)
}
\arguments{
\item{df}{a data frame}

\item{fac1}{name of first factor, must be found in \code{df}}

\item{fac2}{name of second factor, must be found in \code{df}}

\item{new_name}{name of the new factor}
}
\value{
a data frame
}
\description{
Add a crossing of two factors to a data frame
}
\seealso{
Other data utilities: 
\code{\link{data_types}()},
\code{\link{rep_df}()},
\code{\link{to_long_format}()}
}
\concept{data utilities}
