% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-plot.R
\name{add_layer}
\alias{add_layer}
\alias{add_c_layer_line}
\alias{add_c_layer_ribbon}
\alias{add_g_layer_ribbon}
\alias{add_g_layer_fit}
\alias{add_g_layer_signal}
\alias{add_g_layer_faceting}
\title{Add layers to a grouped or componentwise plot}
\usage{
add_c_layer_line(h, df, x_name, group_by, color_by, alpha)

add_c_layer_ribbon(h, df, x_name, group_by, color_by, alpha)

add_g_layer_ribbon(h, df, x_name, group_by, color, alpha)

add_g_layer_fit(h, df, x_name, group_by, color, alpha)

add_g_layer_signal(h, df, x_name, group_by, color)

add_g_layer_faceting(h, facet_by, nrow, ncol)
}
\arguments{
\item{h}{a ggplot object}

\item{df}{a data frame containing the things to add}

\item{x_name}{name of the x-axis variable}

\item{group_by}{name of the grouping factor}

\item{color_by}{name of the coloring factor}

\item{alpha}{opacity}

\item{color}{color}

\item{facet_by}{factor to use for faceting}

\item{nrow}{number of rows, an argument for
\code{\link[ggplot2]{facet_wrap}}}

\item{ncol}{number of columns, an argument for
\code{\link[ggplot2]{facet_wrap}}}
}
\value{
a modified ggplot object
}
\description{
Add layers to a grouped or componentwise plot
}
