% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_sum_arraylist}
\alias{add_sum_arraylist}
\title{Add the sum of all arrays in a list to the list}
\usage{
add_sum_arraylist(x)
}
\arguments{
\item{x}{a list of arrays of shape \code{n} x \code{m}, with
length \code{L}}
}
\value{
a list of arrays of shape \code{n} x \code{m},
with length \code{L + 1}
}
\description{
Add the sum of all arrays in a list to the list
}
