% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postproc-predict.R
\name{add_test_caseIDs}
\alias{add_test_caseIDs}
\title{Add case IDs to test data frame}
\usage{
add_test_caseIDs(X_test, X_data)
}
\arguments{
\item{X_test}{test data frame}

\item{X_data}{data frame}
}
\value{
Updated X_test data frame.
}
\description{
Add case IDs to test data frame
}
