% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_array.R
\name{add_to_columns}
\alias{add_to_columns}
\title{Add a vector to each column of an array}
\usage{
add_to_columns(x, v)
}
\arguments{
\item{x}{an array of shape \code{n} x \code{m}}

\item{v}{a vector of length \code{n}}
}
\value{
an array of shape \code{n} x \code{m}
}
\description{
Add a vector to each column of an array
}
\seealso{
Other array utilities: 
\code{\link{apply_reduce}()},
\code{\link{arr3_select}()},
\code{\link{arr3_to_list}()},
\code{\link{array_to_arraylist}()},
\code{\link{ensure_2dim}()},
\code{\link{make_draw_df}()},
\code{\link{normalize_rows}()},
\code{\link{null_if_all_nan}()},
\code{\link{reduce_rows}()},
\code{\link{repvec}()},
\code{\link{row_vars}()},
\code{\link{select_row}()},
\code{\link{squeeze_second_dim}()}
}
\concept{array utilities}
