% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-misc.R
\name{apply_reduce}
\alias{apply_reduce}
\title{Apply possible reduction to rows of an array}
\usage{
apply_reduce(x, reduce)
}
\arguments{
\item{x}{an array of shape \code{n} x \code{m}}

\item{reduce}{a function or \code{NULL}}
}
\value{
an array of shape \code{1} x \code{m} if \code{reduce} is not
\code{NULL}, otherwise the original array \code{x}
}
\description{
Apply possible reduction to rows of an array
}
\concept{array utilities}
