% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-create_model.R
\name{apply_scaling}
\alias{apply_scaling}
\title{Apply variable scaling}
\usage{
apply_scaling(scaling, x, inverse = FALSE)
}
\arguments{
\item{scaling}{an object of class \linkS4class{lgpscaling}}

\item{x}{object to which apply the scaling (numeric)}

\item{inverse}{whether scaling should be done in inverse direction}
}
\value{
a similar object as \code{x}
}
\description{
Apply variable scaling
}
\seealso{
Other variable scaling functions: 
\code{\link{create_scaling}()}
}
\concept{variable scaling functions}
