% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-approximate.R
\name{approximate_model}
\alias{approximate_model}
\title{Create an approximate GP model}
\usage{
approximate_model(model, bf_options = list(M_bf = 30, L_bf = 3))
}
\arguments{
\item{model}{An object of class \linkS4class{lgpmodel}.}

\item{bf_options}{TODO!}
}
\value{
An object of class \linkS4class{lgpmodel}.
}
\description{
Create an approximate GP model
}
\seealso{
Other main functions: 
\code{\link{create_model}()},
\code{\link{draw_pred}()},
\code{\link{get_draws}()},
\code{\link{lgp}()},
\code{\link{pred}()},
\code{\link{prior_pred}()},
\code{\link{sample_model}()}
}
\concept{main functions}
