% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-arguments.R
\name{argument_check}
\alias{argument_check}
\title{Check if argument is valid}
\usage{
argument_check(arg, allowed)
}
\arguments{
\item{arg}{The given argument. Cannot be a list.}

\item{allowed}{Allowed arguments. Must be have length at least 2.}
}
\value{
Return the index of \code{arg} in \code{allowed} or 
throws an error if argument is not valid.
}
\description{
Replacement for the \code{base} R function \code{match.arg}. Gives more informative errors and requires
exact match. Should only be used as a subroutine of other functions and never directly.
}
