% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-utils.R
\name{arr3_select}
\alias{arr3_select}
\title{Select an element along the first dimension of an array, and only drop
the first dimension}
\usage{
arr3_select(x, idx)
}
\arguments{
\item{x}{an array of shape\code{d1} x \code{d2} x \code{d3}}

\item{idx}{the index to select}
}
\value{
an array of shape \code{d2} x \code{d3}
}
\description{
Select an element along the first dimension of an array, and only drop
the first dimension
}
\seealso{
Other array utilities: 
\code{\link{add_to_columns}()},
\code{\link{apply_reduce}()},
\code{\link{arr3_to_list}()},
\code{\link{array_to_arraylist}()},
\code{\link{ensure_2dim}()},
\code{\link{make_draw_df}()},
\code{\link{normalize_rows}()},
\code{\link{null_if_all_nan}()},
\code{\link{reduce_rows}()},
\code{\link{repvec}()},
\code{\link{row_vars}()},
\code{\link{select_row}()},
\code{\link{squeeze_second_dim}()}
}
\concept{array utilities}
