% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-utils.R
\name{arr3_to_list}
\alias{arr3_to_list}
\title{Format a 3-dimensional array into a list of 2-dimensional arrays}
\usage{
arr3_to_list(x)
}
\arguments{
\item{x}{an array}
}
\value{
a list of arrays
}
\description{
Format a 3-dimensional array into a list of 2-dimensional arrays
}
\seealso{
Other array utilities: 
\code{\link{add_to_columns}()},
\code{\link{apply_reduce}()},
\code{\link{arr3_select}()},
\code{\link{array_to_arraylist}()},
\code{\link{ensure_2dim}()},
\code{\link{make_draw_df}()},
\code{\link{normalize_rows}()},
\code{\link{null_if_all_nan}()},
\code{\link{reduce_rows}()},
\code{\link{repvec}()},
\code{\link{row_vars}()},
\code{\link{select_row}()},
\code{\link{squeeze_second_dim}()}
}
\concept{array utilities}
