% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{array_to_arraylist}
\alias{array_to_arraylist}
\title{Array reshaping}
\usage{
array_to_arraylist(x, L, draws)
}
\arguments{
\item{x}{an array of shape \code{n} x \code{m}}

\item{L}{an integer}

\item{draws}{see the \code{draws} argument of \code{\link{get_posterior_f}}}
}
\value{
a list of length \code{L}, where each element is an array of
shape \code{h} x (\code{m}/\code{L}), where \code{h = length(draws)}
}
\description{
Array reshaping
}
\seealso{
Other array utilities: 
\code{\link{ensure_2dim}()},
\code{\link{null_if_all_nan}()},
\code{\link{reduce_rows}()},
\code{\link{repvec}()},
\code{\link{squeeze_second_dim}()}
}
\concept{array utilities}
