% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_s4_classes.R
\name{as.character,lgpexpr-method}
\alias{as.character,lgpexpr-method}
\title{Convert an \code{lgpexpr} to character}
\usage{
\S4method{as.character}{lgpexpr}(x)
}
\arguments{
\item{object}{an object of class \code{\link{lgpexpr}}}
}
\description{
Convert an \code{lgpexpr} to character
}
