% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_s4_classes.R
\name{as.character,lgpformula-method}
\alias{as.character,lgpformula-method}
\title{Cast \code{lgpformula} to character}
\usage{
\S4method{as.character}{lgpformula}(x)
}
\arguments{
\item{object}{an object of class \code{\link{lgpformula}}}
}
\description{
Cast \code{lgpformula} to character
}
