% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_s4_classes.R
\name{as.character,lgpmodel-method}
\alias{as.character,lgpmodel-method}
\title{Cast \code{lgpmodel} to character}
\usage{
\S4method{as.character}{lgpmodel}(x)
}
\arguments{
\item{object}{an object of class \code{\link{lgpmodel}}}
}
\description{
Cast \code{lgpmodel} to character
}
