% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_s4_classes.R
\name{as.character,lgpproduct-method}
\alias{as.character,lgpproduct-method}
\title{Convert an \code{lgpproduct} to character}
\usage{
\S4method{as.character}{lgpproduct}(x)
}
\arguments{
\item{object}{an object of class \code{lgpproduct}}
}
\description{
Convert an \code{lgpproduct} to character
}
