% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_s4_classes.R
\name{as.character,lgprhs-method}
\alias{as.character,lgprhs-method}
\title{Convert an \code{lgprhs} to character}
\usage{
\S4method{as.character}{lgprhs}(x)
}
\arguments{
\item{object}{an object of class \code{\link{lgprhs}}}
}
\description{
Convert an \code{lgprhs} to character
}
