% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_s4_classes.R
\name{as.character,lgpsum-method}
\alias{as.character,lgpsum-method}
\title{Convert an \code{lgpsum} to character}
\usage{
\S4method{as.character}{lgpsum}(x)
}
\arguments{
\item{object}{an object of class \code{lgpsum}}
}
\description{
Convert an \code{lgpsum} to character
}
