% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_s4_classes.R
\name{as.character,lgpterm-method}
\alias{as.character,lgpterm-method}
\title{Convert an \code{lgpterm} to character}
\usage{
\S4method{as.character}{lgpterm}(x)
}
\arguments{
\item{object}{an object of class \code{lgpterm}}
}
\description{
Convert an \code{lgpterm} to character
}
