% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4-as_character.R
\name{as_character}
\alias{as_character}
\alias{as.character,lgpexpr-method}
\alias{as.character,lgpterm-method}
\alias{as.character,lgprhs-method}
\alias{as.character,lgpformula-method}
\alias{as.character,lgpmodel-method}
\alias{as.character,lgpsim-method}
\title{Character representations of different S4 objects}
\usage{
\S4method{as.character}{lgpexpr}(x)

\S4method{as.character}{lgpterm}(x)

\S4method{as.character}{lgprhs}(x)

\S4method{as.character}{lgpformula}(x)

\S4method{as.character}{lgpmodel}(x)

\S4method{as.character}{lgpsim}(x)
}
\arguments{
\item{x}{an object of some S4 class}
}
\value{
a character representation of the object
}
\description{
Character representations of different S4 objects
}
