% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postproc-main.R
\name{assess_convergence}
\alias{assess_convergence}
\title{Assess convergence of the chains}
\usage{
assess_convergence(fit, skip_generated_quantities = TRUE)
}
\arguments{
\item{fit}{An (incomplete) object of class \code{lgpfit}.}

\item{skip_generated_quantities}{Should F_mean, F_var etc. be ignored}
}
\value{
A data frame with columns \code{c("Rhat", "Bulk_ESS", "Tail_ESS")}.
}
\description{
Assess convergence of the chains
}
