% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-utils.R
\name{average_predictions}
\alias{average_predictions}
\title{Average predictions over samples}
\usage{
average_predictions(LIST)
}
\arguments{
\item{LIST}{a list over samples}
}
\value{
a list
}
\description{
Average predictions over samples
}
