% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{call_fun}
\alias{call_fun}
\title{A linter-friendly way to call a function}
\usage{
call_fun(fun, arg)
}
\arguments{
\item{fun}{a functionn which takes one argument}

\item{arg}{the argument}
}
\value{
Literally just \code{fun(arg)}.
}
\description{
This exists just to allow calling a function which is a slot
of an S4 object, without some linter warnings.
}
