% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgp-formula.R
\name{categorical}
\alias{categorical}
\title{Create a categorical term}
\usage{
categorical(covariate)
}
\arguments{
\item{covariate}{name of a categorical covariate}
}
\value{
an object of class \code{\link{lgpterm}}
}
\description{
Create a categorical term
}
