% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preproc-helper.R
\name{check_data}
\alias{check_data}
\title{Validate the `data` input to \code{lgp} and resolve covariate types}
\usage{
check_data(data, varInfo, verbose)
}
\arguments{
\item{data}{the data frame that was passed to \code{lgp}}

\item{varInfo}{variable type info}

\item{verbose}{can this print some info?}
}
\value{
a list
}
\description{
Validate the `data` input to \code{lgp} and resolve covariate types
}
