% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_dimension_list}
\alias{check_dimension_list}
\title{Check that all listed dimensions are equal}
\usage{
check_dimension_list(dims)
}
\arguments{
\item{dims}{a list of dimensions}
}
\value{
list of errors
}
\description{
Check that all listed dimensions are equal
}
