% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-argcheck.R
\name{check_false}
\alias{check_false}
\title{Check that argument is false or zero}
\usage{
check_false(arg)
}
\arguments{
\item{arg}{the argument}
}
\value{
Returns \code{TRUE} if the check passes.
}
\description{
Check that argument is false or zero
}
