% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preproc-helper.R
\name{check_formula}
\alias{check_formula}
\title{Validate the formula of \code{lgp}}
\usage{
check_formula(formula, data)
}
\arguments{
\item{formula}{the formula that was passed to \code{lgp_model}}

\item{data}{the data frame that was passed to \code{lgp_model}}
}
\value{
nothing
}
\description{
Checks if the input `formula` to \code{lgp_model} 
are valid with the given data
}
