% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preproc-parse_prior.R
\name{check_hyperparameter_names}
\alias{check_hyperparameter_names}
\title{An error message for wrong hyperparameter naming}
\usage{
check_hyperparameter_names(dist, correct)
}
\arguments{
\item{dist}{the distribution}

\item{correct}{the allowed hyperparameter names}
}
\value{
nothing
}
\description{
An error message for wrong hyperparameter naming
}
