% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-misc.R
\name{check_in_data}
\alias{check_in_data}
\title{Check that data contains a variable with a certain name}
\usage{
check_in_data(var_name, data)
}
\arguments{
\item{var_name}{the variable to be searched for}

\item{data}{an object of class \code{data.frame}}
}
\value{
\code{TRUE} if the variable is found
}
\description{
Check that data contains a variable with a certain name
}
