% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-argcheck.R
\name{check_length}
\alias{check_length}
\title{Check if argument has correct length.}
\usage{
check_length(object, expected)
}
\arguments{
\item{object}{Any object.}

\item{expected}{Expected length.}
}
\value{
Returns \code{TRUE} if the object has an allowed length.
}
\description{
Check if argument has correct length.
}
