% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-argcheck.R
\name{check_lengths}
\alias{check_lengths}
\title{Check if arguments have the same length}
\usage{
check_lengths(a, b)
}
\arguments{
\item{a}{Any object.}

\item{b}{Any object.}
}
\value{
Returns \code{TRUE} if the arguments have equal length.
}
\description{
Check if arguments have the same length
}
