% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_s4_classes.R
\name{check_lgpexpr}
\alias{check_lgpexpr}
\title{Check validity of an lgpexpr object}
\usage{
check_lgpexpr(object)
}
\arguments{
\item{object}{an object of class \code{\link{lgpexpr}}}
}
\value{
\code{TRUE} if valid, otherwise reasons for invalidity
}
\description{
Check validity of an lgpexpr object
}
