% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_s4_classes.R
\name{check_lgpformula}
\alias{check_lgpformula}
\title{Check validity of an lgpformula object}
\usage{
check_lgpformula(object)
}
\arguments{
\item{object}{an object of class \code{\link{lgpformula}}}
}
\value{
\code{TRUE} if valid, otherwise reasons for invalidity
}
\description{
Check validity of an lgpformula object
}
