% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_s4_classes.R
\name{check_lgpscaling}
\alias{check_lgpscaling}
\title{Check validitity of an lgpscaling object}
\usage{
check_lgpscaling(object)
}
\arguments{
\item{object}{an object of class \code{\link{lgpscaling}}}
}
\value{
\code{TRUE} if valid, otherwise reasons for invalidity
}
\description{
Check validitity of an lgpscaling object
}
