% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-argcheck.R
\name{check_not_null}
\alias{check_not_null}
\title{Check that argument is not null}
\usage{
check_not_null(arg)
}
\arguments{
\item{arg}{Any object.}
}
\value{
Returns \code{TRUE} if the object is not null.
}
\description{
Check that argument is not null
}
