% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-argcheck.R
\name{check_numeric}
\alias{check_numeric}
\title{Check if a value is numeric and throw error if not}
\usage{
check_numeric(arg, arg_name = NULL, require_positive = FALSE)
}
\arguments{
\item{arg}{A single number}

\item{arg_name}{Argument name}

\item{require_positive}{Should this also check that \code{arg} is positive?}
}
\description{
Check if a value is numeric and throw error if not
}
