% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-data.R
\name{check_term_covariates}
\alias{check_term_covariates}
\title{Helper for converting an lgpterm to numeric representation for Stan}
\usage{
check_term_covariates(covariates, pf)
}
\arguments{
\item{covariates}{a list returned by \code{\link{stan_data_covariates}}}

\item{pf}{a list returned by \code{\link{check_term_factors}}}
}
\value{
two integers
}
\description{
Helper for converting an lgpterm to numeric representation for Stan
}
