% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{check_term_factors}
\alias{check_term_factors}
\title{Helper for converting an lgpterm to numeric representation for Stan}
\usage{
check_term_factors(term)
}
\arguments{
\item{term}{an object of class \linkS4class{lgpterm}}
}
\value{
a named list
}
\description{
Helper for converting an lgpterm to numeric representation for Stan
}
