% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-argcheck.R
\name{check_type}
\alias{check_type}
\title{Check if argument has correct type.}
\usage{
check_type(object, allowed)
}
\arguments{
\item{object}{Any object.}

\item{allowed}{Allowed class names.}
}
\value{
Returns \code{TRUE} if the object has an allowed type.
}
\description{
Check if argument has correct type.
}
