% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preproc-helper.R
\name{check_varInfo}
\alias{check_varInfo}
\title{Check that variable types make sense}
\usage{
check_varInfo(varInfo)
}
\arguments{
\item{varInfo}{a named list}
}
\value{
nothing
}
\description{
Check that variable types make sense
}
