% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-prior.R
\name{check_warp_prior}
\alias{check_warp_prior}
\title{Visual checking of what kind of input warping functions a given prior
allows}
\usage{
check_warp_prior(prior, L = 24, ...)
}
\arguments{
\item{prior}{A valid \code{prior} argument to \code{\link{lgpmodel}}.}

\item{L}{time scale}

\item{...}{other arguments to \code{\link{plot_invgamma}}}
}
\value{
a ggplot object
}
\description{
Visual checking of what kind of input warping functions a given prior
allows
}
