% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-argcheck.R
\name{checks}
\alias{checks}
\alias{check_type}
\alias{check_function}
\alias{check_numeric}
\alias{check_positive}
\alias{check_non_negative}
\alias{check_positive_all}
\alias{check_non_negative_all}
\alias{check_integer_all}
\alias{check_interval}
\alias{check_not_null}
\alias{check_null}
\alias{check_length}
\alias{check_length_1_or}
\alias{check_length_geq}
\alias{check_lengths}
\alias{check_false}
\alias{check_in_data}
\alias{check_df_with}
\alias{check_all_leq}
\alias{check_not_named}
\alias{check_named}
\alias{check_dim}
\alias{check_unique}
\title{Argument validation}
\usage{
check_type(arg, allowed)

check_function(arg, arg_name)

check_numeric(arg)

check_positive(arg)

check_non_negative(arg)

check_positive_all(arg)

check_non_negative_all(arg)

check_integer_all(arg)

check_interval(arg, lower, upper)

check_not_null(arg)

check_null(arg, reason = NULL)

check_length(arg, expected)

check_length_1_or(arg, expected)

check_length_geq(arg, min_length)

check_lengths(a, b)

check_false(arg)

check_in_data(var_name, data, arg_name)

check_df_with(arg, var_names)

check_all_leq(arg, maximums)

check_not_named(arg)

check_named(arg)

check_dim(arg, D)

check_unique(arg)
}
\arguments{
\item{arg}{An argument to check.}

\item{allowed}{Allowed class names.}

\item{arg_name}{name of the data frame}

\item{lower}{interval minimum}

\item{upper}{interval maximum}

\item{reason}{explanation why the argument should be NULL}

\item{expected}{Expected length.}

\item{min_length}{Minimum allowed length.}

\item{a}{An argument to check.}

\item{b}{An argument to check.}

\item{var_name}{the variable to be searched for}

\item{data}{a data frame}

\item{var_names}{required variable names in a data frame}

\item{maximums}{maximum allowed values for \code{arg}}

\item{D}{expected number of dimensions}
}
\value{
\code{TRUE} if the check passes, throw error otherwise.
}
\description{
Argument validation
}
\section{Functions}{
\itemize{
\item \code{check_type}: checks that argument has expected type

\item \code{check_function}: checks that argument is a function

\item \code{check_numeric}: checks that argument is numeric

\item \code{check_positive}: checks that argument is positive

\item \code{check_non_negative}: checks that argument is non-negative

\item \code{check_positive_all}: checks that argument has only positive values

\item \code{check_non_negative_all}: checks that argument has only non-negative values

\item \code{check_integer_all}: checks that argument has only integer values

\item \code{check_interval}: checks that value is in some interval (strictly)

\item \code{check_not_null}: checks that argument is not NULL

\item \code{check_null}: checks that argument is NULL

\item \code{check_length}: checks that argument has expected length

\item \code{check_length_1_or}: checks that argument has length 1 or required other value

\item \code{check_length_geq}: checks that argument has at least some length

\item \code{check_lengths}: checks that arguments have same length

\item \code{check_false}: checks that argument is FALSE

\item \code{check_in_data}: checks that data frame contains a variable

\item \code{check_df_with}: checks that argument is a \code{data.frame} and contains
required variables

\item \code{check_all_leq}: checks that argument has values less than equal to
given maximums, elementwise

\item \code{check_not_named}: checks that object has no names

\item \code{check_named}: checks that object has names

\item \code{check_dim}: checks that argument has expected number of dimensions

\item \code{check_unique}: checks that vector has only unique values
}}

\seealso{
Other argument checks: 
\code{\link{check_allowed}()}
}
\concept{argument checks}
