% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-utils.R
\name{choose_stan_model}
\alias{choose_stan_model}
\title{Choose the correct stan model based on given options}
\usage{
choose_stan_model(info)
}
\arguments{
\item{info}{model info list}
}
\value{
an object of class \code{stanmodel}
}
\description{
Choose the correct stan model based on given options
}
