% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-utils.R
\name{class_info}
\alias{class_info}
\title{Class info for show methods}
\usage{
class_info(class_name)
}
\arguments{
\item{class_name}{class name}
}
\value{
a string
}
\description{
Class info for show methods
}
