% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-misc.R
\name{class_info_pred}
\alias{class_info_pred}
\title{Class info for show methods of pred objects}
\usage{
class_info_pred(class_name, num_comps, D)
}
\arguments{
\item{class_name}{class name}

\item{num_comps}{number of components}

\item{D}{dimensions}
}
\value{
a string
}
\description{
Class info for show methods of pred objects
}
