% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-plot.R
\name{color_palette}
\alias{color_palette}
\alias{fill_palette}
\title{A color palettes}
\usage{
color_palette(n)

fill_palette(n)
}
\arguments{
\item{n}{an integer from 1 to 5}
}
\value{
an array of \code{n} hex values
}
\description{
A color palettes
}
\seealso{
Other color utilities: 
\code{\link{colorset}()},
\code{\link{plot_color_palette}()},
\code{\link{scale_color}()}
}
\concept{color utilities}
