% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-colors.R
\name{color_set}
\alias{color_set}
\title{Package colors}
\usage{
color_set(name)
}
\arguments{
\item{name}{Color name.}
}
\value{
A hex value of the color.
}
\description{
Here is defined a unified set of colors to be used throughout the package.
Plot commands which specify the \code{col} argument should use a
color that is specified here.
}
