% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-colors.R
\name{colorset}
\alias{colorset}
\title{Plot colors to use}
\usage{
colorset(main, sub = "mid")
}
\arguments{
\item{main}{Color name. Must be a valid \code{bayesplot} color name.}

\item{variant}{Must be one of {"light", "light_highlight", "mid",
"mid_highlight", "dark", "dark_highlight"}.}
}
\value{
A hex value of the color.
}
\description{
Plot commands which specify the \code{col} argument should use a
color that is specified here.
}
