% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-components.R
\name{component_index_to_covariate_index}
\alias{component_index_to_covariate_index}
\title{Component index to covariate index}
\usage{
component_index_to_covariate_index(D, idx)
}
\arguments{
\item{D}{integer vector of length 6}

\item{idx}{integer}
}
\value{
an integer
}
\description{
Component index to covariate index
}
