% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-components.R
\name{component_index_to_type}
\alias{component_index_to_type}
\title{Component index to component type}
\usage{
component_index_to_type(D, idx)
}
\arguments{
\item{D}{integer vector of length 6}

\item{idx}{integer}
}
\value{
an integer
}
\description{
Component index to component type
}
