% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-lgpmodel.R
\name{component_info}
\alias{component_info}
\alias{component_names}
\title{Information about the model components}
\usage{
component_info(object)

component_names(object)
}
\arguments{
\item{object}{an object of class \linkS4class{lgpmodel} or
\linkS4class{lgpfit}}
}
\value{
a list or a string
}
\description{
Information about the model components
}
