% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior.R
\name{compute_gp_posteriors}
\alias{compute_gp_posteriors}
\title{Compute GP posteriors}
\usage{
compute_gp_posteriors(kernels, y, delta, sigma, STREAM)
}
\arguments{
\item{kernels}{a list with length \code{num_draws}}

\item{y}{response variable vector of length \code{num_obs}}

\item{delta}{jitter to ensure positive definite matrices}

\item{sigma}{a vector with length \code{num_draws}}

\item{STREAM}{external pointer}
}
\value{
A list.
}
\description{
Compute GP posteriors
}
\seealso{
Other GP posterior computation functions: 
\code{\link{compute_kernel_matrices}()},
\code{\link{format_X_pred}()},
\code{\link{get_draws_kernel_pars}()},
\code{\link{posterior_predict}()}
}
\concept{GP posterior computation functions}
