% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-kernels.R
\name{compute_K_beta}
\alias{compute_K_beta}
\title{Compute the multiplier matrix K_beta (to eneable heterogeneous disease effect)}
\usage{
compute_K_beta(beta, row_to_caseID_1, row_to_caseID_2)
}
\arguments{
\item{beta}{a row vector of length \code{N_cases}}

\item{row_to_caseID_1}{mapping from row index to case ID}

\item{row_to_caseID_2}{mapping from row index to case ID}
}
\value{
a matrix
}
\description{
Compute the multiplier matrix K_beta (to eneable heterogeneous disease effect)
}
