% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior.R
\name{compute_kernel_matrices}
\alias{compute_kernel_matrices}
\title{Compute kernel matrices}
\usage{
compute_kernel_matrices(model, pred, theta, STREAM)
}
\arguments{
\item{model}{An object of class \linkS4class{lgpmodel}.}

\item{pred}{object returned by \code{\link{format_X_pred}}}

\item{theta}{a list with length \code{num_draws}}

\item{STREAM}{external pointer}
}
\value{
A list.
}
\description{
Compute kernel matrices
}
\seealso{
Other GP posterior computation functions: 
\code{\link{compute_gp_posteriors}()},
\code{\link{format_X_pred}()},
\code{\link{get_draws_kernel_pars}()},
\code{\link{posterior_predict}()}
}
\concept{GP posterior computation functions}
