% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior.R
\name{compute_kernel_matrices_const}
\alias{compute_kernel_matrices_const}
\title{Compute constant kernel matrices}
\usage{
compute_kernel_matrices_const(model, x1, x2, STREAM = get_stream())
}
\arguments{
\item{model}{An object of class \linkS4class{lgpmodel}.}

\item{x1}{A list defining the first input argument. Can be
created by \code{\link{create_kernel_input}}.}

\item{x2}{A list defining the second input argument. Can be
created by \code{\link{create_kernel_input}}.}

\item{STREAM}{External pointer. By default obtained with
\code{\link[rstan]{get_stream}}.}
}
\value{
A list of kernel matrices with length equal to number of components.
}
\description{
Compute constant kernel matrices
}
\seealso{
Other GP posterior computation functions: 
\code{\link{compute_gp_posteriors}()},
\code{\link{compute_kernel_matrices}()},
\code{\link{create_kernel_input}()},
\code{\link{get_draws_kernel_pars}()},
\code{\link{posterior_predict_kernels}()},
\code{\link{posterior_predict}()}
}
\concept{GP posterior computation functions}
