% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict-main.R
\name{compute_lppd}
\alias{compute_lppd}
\title{Compute log-posterior predictive density at test points}
\usage{
compute_lppd(PRED, test_data)
}
\arguments{
\item{PRED}{predictions}

\item{test_data}{test data}
}
\value{
A data frame.
}
\description{
Compute log-posterior predictive density at test points
}
