% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postproc-relevances.R
\name{compute_noise_level}
\alias{compute_noise_level}
\title{Determine noise level}
\usage{
compute_noise_level(pars, model, noise_method)
}
\arguments{
\item{pars}{A data frame representing one parameter sample,
i.e one row of \code{as.data.frame(stanfit)}, where stanfit
is an object of class \code{stanfit}}

\item{model}{An object of class \code{lgpmodel}}

\item{noise_method}{Noise level estimation method.}
}
\value{
a value between 0 and 1
}
\description{
Determine noise level
}
