% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict-main.R
\name{compute_predicted_components}
\alias{compute_predicted_components}
\title{Compute component-wise predictions at test points}
\usage{
compute_predicted_components(KK, KKs, KKss, y_data, sigma_n, DELTA)
}
\arguments{
\item{KK}{Kernel matrices data vs. data.}

\item{KKs}{Kernel matrices test vs. data.}

\item{KKss}{Kernel matrices test vs. test.}

\item{y_data}{Response variable.}

\item{sigma_n}{Noise standard deviation parameter.}

\item{DELTA}{Diagonal jitter that ensures pos. def. kernel.}
}
\value{
A list containing predicted means and variances.
}
\description{
Used by \code{\link{compute_predictions}}.
}
