% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postproc-predict.R
\name{compute_predictions}
\alias{compute_predictions}
\title{Compute component-wise predictions at test points}
\usage{
compute_predictions(X_data, y_data, X_test, params, D, info, cnames, TSCL,
  handle_extra = "warning")
}
\arguments{
\item{X_data}{Covariate matrix (data points).}

\item{y_data}{Response variable (data points).}

\item{X_test}{Covariate matrix (test points).}

\item{params}{Kernel function and other hyperparameters}

\item{D}{a vector of length 6}

\item{info}{other model info}

\item{cnames}{Names of the model components.}

\item{TSCL}{time scaling function and its inverse}

\item{handle_extra}{What to do if test data contains individuals that are
not in the training data? Must be 'silent', 'warning' or 'error'.}
}
\value{
A list.
}
\description{
Used by \code{\link{lgp_predict}}.
}
