% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postproc-relevances.R
\name{compute_relevances}
\alias{compute_relevances}
\title{Compute component relevances and estimate amount of noise
(one MCMC sample)}
\usage{
compute_relevances(pars, model, method, noise_method)
}
\arguments{
\item{pars}{A data frame representing one parameter sample,
i.e one row of \code{as.data.frame(stanfit)}, where stanfit
is an object of class \code{stanfit}}

\item{model}{An object of class \code{lgpmodel}}

\item{method}{Relevance determination method. Must be either
\code{"f_mean"} or \code{"alpha"}.}

\item{noise_method}{Noise level estimation method.}
}
\value{
a matrix of size \code{1} x \code{n_components + 1}
}
\description{
Compute component relevances and estimate amount of noise
(one MCMC sample)
}
