% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postproc-relevances.R
\name{compute_relevances_alpha}
\alias{compute_relevances_alpha}
\title{The alpha relevance determination method}
\usage{
compute_relevances_alpha(pars, model)
}
\arguments{
\item{pars}{A data frame representing one parameter sample,
i.e one row of \code{as.data.frame(stanfit)}, where stanfit
is an object of class \code{stanfit}}

\item{model}{An object of class \code{lgpmodel}}
}
\value{
a vector of length \code{n_components}
}
\description{
The alpha relevance determination method
}
