% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postproc-relevances.R
\name{compute_relevances_fmean}
\alias{compute_relevances_fmean}
\title{The f_mean relevance determination method}
\usage{
compute_relevances_fmean(pars, model)
}
\arguments{
\item{pars}{A data frame representing one parameter sample,
i.e one row of \code{as.data.frame(stanfit)}, where stanfit
is an object of class \code{stanfit}}

\item{model}{An object of class \code{lgpmodel}}
}
\value{
a matrix of size \code{1} x \code{n_components + 1}
}
\description{
The f_mean relevance determination method
}
