% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compute_teff_obs}
\alias{compute_teff_obs}
\title{Compute observed effect times using other variables}
\usage{
compute_teff_obs(id, age, dis_age)
}
\arguments{
\item{id}{id variable, a factor}

\item{age}{age variable on original scale, a vector}

\item{dis_age}{disease-related age variable on original scale,
a vector}
}
\value{
a named vector with length equal to number of levels in id
}
\description{
Compute observed effect times using other variables
}
