% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethodsModel.R
\name{const_kernels}
\alias{const_kernels}
\alias{const_kernels.decompose}
\alias{const_kernels.decompositions}
\alias{const_kernels.reconstruct}
\title{Compute constant kernel matrices which don't depend on parameters}
\usage{
const_kernels(object, STREAM = get_stream())

const_kernels.decompose(object, STREAM = get_stream())

const_kernels.decompositions(object, STREAM = get_stream())

const_kernels.reconstruct(decompositions, STREAM = get_stream())
}
\arguments{
\item{object}{an object of class \linkS4class{lgpmodel} or
\linkS4class{lgpfit}}

\item{STREAM}{external pointer}

\item{decompositions}{a list returned by \code{const_kernel.decompositions}}
}
\value{
a list of matrices
}
\description{
\itemize{
   \item \code{const_kernels} computes all constant kernels
   \item \code{const_kernels.decompositions} computes their exact
   rank decompositions
}
}
