% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethodsModel.R
\name{covariate_info}
\alias{covariate_info}
\alias{covariate_info.cat}
\alias{covariate_info.cont}
\title{Information about covariates used in a model}
\usage{
covariate_info(object)

covariate_info.cat(object)

covariate_info.cont(object)
}
\arguments{
\item{object}{an object of class \linkS4class{lgpmodel} or
\linkS4class{lgpfit}}
}
\value{
a list or a string
}
\description{
Information about covariates used in a model
}
