% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stanfunctions.R
\name{cpp_bfa_lambda}
\alias{cpp_bfa_lambda}
\title{The m'th eigenvalue of the Dirichlet boundary value problem}
\usage{
cpp_bfa_lambda(m, L, STREAM = get_stream())
}
\arguments{
\item{m}{index of the eigenfunction (non-negative integer)}

\item{L}{domain width (positive real number)}

\item{STREAM}{an external pointer}
}
\value{
a number
}
\description{
This is a wrapper for \code{STAN_bfa_lambda}, with some input
validation added.
}
\seealso{
Other basisfunction approximation functions: 
\code{\link{cpp_bfa_phi}()},
\code{\link{cpp_spd_eq}()}
}
\concept{basisfunction approximation functions}
