% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stanfunctions.R
\name{cpp_gp_posterior}
\alias{cpp_gp_posterior}
\title{Compute all componentwise GP posteriors of a model}
\usage{
cpp_gp_posterior(KX, KX_s, KX_ss, y, delta, sigma, STREAM = get_stream())
}
\arguments{
\item{KX}{list of kernel matrices fo size \code{num_obs} x \code{num_obs}}

\item{KX_s}{list of kernel matrices fo size \code{num_out} x \code{num_obs}}

\item{KX_ss}{list of kernel matrices fo size \code{num_out} x \code{num_out}}

\item{y}{response vector}

\item{delta}{jitter}

\item{sigma}{Gaussian noise magnitude parameter}

\item{STREAM}{an external pointer}
}
\value{
an array of kernel matrices
}
\description{
This is a wrapper for \code{STAN_gp_posterior}
}
