% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-stan_wrappers.R
\name{cpp_kernel_all}
\alias{cpp_kernel_all}
\title{Compute all kernel matrices of a model}
\usage{
cpp_kernel_all(
  n1,
  n2,
  K_const,
  components,
  x1,
  x2,
  x1_unnorm,
  x2_unnorm,
  alpha,
  ell,
  wrp,
  beta,
  teff,
  vm_params,
  idx1_expand,
  idx2_expand,
  teff_zero,
  STREAM = get_stream()
)
}
\arguments{
\item{n1}{integer}

\item{n2}{integer}

\item{K_const}{list of constant matrices}

\item{components}{an integer array}

\item{x1}{continuous covariates}

\item{x2}{continuous covariates}

\item{x1_unnorm}{unnormalized versions of the continuous covariates}

\item{x2_unnorm}{unnormalized versions of the continuous covariates}

\item{alpha}{a vector of magnitude parameters}

\item{ell}{a vector of lengthscale parameters or NaN}

\item{wrp}{the input warping parameter or NaN}

\item{beta}{a vector of heterogeneity parameters or NaN}

\item{teff}{a vector of effect time parameters or NaN}

\item{vm_params}{variance mask parameters}

\item{idx1_expand}{expansion of beta and/or teff}

\item{idx2_expand}{expansion of beta and/or teff}

\item{teff_zero}{observed effect times}

\item{STREAM}{an external pointer}
}
\value{
an array of kernel matrices
}
\description{
This is a wrapper for \code{STAN_kernel_all}
}
