% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stanfunctions_bfa.R
\name{cpp_multi_normal_bfa_logpdf}
\alias{cpp_multi_normal_bfa_logpdf}
\title{Multivariate normal density using basis function approximation}
\usage{
cpp_multi_normal_bfa_logpdf(y, V, D_diag, sigma, STREAM = get_stream())
}
\arguments{
\item{y}{a vector of length \code{n}}

\item{V}{a matrix of size \code{n} x \code{R*M}}

\item{D_diag}{a vector of length \code{R*M}}

\item{sigma}{noise standard deviation, a positive real number}

\item{STREAM}{an external pointer}
}
\value{
a real number
}
\description{
This is a wrapper for \code{STAN_multi_normal_bfa_logpdf},
with some input validation added.
}
\seealso{
Other basis function approximation functions: 
\code{\link{cpp_lambda}()},
\code{\link{cpp_phi}()},
\code{\link{cpp_spd_eq}()}
}
\concept{basis function approximation functions}
