% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stanfunctions_bfa.R
\name{cpp_phi}
\alias{cpp_phi}
\title{The m'th eigenfunction of the Dirichlet boundary value problem}
\usage{
cpp_phi(x, m, L, STREAM = get_stream())
}
\arguments{
\item{x}{a numeric vector}

\item{m}{index of the eigenfunction (non-negative integer)}

\item{L}{domain width (positive real number)}

\item{STREAM}{an external pointer}
}
\value{
with length equal to that of \code{x}
}
\description{
This is a wrapper for \code{STAN_phi}, with some input
validation added.
}
\seealso{
Other basis function approximation functions: 
\code{\link{cpp_lambda}()},
\code{\link{cpp_multi_normal_bfa_logpdf}()},
\code{\link{cpp_spd_eq}()}
}
\concept{basis function approximation functions}
