% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stanfunctions_bfa.R
\name{cpp_spd_eq}
\alias{cpp_spd_eq}
\title{Spectral density function of the exponentiated quadratic kernel}
\usage{
cpp_spd_eq(w, ell, STREAM = get_stream())
}
\arguments{
\item{w}{the frequency, real number}

\item{ell}{lengthscale of the kernel}

\item{STREAM}{an external pointer}
}
\value{
a real number
}
\description{
This is a wrapper for \code{STAN_spd_eq}, with some input
validation added.
}
\seealso{
Other basis function approximation functions: 
\code{\link{cpp_lambda}()},
\code{\link{cpp_multi_normal_bfa_logpdf}()},
\code{\link{cpp_phi}()}
}
\concept{basis function approximation functions}
