% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stanfunctions.R
\name{cpp_var_mask}
\alias{cpp_var_mask}
\title{Variance masking function}
\usage{
cpp_var_mask(x, a, STREAM = get_stream())
}
\arguments{
\item{x}{a numeric vector}

\item{a}{positive steepness parameter}

\item{STREAM}{an external pointer}
}
\value{
a vector
}
\description{
This is a wrapper for \code{STAN_var_mask}, with some input
validation added.
}
