% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aa-preproc-old.R
\name{create_covariates_stan}
\alias{create_covariates_stan}
\title{Create the covariate matrix that is given to stan}
\usage{
create_covariates_stan(data, varInfo, types, formula, verbose)
}
\arguments{
\item{data}{the data frame that was passed to \code{lgp}}

\item{varInfo}{original variable type info}

\item{types}{the types returned by \code{\link{check_data}}}

\item{formula}{the model formula}

\item{verbose}{can this print some info?}
}
\value{
a list
}
\description{
Create the covariate matrix that is given to stan
}
