% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-helper.R
\name{create_data_plot_df}
\alias{create_data_plot_df}
\title{Create a plotting data frame for ggplot}
\usage{
create_data_plot_df(data, hl_1, hl_2, hl_cont)
}
\arguments{
\item{data}{a data frame}

\item{hl_1}{highlighting by color}

\item{hl_2}{highlighting by linestyle}

\item{hl_cont}{highlighting continuous}
}
\value{
an extended data frame
}
\description{
A helper function for \code{plot_data}.
}
