% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-utils.R
\name{create_example_fit}
\alias{create_example_fit}
\title{Create an example fit object}
\usage{
create_example_fit(N = 4, t = 10 * c(1, 2, 3, 4, 5), iter = 100, chains = 1)
}
\arguments{
\item{N}{number of individuals}

\item{t}{time points}

\item{iter}{number of iterations}

\item{chains}{number of chains}
}
\value{
an object of class \code{\link{lgpfit}}
}
\description{
Create an example fit object
}
