% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-utils.R
\name{create_grouping_factor}
\alias{create_grouping_factor}
\title{Create grouping factor that can be added to a data frame}
\usage{
create_grouping_factor(x, group_by)
}
\arguments{
\item{x}{A data frame of prediction points, containing at least the
variables used as covariates in the model. This must be specified if
\code{pred} is not \code{NULL}.}

\item{group_by}{name of the grouping variable (use \code{group_by=NA}
to avoid grouping)}
}
\value{
a factor
}
\description{
used e.g. by \code{\link{plot_pred}} and
\code{\link{new_x}}
}
