% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors-parse.R
\name{create_hyper_beta}
\alias{create_hyper_beta}
\title{Create the hyper_beta input to Stan}
\usage{
create_hyper_beta(prior, num_heter)
}
\arguments{
\item{prior}{A named list, defining the prior distribution of model
(hyper)parameters.}

\item{num_heter}{the \code{num_heter} input created for Stan}
}
\value{
an array of size (0, 2) or (1, 2)
}
\description{
Create the hyper_beta input to Stan
}
