% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-idx_expand.R
\name{create_idx_expand}
\alias{create_idx_expand}
\title{Create idx_expand input for Stan}
\usage{
create_idx_expand(components, x_cat, x_cont_mask)
}
\arguments{
\item{components}{the \code{components} input for Stan}

\item{x_cat}{the \code{x_cat} input for Stan}

\item{x_cont_mask}{the \code{x_cont_mask} input for Stan}

\item{return}{the \code{idx_expand} input for Stan}
}
\description{
Create idx_expand input for Stan
}
