% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-expand.R
\name{create_idx_expand_helper}
\alias{create_idx_expand_helper}
\title{Helper function for creating the idx_expand input for Stan}
\usage{
create_idx_expand_helper(x_cat, x_cont_mask)
}
\arguments{
\item{x_cat}{the \code{x_cat} input for Stan}

\item{x_cont_mask}{the \code{x_cont_mask} input for Stan}

\item{return}{a vector of length \code{n_obs}}
}
\description{
Helper function for creating the idx_expand input for Stan
}
