% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-expand.R
\name{create_idx_expand_picker}
\alias{create_idx_expand_picker}
\title{Helper function for creating the idx_expand input for Stan}
\usage{
create_idx_expand_picker(components, x_cat)
}
\arguments{
\item{components}{the \code{components} input for Stan}

\item{x_cat}{the \code{x_cat} input for Stan}
}
\value{
a vector of length \code{n_obs}
}
\description{
Helper function for creating the idx_expand input for Stan
}
