% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior.R
\name{create_kernel_input}
\alias{create_kernel_input}
\title{Kernel input information to Stan format}
\usage{
create_kernel_input(model, X_pred)
}
\arguments{
\item{model}{An object of class \linkS4class{lgpmodel}.}

\item{X_pred}{A data frame of points where predictions are computed.
The function \code{\link{new_x}} can help in creating it.}
}
\value{
A list.
}
\description{
Kernel input information to Stan format
}
\seealso{
Other GP posterior computation functions: 
\code{\link{compute_gp_posteriors}()},
\code{\link{compute_kernel_matrices_const}()},
\code{\link{compute_kernel_matrices}()},
\code{\link{get_draws_kernel_pars}()},
\code{\link{posterior_predict_kernels}()},
\code{\link{posterior_predict}()}
}
\concept{GP posterior computation functions}
