% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-create_model.R
\name{create_model.approx_options}
\alias{create_model.approx_options}
\title{Parse the given approximation options}
\usage{
create_model.approx_options(approx)
}
\arguments{
\item{approx}{A named list with the following possible fields:
\itemize{
  \item \code{num_bf} Number of basis functions (0 = no approximation).
  \item \code{scale_bf} Scale of the domain to be used in basis
  function approximation. Has no effect if \code{num_bf = 0}.
}
If \code{approx} is \code{NULL}, default options are used. The defaults
are equivalent to
\code{options = list(
  num_bf = 0,
  scale_bf = 1.5
)
}.}
}
\value{
a named list of parsed options
}
\description{
Parse the given approximation options
}
