% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-create_model-prior.R
\name{create_model.prior}
\alias{create_model.prior}
\title{Parse given prior}
\usage{
create_model.prior(prior, stan_input, verbose)
}
\arguments{
\item{prior}{A named list, defining the prior distribution of model
(hyper)parameters. See the "Defining priors" section below
(\code{\link{lgp}}).}

\item{stan_input}{a list of stan input fields}

\item{verbose}{Should some informative messages be printed?}
}
\value{
a named list of parsed options
}
\description{
Parse given prior
}
\seealso{
Other internal model creation functions: 
\code{\link{create_model.covs_and_comps}()},
\code{\link{create_model.formula}()},
\code{\link{create_model.likelihood}()}
}
\concept{internal model creation functions}
