% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethodsModel.R
\name{create_plot_df}
\alias{create_plot_df}
\title{Helper function for plots}
\usage{
create_plot_df(object, x = "age", group_by = "id")
}
\arguments{
\item{object}{an object of class \linkS4class{lgpmodel} or
\linkS4class{lgpfit}}

\item{x}{x-axis variable name or a vector to be used as x-axis values}

\item{group_by}{grouping variable name or a factor to be used for grouping
the plot into panels}
}
\value{
a data frame
}
\description{
Helper function for plots
}
