% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-posterior-helper.R
\name{create_predictions_plot_df2}
\alias{create_predictions_plot_df2}
\title{Create a plotting data frame for ggplot}
\usage{
create_predictions_plot_df2(model, PRED, scale_f = TRUE, mode, n_sds)
}
\arguments{
\item{model}{An object of class \code{lgpmodel}.}

\item{PRED}{Predictions computed using \code{lgp_predict}.}

\item{scale_f}{Should the predictions be scaled back to the original
data scale?}

\item{mode}{mode}

\item{n_sds}{number of standard deviations for the uncertainty band width}
}
\value{
a data frame
}
\description{
A helper function for \code{plot_predictions}.
}
