% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-data.R
\name{create_scaling}
\alias{create_scaling}
\title{Create the function that does a standardizing transform and its inverse}
\usage{
create_scaling(y)
}
\arguments{
\item{y}{response variable measurements}
}
\value{
an object of class \linkS4class{lgpscaling}
}
\description{
Create the function that does a standardizing transform and its inverse
}
