% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-simdata.R
\name{create_simdata_plot_df}
\alias{create_simdata_plot_df}
\title{Create a plotting data frame for ggplot}
\usage{
create_simdata_plot_df(simData)
}
\arguments{
\item{simData}{An object created using \code{simulate_data}.}
}
\value{
a data frame
}
\description{
A helper function for \code{plot_simdata_by_component}.
}
