% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preproc-main.R
\name{create_stan_input}
\alias{create_stan_input}
\title{Create input for Stan}
\usage{
create_stan_input(
  formula,
  data,
  prior,
  likelihood,
  varInfo,
  standardize,
  uncertain_effect_time,
  equal_effect,
  C_hat,
  DELTA,
  sample_F,
  verbose,
  variance_mask,
  N_trials,
  skip_gen_quant
)
}
\arguments{
\item{formula}{A formula specified using the common \code{\link{formula}}
syntax, such as \code{y ~ x1 + x2:z1 + } \code{x2:z2 + z2}.
\itemize{
  \item The formula must contain exatly one tilde (\code{~}), with response
  variable on the left-hand side and model terms on the right-hand side.
  \item Terms are be separated by a plus (\code{+}) sign.
  \item Terms can consits of a single variable name, such as \code{x}, or
  an interaction of two variables, such as \code{x:z}.
  \item In single-variable terms, the variable can be either continuous or
  categorical, whereas in interaction terms the variable
  on the left-hand side of the colon (\code{:}) has to be continuous and the
  one on the right-hand side has to be categorical (a factor).
  \item All variables appearing in \code{formula} must be
  found in \code{data}.
}}

\item{data}{A \code{data.frame} where each column corresponds to one
variable, and each row is one observation. Continuous covariates must have
type \code{"numeric"} and categorical ones must have type \code{"factor"}.
Missing values should be indicated with \code{NaN} or \code{NA}.}

\item{prior}{A named list, defining the prior distribution of model
(hyper)parameters. It is recommended to first create this using the function
\code{\link{prior_default}}, and then possibly modify it.}

\item{likelihood}{Determines the observation model. Must be either
\code{"gaussian"} (default), \code{"poisson"}, \code{"nb"} (negative
binomial) or \code{"binomial"}. To use Bernoulli likelihood, use
\code{likelihood="binomial"} and set \code{num_trials} as a vector of ones.}

\item{varInfo}{Variable type info.}

\item{standardize}{Should the response variable be standardized?}

\item{verbose}{Should more verbose output be printed?}
}
\value{
A list containing the data to be given to \code{rstan::sampling},
some info about preprocessing and all the information about scaling the
inputs and response, and updated variable type info.
}
\description{
Parses the \code{formula} and \code{data} input to
\code{\link{lgp_model}}. Also performs many input checks.
}
