% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict-main.R
\name{create_test_points}
\alias{create_test_points}
\title{Create a matrix of test points}
\usage{
create_test_points(object, t_test)
}
\arguments{
\item{object}{An object of class \code{lgpmodel} or \code{lgpfit}}

\item{t_test}{Test time points (will be same for each individual).}
}
\value{
A data frame.
}
\description{
Create a matrix of test points
}
