% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict-helper.R
\name{create_X_star}
\alias{create_X_star}
\title{Create X_star}
\usage{
create_X_star(X, D, t_test, SCL, X_notnan)
}
\arguments{
\item{X}{covariate matrix}

\item{D}{covariate type information}

\item{t_test}{Test time points (will be same for each individual).}

\item{SCL}{time scaling function and its inverse}

\item{X_notnan}{indicates where \code{X_diseaseAge} is not \code{NaN}}
}
\value{
A data frame.
}
\description{
Create X_star
}
