% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate-main.R
\name{create_y}
\alias{create_y}
\title{Generate noisy observations}
\usage{
create_y(noise_type, f, snr, phi)
}
\arguments{
\item{noise_type}{Either "Gaussian", "Poisson" or "NB" (negative binomial).}

\item{f}{The underlying signal.}

\item{snr}{The desired signal-to-noise ratio. This argument is valid
only with \cr
\code{noise_type = "Gaussian"}.}

\item{phi}{The dispersion parameter for negative binomial noise.}
}
\value{
A list \code{out}, where
\itemize{
  \item \code{out$g} is \code{f} mapped through an inverse link function and
  \item \code{out$y} is the noisy response variable.
}
}
\description{
Generate noisy observations
}
