% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethodsModel.R
\name{data_getters}
\alias{data_getters}
\alias{get_y}
\alias{get_x_cat}
\alias{get_x_cont}
\title{Functions that access original data stored in a model object}
\usage{
get_y(object, original = TRUE)

get_x_cat(object)

get_x_cont(object, original = TRUE, mask_with = NaN)
}
\arguments{
\item{object}{an object of class \linkS4class{lgpmodel} or
\linkS4class{lgpfit}}

\item{original}{should the variable be in its original form}

\item{mask_with}{value to use to mask the originally missing
values in the data}
}
\value{
\itemize{
  \item \code{get_y} gets response variable measurements
  (on their original scaled if \code{original} is \code{TRUE})
  \item \code{get_x_cat} gets the categorical covariate matrix
  \item \code{get_x_cont} gets the continuous covariate matrix (with
  variables on original scale and \code{NaN}s in their original locations
  if \code{original} is \code{TRUE})
}
}
\description{
Functions that access original data stored in a model object
}
\seealso{
Other lgpmodel accessors: 
\code{\link{model_getters}}
}
\concept{lgpmodel accessors}
