% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_data.R
\name{data_info}
\alias{data_info}
\alias{data_info.fac}
\alias{data_info.num}
\alias{data_info.var}
\title{Information a about data frame}
\usage{
data_info(data, digits = 2)

data_info.fac(data)

data_info.num(data, digits)

data_info.var(var, digits)
}
\arguments{
\item{data}{a data frame}

\item{digits}{number of digits to show for floats}

\item{var}{a numeric vector or a factor}
}
\value{
a named list
}
\description{
Information a about data frame
}
\seealso{
Other data utilities: 
\code{\link{add_dis_age}()},
\code{\link{add_factor_crossing}()},
\code{\link{add_factor}()},
\code{\link{data_types}()},
\code{\link{get_teff_obs}()},
\code{\link{new_x}()},
\code{\link{pick_one_row_each}()},
\code{\link{rep_df}()},
\code{\link{select_factors_and}()},
\code{\link{split}()},
\code{\link{to_long_format}()},
\code{\link{valid_data}}
}
\concept{data utilities}
