% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-misc.R
\name{data_types}
\alias{data_types}
\title{Get type of each data column}
\usage{
data_types(data)
}
\arguments{
\item{data}{a data frame}
}
\value{
a named list
}
\description{
Get type of each data column
}
\seealso{
Other data utilities: 
\code{\link{add_factor_crossing}()},
\code{\link{rep_df}()},
\code{\link{to_long_format}()}
}
\concept{data utilities}
