% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-utils.R
\name{default_vm_params}
\alias{default_vm_params}
\title{Default variance masking function parameters are defined here}
\usage{
default_vm_params()
}
\value{
two numbers
}
\description{
Default variance masking function parameters are defined here
}
