% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-inv_gamma.R
\name{dinvgamma_stanlike}
\alias{dinvgamma_stanlike}
\title{Inverse gamma distribution pdf with the same parametrization
as in Stan}
\usage{
dinvgamma_stanlike(x, alpha, beta, log = FALSE)
}
\arguments{
\item{x}{point where to compute the density}

\item{alpha}{positive real number}

\item{beta}{positive real number}

\item{log}{should logarithm of the density be returned?}
}
\value{
density or log-density at \code{x}
}
\description{
See
\href{https://mc-stan.org/docs/2_21/functions-reference/inverse-gamma-distribution.html}
{here}.
}
