% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate-create_F.R
\name{disease_peaks}
\alias{disease_peaks}
\title{Draw disease component from a parameteric form}
\usage{
disease_peaks(X_id, X_disAge)
}
\arguments{
\item{X_id}{the id covariate}

\item{X_disAge}{the diseaseAge covariate}
}
\value{
a vector
}
\description{
Draw disease component from a parameteric form
}
