% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-misc.R
\name{dollar}
\alias{dollar}
\title{A safer alternative for the dollar operator}
\usage{
dollar(object, var_name)
}
\arguments{
\item{object}{a list or data frame}

\item{var_name}{name of the variable to access}
}
\value{
Returns \code{object[[var_name, exact = TRUE]]} if variable
exists.
}
\description{
Requires exact match and throws error if \code{var_name} is
not in \code{names(object)}.
}
\seealso{
Other list utilities: 
\code{\link{list_if_named}()},
\code{\link{list_to_matrix}()},
\code{\link{matrix_to_list}()},
\code{\link{simplify_list}()}
}
\concept{list utilities}
