% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-draw_pred.R
\name{draw_pred}
\alias{draw_pred}
\title{Draw pseudo-observations from the predictive distribution}
\usage{
draw_pred(fit, pred = NULL)
}
\arguments{
\item{fit}{An object of class \linkS4class{lgpfit} that can been created
with \code{sample_f=TRUE}.}

\item{pred}{An object of class \linkS4class{Prediction} containing the draws
of \code{f}. If \code{NULL}, this is obtained using \code{get_pred(fit)}.}
}
\value{
An array with shape \eqn{S x P}, where \eqn{S} is the number of
parameter draws that \code{fit} contains, and \eqn{P} is the number of
points where the predictive distribution draws are obtained. Each row
\eqn{s = 1, \ldots, S} of the output is one vector drawn from the predictive
distribution, given parameter draw \eqn{s}. If \code{fit} was created using
\code{prior_only=TRUE}, then these are draws from the prior predictive
distribution, otherwise from the posterior predictive distribution.
}
\description{
This doesn't work for \linkS4class{GaussianPrediction} objects
because for memory reasons they don't store the
covariance matrices of the predictive distributions.
}
\seealso{
Other main functions: 
\code{\link{create_model}()},
\code{\link{get_draws}()},
\code{\link{lgp}()},
\code{\link{pred}()},
\code{\link{prior_pred}()},
\code{\link{sample_model}()}
}
\concept{main functions}
