% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate-create_F.R
\name{drawLatentComponents}
\alias{drawLatentComponents}
\title{Draw realizations of multivariate normals}
\usage{
drawLatentComponents(KK)
}
\arguments{
\item{KK}{3D matrix where \code{KK[,,j]} is the \code{j}th kernel matrix}
}
\value{
a matrix \code{FFF}
}
\description{
Draw realizations of multivariate normals
}
