% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-misc.R
\name{enclose_fun}
\alias{enclose_fun}
\title{Paste function and argument enclosed in parentheses}
\usage{
enclose_fun(s, fun)
}
\arguments{
\item{s}{argument name, a string}

\item{fun}{function name, a string}
}
\value{
a string
}
\description{
Paste function and argument enclosed in parentheses
}
\seealso{
Other function utilities: 
\code{\link{call_fun}()}
}
\concept{function utilities}
