% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-utils.R
\name{ensure_2dim}
\alias{ensure_2dim}
\title{Ensure that v is 2-dimensional}
\usage{
ensure_2dim(v)
}
\arguments{
\item{v}{a vector of length \code{m} or an array of size \code{n} x \code{m}}
}
\value{
returns a 2-dimensional array
}
\description{
Ensure that v is 2-dimensional
}
