% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ensure_len}
\alias{ensure_len}
\title{Ensure vector has expected length}
\usage{
ensure_len(v, len)
}
\arguments{
\item{v}{original vector or just one value that is replicated}

\item{len}{the expected length}
}
\value{
a vector of length \code{len}
}
\description{
Ensure vector has expected length
}
