% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgp-formula.R
\name{ensure_lgpsum}
\alias{ensure_lgpsum}
\title{Promote lgpterm or lgpproduct to lgpsum}
\usage{
ensure_lgpsum(x)
}
\arguments{
\item{x}{an object of class \code{\link{lgpterm}}, \code{\link{lgpproduct}}
or \code{\link{lgpsum}}}
}
\value{
an object of class \code{\link{lgpterm}}
}
\description{
Promote lgpterm or lgpproduct to lgpsum
}
