% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-utils.R
\name{example_call}
\alias{example_call}
\title{Returns a valid example call of the \code{lgp} function
witi valid data input}
\usage{
example_call()
}
\value{
a string
}
\description{
Returns a valid example call of the \code{lgp} function
witi valid data input
}
