% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-utils.R
\name{extract_components_onesample}
\alias{extract_components_onesample}
\title{Extract inferred components for one sample}
\usage{
extract_components_onesample(fit, sample_idx)
}
\arguments{
\item{fit}{an object of class \code{lgpfit}}

\item{sample_idx}{sample index}
}
\value{
a list
}
\description{
Extract inferred components for one sample
}
