% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-utils.R
\name{extract_t_effect_samples}
\alias{extract_t_effect_samples}
\title{Extract samples of T_effect}
\usage{
extract_t_effect_samples(fit)
}
\arguments{
\item{fit}{an object of class \code{lgpfit}}
}
\value{
a matrix
}
\description{
Extract samples of T_effect
}
