% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-utils.R
\name{field_must_exist}
\alias{field_must_exist}
\title{Access list field and throw error if it is NULL}
\usage{
field_must_exist(fields, name)
}
\arguments{
\item{fields}{list}

\item{name}{field name}
}
\value{
Returns \code{fields[[name]]} unless it is \code{NULL}
}
\description{
Access list field and throw error if it is NULL
}
