% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-prior.R
\name{fill_prior}
\alias{fill_prior}
\title{Fill a partially defined prior}
\usage{
fill_prior(prior)
}
\arguments{
\item{prior}{A named list, defining the prior distribution of model
(hyper)parameters.}
}
\value{
a named list
}
\description{
Fill a partially defined prior
}
