% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-lgpfit.R
\name{fit_summary}
\alias{fit_summary}
\title{Posterior summary}
\usage{
fit_summary(fit)
}
\arguments{
\item{fit}{an object of class \linkS4class{lgpfit}}
}
\value{
a character representation
}
\description{
Posterior summary
}
\seealso{
Other fit postprocessing functions: 
\code{\link{get_draws}()},
\code{\link{get_posterior_f}()},
\code{\link{scale_f_total}()}
}
\concept{fit postprocessing functions}
