% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-fp_input.R
\name{fp_input}
\alias{fp_input}
\title{Create input for Stan model that evaluates function posteriors}
\usage{
fp_input(fit, x, reduce, draws, refresh)
}
\arguments{
\item{fit}{An object of class \linkS4class{lgpfit}.}

\item{x}{A data frame of points where the predictions are computed.
The function \code{\link{new_x}} can help in creating it.}

\item{reduce}{Reduction for parameters draws. Can be a function that
is applied to reduce all parameter draws into one parameter set, or
NULL (no reduction). Has no effect if \code{draws} is specified.}

\item{draws}{Indices of parameter draws to use, or \code{NULL} to use all
draws.}

\item{refresh}{How often to print progress? Has no effect if \code{verbose}
is \code{FALSE}.}
}
\value{
a list that can be used as data for a stan model
}
\description{
Gives a warning if \code{x} is \code{NULL}, but in that
case \code{x} are taken from the Stan input of the original model. This
option is just for debugging purposes.
}
