% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-utils.R
\name{full_model}
\alias{full_model}
\title{Create a full model with all covariates included}
\usage{
full_model(data, ...)
}
\arguments{
\item{data}{a data frame}

\item{...}{additional parameters to \code{\link{lgp_model}}}
}
\value{
a ggplot object
}
\description{
Create a full model with all covariates included
}
