% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-utils.R
\name{full_model_formula}
\alias{full_model_formula}
\title{Get formula of a full model with all covariates included}
\usage{
full_model_formula(data)
}
\arguments{
\item{data}{a data frame, where the response variable is the last column}
}
\value{
a formula
}
\description{
Get formula of a full model with all covariates included
}
