% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-lgpr.R, R/methods-FunctionDraws.R
\docType{class}
\name{FunctionDraws-class}
\alias{FunctionDraws-class}
\alias{FunctionDraws}
\alias{show,FunctionDraws-method}
\alias{plot,FunctionDraws,missing-method}
\title{An S4 class to represent general additive function (component) draws}
\usage{
\S4method{show}{FunctionDraws}(object)

\S4method{plot}{FunctionDraws,missing}(x, y, ...)
}
\arguments{
\item{object}{\linkS4class{FunctionDraws} object for which to apply a class
method.}

\item{x}{a \linkS4class{FunctionDraws} object to visualize}

\item{y}{unused argument}
}
\description{
An S4 class to represent general additive function (component) draws
}
\section{Methods (by generic)}{
\itemize{
\item \code{show}: Print a summary about the object.

\item \code{plot}: Visualization.
Optional arguments (\code{...}) are passed to
\code{\link{plot_FunctionDraws}}.
}}

\section{Slots}{

\describe{
\item{\code{components}}{Data frame representing the draws of each model component.}

\item{\code{x}}{The evaluation points (values of covariates) where the function
draws are evaluated.}

\item{\code{model}}{The \linkS4class{lgpmodel} for which these draws are
computed. Contains important information about how to transform the
total draws through an inverse link function.}
}}

\seealso{
\linkS4class{FunctionPosterior}
}
