% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-all_classes.R, R/S4-FunctionPosterior.R
\docType{class}
\name{FunctionPosterior-class}
\alias{FunctionPosterior-class}
\alias{FunctionPosterior}
\alias{show,FunctionPosterior-method}
\title{An S4 class to represent posterior distributions of an additive marginal GP}
\usage{
\S4method{show}{FunctionPosterior}(object)
}
\arguments{
\item{object}{\linkS4class{FunctionPosterior} object for which to apply a
class method.}
}
\description{
An S4 class to represent posterior distributions of an additive marginal GP
}
\section{Methods (by generic)}{
\itemize{
\item \code{show}: Print a summary about the object.
}}

\section{Slots}{

\describe{
\item{\code{components}}{Data frame representing the posterior distribution of
each model component (on normalized scale).}

\item{\code{total}}{Data frame representing the posterior distribution of the sum of
the components (on normalized scale).}

\item{\code{x}}{The evaluation points (values of covariates) where the posteriors
are evaluated.}

\item{\code{model}}{The \linkS4class{lgpmodel} for which these posteriors are
computed. Contains important information about how to scale the total
posterior from normalized scale to the original scale.}
}}

\seealso{
\linkS4class{FunctionDraws}
}
