% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-all_classes.R
\docType{class}
\name{GaussianPrediction-class}
\alias{GaussianPrediction-class}
\alias{GaussianPrediction}
\title{An S4 class to represent posterior distributions of an additive GP}
\description{
An S4 class to represent posterior distributions of an additive GP
}
\section{Slots}{

\describe{
\item{\code{component}}{data frame representing the posterior distribution of
each model component}

\item{\code{data}}{frame representing the posterior distribution of the sum of
the components}

\item{\code{f_std}}{signal standard deviation (on normalized scale)}

\item{\code{y_mean}}{predictive mean (on original data scale)}

\item{\code{y_std}}{predictive standard deviation (on original data scale)}
}}

\seealso{
\linkS4class{Prediction}
}
