% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-utils.R
\name{get_case_ids}
\alias{get_case_ids}
\title{Get case ids in original data}
\usage{
get_case_ids(fit)
}
\arguments{
\item{fit}{an object of class \code{lgpfit}}
}
\value{
a character vector
}
\description{
Get case ids in original data
}
