% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-model.R
\name{get_chat}
\alias{get_chat}
\title{Get the used GP mean vector}
\usage{
get_chat(object)
}
\arguments{
\item{object}{an object of class \linkS4class{lgpmodel} or
\linkS4class{lgpfit}}
}
\value{
a vector with length equal to number of observations
}
\description{
Get the used GP mean vector
}
