% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-components.R
\name{get_component_plot_titles}
\alias{get_component_plot_titles}
\title{Get inferred components}
\usage{
get_component_plot_titles(corrected = TRUE, sample_idx = NULL)
}
\arguments{
\item{corrected}{Should this obtain the covariate-effect corrected components?}

\item{sample_idx}{Integer or NULL}
}
\value{
a list
}
\description{
Get inferred components
}
