% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-lgpmodel.R
\name{get_covariate_names}
\alias{get_covariate_names}
\title{Get covariate names}
\usage{
get_covariate_names(object, type = "all")
}
\arguments{
\item{object}{an object of class \linkS4class{lgpmodel} or
\linkS4class{lgpfit}}

\item{type}{What types of covariates to include. Must be either
"all", "categorical" or "continuous".}
}
\value{
A string where names are separated by a comma.
}
\description{
Get covariate names
}
