% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-utils.R
\name{get_cutpoints}
\alias{get_cutpoints}
\title{Get samples of cutpoints for ordinal model}
\usage{
get_cutpoints(fit)
}
\arguments{
\item{fit}{An \code{lgpfit} object}
}
\value{
A data frame.
}
\description{
Get samples of cutpoints for ordinal model
}
