% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preproc-disease_info.R
\name{get_diseased_info}
\alias{get_diseased_info}
\title{Get some variables related to diseased individuals}
\usage{
get_diseased_info(D, X, X_notnan, uncertain_diagnosis, equal_effect, TSCL)
}
\arguments{
\item{D}{an integer vector of length 6}

\item{X}{the design matrix}

\item{X_notnan}{a binary vector of length n}

\item{uncertain_diagnosis}{Boolean value}

\item{equal_effect}{Boolean value}

\item{TSCL}{time scaling function and its inverse}
}
\value{
a list
}
\description{
Get some variables related to diseased individuals
}
