% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-pred_input.R
\name{get_draw_arr}
\alias{get_draw_arr}
\alias{get_draw_arr_vec}
\title{Get an array of draws formatted suitably for Stan input}
\usage{
get_draw_arr(fit, draws, reduce, par_name, S, D)

get_draw_arr_vec(fit, draws, reduce, par_name, S, B, V)
}
\arguments{
\item{fit}{An object of class \linkS4class{lgpfit}.}

\item{draws}{Indices of parameter draws to use, or \code{NULL} to use all
draws.}

\item{reduce}{Reduction for parameters draws. Can be a function that
is applied to reduce all parameter draws into one parameter set, or
NULL (no reduction). Has no effect if \code{draws} is specified.}

\item{par_name}{name of parameter (group)}

\item{S}{number of parameter sets (first dimension of output array)}

\item{D}{number of parameters (second dimension of output array)}

\item{V}{vector length for vector params (possible third dimension of
output array)}
}
\value{
an array
}
\description{
Get an array of draws formatted suitably for Stan input
}
