% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethodsFit.R
\name{get_draws}
\alias{get_draws}
\alias{get_draws_catch}
\alias{get_num_draws}
\title{Extract posterior draws}
\usage{
get_draws(fit, stack_chains = FALSE, ...)

get_draws_catch(fit, stack_chains = FALSE, ...)

get_num_draws(fit)
}
\arguments{
\item{fit}{an object of class \linkS4class{lgpfit}}

\item{stack_chains}{should this return a 2-dimensional array?}

\item{...}{additional keyword arguments to \code{rstan::extract}}
}
\value{
a 3 or 2-dimensional array
}
\description{
These functions use \code{\link[rstan]{extract}} with
\code{permuted = FALSE} and \code{inc_warmup = FALSE}.
The \code{get_draws_catch} function calls \code{get_draws} but catches
errors and returns NULL if an error occurs.
}
\seealso{
Other fit postprocessing functions: 
\code{\link{fit_summary}()},
\code{\link{get_f}()},
\code{\link{get_y_rng}()},
\code{\link{scale_f_total}()}
}
\concept{fit postprocessing functions}
