% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-utils.R
\name{get_ell_shared_samples}
\alias{get_ell_shared_samples}
\title{Extract samples of the shared age lengthscale}
\usage{
get_ell_shared_samples(fit)
}
\arguments{
\item{fit}{an object of class 'lgpfit'}
}
\value{
a vector
}
\description{
Extract samples of the shared age lengthscale
}
