% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-utils.R
\name{get_ell_smooth}
\alias{get_ell_smooth}
\title{A convenience function used in postproc-main.R}
\usage{
get_ell_smooth(ell_smooth, ell_smooth_multip, ell_smp)
}
\arguments{
\item{ell_smooth}{a character or numeric argument}

\item{ell_smooth_multip}{numeric}

\item{ell_smp}{numeric}
}
\value{
a number
}
\description{
A convenience function used in postproc-main.R
}
