% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postproc-helper.R
\name{get_function_component_samples}
\alias{get_function_component_samples}
\title{Get values of sampled function components at data points}
\usage{
get_function_component_samples(fit)
}
\arguments{
\item{fit}{An (incomplete) object of class \code{lgpfit}.}
}
\value{
An array of size \code{n_samples} x \code{n_data} x \code{n_components+2}
}
\description{
Get values of sampled function components at data points
}
