% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postproc-helper.R
\name{get_function_components}
\alias{get_function_components}
\title{Get all function components}
\usage{
get_function_components(fit)
}
\arguments{
\item{fit}{An \code{lgpfit} object.}
}
\value{
A data frame. See description.
}
\description{
Get samples or analytically computed means of all 
function components. Returns a data frame \code{df} of size 
\code{n_samples} \code{ x } \code{n_data} \code{ x } 
\code{n_components+2}, so that 
\itemize{
  \item \code{df[ , , n_components+1]} is the sum of all f_j
  \item \code{df[ , , n_components+2]} is a possible transformation
}
}
