% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postproc-helper.R
\name{get_function_components_from_df}
\alias{get_function_components_from_df}
\title{Get values of function components at data points, for one MCMC sample}
\usage{
get_function_components_from_df(pars, model)
}
\arguments{
\item{pars}{A data frame representing one parameter sample,
i.e one row of \code{as.data.frame(stanfit)}, where stanfit
is an object of class \code{stanfit}}

\item{model}{An object of class \code{lgpmodel}}
}
\value{
A matrix of size \code{n_data} x \code{n_components+2}
}
\description{
Get values of function components at data points, for one MCMC sample
}
