% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postproc-helper.R
\name{get_function_components_from_df_all}
\alias{get_function_components_from_df_all}
\title{Get values of function components at data points}
\usage{
get_function_components_from_df_all(df, model)
}
\arguments{
\item{df}{A \code{stanfit} object as data frame,
obtained as \code{as.data.frame(stanfit)}}

\item{model}{An object of class \code{lgpmodel}}
}
\value{
An array of size 
\code{n_samples} x \code{n_data} x \code{n_components+2}
}
\description{
Get values of function components at data points
}
