% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postproc-helper.R
\name{get_g_from_f}
\alias{get_g_from_f}
\title{Get signal on data scale from process f}
\usage{
get_g_from_f(f, model)
}
\arguments{
\item{f}{A vector}

\item{model}{an object of class \code{lgpmodel}}
}
\value{
A vector g
}
\description{
Get signal on data scale from process f
}
