% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-components.R
\name{get_inferred_components}
\alias{get_inferred_components}
\title{Get inferred components}
\usage{
get_inferred_components(fit, sample_idx = NULL)
}
\arguments{
\item{fit}{An object of class \code{lgpfit}.}

\item{sample_idx}{If given, only one sample will be obtained else the average
components over all samples.}
}
\value{
a list
}
\description{
Get inferred components
}
