% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgp-model.R
\name{get_integer_encoding}
\alias{get_integer_encoding}
\title{Get integer encoding of model components}
\usage{
get_integer_encoding(model)
}
\arguments{
\item{model}{an object of class \linkS4class{lgpmodel}}
}
\value{
A data frame.
}
\description{
Get integer encoding of model components
}
