% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preproc-helper.R
\name{get_model_dims}
\alias{get_model_dims}
\title{Set a lot of generic variables that the Stan model needs as input}
\usage{
get_model_dims(X, D, likelihood)
}
\arguments{
\item{X}{the design matrix}

\item{D}{a vector of length 6}

\item{likelihood}{the `likelihood` input to \code{lgp}}
}
\value{
a list
}
\description{
Set a lot of generic variables that the Stan model needs as input
}
