% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_num_ns}
\alias{get_num_ns}
\title{Get number of nonstationary model components}
\usage{
get_num_ns(stan_input)
}
\arguments{
\item{stan_input}{a list containing an element named \code{components}}
}
\value{
an integer
}
\description{
Get number of nonstationary model components
}
