% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-fit_info.R
\name{get_num_paramsets}
\alias{get_num_paramsets}
\title{Decide number of output param sets}
\usage{
get_num_paramsets(fit, draws, reduce)
}
\arguments{
\item{fit}{an object of class \linkS4class{lgpfit}}

\item{draws}{Indices of parameter draws to return use. All post-warmup
draws are returned if this is \code{NULL}.}

\item{reduce}{Function used to reduce all parameter draws into
one set of parameters. Ignored if \code{NULL}, or if \code{draws} is not
\code{NULL}.}
}
\value{
an integer
}
\description{
Decide number of output param sets based on total number of
posterior draws, possible reduction and subset of draw indices
}
