% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-utils.R
\name{get_obs_onset_times}
\alias{get_obs_onset_times}
\title{Extract observed disease onset times from diseaseAge covariate vector}
\usage{
get_obs_onset_times(id, age, disAge)
}
\arguments{
\item{id}{the id covariate, vector of length \code{n}}

\item{age}{the age covariate, vector of length \code{n}}

\item{disAge}{the observed disease-related age covariate, vector of length \code{n}}
}
\value{
vector of observed onset times
}
\description{
Extract observed disease onset times from diseaseAge covariate vector
}
