% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-utils.R
\name{get_observed_effect_times}
\alias{get_observed_effect_times}
\title{Get observed effect times from a data frame}
\usage{
get_observed_effect_times(data, age_variable, disage_variable, id_variable)
}
\arguments{
\item{data}{a data frame}

\item{age_variable}{age variable name}

\item{disage_variable}{disease-related age variable name}

\item{id_variable}{id variable name}
}
\value{
a named vector
}
\description{
Get observed effect times from a data frame
}
