% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preproc-disease_info.R
\name{get_onset_info}
\alias{get_onset_info}
\title{Get disease onset info}
\usage{
get_onset_info(D, X, MAPS, TSCL)
}
\arguments{
\item{D}{an integer vector of length 6}

\item{X}{the design matrix}

\item{MAPS}{mappings created by \code{get_case_row_mappings}}

\item{TSCL}{time scaling function and its inverse}
}
\value{
two vectors of length \code{N_cases}
}
\description{
This returns
\itemize{
  \item a vector of observed onsets
  \item mapping from case ID to average sampling interval before
  the observed disease onset
}
}
