% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-lgpfit.R
\name{get_posterior_f}
\alias{get_posterior_f}
\title{Extract posterior of the function f and its components}
\usage{
get_posterior_f(fit, draws = NULL)
}
\arguments{
\item{fit}{an object of class \linkS4class{lgpfit}}

\item{draws}{Indices of posterior draws for which to get \code{f}. This can
be a single integer, a vector of indices, or \code{NULL} (default). In the
latter case all draws are obtained.}
}
\value{
Returns a list with names \code{num_draws} and \code{f}.
The latter is a named list of which has length equal to the number of
components plus one. Let \code{S = length(draws)}. Each list element is
\itemize{
  \item Array of size \code{S} x \code{num_obs}, where each row is one
  posterior draw of the function f, \code{is_sampled(model)} is \code{TRUE}.
  \item A list with fields \code{mean} and \code{std}, if
  \code{is_sampled(model)} is \code{FALSE}. Both fields are arrays of size
  \code{S} x \code{num_obs}. These are the analytically computed means and
  standard deviations for each posterior draw.
}
}
\description{
Extract posterior of the function f and its components
}
\seealso{
Other fit postprocessing functions: 
\code{\link{fit_summary}()},
\code{\link{get_draws}()},
\code{\link{scale_f_total}()}
}
\concept{fit postprocessing functions}
