% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-post_post.R
\name{get_pred}
\alias{get_pred}
\title{Posterior into posterior predictive}
\usage{
get_pred(post)
}
\arguments{
\item{post}{an object of class \linkS4class{gpposterior}}
}
\value{
TODO
}
\description{
Posterior into posterior predictive
}
