% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-pred_get.R
\name{get_pred_sampled}
\alias{get_pred_sampled}
\alias{get_pred.sampled.f_comp}
\alias{get_pred.sampled.f}
\alias{get_pred.sampled.h}
\title{Get predictions for a model where the latent function f was sampled}
\usage{
get_pred.sampled.f_comp(fit, draws, reduce)

get_pred.sampled.f(fit, draws, reduce)

get_pred.sampled.h(fit, draws, reduce)
}
\arguments{
\item{fit}{an object of class \linkS4class{lgpfit}}

\item{draws}{Indices of parameter draws to return use. All post-warmup
draws are returned if this is \code{NULL}.}

\item{reduce}{Function used to reduce all parameter draws into
one set of parameters. Ignored if \code{NULL}, or if \code{draws} is not
\code{NULL}.}
}
\value{
an array of shape \code{num_draws} x \code{num_obs} (actually a
list of such arrays for \code{get_pred.sampled.f_comp})
}
\description{
These are helper functions for \code{\link{get_pred.sampled}}.
\itemize{
  \item Function \code{get_pred.sampled.f_comp} gets the draws of each
  component of \code{f} on the normalized scale.
  \item Function \code{get_pred.sampled.f} gets draws of the total \code{f}
  on the normalized scale.
  \item Function \code{get_pred.sampled.h} gets draws of the total \code{f},
  adds \code{c_hat} to each draw, and then maps through the inverse
  link function. It applies \code{reduce} only after this transformation.
}
}
\seealso{
Other prediction extraction functions: 
\code{\link{get_pred_gaussian}},
\code{\link{get_pred}()}
}
\concept{prediction extraction functions}
