% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postproc-helper.R
\name{get_predicted}
\alias{get_predicted}
\title{A helper function}
\usage{
get_predicted(fit)
}
\arguments{
\item{fit}{An (incomplete) object of class \code{lgpfit}.}
}
\value{
a list
}
\description{
A helper function
}
