% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preproc-parse_prior.R
\name{get_prior_params}
\alias{get_prior_params}
\title{Get prior parameters}
\usage{
get_prior_params(dist, add_correct)
}
\arguments{
\item{dist}{the distribution}

\item{add_correct}{additional correct parameter names}
}
\value{
a hyperparameter vector of length 2
}
\description{
Get prior parameters
}
