% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preproc-parse_prior.R
\name{get_prior_type}
\alias{get_prior_type}
\title{A dictionary from distribution names to integer encoding}
\usage{
get_prior_type(type)
}
\arguments{
\item{type}{type of the distribution as a string}
}
\value{
an integer
}
\description{
A dictionary from distribution names to integer encoding
}
