% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preproc-helper.R
\name{get_response}
\alias{get_response}
\title{Get the (scaled) response variable}
\usage{
get_response(data, varInfo, standardize, LH)
}
\arguments{
\item{data}{the data frame given as input to \code{lgp}}

\item{varInfo}{variable type info}

\item{standardize}{should the response be standardized to unit variance and zero mean}

\item{LH}{likelihood as integer}
}
\value{
a list with the (scaled) response variable
}
\description{
Gets and possibly scales the response variable.
}
