% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postproc-helper.R
\name{get_runtime}
\alias{get_runtime}
\title{Get average runtime of a chain}
\usage{
get_runtime(object)
}
\arguments{
\item{object}{An object of class \code{lgpfit}.}
}
\value{
Average runtimes for warmup and sampling
}
\description{
Get average runtime of a chain
}
