% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-misc_methods.R
\name{get_sim_components}
\alias{get_sim_components}
\title{Get simulated components from an lgpsim object}
\usage{
get_sim_components(simdata)
}
\arguments{
\item{simdata}{an object of class \linkS4class{lgpsim}}
}
\value{
a data frame
}
\description{
Get simulated components from an lgpsim object
}
