% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-fit_info.R
\name{get_stan_fit}
\alias{get_stan_fit}
\title{Get the stanfit object from an lgpfit}
\usage{
get_stan_fit(fit)
}
\arguments{
\item{fit}{an object of class \linkS4class{lgpfit}}
}
\description{
Get the stanfit object from an lgpfit
}
\seealso{
Other fit postprocessing functions: 
\code{\link{fit_summary}()},
\code{\link{get_draws}()}
}
\concept{fit postprocessing functions}
