% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-utils.R
\name{get_stan_model}
\alias{get_stan_model}
\title{Get one of the stan models of the package}
\usage{
get_stan_model(idx = 1)
}
\arguments{
\item{idx}{model index (default = 1)}
}
\value{
an object of class stanmodel
}
\description{
Get one of the stan models of the package
}
