% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_teff_obs}
\alias{get_teff_obs}
\title{Infer observed effect times from a data frame}
\usage{
get_teff_obs(data, age_variable, disage_variable, id_variable)
}
\arguments{
\item{data}{a data frame}

\item{age_variable}{age variable name}

\item{disage_variable}{disease-related age variable name}

\item{id_variable}{id variable name}
}
\value{
a data frame
}
\description{
Infer observed effect times from a data frame
}
