% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preproc-parse_prior.R
\name{get_transform_type}
\alias{get_transform_type}
\title{A dictionary from transform names to integer encoding}
\usage{
get_transform_type(type)
}
\arguments{
\item{type}{Type of the transform as a string. Allowed arguments
are "none" or "square". If NULL, "none" is used.}
}
\value{
an integer (0, 1 or 2)
}
\description{
A dictionary from transform names to integer encoding
}
