% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgpmodel-class.R
\name{get_y}
\alias{get_y}
\alias{get_y_name}
\title{Get response variable measurements}
\usage{
get_y(object, original = TRUE)

get_y_name(object)
}
\arguments{
\item{object}{an object of class \linkS4class{lgpmodel} or
\linkS4class{lgpfit}}

\item{original}{should the measuments be on their original scale?}
}
\value{
a vector
}
\description{
Get response variable measurements
}
