% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-post.R
\name{get_y_rng}
\alias{get_y_rng}
\title{Extract posterior or prior predictive distribution draws}
\usage{
get_y_rng(fit, original_scale = TRUE)
}
\arguments{
\item{fit}{an object of class \linkS4class{lgpfit}}

\item{original_scale}{should the draws be scaled back to original y scale
(only has effect if likelihood is "gaussian", when data has been normalized
to zero mean and unit variance)}
}
\value{
an array of shape \code{num_draws} x \code{num_obs}
}
\description{
Extract posterior or prior predictive distribution draws
}
\seealso{
Other fit postprocessing functions: 
\code{\link{fit_summary}()},
\code{\link{get_draws}()}
}
\concept{fit postprocessing functions}
