% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgp-formula.R
\name{gp}
\alias{gp}
\title{Create a stationary GP term}
\usage{
gp(covariate)
}
\arguments{
\item{covariate}{name of a continuous covariate}
}
\value{
an object of class \code{\link{lgpterm}}
}
\description{
Create a stationary GP term
}
