% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgp-formula.R
\name{gp_ns}
\alias{gp_ns}
\title{Create a nonstationary GP term}
\usage{
gp_ns(covariate)
}
\arguments{
\item{covariate}{name of a continuous covariate}
}
\value{
an object of class \code{\link{lgpterm}}
}
\description{
Create a nonstationary GP term
}
