% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{gp_posterior}
\alias{gp_posterior}
\title{Compute all componentwise GP posteriors of a model}
\usage{
gp_posterior(KX, y, delta, sigma, STREAM = get_stream())
}
\arguments{
\item{KX}{list of kernel matrices}

\item{y}{response vector}

\item{delta}{jitter}

\item{sigma}{Gaussian noise magnitude parameter}

\item{STREAM}{an external pointer}
}
\value{
an array of kernel matrices
}
\description{
This is a wrapper for \code{STAN_gp_posterior}
}
