% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior.R
\name{gp_posteriors}
\alias{gp_posteriors}
\alias{gp_kernels}
\title{Compute total and componentwise GP posteriors at given points}
\usage{
gp_posteriors(fit, x, STREAM = get_stream())

gp_kernels(fit, X_pred, STREAM = get_stream())
}
\arguments{
\item{fit}{An object of class \linkS4class{lgpfit}.}

\item{x}{A data frame of points where the posteriors are computed.
The function \code{\link{new_x}} provides an easy way to create it.}

\item{STREAM}{External pointer. By default obtained with
\code{\link[rstan]{get_stream}}.}
}
\value{
See description.
}
\description{
\itemize{
  \item Total and componentwise GP posteriors at test points can be
  computed using \code{gp_posteriors}. This function requires
  the observation model to be Gaussian. The actual work is done
  in \code{\link{gp_posteriors_compute}}. Return?
  \item All kernel matrices required for computing posterior distributions
  at test points are computed by \code{gp_kernels}. It returns a list
  with fields named \code{data_vs_data}, \code{pred_vs_data} and
  \code{pred_vs_pred}. Each of them is a list with length equal to the
  number of posterior draws. The actual work is done
  in \code{\link{kernel_matrices}}
}
}
\seealso{
Other GP posterior computation functions: 
\code{\link{gp_posteriors_compute}()}
}
\concept{GP posterior computation functions}
