% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior.R
\name{gp_posteriors_compute}
\alias{gp_posteriors_compute}
\title{Compute GP posteriors}
\usage{
gp_posteriors_compute(kernels, y, sigma, delta, STREAM = get_stream())
}
\arguments{
\item{kernels}{a list of arrays returned by \code{\link{kernel_matrices}}}

\item{y}{response variable vector of length \code{num_obs}}

\item{sigma}{a vector with length \code{num_draws}}

\item{delta}{jitter to ensure positive definite matrices}

\item{STREAM}{External pointer. By default obtained with
\code{\link[rstan]{get_stream}}.}
}
\value{
A list.
}
\description{
Compute GP posteriors
}
\seealso{
Other GP posterior computation functions: 
\code{\link{gp_posteriors}()}
}
\concept{GP posterior computation functions}
