% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-all_classes.R
\docType{class}
\name{GPDraws-class}
\alias{GPDraws-class}
\alias{GPDraws}
\title{An S4 class to represent posterior or prior function (component)
draws of an additive latent GP model}
\description{
An S4 class to represent posterior or prior function (component)
draws of an additive latent GP model
}
\section{Slots}{

\describe{
\item{\code{components}}{Data frame representing the draws of each model component.}

\item{\code{x}}{The evaluation points (values of covariates) where the function
draws are evaluated.}

\item{\code{model}}{The \linkS4class{lgpmodel} for which these draws are
computed. Contains important information about how to transform the
total draws through an inverse link function.}
}}

\seealso{
\linkS4class{GPPosterior}
}
