% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postproc-helper.R
\name{hyperparam_estimate}
\alias{hyperparam_estimate}
\title{Get a posterior estimate of model (hyper)parameters}
\usage{
hyperparam_estimate(object, type = "mean")
}
\arguments{
\item{object}{An (incomplete) object of class \code{lgpfit}.}

\item{type}{Must be "mean" or "median".}
}
\value{
a data frame
}
\description{
Get a posterior estimate of model (hyper)parameters
}
