% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postproc-helper.R
\name{hyperparam_samples}
\alias{hyperparam_samples}
\title{Get a set of model (hyper)parameter samples}
\usage{
hyperparam_samples(object, samples = NULL)
}
\arguments{
\item{object}{An (incomplete) object of class \code{lgpfit}.}

\item{samples}{Sample indices. If NULL, all samples are taken.}
}
\value{
a data frame
}
\description{
Get a set of model (hyper)parameter samples
}
