% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{idx_expand}
\alias{idx_expand}
\alias{create_idx_expand}
\alias{create_idx_expand_picker}
\alias{map_factor_to_caseid}
\alias{map_caseid_to_row}
\title{Creating the idx_expand input for Stan}
\usage{
create_idx_expand(components, x_cat, x_cont_mask)

create_idx_expand_picker(components, x_cat)

map_factor_to_caseid(x_fac, x_cont_mask)

map_caseid_to_row(x_fac, map)
}
\arguments{
\item{components}{the \code{components} input for Stan}

\item{x_cat}{the \code{x_cat} input for Stan}

\item{x_cont_mask}{the \code{x_cont_mask} input for Stan}

\item{x_fac}{object returned by \code{\link{create_idx_expand_picker}}}

\item{map}{object returned by \code{\link{map_factor_to_caseid}}}
}
\value{
the \code{idx_expand} input for Stan and a mapping data frame
}
\description{
Creating the idx_expand input for Stan
}
