% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-components.R
\name{idx_to_cont_index}
\alias{idx_to_cont_index}
\title{Component index to how manyth continuous covariate it is}
\usage{
idx_to_cont_index(D, idx)
}
\arguments{
\item{D}{integer vector of length 6}

\item{idx}{an integer}
}
\value{
an integer
}
\description{
Component index to how manyth continuous covariate it is
}
