% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-invgamma.R
\name{invgamma}
\alias{invgamma}
\alias{dinvgamma_stanlike}
\alias{dinvgamma_stanlike,}
\alias{qinvgamma_stanlike,}
\alias{plot_invgamma}
\alias{qinvgamma_stanlike}
\title{Functions related to the inverse gamma distribution}
\usage{
dinvgamma_stanlike(x, alpha, beta, log = FALSE)

qinvgamma_stanlike(p, alpha, beta)

plot_invgamma(
  alpha,
  beta,
  by = 0.01,
  log = FALSE,
  IQR = 0.95,
  return_quantiles = FALSE,
  linecolor = colorset("red", "dark"),
  fillcolor = colorset("red", "mid")
)
}
\arguments{
\item{x}{point where to compute the density}

\item{alpha}{positive real number}

\item{beta}{positive real number}

\item{log}{is log-scale used?}

\item{p}{quantile (must be between 0 and 1)}

\item{by}{grid size}

\item{IQR}{inter-quantile range width for \code{\link{plot_invgamma}}}

\item{return_quantiles}{\code{\link{plot_invgamma}} return a list}

\item{linecolor}{line color}

\item{fillcolor}{fill color}
}
\value{
density/quantile value or ggplot object
}
\description{
Using the same parametrization as Stan. More info
\href{https://mc-stan.org/docs/2_24/functions-reference/inverse-gamma-distribution.html}{here}.
}
