% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-Kernel.R
\name{Kernel-classes}
\alias{Kernel-classes}
\title{Kernel description classes}
\description{
Kernel description classes
}
\section{Methods}{

\code{$desc()} Get kernel description.

\code{$print()} Print the kernel description.

\code{$parameters()} Get all \code{Parameter}s of the kernel as a list.

\code{$variables()} Get all input \code{Variable}s of the kernel as a list.
}

