% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-kernel_reference.R
\name{kernel}
\alias{kernel}
\title{Compute a kernel matrix (covariance matrix)}
\arguments{
\item{x1}{vector of length \eqn{n}}

\item{x2}{vector of length \eqn{m}}

\item{alpha}{marginal std (default = 1)}

\item{ell}{lengthscale}
}
\value{
A matrix of size \eqn{n} x \eqn{m}.
}
\description{
These have \code{STAN_kernel_*} counterparts. These R versions
are provided for reference and are not optimized for speed. These are
used when generating simulated data, and not during model inference.
}
