% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stanfunctions.R
\name{kernel_all}
\alias{kernel_all}
\title{Compute all kernel matrices of a model}
\usage{
kernel_all(
  K_const,
  components,
  x1,
  x2,
  x1_unnorm,
  x2_unnorm,
  alpha,
  ell,
  wrp,
  beta,
  teff,
  vm_params,
  idx1_expand,
  idx2_expand,
  teff_obs,
  STREAM = get_stream()
)
}
\arguments{
\item{K_const}{list of constant matrices}

\item{components}{an integer array}

\item{x1}{continuous covariates}

\item{x2}{continuous covariates}

\item{x1_unnorm}{unnormalized versions of the continuous covariates}

\item{x2_unnorm}{unnormalized versions of the continuous covariates}

\item{alpha}{magnitude parameters}

\item{ell}{lengthscale parameters}

\item{wrp}{input warping parameters}

\item{beta}{heterogeneity parameters}

\item{teff}{effect time parameters}

\item{vm_params}{variance mask parameters}

\item{idx1_expand}{expansion of beta and/or teff}

\item{idx2_expand}{expansion of beta and/or teff}

\item{teff_obs}{observed effect times}

\item{STREAM}{an external pointer}
}
\value{
an array of kernel matrices
}
\description{
This is a wrapper for \code{STAN_kernel_all}
}
