% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-kernels.R
\name{kernel_cat}
\alias{kernel_cat}
\title{Compute a categorical kernel matrix}
\usage{
kernel_cat(x1, x2, alpha = 1)
}
\arguments{
\item{x1}{(integer) vector of length n}

\item{x2}{(integer) vector of length m}

\item{alpha}{marginal std (default = 1)}
}
\value{
A (binary) kernel matrix of size n x m
}
\description{
Compute a categorical kernel matrix
}
