% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan-functions.R
\name{kernel_const_all}
\alias{kernel_const_all}
\title{Compute all constant kernel matrices of a model}
\usage{
kernel_const_all(
  x1_cat,
  x2_cat,
  x1_cont_mask,
  x2_cont_mask,
  x_cat_num_levels,
  components,
  STREAM = get_stream()
)
}
\arguments{
\item{x1_cat}{categorical covariates}

\item{x2_cat}{categorical covariates}

\item{x1_cont_mask}{continuous covariate masks}

\item{x2_cont_mask}{continuous covariate masks}

\item{x_cat_num_levels}{number of levels for each categorical covariate}

\item{components}{an integer array}

\item{STREAM}{an external pointer}
}
\value{
an array of kernel matrices
}
\description{
This is a wrapper for \code{STAN_kernel_const_all}
}
