% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethodsModel.R
\name{kernel_decomposition}
\alias{kernel_decomposition}
\title{Basis function decomposition of the full additive kernel matrix}
\usage{
kernel_decomposition(
  object,
  alpha,
  ell,
  num_basisfun = 10,
  width_basisfun = 2.5,
  STREAM = get_stream()
)
}
\arguments{
\item{object}{an object of class \linkS4class{lgpmodel} or
\linkS4class{lgpfit}}

\item{alpha}{vector of magnitude params}

\item{ell}{vector of lengthscale params}

\item{num_basisfun}{number of basis functions to use}

\item{width_basisfun}{basis function domain size (L)}

\item{STREAM}{external pointer}
}
\value{
a list containing matrices \code{V} and \code{D}
}
\description{
Basis function decomposition of the full additive kernel matrix
}
