% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-pred.R
\name{kernel_matrices}
\alias{kernel_matrices}
\alias{kernel_matrices_const}
\alias{kernel_matrices_create_input}
\title{Kernel matrix computation using the C++ functions of the package namespace}
\usage{
kernel_matrices(model, x1, x2, theta, verbose, STREAM = get_stream())

kernel_matrices_const(model, x1, x2, STREAM = get_stream())

kernel_matrices_create_input(model, x)
}
\arguments{
\item{model}{An object of class \linkS4class{lgpmodel}.}

\item{x1}{A list defining the first input argument. Can be
created by \code{\link{kernel_matrices_create_input}}.}

\item{x2}{A list defining the second input argument. Can be
created by \code{\link{kernel_matrices_create_input}}.}

\item{theta}{A list of kernel parameter sets.}

\item{verbose}{Should more information and a progress bar be printed?}

\item{STREAM}{External pointer. By default obtained with
\code{\link[rstan]{get_stream}}.}

\item{x}{A data frame of points where the predictions are computed.
The function \code{\link{new_x}} can help in creating it.}
}
\description{
\itemize{
  \item Kernel matrices using parameter values defined by each posterior
  draw can be computed using \code{kernel_matrices}. It returns a
  4-dimensional array of shape \code{num_draws} x \code{num_comps} x
  \code{n1} x \code{n2}.
  \item The constant kernel matrices that do not depend on parameters
  can be computed using \code{kernel_matrices_const}. It returns a list
  of kernel matrices with length equal to number of model components.
  \item Input information can be translated from a data frame to the
  needed format using \code{kernel_matrices_create_input}.
}
}
\seealso{
Other prediction functions: 
\code{\link{pred.latent_compute}()},
\code{\link{pred.marginal_compute}()},
\code{\link{pred_kernels}()},
\code{\link{pred}()}
}
\concept{prediction functions}
