% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-kernels.R
\name{kernel_se}
\alias{kernel_se}
\title{Compute a squared exponential kernel matrix}
\usage{
kernel_se(x1, x2, alpha = 1, ell = 1)
}
\arguments{
\item{x1}{vector of length n}

\item{x2}{vector of length m}

\item{alpha}{marginal std (default = 1)}

\item{ell}{lengthscale (default = 1)}
}
\value{
A kernel matrix of size n x m
}
\description{
Compute a squared exponential kernel matrix
}
