% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postproc-helper.R
\name{kernel_smoothing}
\alias{kernel_smoothing}
\title{Estimate conditional mean time profile using gaussian kernel smoothing}
\usage{
kernel_smoothing(v, t, t_out, ell)
}
\arguments{
\item{v}{a vector of length \code{n} to be smoothed}

\item{t}{vector of \code{n} time points corresponding to \code{y}}

\item{t_out}{the set of \code{p} time points where the smoothing should be evaluated}

\item{ell}{kernel lengthscale}
}
\value{
a vector of length \code{p}
}
\description{
Estimate conditional mean time profile using gaussian kernel smoothing
}
