% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-kernel.R
\name{kernels_fpost}
\alias{kernels_fpost}
\title{Kernel matrices for function posterior computation}
\usage{
kernels_fpost(
  fit,
  x,
  reduce,
  draws,
  verbose,
  debug_dims,
  STREAM = get_stream()
)
}
\arguments{
\item{fit}{An object of class \linkS4class{lgpfit}.}

\item{x}{A data frame of points where the predictions are computed.
The function \code{\link{new_x}} can help in creating it.}

\item{reduce}{Reduction for parameters draws. Can be a function that
is applied to reduce all parameter draws into one parameter set, or
\code{NULL} (no reduction). Has no effect if \code{draws} is specified.}

\item{draws}{Indices of parameter draws to use, or \code{NULL} to use all
draws.}

\item{verbose}{Should more some informational messages be printed?}

\item{debug_dims}{Should this print dimensions of some variables.
Can be used for debugging \code{\link{kernels_fpost}}.}

\item{STREAM}{an external pointer}
}
\description{
Kernel matrices for function posterior computation
}
