% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{large_data_msg}
\alias{large_data_msg}
\title{Display a runtime estimation message if starting to analyse a large data set}
\usage{
large_data_msg(num_obs, threshold)
}
\arguments{
\item{num_obs}{number of observations}

\item{threshold}{threshold for number of observations}
}
\value{
nothing
}
\description{
Display a runtime estimation message if starting to analyse a large data set
}
