% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R, R/methods-lgpmodel.R
\docType{class}
\name{LatentGPModelApprox-class}
\alias{LatentGPModelApprox-class}
\alias{LatentGPModelApprox}
\alias{get_stanmodel,LatentGPModelApprox-method}
\title{An S4 class to represent an approximate latent GP model}
\usage{
\S4method{get_stanmodel}{LatentGPModelApprox}(object)
}
\arguments{
\item{object}{The object for which to apply a class method.}
}
\description{
An S4 class to represent an approximate latent GP model
}
\section{Methods (by generic)}{
\itemize{
\item \code{get_stanmodel}: Get name of corresponding 'Stan' model.
}}

