% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\docType{class}
\name{LatentGPModelParams-class}
\alias{LatentGPModelParams-class}
\alias{LatentGPModelParams}
\title{An S4 class to represent a set of latent GP model parameters}
\description{
An S4 class to represent a set of latent GP model parameters
}
\section{Slots}{

\describe{
\item{\code{sigma}}{Gaussian observation model param}

\item{\code{phi}}{Negative binomial observation model param}

\item{\code{sigma}}{Beta-binomial observation model param}
}}

