% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgp-main.R
\name{lgp}
\alias{lgp}
\title{The main function of the `lgpr` package}
\usage{
lgp(
  formula,
  data,
  likelihood = "Gaussian",
  prior = prior_default(),
  uncertain_effect_time = FALSE,
  equal_effect = TRUE,
  id_variable = "id",
  time_variable = "age",
  disAge_variable = NULL,
  continuous_vars = NULL,
  categorical_vars = NULL,
  offset_vars = NULL,
  C_hat = NULL,
  DELTA = 1e-08,
  sample_F = NULL,
  parallel = FALSE,
  skip_postproc = FALSE,
  threshold = 0.95,
  variance_mask = TRUE,
  N_trials = NULL,
  norm_factors = NULL,
  relevance_method = "f_mean",
  verbose = FALSE,
  ...
)
}
\arguments{
\item{formula}{A formula of the form
\code{y ~ } \code{x1 + x2 + x3}
defining the response variable \code{y} and covariates \code{xi}. The formula must
contain exatly one tilde (~), with response variable on the left-hand side and 
covariates on the right-hand side. Covariates should be separated by a plus (+)
sign. All variables that appear in the formula must exist as columns of \code{data}.
Note that effects of categorical covariates are by default defined
as interactions with \code{time_variable}. If you wish to change
this, see the argument \code{offset_vars}. The subject identifier
variable cannot currently be included in \code{offset_vars}. If you wish
to model the effect of \code{id_variable} as a constant offset,
you can create another covariate with the same values and
use it in your \code{formula} and \code{offset_vars} instead.}

\item{data}{A data frame containing the variables given in \code{formula} as columns.}

\item{likelihood}{Determines the observation model. Must be either \code{"Gaussian"} 
(default), \code{"Poisson"}, \code{"NB"} (negative binomial) or \code{"binomial"}. To
use Bernoulli likelihood, use \code{likelihood="binomial"} and set \code{N_trials} 
as a vector of ones.}

\item{prior}{A named list, defining the prior distribution of model (hyper)parameters.
It is recommended to first create this using the function \code{\link{prior_default}},
and then possibly modify it.}

\item{uncertain_effect_time}{Do we wish to model uncertainty in the disease effect time?}

\item{equal_effect}{Is the disease effect assumed to be equally strong for all diseased 
individuals?}

\item{id_variable}{Name of the unique subject identifier variable (default = \code{"id"}).}

\item{time_variable}{Name of the time variable (default = \code{"age"}).}

\item{disAge_variable}{Name of the disease-related age variable. If {NULL} (default),
this will be chosen to be \code{"diseaseAge"}, if such covariate is found in the data.}

\item{continuous_vars}{Names of other continuous covariates. If \code{NULL}, the
remaining covariates that have floating point values are interpreted as continuous.}

\item{categorical_vars}{Names of categorical covariates that interact with the time variable.
If \code{NULL} (default), the remaining covariates that have integer values are interpreted 
as categorical.}

\item{offset_vars}{Names of the categorical covariates that are treated as 
time-independent group offsets. If \code{NULL} (default), no variables are interpreted as such
covariates.}

\item{C_hat}{The constant GP mean. By default this is NULL, and 
 set to
 \itemize{
 \item \code{C_hat = 0}, if \code{likelihood} is \code{"Gaussian"}, because with 
 Gaussian likelihood the 
 response variable is by default centered to have zero mean.
\item \code{C_hat = } \code{log(mean(y))} if \code{likelihood} is \code{"Poisson"}
or \code{"NB"},
\item \code{C_hat = } \code{log(p/(1-p))}, where \code{p = mean(y/N_trials)} if 
\code{likelihood} is \code{"binomial"}
 }
Above, \code{y} denotes the response variable.}

\item{DELTA}{the amount of added jitter to ensure positive definiteness of the kernel}

\item{sample_F}{Determines if the function values are be sampled (must be \code{TRUE} if
likelihood is not \code{"Gaussian"}).}

\item{parallel}{Determines if the chain will be run in parallel (\code{default = FALSE}).
If \code{TRUE}, then Stan is run by first defining 
\code{options(mc.cores = } \code{parallel::detectCores())}.}

\item{skip_postproc}{In this mode the postprocessing after running Stan is skipped.}

\item{threshold}{Component selection threshold for relevance sum.}

\item{variance_mask}{Should a variance mask be used to force disease component
variance to zero before disease onset?}

\item{N_trials}{This argument (number of trials) is only needed when likelihood is binomial.
Must have length one or equal to number of data points. Setting \code{N_trials=1} corresponds to 
Bernoulli observation model.}

\item{norm_factors}{Normalizing factors for each data point. If not \code{NULL}, this must be a 
vector with length \code{dim(data)[1]}, specifying a multiplicative factor that scales
the signal for the obsevation model at each data point. Setting this to \code{NULL} is
effectively same as setting this to a vector of ones. With Gaussian likelihood, do not
use this, normalize the data beforehand instead.}

\item{relevance_method}{Component relevance determination method. 
Must be either \code{"f_mean"} or \code{"alpha"}.}

\item{verbose}{Should more verbose output be printed?}

\item{...}{Optional arguments passed to \code{rstan::sampling}, for example \code{iter},
\code{chains} or \code{control}. See \code{\link[rstan]{sampling}} for the possible arguments.}
}
\value{
An object of class \code{lgpfit}.
}
\description{
This is a wrapper for both \code{\link{lgp_model}}
 and \code{\link{lgp_fit}}. It first creates an 
 \code{lgpmodel} object and then fits the model, 
 finally returning an \code{lgpfit} object. Note that the
 covariate types are automatically inferred from the given 
 \code{data}. If you wish to change these, see the
 arguments
 \itemize{
    \item \code{id_variable}
    \item \code{time_variable}
    \item \code{disAge_variable}
    \item \code{continuous_vars} and
    \item \code{categorical_vars}.
 }
}
