% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-utils.R
\name{lgp_component_names}
\alias{lgp_component_names}
\title{Get names of model components}
\usage{
lgp_component_names(stan_dat)
}
\arguments{
\item{stan_dat}{The data that was passed to \code{rstan::sampling}}
}
\value{
names of model components
}
\description{
Get names of model components
}
