% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-utils.R
\name{lgp_covariate_names}
\alias{lgp_covariate_names}
\title{Get names of model covariates}
\usage{
lgp_covariate_names(stan_dat)
}
\arguments{
\item{stan_dat}{The data that was passed to \code{rstan::sampling}}
}
\value{
names of model components
}
\description{
Get names of model covariates
}
