% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgp-main.R
\name{lgp_fit}
\alias{lgp_fit}
\title{Fit an lgp model}
\usage{
lgp_fit(
  model,
  threshold = 0.95,
  parallel = FALSE,
  skip_postproc = FALSE,
  relevance_method = "f_mean",
  verbose = FALSE,
  ...
)
}
\arguments{
\item{model}{An object of class \code{lgpmodel}.}

\item{threshold}{Component selection threshold for relevance sum.}

\item{parallel}{Determines if the chain will be run in parallel (\code{default = FALSE}).
If \code{TRUE}, then Stan is run by first defining 
\code{options(mc.cores = } \code{parallel::detectCores())}.}

\item{skip_postproc}{In this mode the postprocessing after running Stan is skipped.}

\item{relevance_method}{Component relevance determination method. 
Must be either \code{"f_mean"} or \code{"alpha"}.}

\item{verbose}{should some output be printed?}

\item{...}{Optional arguments passed to \code{rstan::sampling}, for example \code{iter},
\code{chains} or \code{control}. See \code{\link[rstan]{sampling}} for the possible arguments.}
}
\value{
An object of class \code{lgpfit}.
}
\description{
Samples the posterior of an additive Gaussian process regression model
using \code{\link{rstan}}.
}
\seealso{
For the possible additional arguments, see \code{\link[rstan]{sampling}}.
For creating the \code{lgpmodel} input, see \code{\link{lgp_model}}.
}
