% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgp-formula.R
\name{lgp_formula}
\alias{lgp_formula}
\title{Create a model formula}
\usage{
lgp_formula(formula)
}
\arguments{
\item{formula}{an object of class \code{formula}}
}
\value{
an object of class \code{\link{lgpformula}}
}
\description{
Create a model formula
}
