% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgp-main.R
\name{lgp_model}
\alias{lgp_model}
\title{Create an lgp model}
\usage{
lgp_model(formula, data, likelihood = "Gaussian",
  prior = prior_default(likelihood), uncertain_diagnosis = FALSE,
  equal_effect = TRUE, C_hat = NULL, DELTA = 1e-12,
  sample_F = (likelihood != "Gaussian"), t_test = NULL,
  id_variable = "id", time_variable = "age", disAge_variable = NULL,
  continuous_vars = NULL, categorical_vars = NULL,
  offset_vars = NULL, variance_mask = FALSE,
  cat_interact_kernel_type = "categorical")
}
\arguments{
\item{formula}{A formula of the form
\code{y ~ } \code{x1 + x2 + x3}
defining the response variable \code{y} and covariates \code{xi}. All variables that
appear in the formula must exist as columns of \code{data}.}

\item{data}{A data frame containing (at least) the variables given in \code{formula}.}

\item{likelihood}{Either \code{"Gaussian"} (default), \code{"Poisson"} or \code{"NB"}.}

\item{prior}{Prior distribution. Can be created for example using the function 
\code{\link{prior_default}}.}

\item{uncertain_diagnosis}{Do we wish to model uncertainty in the disease onset?}

\item{equal_effect}{Is the disease effect assumed to be equally strong for all diseased individuals?}

\item{C_hat}{This can only be given if likelihood is not Gaussian. The signal \code{f} 
will the be transformed so that \code{g = } \code{exp(C_hat + f)}. If \code{NULL}, it will be
set to \code{C_hat = } \code{log(mean(y))}, where \code{y} is the response variable.}

\item{DELTA}{the amount of added jitter to ensure positive definiteness of the kernel}

\item{sample_F}{Determines if the function values are be sampled (must be \code{TRUE} if
likelihood is not Gaussian).}

\item{t_test}{Optional test time points. Should only be used if \code{sample_F = TRUE}.
Otherwise use \code{\link{lgp_predict}} after fitting the model.}

\item{id_variable}{Name of the unique subject identifier variable.}

\item{time_variable}{Name of the time variable.}

\item{disAge_variable}{Name of the disease-related age variable. If {NULL}, this
will be chosen to be "diseaseAge", if such covariate is found in the data.}

\item{continuous_vars}{Names of other continuous covariates. If \code{NULL}, the
remaining covariates that have floating point values are interpreted as continuous.}

\item{categorical_vars}{Names of categorical covariates that interact with the time variable.
If \code{NULL}, the remaining covariates that have integer values are interpreted 
as categorical.}

\item{offset_vars}{Names of the categorical covariates that are treated as 
time-independent group offsets. If \code{NULL}, no variables are interpreted as such
covariates.}

\item{variance_mask}{Should a variance mask be used to force disease component
variance to zero before disease onset?}

\item{cat_interact_kernel_type}{Kernel type for categorical variables (other than id). 
Possible options are \code{"categorical"} (default) and \code{"binary"} (mask kernel where
only category \code{"1"} will have an effect).}
}
\value{
An object of class \code{lgpmodel}.
}
\description{
Creates an object of class \code{lgpmodel}
}
\seealso{
For fitting the model, see \code{\link{lgp_fit}}.
}
