% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgp-main.R
\name{lgp_test}
\alias{lgp_test}
\title{Compute predictions and log-posterior predictive density at test points}
\usage{
lgp_test(fit, test_data, plot = FALSE, reduction = "mean",
  samples = "mean")
}
\arguments{
\item{fit}{an object of class \code{lgpfit}}

\item{test_data}{a test data matrix}

\item{plot}{should this plot the data and predictions?}

\item{reduction}{must be either "mean", "sum" or "none"}

\item{samples}{Sample indices.}
}
\value{
a ggplot object or lppd
}
\description{
This is a convenience function that wraps \code{\link{lgp_predict}},
\code{\link{compute_lppd}} and \code{\link{plot_predictions}}.
}
