% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_s4_classes.R
\docType{class}
\name{lgpfit-class}
\alias{lgpfit-class}
\alias{lgpfit}
\title{An S4 class to represent the output of the \code{lgp} function}
\description{
An S4 class to represent the output of the \code{lgp} function
}
\section{Slots}{

\describe{
\item{\code{stan_fit}}{An object of class \code{stanfit}.}

\item{\code{model}}{An object of class \code{lgpmodel}.}
}}

\seealso{
The main function \code{\link{lgp}}.
}
