% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-classes.R
\docType{class}
\name{lgpmodel-class}
\alias{lgpmodel-class}
\alias{lgpmodel}
\title{An S4 class to represent an lgp model}
\description{
An S4 class to represent an lgp model
}
\section{Slots}{

\describe{
\item{\code{data}}{The original unmodified data frame.}

\item{\code{stan_dat}}{The data to be given as input to \code{rstan::sampling}.}

\item{\code{scalings}}{Preprocessing scaling functions and their inverse operations.}

\item{\code{info}}{Model info.}
}}

