% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-model_info.R
\name{lgpmodel_methods}
\alias{lgpmodel_methods}
\alias{model_summary}
\alias{param_summary}
\title{Methods to access and print information about a model}
\usage{
model_summary(object)

param_summary(object, digits = 3)
}
\arguments{
\item{object}{an object of class \linkS4class{lgpmodel} or
\linkS4class{lgpfit}}

\item{digits}{number of digits to show for floating point numbers}
}
\description{
Methods to access and print information about a model
}
\section{Functions}{
\itemize{
\item \code{model_summary}: Print a model summary. Returns \code{object}
invisibly.

\item \code{param_summary}: Parameter summary (bounds and priors). Returns
a \code{data.frame}.
}}

