% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R, R/methods-lgpparams.R
\docType{class}
\name{lgpparams-class}
\alias{lgpparams-class}
\alias{lgpparams}
\alias{num_paramsets,lgpparams-method}
\title{An S4 class to represent a set of model parameters}
\usage{
\S4method{num_paramsets}{lgpparams}(object)
}
\description{
An S4 class to represent a set of model parameters
}
\section{Methods (by generic)}{
\itemize{
\item \code{num_paramsets}: Get number of parameter sets.
}}

\section{Slots}{

\describe{
\item{\code{num_paramsets}}{number of parameter sets}

\item{\code{origin}}{how the parameters were obtained}

\item{\code{alpha}}{marginal standard deviation parameters}

\item{\code{ell}}{lengthscale parameters}

\item{\code{wrp}}{input warping steepness parameters}

\item{\code{beta}}{heterogeneity parameters}

\item{\code{xunc}}{covariate uncertainty parameters}
}}

