% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_s4_classes.R
\docType{class}
\name{lgpproduct-class}
\alias{lgpproduct-class}
\alias{lgpproduct}
\title{An S4 class to represent a product of lgp terms}
\description{
An S4 class to represent a product of lgp terms
}
\section{Slots}{

\describe{
\item{\code{factors}}{a list of at most two \code{\link{lgpterm}}s}
}}

