% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgpr.R
\docType{package}
\name{lgpr-package}
\alias{lgpr-package}
\alias{lgpr}
\title{The 'lgpr' package.}
\description{
Longitudinal Gaussian Process regression.
The package features additive Gaussian process modeling of longitudinal
data with with interpretable covariate effects and covariate relevance
assesment. Models can include non-stationary, heterogeneous and temporally
uncertain effects. Bayesian inference of the model (hyper)parameters using
\code{\link[rstan]{rstan}}. Functions for visualizing longitudinal data,
posterior draws and model predictions are also provided.
}
\section{Basic usage}{

\itemize{
\item See documentation of the main function \code{\link{lgp}}.
 \item See tutorials at
 \url{https://jtimonen.github.io/lgpr-usage/index.html}
}
}

\section{Citation}{


\emph{An interpretable probabilistic machine learning method for
heterogeneous longitudinal studies}. Juho Timonen, Henrik Mannerstrom,
Aki Vehtari and Harri Lahdesmaki, 2019.
\url{https://arxiv.org/abs/1912.03549}
}

\references{
\enumerate{
  \item Carpenter, B. et al. (2017).
  \emph{Stan: A probabilistic programming language}. Journal of Statistical
   Software 76(1).
  \item Jonah Gabry, Ben Goodrich and Martin Lysy (2019).
  \emph{rstantools: Tools for Developing R Packages Interfacing with 'Stan'}.
  R package version 2.0.0.
  \item Gabry, J. and Mahr, T. (2019).
  \emph{bayesplot: Plotting for Bayesian Models}. R package version 1.7.0,
   http://mc-stan.org/bayesplot.
  \item Stan Development Team (2019). \emph{RStan: the R interface to Stan.}
  R package version 2.19.2. http://mc-stan.org/.
}
}
\author{
Juho Timonen (first.last at iki.fi)
}
\keyword{Gaussian}
\keyword{Stan,}
\keyword{covariate}
\keyword{data,}
\keyword{interpretable}
\keyword{longitudinal}
\keyword{models}
\keyword{processes,}
\keyword{relevances,}
