% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\docType{class}
\name{lgprhs-class}
\alias{lgprhs-class}
\alias{lgprhs}
\title{An S4 class to represent the right-hand side of an lgp formula}
\description{
An S4 class to represent the right-hand side of an lgp formula
}
\section{Slots}{

\describe{
\item{\code{summands}}{a list of one or more \linkS4class{lgpterm}s}
}}

\seealso{
See \code{\link{operations}} for performing arithmetics
on \linkS4class{lgprhs}, \linkS4class{lgpterm} and \linkS4class{lgpexpr}
objects.
}
