% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-all_classes.R
\docType{class}
\name{lgpscaling-class}
\alias{lgpscaling-class}
\alias{lgpscaling}
\title{An S4 class to represent variable scaling and its inverse}
\description{
An S4 class to represent variable scaling and its inverse
}
\section{Slots}{

\describe{
\item{\code{fun}}{function that performs normalization}

\item{\code{fun_inv}}{inverse function of \code{fun}}

\item{\code{var_name}}{variable name}
}}

