% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_s4_classes.R
\docType{class}
\name{lgpsum-class}
\alias{lgpsum-class}
\alias{lgpsum}
\title{An S4 class to represent the right-hand side of an lgp formula}
\description{
An S4 class to represent the right-hand side of an lgp formula
}
\section{Slots}{

\describe{
\item{\code{summands}}{a list of one or more \code{\link{lgpproduct}}s}
}}

