% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_s4_classes.R
\docType{class}
\name{lgpterm-class}
\alias{lgpterm-class}
\alias{lgpterm}
\title{An S4 class to represent an lgp term with a single covariate}
\description{
An S4 class to represent an lgp term with a single covariate
}
\section{Slots}{

\describe{
\item{\code{covariate}}{name of a covariate}

\item{\code{type}}{covariate type}

\item{\code{fun}}{kernel function name}
}}

