% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-utils.R
\name{likelihood_as_int}
\alias{likelihood_as_int}
\title{Convert likelihood string to Stan encoding}
\usage{
likelihood_as_int(likelihood)
}
\arguments{
\item{likelihood}{a string}
}
\value{
an integer
}
\description{
Convert likelihood string to Stan encoding
}
