% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-utils.R
\name{likelihood_as_str}
\alias{likelihood_as_str}
\title{Convert the Stan likelihood encoding to a string}
\usage{
likelihood_as_str(index)
}
\arguments{
\item{index}{an integer}
}
\value{
a string
}
\description{
Convert the Stan likelihood encoding to a string
}
