% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-utils.R
\name{likelihood_encoding}
\alias{likelihood_encoding}
\alias{likelihood_list}
\alias{likelihood_as_str}
\alias{likelihood_as_int}
\alias{is_bin_or_bb}
\alias{is_pois_or_nb}
\title{Integer encoding of likelihod function names}
\usage{
likelihood_list()

likelihood_as_str(index)

likelihood_as_int(likelihood)

is_bin_or_bb(likelihood)

is_pois_or_nb(likelihood)
}
\arguments{
\item{index}{an integer}

\item{likelihood}{a string}
}
\description{
\itemize{
  \item \code{likelihood_as_int} converts likelihood name to Stan encoding
  \item \code{likelihood_as_str} converts the Stan likelihood encoding
  to a string
  \item \code{likelihood_list} returns the available likelihood names
}
}
