% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{line_alpha_fun}
\alias{line_alpha_fun}
\title{Line alpha decider function}
\usage{
line_alpha_fun(num_lines)
}
\arguments{
\item{num_lines}{number of lines to draw}
}
\value{
a value between 0 and 1
}
\description{
A function that takes as parameter the number of lines and
returns a line alpha value (between 0 and 1).
}
