% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{link}
\alias{link}
\alias{link_inv}
\title{Link functions and their inverses}
\usage{
link(x, likelihood)

link_inv(x, likelihood)
}
\arguments{
\item{x}{the input}

\item{likelihood}{name of the likelihood model}
}
\value{
transformed input
}
\description{
Link functions and their inverses
}
