% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_list.R
\name{list_if_named}
\alias{list_if_named}
\title{Wrap list into a list of length 1 if the original list is named}
\usage{
list_if_named(x)
}
\arguments{
\item{x}{a list}
}
\value{
a list with no names
}
\description{
Wrap list into a list of length 1 if the original list is named
}
\seealso{
Other list utilities: 
\code{\link{dollar}()},
\code{\link{list_to_matrix}()},
\code{\link{matrix_to_list}()}
}
\concept{list utilities}
