% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict-main.R
\name{log_gaussian_density}
\alias{log_gaussian_density}
\title{Compute log-density for gaussian distribution}
\usage{
log_gaussian_density(x, mu, s2)
}
\arguments{
\item{x}{point x}

\item{mu}{mean}

\item{s2}{variance}
}
\value{
a number
}
\description{
Compute log-density for gaussian distribution
}
