% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_array.R
\name{make_draw_df}
\alias{make_draw_df}
\title{Helper function for creating plot data frames from prediction objects}
\usage{
make_draw_df(x)
}
\arguments{
\item{x}{an array with \code{num_draws} rows}
}
\value{
a data frame with \code{num_draws} columns
}
\description{
Helper function for creating plot data frames from prediction objects
}
\seealso{
Other array utilities: 
\code{\link{add_to_columns}()},
\code{\link{apply_reduce}()},
\code{\link{arr3_select}()},
\code{\link{arr3_to_list}()},
\code{\link{array_to_arraylist}()},
\code{\link{ensure_2dim}()},
\code{\link{normalize_rows}()},
\code{\link{null_if_all_nan}()},
\code{\link{reduce_rows}()},
\code{\link{repvec}()},
\code{\link{row_vars}()},
\code{\link{select_row}()},
\code{\link{squeeze_second_dim}()}
}
\concept{array utilities}
