% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\docType{class}
\name{MarginalGPModelParams-class}
\alias{MarginalGPModelParams-class}
\alias{MarginalGPModelParams}
\title{An S4 class to represent a set of marginal GP model parameters}
\description{
An S4 class to represent a set of marginal GP model parameters
}
\section{Slots}{

\describe{
\item{\code{sigma}}{noise standard deviation}
}}

