% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgp-formula.R
\name{mask}
\alias{mask}
\title{Create a mask term}
\usage{
mask(covariate)
}
\arguments{
\item{covariate}{name of a categorical covariate}
}
\value{
an object of class \code{\link{lgpterm}}
}
\description{
Create a mask term
}
