% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-utils.R
\name{matrix_to_df}
\alias{matrix_to_df}
\title{Matrix to data frame without editing column names}
\usage{
matrix_to_df(M)
}
\arguments{
\item{M}{a matrix}
}
\value{
a data frame
}
\description{
Matrix to data frame without editing column names
}
