% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-utils.R
\name{matrix_to_list}
\alias{matrix_to_list}
\title{Matrix rows to a list}
\usage{
matrix_to_list(x)
}
\arguments{
\item{x}{a matrix or array with \code{m} rows and \code{n} columns}
}
\value{
an unnamed list of length \code{m} where each element is a
vector of length \code{n}
}
\description{
Matrix rows to a list
}
