% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-test_data.R
\name{minimal_stan_data}
\alias{minimal_stan_data}
\title{Create minimal data for Stan}
\usage{
minimal_stan_data()
}
\value{
A list of data suitable for testing \code{lgp.stan}
}
\description{
Create minimal data for Stan
}
