% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{minus}
\alias{minus}
\alias{minus.append}
\alias{minus.prepend}
\title{Add minus to a string depending on options}
\usage{
minus.append(str, val)

minus.prepend(str, prepend)
}
\arguments{
\item{str}{a string}

\item{val}{a value to append with minus (will not be appended if value
is zero)}

\item{prepend}{should a minus be prepended (true if this is not zero)}
}
\description{
Add minus to a string depending on options
}
