% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-lgpmodel.R
\name{model_getters}
\alias{model_getters}
\alias{get_covariate_names}
\alias{get_stan_input}
\alias{get_component_info}
\alias{get_component_names}
\alias{get_num_obs}
\alias{is_f_sampled}
\title{Functions that access model properties}
\usage{
get_covariate_names(object, type = "all")

get_stan_input(object)

get_component_info(object)

get_component_names(object)

get_num_obs(object)

is_f_sampled(object)
}
\arguments{
\item{object}{an object of class \linkS4class{lgpmodel} or
\linkS4class{lgpfit}}

\item{type}{What types of covariates to include. Must be either
"all", "categorical" or "continuous".}
}
\value{
\itemize{
  \item \code{get_covariate_names} returns a string where names
  are separated by a comma
  \item \code{get_stan_input} returns a list
  \item \code{get_component_info} returns a data frame
  \item \code{get_component_names} returns an array of names
  \item \code{get_num_obs} returns an integer
  \item \code{is_f_sampled} returns a boolean value
}
}
\description{
Functions that access model properties
}
