% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-utils.R
\name{model_info}
\alias{model_info}
\title{Get model info}
\usage{
model_info(object, print = TRUE)
}
\arguments{
\item{object}{an object of class \code{lgpmodel} or \code{lgpfit}}

\item{print}{should this print the info?}
}
\value{
the info as a string
}
\description{
Get model info
}
