% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgpmodel-class.R
\name{model_summaries}
\alias{model_summaries}
\alias{model_summary}
\alias{param_summary}
\title{Model summary}
\usage{
model_summary(object)

param_summary(object, digits = 3)
}
\arguments{
\item{object}{a model or fit}

\item{digits}{number of digits to show for floating point numbers}
}
\description{
Model summary
}
\section{Functions}{
\itemize{
\item \code{model_summary}: Prints a model summary. Returns \code{object} 
invisibly.

\item \code{param_summary}: Get parameter summary (bounds and priors).
Returns a \code{data.frame}.
}}

