% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-lgpmodel.R
\name{model_summary}
\alias{model_summary}
\alias{model_summary_brief}
\alias{print_stan_input}
\title{Print a model summary}
\usage{
model_summary(object)

model_summary_brief(object)

print_stan_input(object)
}
\arguments{
\item{object}{an object of class \linkS4class{lgpmodel} or
\linkS4class{lgpfit}}
}
\value{
\itemize{
  \item \code{\link{model_summary_brief}} returns a string
  \item \code{\link{model_summary}} prints the summary and
returns \code{object} invisibly
  \item \code{\link{print_stan_input}} prints the stan input
  and returns \code{object} invisibly
}
}
\description{
Print a model summary
}
