% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate-create_F.R
\name{nameComponents}
\alias{nameComponents}
\title{Create names for all components based on covariate names and types}
\usage{
nameComponents(types, names)
}
\arguments{
\item{types}{vector of covariate types}

\item{names}{names of the covariates}
}
\value{
a vector of component names
}
\description{
Create names for all components based on covariate names and types
}
