% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-datautil.R
\name{new_x}
\alias{new_x}
\title{Create prediction points}
\usage{
new_x(data, x_values, group_by = "id", x = "age", x_ns = NULL)
}
\arguments{
\item{data}{a data frame}

\item{x_values}{the values of \code{x} to set for each individual}

\item{group_by}{name of the grouping varible, must be a factor
in \code{data} (or use \code{group_by=NA} to create a dummy grouping
factor which has only one value)}

\item{x}{of the variable along which to extend,
must be a numeric in \code{data}}

\item{x_ns}{of a nonstationary variable derived from \code{x},
must be a numeric in \code{data}}
}
\value{
a data frame containing the following columns
\itemize{
 \item all factors in the original \code{data}
 \item \code{x}
 \item \code{x_ns} (unless it is NULL)
}
}
\description{
Replaces a continuous variable \code{x} in the data frame, and
possibly another continuous variable \code{x_ns} derived from it, with new
values, for each level of a grouping factor (usually id)
}
\seealso{
Other data utilities: 
\code{\link{add_dis_age}()},
\code{\link{add_factor_crossing}()},
\code{\link{add_factor}()},
\code{\link{data_types}()},
\code{\link{get_teff_obs}()},
\code{\link{pick_one_row_each}()},
\code{\link{rep_df}()},
\code{\link{select_factors_and}()},
\code{\link{split}()},
\code{\link{to_long_format}()}
}
\concept{data utilities}
