% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relevances.R
\name{noise_proportion}
\alias{noise_proportion}
\title{Determine noise proportion for each sample}
\usage{
noise_proportion(f_total, y, likelihood)
}
\arguments{
\item{f_total}{an array of shape \code{num_samples} x \code{num_obs}}

\item{y}{an array of shape \code{1} x \code{num_obs}}

\item{likelihood}{used likelihood model name}
}
\value{
a value between 0 and 1
}
\description{
Determine noise proportion for each sample
}
