% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{normalize_rows}
\alias{normalize_rows}
\title{Normalize matrix or data frame rows so that they sum to 1}
\usage{
normalize_rows(x)
}
\arguments{
\item{x}{an array of size \code{n} x \code{m}}
}
\value{
an array of size \code{n} x \code{m}
}
\description{
Normalize matrix or data frame rows so that they sum to 1
}
\seealso{
Other array utilities: 
\code{\link{array_to_arraylist}()},
\code{\link{ensure_2dim}()},
\code{\link{null_if_all_nan}()},
\code{\link{reduce_rows}()},
\code{\link{repvec}()},
\code{\link{row_vars}()},
\code{\link{select_row}()},
\code{\link{squeeze_second_dim}()}
}
\concept{array utilities}
