% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-utils.R
\name{null_if_all_nan}
\alias{null_if_all_nan}
\title{Return NULL if vector contains only NaN values}
\usage{
null_if_all_nan(x)
}
\arguments{
\item{x}{a vector}
}
\value{
\code{x} unchanged or \code{NULL}
}
\description{
Return NULL if vector contains only NaN values
}
