% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{object_to_model}
\alias{object_to_model}
\title{Helper function for generic functions}
\usage{
object_to_model(object)
}
\arguments{
\item{object}{an object of class \linkS4class{lgpmodel} or
\linkS4class{lgpfit}}
}
\value{
an object of class \linkS4class{lgpmodel}
}
\description{
Helper function for generic functions that work on
both of \linkS4class{lgpmodel} and \linkS4class{lgpfit} class objects.
}
