% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate-create_X.R
\name{onsetsToDiseaseAge}
\alias{onsetsToDiseaseAge}
\title{Compute the disease-related ages}
\usage{
onsetsToDiseaseAge(onsets, age, k)
}
\arguments{
\item{onsets}{true disease onsets, a vector of length \code{N}}

\item{age}{the age covariate, a vector of length \code{N*k}}

\item{k}{number of measurements per individual}
}
\value{
the diseaseAge covariate, a vector of length \code{N*k}
}
\description{
Compute the disease-related ages
}
