% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-lgpformula.R
\name{operations}
\alias{operations}
\alias{+,lgprhs,lgprhs-method}
\alias{+,lgpterm,lgpterm-method}
\alias{+,lgprhs,lgpterm-method}
\alias{*,lgpterm,lgpterm-method}
\title{Operations on formula terms and expressions}
\usage{
\S4method{+}{lgprhs,lgprhs}(e1, e2)

\S4method{+}{lgpterm,lgpterm}(e1, e2)

\S4method{+}{lgprhs,lgpterm}(e1, e2)

\S4method{*}{lgpterm,lgpterm}(e1, e2)
}
\arguments{
\item{e1}{The first sum, term or expression}

\item{e2}{The second sum, term or expression}
}
\value{
The behaviour and return type depend on the types of \code{e1} and \code{e2}.
You can
\itemize{
  \item sum (\code{+}) two \linkS4class{lgprhs}'s to yield an
  \linkS4class{lgprhs}
  \item sum (\code{+}) two \linkS4class{lgpterm}'s to yield an
  \linkS4class{lgprhs}
  \item sum (\code{+}) an \linkS4class{lgprhs} and an \linkS4class{lgpterm}
  to yield an \linkS4class{lgprhs}
  \item multiply (\code{*}) two \linkS4class{lgpterm}'s to yield
  an \linkS4class{lgpterm}
}
}
\description{
Operations on formula terms and expressions
}
