% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relevances.R
\name{p_explained}
\alias{p_explained}
\alias{residuals}
\alias{p_explained_noise}
\alias{p_explained_components}
\alias{component_variances}
\title{Determining the variance explained by noise and the signal components}
\usage{
residuals(fit)

p_explained_noise(fit)

p_explained_components(fit)

component_variances(fit)
}
\arguments{
\item{fit}{an object of class \code{lgpfit}}
}
\value{
an array or vector
}
\description{
\itemize{
  \item \code{residuals} computes the residuals in each sample
  \item \code{p_explained_noise} computes the noise proportion in each sample
  \item \code{p_explained_components} divides the signal proportion for
  each component
  \item \code{component_variances} computes variances in each
  sample for each component and returns a \code{data.frame}
}
}
