% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-model_methods.R
\name{param_summary}
\alias{param_summary}
\alias{prior_summary}
\title{Parameter summary (priors etc.)}
\usage{
param_summary(object, digits = 3)

prior_summary(object, digits = 3)
}
\arguments{
\item{object}{an object of class \linkS4class{lgpmodel} or
\linkS4class{lgpfit}}

\item{digits}{number of digits to show for floating point numbers}
}
\value{
data frame
}
\description{
Parameter summary (priors etc.)
}
