% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{parse_covs_and_comps}
\alias{parse_covs_and_comps}
\title{Parse the covariates and model components from given data and formula}
\usage{
parse_covs_and_comps(data, model_formula, x_cont_scl)
}
\arguments{
\item{data}{A \code{data.frame} where each column corresponds to one
variable, and each row is one observation. Continuous covariates and the
response variable must have type \code{"numeric"} and categorical covariates
must have type \code{"factor"}. Missing values should be indicated with
\code{NaN} or \code{NA}. The response variable cannot contain missing
values.}

\item{model_formula}{An object of class \code{lgpformula}.}

\item{x_cont_scl}{Information on how to scale the continuous covariates.
This can either be
\itemize{
  \item an existing list of objects with class \linkS4class{lgpscaling}, or
  \item \code{NA}, in which case such list is created by computing mean
  and standard deviation from \code{data}
}}
}
\value{
parsed input to stan and covariate scaling
}
\description{
Parse the covariates and model components from given data and formula
}
