% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-data.R
\name{parse_data}
\alias{parse_data}
\title{Parse the given data}
\usage{
parse_data(data, formula)
}
\arguments{
\item{data}{A \code{data.frame} where each column corresponds to one
variable, and each row is one observation. Continuous covariates must have
type \code{"numeric"} and categorical ones must have type \code{"factor"}.
Missing values should be indicated with \code{NaN} or \code{NA}.}
}
\value{
a named list of parsed options
}
\description{
Parse the given data
}
