% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-options.R
\name{parse_disease_options}
\alias{parse_disease_options}
\title{Parse the given modeling options for the disease component}
\usage{
parse_disease_options(disease_options = NULL)
}
\arguments{
\item{disease_options}{A named list with the following possible fields:
\itemize{
  \item \code{heterogeneous} Is the disease effect magnitude assumed to be
  heterogeneous over the diseased individuals?
  \item \code{uncertain} Do we wish to model uncertainty in the disease
effect time?
  \item \code{vm_params} Parameters of the variance masking function. Use
  \code{NA} to avoid using variance masking at all.
}}
}
\value{
a named list of parsed options
}
\description{
Parse the given modeling options for the disease component
}
