% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula.R
\name{parse_expr}
\alias{parse_expr}
\title{Parse a string representation of one formula expression}
\usage{
parse_expr(expr)
}
\arguments{
\item{expr}{the expression as a string, e.g. \code{"gp(x)"}}
}
\value{
an object of class \linkS4class{lgpexpr}
}
\description{
Parse a string representation of one formula expression
}
