% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-formula.R
\name{parse_formula}
\alias{parse_formula}
\title{Create a model formula}
\usage{
parse_formula(formula)
}
\arguments{
\item{formula}{A formula specified using the common \code{\link{formula}}
syntax, such as \code{y ~ x1 + x2:z1 + } \code{x2:z2 + z2}.
\itemize{
  \item The formula must contain exatly one tilde (\code{~}), with response
  variable on the left-hand side and model terms on the right-hand side.
  \item Terms are be separated by a plus (\code{+}) sign.
  \item Terms can consits of a single variable name, such as \code{x}, or
  an interaction of two variables, such as \code{x:z}.
  \item In single-variable terms, the variable can be either continuous or
  categorical, whereas in interaction terms the variable
  on the left-hand side of the colon (\code{:}) has to be continuous and the
  one on the right-hand side has to be categorical (a factor).
  \item All variables appearing in \code{formula} must be
  found in \code{data}.
}}
}
\value{
an object of class \code{\link{lgpformula}}
}
\description{
Create a model formula
}
