% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-parse_formula.R
\name{parse_formula_advanced}
\alias{parse_formula_advanced}
\title{Parse a model formula that uses the advanced formula syntax}
\usage{
parse_formula_advanced(formula)
}
\arguments{
\item{formula}{The model formula, where
\itemize{
  \item it must contain exatly one tilde (\code{~}), with response
  variable on the left-hand side and model terms on the right-hand side
  \item terms are be separated by a plus (\code{+}) sign
  \item all variables appearing in \code{formula} must be
  found in \code{data}
}
See the "Model formula syntax" section below (\code{\link{lgp}}) for
instructions on how to specify the model terms.}
}
\value{
an object of class \linkS4class{lgpformula}
}
\description{
Parse a model formula that uses the advanced formula syntax
}
\seealso{
Other advanced formula parsers: 
\code{\link{parse_expr}()},
\code{\link{parse_rhs}()},
\code{\link{parse_term}()},
\code{\link{rhs_variables}()},
\code{\link{term_variables}()}
}
\concept{advanced formula parsers}
