% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-parse_prior.R
\name{parse_prior}
\alias{parse_prior}
\title{Parse given prior}
\usage{
parse_prior(prior, stan_input, verbose)
}
\arguments{
\item{prior}{A named list, defining the prior distribution of model
(hyper)parameters. See the "Defining priors" section below
(\code{\link{lgp}}).}

\item{stan_input}{a list of stan input fields}

\item{verbose}{Should some informative messages be printed?}
}
\value{
a named list of parsed options
}
\description{
Parse given prior
}
\seealso{
Other internal model creation functions: 
\code{\link{parse_covs_and_comps}()},
\code{\link{parse_formula}()},
\code{\link{parse_y}()}
}
\concept{internal model creation functions}
