% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preproc-parse_prior.R
\name{parse_prior_distribution}
\alias{parse_prior_distribution}
\title{Turn a list describing a prior distribution into vectors to be given to Stan}
\usage{
parse_prior_distribution(dist, add_correct = NULL)
}
\arguments{
\item{dist}{a list with field type, and possibly others}

\item{add_correct}{additional correct parameter names}
}
\value{
a list with two vectors to be given to Stan
}
\description{
Turn a list describing a prior distribution into vectors to be given to Stan
}
