% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-priors_parse.R
\name{parse_prior_full}
\alias{parse_prior_full}
\alias{parse_prior_full_pos}
\alias{parse_prior_full_unit}
\alias{parse_prior_full_teff}
\title{Parse a fully defined prior}
\usage{
parse_prior_full(prior, stan_input, obs_model)

parse_prior_full_pos(prior, stan_input, obs_model)

parse_prior_full_unit(prior, stan_input, obs_model)

parse_prior_full_teff(prior, stan_input)
}
\arguments{
\item{prior}{A named list, defining the prior distribution of model
(hyper)parameters. See the "Defining priors" section below
(\code{\link{lgp}}).}

\item{stan_input}{a list of stan input fields}

\item{obs_model}{observation model as integer}
}
\value{
a named list to be given to Stan
}
\description{
Parse a fully defined prior
}
\seealso{
Other prior parsers: 
\code{\link{parse_prior_single}()},
\code{\link{position_hyper_params}()},
\code{\link{prior_to_num}()}
}
\concept{prior parsers}
