% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors_parse.R
\name{parse_prior_single}
\alias{parse_prior_single}
\title{Parse the given prior for a single parameter type}
\usage{
parse_prior_single(desc, num)
}
\arguments{
\item{desc}{An unnamed list of prior descriptions.}

\item{num}{number of parameters of this type}
}
\value{
a named list of parsed options
}
\description{
Parse the given prior for a single parameter type
}
\seealso{
Other prior parsers: 
\code{\link{parse_prior_full}()},
\code{\link{position_hyper_params}()},
\code{\link{prior_to_num}()}
}
\concept{prior parsers}
