% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preproc-parse_prior.R
\name{parse_prior_t_effect}
\alias{parse_prior_t_effect}
\title{Turn a list describing an effect time distribution into
things to be given to Stan}
\usage{
parse_prior_t_effect(dist, N_cases, T_observed, T_last, UNCRT)
}
\arguments{
\item{dist}{This is \code{prior$effect}, where \code{prior} is an argument of
\code{lgp_model}}

\item{N_cases}{number of case individuals}

\item{T_observed}{observed disease onsets / initiation times}

\item{T_last}{last time point for each diseased individual}

\item{UNCRT}{0 or 1}
}
\value{
a list with things to be given to Stan
}
\description{
Turn a list describing an effect time distribution into
things to be given to Stan
}
