% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula.R
\name{parse_whitespace}
\alias{parse_whitespace}
\title{Remove quotes and whitespace from a formula}
\usage{
parse_whitespace(rhs)
}
\arguments{
\item{rhs}{The formula right-hand side in text format.}
}
\value{
a character string
}
\description{
Remove quotes and whitespace from a formula
}
\seealso{
Other model formula functions: 
\code{\link{formula_to_advanced}()},
\code{\link{is_advanced_formula}()},
\code{\link{parse_expr}()},
\code{\link{parse_formula}()},
\code{\link{parse_rhs}()},
\code{\link{parse_term}()},
\code{\link{rhs_to_terms}()},
\code{\link{rhs_variables}()},
\code{\link{term_variables}()}
}
\concept{model formula functions}
