% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-parse_y.R
\name{parse_y_helper}
\alias{parse_y_helper}
\alias{parse_y.marginal}
\alias{parse_y.latent}
\title{Subroutines for parsing the response variable and its likelihood model}
\usage{
parse_y.marginal(Y_RAW, y_name)

parse_y.latent(Y_RAW, y_name, LH, c_hat, num_trials)
}
\arguments{
\item{Y_RAW}{raw response variable taken from the user-supplied data frame}

\item{y_name}{Name of response variable}

\item{LH}{integer encoding of likelihood}

\item{c_hat}{The GP mean. This should only be given if \code{sample_f} is
\code{TRUE}, otherwise the GP will always have zero mean. If \code{sample_f}
is \code{TRUE}, the given \code{c_hat} can be a vector of length
\code{dim(data)[1]}, or a real number defining a constant GP mean. If not
specified and \code{sample_f} is \code{TRUE}, \code{c_hat} is set to
 \itemize{
   \item \code{c_hat = mean(y)}, if \code{likelihood} is \code{"gaussian"},
   \item \code{c_hat = } \code{log(mean(y))} if \code{likelihood} is
   \code{"poisson"} or \code{"nb"},
   \item \code{c_hat = } \code{log(p/(1-p))}, where
   \code{p = mean(y/num_trials)} if \code{likelihood} is \code{"binomial"}
   or \code{"bb"},
 }
where \code{y} denotes the response variable measurements.}

\item{num_trials}{This argument (number of trials) is only needed when
likelihood is \code{"binomial"} or \code{"bb"}. Must have length one or
equal to the number of data points. Setting \code{num_trials=1} and
\code{likelihood="binomial"} corresponds to Bernoulli observation model.}
}
\description{
Subroutines for parsing the response variable and its likelihood model
}
