% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_data.R
\name{pick_one_row_each}
\alias{pick_one_row_each}
\title{For each unique value of a factor, pick one row from data}
\usage{
pick_one_row_each(data, fac)
}
\arguments{
\item{data}{a data frame}

\item{fac}{name of a factor in \code{data}}
}
\value{
a data frame
}
\description{
For each unique value of a factor, pick one row from data
}
\seealso{
Other data utilities: 
\code{\link{add_dis_age}()},
\code{\link{add_factor_crossing}()},
\code{\link{add_factor}()},
\code{\link{data_types}()},
\code{\link{get_teff_obs}()},
\code{\link{new_x}()},
\code{\link{rep_df}()},
\code{\link{select_factors_and}()},
\code{\link{split}()},
\code{\link{to_long_format}()},
\code{\link{valid_data}}
}
\concept{data utilities}
