% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethodsFit.R
\name{plot,lgpfit,missing-method}
\alias{plot,lgpfit,missing-method}
\title{Visualize parameter draws}
\usage{
\S4method{plot}{lgpfit,missing}(x, y, ...)
}
\arguments{
\item{x}{an object of class \linkS4class{lgpfit}}

\item{y}{not used}

\item{...}{keyword arguments passed to \code{\link{plot_draws}}}
}
\value{
a \code{ggplot} object
}
\description{
Visualize parameter draws
}
\seealso{
Other model fit visualization functions: 
\code{\link{plot_draws}()},
\code{\link{plot_fit}()}
}
\concept{model fit visualization functions}
