% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-lgpsim.R
\name{plot,lgpsim,missing-method}
\alias{plot,lgpsim,missing-method}
\title{Visualize a simulated longitudinal data set}
\usage{
\S4method{plot}{lgpsim,missing}(x, y, ...)
}
\arguments{
\item{x}{an object of class \linkS4class{lgpsim}}

\item{y}{not used}

\item{...}{keyword arguments passed to \code{\link{sim_plot}}}
}
\value{
a \code{ggplot} object
}
\description{
Visualize a simulated longitudinal data set
}
