% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgpr.R
\name{plot.lgpr}
\alias{plot.lgpr}
\title{Plot method for lgpr object}
\usage{
\method{plot}{lgpr}(x, index = 1, normalize = T, digits = 3, label = NULL, ...)
}
\arguments{
\item{x}{a lgpr object}

\item{index}{the index of the prediction to be plotted. Default is 1.}

\item{normalize}{a logical stating if the density (y-axis) should be
normalized (values are divided by the maximum value). Default is TRUE}

\item{digits}{an integer defining the precision of the round function.
Default is 3.}

\item{label}{a character defining the label of the x-axis. Default is NULL.}

\item{...}{...}
}
\value{
a ggplot2 graphical object.
}
\description{
Plot method for lgpr object
}
\author{
David Senhora Navega
}
