% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot}
\alias{plot}
\title{Visualize S4 class objects using the plot generic}
\usage{
\S4method{plot}{lgpfit,ANY}(fit, x = 1, y = 1, color_scheme = "red")
}
\arguments{
\item{fit}{an object of class \linkS4class{lgpfit}}

\item{x}{does nothing}

\item{y}{does nothing}

\item{color_scheme}{bayesplot color scheme}
}
\value{
a ggplot object
}
\description{
Visualize S4 class objects using the plot generic
}
