% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-plot.R
\name{plot_api_g_add_effect_times}
\alias{plot_api_g_add_effect_times}
\title{Add real or observed disease effect times to a grouped plot}
\usage{
plot_api_g_add_effect_times(h, teff, group_by, linecolor, linetype, lwd)
}
\arguments{
\item{h}{a \code{ggplot} object}

\item{teff}{a named vector}

\item{group_by}{name of grouping variable}

\item{linecolor}{line color}

\item{linetype}{line type}

\item{lwd}{line width}
}
\value{
updated \code{ggplot} object
}
\description{
Add real or observed disease effect times to a grouped plot
}
\seealso{
Other plot_api_g functions: 
\code{\link{plot_api_g_create_aes}()},
\code{\link{plot_api_g_create}()}
}
\concept{plot_api_g functions}
