% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_api.R
\name{plot_api_g_create_aes}
\alias{plot_api_g_create_aes}
\title{Helper function for plot_grouped}
\usage{
plot_api_g_create_aes(data, type_name)
}
\arguments{
\item{data}{a data frame with three columns}

\item{type_name}{name of a factor that is to be distinguished by
plot marker and/or type}
}
\value{
a ggplot aes object
}
\description{
Helper function for plot_grouped
}
\seealso{
Other plot_api_g functions: 
\code{\link{plot_api_g_add_effect_times}()},
\code{\link{plot_api_g_create}()}
}
\concept{plot_api_g functions}
