% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-samples.R
\name{plot_beta}
\alias{plot_beta}
\title{Visualize posterior samples of individual-specific disease effect magnitude
parameters}
\usage{
plot_beta(fit, color_scheme = "red", threshold = 0.5)
}
\arguments{
\item{fit}{An object of class \code{lgpfit}.}

\item{color_scheme}{Name of bayesplot color scheme.}

\item{threshold}{Threshold for median.}
}
\value{
a ggplot object
}
\description{
Can only be used if the disease effect was modeled
heterogeneously.
}
\seealso{
Other model fit visualization functions: 
\code{\link{plot,lgpfit,missing-method}},
\code{\link{plot_effect_times}()},
\code{\link{plot_fit}()},
\code{\link{plot_posterior_warp}()},
\code{\link{plot_posterior}()}
}
\concept{model fit visualization functions}
