% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_color_palette}
\alias{plot_color_palette}
\title{Visualize a color palette}
\usage{
plot_color_palette(n)
}
\arguments{
\item{n}{an integer from 1 to 6}
}
\value{
a \code{ggplot} object
}
\description{
Visualize a color palette
}
\seealso{
Other color utilities: 
\code{\link{color_palette}()},
\code{\link{colorset}()}
}
\concept{color utilities}
