% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-components.R
\name{plot_component}
\alias{plot_component}
\title{Visualize one (average) inferred component evaluated at data points}
\usage{
plot_component(fit, idx, sample_idx = NULL, linealpha = 1,
  linetype = 3)
}
\arguments{
\item{fit}{An object of class \code{lgpfit}.}

\item{idx}{Index of component to be plotted.}

\item{sample_idx}{If given, only one sample will be plotted, else the average
components over all samples.}

\item{linealpha}{line alpha}

\item{linetype}{line type}
}
\value{
a ggplot object
}
\description{
Visualize one (average) inferred component evaluated at data points
}
