% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-components.R
\name{plot_component_create_df}
\alias{plot_component_create_df}
\title{Create dataframe for componentwise plotting}
\usage{
plot_component_create_df(FFF, groupvar, colorvar, timevar)
}
\arguments{
\item{FFF}{a matrix of size n_samples x n_data}

\item{groupvar}{grouping variable}

\item{colorvar}{coloring variable}

\item{timevar}{time variable}
}
\value{
a data frame
}
\description{
Create dataframe for componentwise plotting
}
