% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-components.R
\name{plot_component_create_gg}
\alias{plot_component_create_gg}
\title{Subroutine for plot_component}
\usage{
plot_component_create_gg(EFF, id, age, idx, D, covariate_names, X)
}
\arguments{
\item{EFF}{a matrix of size n_samples x n_data}

\item{id}{id covariate}

\item{age}{age covariate}

\item{idx}{covariate index}

\item{D}{an integer vector of length 6}

\item{covariate_names}{covariate names}

\item{X}{covariate matrix}
}
\value{
a ggplot object
}
\description{
Subroutine for plot_component
}
