% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-posterior-main.R
\name{plot_components_posterior}
\alias{plot_components_posterior}
\title{Deprecated function. Use plot_posterior_components instead.}
\usage{
plot_components_posterior(fit, ...)
}
\arguments{
\item{fit}{fit}

\item{...}{other arguments}
}
\value{
Nothing. Throws an error when called.
}
\description{
Deprecated function. Use plot_posterior_components instead.
}
