% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-components.R
\name{plot_components_posterior_sub2}
\alias{plot_components_posterior_sub2}
\title{Helper for \code{\link{plot_components_posterior}}}
\usage{
plot_components_posterior_sub2(fit, PRED, sample_idx, time_is_xvar, n_sd, ...)
}
\arguments{
\item{fit}{An object of class \code{lgpfit}.}

\item{PRED}{object returned by \code{\link{lgp_predict}}}

\item{sample_idx}{Which sample to plot.}

\item{time_is_xvar}{is the time variable the x-axis variable
in all subplots?}

\item{n_sd}{number of standard deviations (ribbon width)}

\item{...}{additional arguments for \code{\link{plot_components}}}
}
\value{
an object returned by ggpubr::ggarrange or a list
}
\description{
Helper for \code{\link{plot_components_posterior}}
}
