% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-spaghetti.R
\name{plot_data}
\alias{plot_data}
\title{A spaghetti plot of longitudinal data.}
\usage{
plot_data(data, highlight_col = NULL, highlight_lty = NULL,
  highlight_cont = NULL, response = "y", palette = "Set1",
  psize = 2, title = NULL)
}
\arguments{
\item{data}{A data frame. Must contain columns \code{id} and \code{age}.}

\item{highlight_col}{Name of a categorical covariate to be highlighted with color.}

\item{highlight_lty}{Name of a categorical covariate to be highlighted with line style.}

\item{highlight_cont}{Name of a continuous covariate to be highlighted.}

\item{response}{Name of the response variable (\code{default ="y"}).}

\item{palette}{Name of a ColorBrewer palette.}

\item{psize}{Point}

\item{title}{optional prefix to plot title}
}
\value{
a ggplot object
}
\description{
A spaghetti plot of longitudinal data.
}
