% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-data.R
\name{plot_data_add_fit}
\alias{plot_data_add_fit}
\title{Add model fit to data plot}
\usage{
plot_data_add_fit(h, fit)
}
\arguments{
\item{h}{the current \code{ggplot} object}

\item{fit}{An object of class \linkS4class{lgpfit}. If this is provided,
also the model fit is plotted.}
}
\value{
a modified \code{ggplot} object
}
\description{
Add model fit to data plot
}
