% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-plot_data.R
\name{plot_data_add_highlight_factor}
\alias{plot_data_add_highlight_factor}
\title{Add factor to data frame for highlighting in plot}
\usage{
plot_data_add_highlight_factor(df, group_by, highlight)
}
\arguments{
\item{df}{data frame}

\item{group_by}{Name of grouping variable (must be a factor).}

\item{highlight}{Value of category of the \code{group_by} variable
that is highlighted. Can only be used if \code{color_by} is \code{NULL}.}
}
\value{
a list
}
\description{
Add factor to data frame for highlighting in plot
}
