% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-plot_data.R
\name{plot_data_aes}
\alias{plot_data_aes}
\title{Create aes for plot_data}
\usage{
plot_data_aes(x_name, y_name, group_by, color_by)
}
\arguments{
\item{x_name}{Name of x-axis variable.}

\item{y_name}{Name of the y-axis variable.}

\item{group_by}{Name of grouping variable (must be a factor).}

\item{color_by}{Name of coloring variable (must be a factor).}
}
\value{
an \code{aes} object
}
\description{
Create aes for plot_data
}
