% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-spaghetti.R
\name{plot_data_hl_cont}
\alias{plot_data_hl_cont}
\title{A spaghetti plot of longitudinal data, highlighting a continuous covariate.}
\usage{
plot_data_hl_cont(data, highlight = NULL, response = "y",
  id_variable = "id", time_variable = "age", psize = 2, lwd = 0.5,
  colgrad = ggplot2::scale_colour_gradient2())
}
\arguments{
\item{data}{A data frame.}

\item{highlight}{Name of a continuous covariate to be highlighted with color.}

\item{response}{Name of the response variable.}

\item{id_variable}{Name of id variable.}

\item{time_variable}{Name of time variable.}

\item{psize}{point size}

\item{lwd}{line width}

\item{colgrad}{color gradient}
}
\value{
a ggplot object
}
\description{
A spaghetti plot of longitudinal data, highlighting a continuous covariate.
}
