% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-data.R
\name{plot_data_hl_individual}
\alias{plot_data_hl_individual}
\title{A spaghetti plot of longitudinal data, highlighting one individual.}
\usage{
plot_data_hl_individual(data, highlight = 1, response = "y",
  id_variable = "id", time_variable = "age", psize = 2, lwd = 0.5)
}
\arguments{
\item{data}{A data frame.}

\item{highlight}{Number indicating the individual to highlight.}

\item{response}{Name of the response variable.}

\item{id_variable}{Name of id variable.}

\item{time_variable}{Name of time variable.}

\item{psize}{point size}

\item{lwd}{line width}
}
\value{
a ggplot object
}
\description{
A spaghetti plot of longitudinal data, highlighting one individual.
}
