% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-datautil.R
\name{plot_data_num_colors}
\alias{plot_data_num_colors}
\title{Get number of distinct colors needsd by plot_data}
\usage{
plot_data_num_colors(data, color_by)
}
\arguments{
\item{data}{A data frame.}

\item{color_by}{Name of coloring variable (must be a factor).}
}
\value{
a list
}
\description{
Get number of distinct colors needsd by plot_data
}
