% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-plot_data.R
\name{plot_data_titles}
\alias{plot_data_titles}
\title{Create titles for plot_data}
\usage{
plot_data_titles(main, sub, data, group_by)
}
\arguments{
\item{main}{main plot title}

\item{sub}{plot subtitle}

\item{data}{A data frame.}

\item{group_by}{Name of grouping variable (must be a factor).}
}
\value{
a list
}
\description{
Create titles for plot_data
}
