% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-samples.R
\name{plot_effect_times}
\alias{plot_effect_times}
\title{Visualize posterior uncertainty in the disease effect times}
\usage{
plot_effect_times(fit, color_scheme = "red", prob = 1,
  prob_outer = 1, point_est = "none")
}
\arguments{
\item{fit}{An object of class \code{lgpfit}.}

\item{color_scheme}{Name of bayesplot color scheme.}

\item{prob}{Inner interval}

\item{prob_outer}{Outer interval}

\item{point_est}{Point estimate type}
}
\value{
a ggplot object
}
\description{
Can only be used if the uncertainty of effect time was modeled.
}
