% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-plot_pred.R
\name{plot_f.create}
\alias{plot_f.create}
\alias{plot_f.create.df_line}
\alias{plot_f.create.df_ribbon}
\alias{plot_f.create.line_arr}
\title{Helper functions for plot_f}
\usage{
plot_f.create.df_line(fit, pred, df_base, comp_idx)

plot_f.create.df_ribbon(fit, pred, df_base, comp_idx, MULT_STD)

plot_f.create.line_arr(fit, pred, comp_idx)
}
\arguments{
\item{fit}{An object of class \linkS4class{lgpfit}.}

\item{pred}{An object of class \linkS4class{FunctionPosteriors} or
\linkS4class{FunctionDraws}.}

\item{df_base}{a data frame with two columns}

\item{comp_idx}{Index of component to plot. The total sum is plotted
if this is \code{NULL}.}

\item{MULT_STD}{a multiplier for standard deviation}
}
\value{
a data frame
}
\description{
Helper functions for plot_f
}
