% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-plot_pred.R
\name{plot_f.get_color_fac}
\alias{plot_f.get_color_fac}
\title{Helper function for plot_f}
\usage{
plot_f.get_color_fac(fit, pred, x, color_by)
}
\arguments{
\item{fit}{An object of class \linkS4class{lgpfit}.}

\item{pred}{An object of class \linkS4class{GaussianPrediction} or
\linkS4class{Prediction}. It is not recommended to set \code{pred=NULL},
as the \code{pred} needs to be then computed inside this function by
calling \code{\link{pred}}.}

\item{x}{Deprecated argument. This is now taken from the \code{pred}
object to ensure compatibility.}

\item{color_by}{name of coloring factor}
}
\value{
a factor
}
\description{
Helper function for plot_f
}
