% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-lgpfit.R
\name{plot_fit_create_df}
\alias{plot_fit_create_df}
\title{Helper function}
\usage{
plot_fit_create_df(df, list_fit)
}
\arguments{
\item{df}{a data frame with group_by factor and x-variable}

\item{list_fit}{a list with fields mean and variance}
}
\value{
a data frame
}
\description{
Helper function
}
