% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_helper.R
\name{plot_fit_create_df_ribbon}
\alias{plot_fit_create_df_ribbon}
\title{Helper function}
\usage{
plot_fit_create_df_ribbon(df, f_mean, f_std, M)
}
\arguments{
\item{df}{a data frame with group_by factor and x-variable}

\item{f_mean}{an array of size \code{1} x \code{num_obs}}

\item{f_std}{an array of size \code{1} x \code{num_obs}}

\item{M}{multiplier for std}
}
\value{
a data frame
}
\description{
Helper function
}
