% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-FunctionPosterior.R
\name{plot_FunctionPosterior}
\alias{plot_FunctionPosterior}
\title{Visualize function posterior distributions}
\usage{
plot_FunctionPosterior(
  fp,
  paramset = NULL,
  component = NULL,
  group_by = "id",
  color_by = NULL,
  t_name = "age",
  MULT_STD = 2,
  alpha_err = 0.2,
  alpha = 1,
  no_err = FALSE,
  no_line = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{fp}{An object of class \linkS4class{FunctionPosterior}.}

\item{paramset}{index of parameter set to pick (NULL = all, but only
posterior means will be plotted)}

\item{component}{name of component to pick (NULL = all)}

\item{group_by}{name of the grouping factor (use \code{group_by=NULL}
to avoid grouping)}

\item{color_by}{name of coloring factor (use \code{color_by=NULL}
to avoid coloring)}

\item{t_name}{name of the x-axis variable}

\item{MULT_STD}{a multiplier for standard deviation}

\item{alpha_err}{ribbon opacity}

\item{alpha}{line opacity}

\item{no_err}{hide error bar even when it would normally be plotted?}

\item{no_line}{hide line even when it would normally be plotted?}

\item{verbose}{Can this print any messages?}
}
\value{
a \code{\link[ggplot2]{ggplot}} object
}
\description{
Visualize function posterior distributions
}
