% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-GaussianPrediction.R
\name{plot_GaussianPrediction}
\alias{plot_GaussianPrediction}
\title{Not implemented}
\usage{
plot_GaussianPrediction(object)
}
\arguments{
\item{object}{object to plot}
}
\description{
Not implemented
}
