% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-lgpfit.R
\name{plot_inputwarp}
\alias{plot_inputwarp}
\title{Visualize the input warping function for different parameter samples}
\usage{
plot_inputwarp(fit, p = 300, color_scheme = "red", b = 0, c = 1)
}
\arguments{
\item{fit}{An object of class \code{lgpfit}.}

\item{p}{number of plot points}

\item{color_scheme}{Name of bayesplot color scheme.}

\item{b}{location of the effective time window (default = 0)}

\item{c}{maximum range (default = 1)}
}
\value{
a ggplot object
}
\description{
Visualize the input warping function for different parameter samples
}
