% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-classes.R
\name{plot,lgpfit,ANY-method}
\alias{plot,lgpfit,ANY-method}
\title{Visualize a fitted `lgpfit` object}
\usage{
\S4method{plot}{lgpfit,ANY}(fit, x = 1, y = 1, color_scheme = "red")
}
\arguments{
\item{fit}{an object of class \code{lgpfit}}

\item{x}{does nothing}

\item{y}{does nothing}

\item{color_scheme}{bayesplot color scheme}
}
\value{
a ggplot object
}
\description{
Visualize a fitted `lgpfit` object
}
