% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_panel}
\alias{plot_panel}
\title{Visuzalize longitudinal data in panels}
\usage{
plot_panel(
  df_data,
  df_signal = NULL,
  df_fit = NULL,
  df_ribbon = NULL,
  i_test = NULL,
  linecolors = color_palette(2),
  vlinecolor = colorset("gray", "mid_highlight"),
  ribbon_color = colorset("red", "light_highlight"),
  nrow = NULL,
  ncol = NULL,
  y_transform = function(x) x,
  teff_true = NULL,
  teff_obs = NULL,
  teff_fit = list(),
  fit_alpha = 1,
  ribbon_alpha = 1
)
}
\arguments{
\item{df_data}{a data frame}

\item{df_signal}{a data frame or \code{NULL}}

\item{df_fit}{a data frame or \code{NULL}}

\item{df_ribbon}{a data frame or \code{NULL}}

\item{i_test}{indices of test points}

\item{linecolors}{two line colors (signal and fit)}

\item{vlinecolor}{color of vertical lines (true and obs. effect time)}

\item{ribbon_color}{ribbon color}

\item{nrow}{number of rows, an argument for \code{ggplot2::facet_wrap}}

\item{ncol}{number of columns, an argument for \code{ggplot2::facet_wrap}}

\item{y_transform}{a function}

\item{teff_true}{a named vector}

\item{teff_obs}{a named vector}

\item{teff_fit}{a list of named vectors}

\item{fit_alpha}{a value between 0 and 1}

\item{ribbon_alpha}{a value between 0 and 1}
}
\value{
a \code{ggplot} object
}
\description{
Creates plots where each observation unit has a separate panel
}
