% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-panel.R
\name{plot_panel_add_effect_times}
\alias{plot_panel_add_effect_times}
\title{Add real or observed disease effect times to plot}
\usage{
plot_panel_add_effect_times(h, teff, group_by, linecolor, linetype, lwd)
}
\arguments{
\item{h}{a \code{ggplot} object}

\item{teff}{a named vector}

\item{group_by}{name of grouping variable}

\item{linecolor}{line color}

\item{linetype}{line type}

\item{lwd}{line width}
}
\value{
updated \code{ggplot} object
}
\description{
Add real or observed disease effect times to plot
}
