% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-panel.R
\name{plot_panel_add_faceting}
\alias{plot_panel_add_faceting}
\title{Add faceting to the panel plot}
\usage{
plot_panel_add_faceting(h, group_by, nrow, ncol)
}
\arguments{
\item{h}{a \code{ggplot} object}

\item{group_by}{name of grouping variable}
}
\value{
a \code{ggplot} object
}
\description{
Add faceting to the panel plot
}
