% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_panel_create}
\alias{plot_panel_create}
\title{Initialize the panel plot}
\usage{
plot_panel_create(df_data, i_test, y_transform, nrow, ncol)
}
\arguments{
\item{df_data}{a data frame}

\item{i_test}{indices of test points}

\item{y_transform}{a function}

\item{nrow}{number of rows, an argument for \code{ggplot2::facet_wrap}}

\item{ncol}{number of columns, an argument for \code{ggplot2::facet_wrap}}
}
\value{
a \code{ggplot} object
}
\description{
Initialize the panel plot
}
