% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_panel_create_aes}
\alias{plot_panel_create_aes}
\title{Helper function}
\usage{
plot_panel_create_aes(data, type_name)
}
\arguments{
\item{data}{a data frame with three columns}

\item{type_name}{name of a factor that is to be distinguished by
plot marker and/or type}
}
\value{
an aes object
}
\description{
Helper function
}
