% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-panel.R
\name{plot_panel_create_fit_df}
\alias{plot_panel_create_fit_df}
\title{Data for adding fit to the panel plot}
\usage{
plot_panel_create_fit_df(h, fit)
}
\arguments{
\item{h}{a \code{ggplot} object}

\item{fit}{a list}
}
\value{
a data frame
}
\description{
Data for adding fit to the panel plot
}
