% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-lgpfit_plot.R
\name{plot_posterior}
\alias{plot_posterior}
\title{Visualize posterior distribution of sampled parameters}
\usage{
plot_posterior(
  fit,
  type = "intervals",
  regex_pars = c("alpha", "ell", "wrp", "sigma", "phi"),
  ...
)
}
\arguments{
\item{fit}{an object of class \linkS4class{lgpfit}}

\item{type}{plot type}

\item{regex_pars}{regex for parameter names to plot}

\item{...}{other arguments to bayesplot functions}
}
\value{
a \code{ggplot} object
}
\description{
Visualize posterior distribution of sampled parameters
}
\seealso{
Other model fit visualization functions: 
\code{\link{plot,lgpfit,missing-method}},
\code{\link{plot_beta}()},
\code{\link{plot_effect_times}()},
\code{\link{plot_fit}()},
\code{\link{plot_posterior_warp}()}
}
\concept{model fit visualization functions}
