% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-predict.R
\name{plot_posterior_components}
\alias{plot_posterior_components}
\title{Plot posterior of the components of f}
\usage{
plot_posterior_components(fit, PRED = NULL, color_scheme = "red",
  alpha = 0.1, alpha_line = 1, plot_uncertainty = TRUE,
  title = NULL, ylim = NULL, n_sds = 2, original_y_scale = FALSE)
}
\arguments{
\item{fit}{An object of class \code{lgpfit}.}

\item{PRED}{Predictions computed using \code{lgp_predict}.}

\item{color_scheme}{Name of bayesplot color scheme or a list with fieds 'dark' and 'light'.}

\item{alpha}{Ribbon fill opacity.}

\item{alpha_line}{Line opacity}

\item{plot_uncertainty}{Should an uncertainty ribbon be plotted?}

\item{title}{optional prefix to plot title}

\item{ylim}{y axis limits}

\item{n_sds}{number of standard deviations for the uncertainty band width}

\item{original_y_scale}{should the predictions be scaled back to original data scale}
}
\value{
a ggplot object
}
\description{
Plot posterior of the components of f
}
