% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-posterior-helper.R
\name{plot_posterior_components_sub1}
\alias{plot_posterior_components_sub1}
\title{Helper for \code{\link{plot_posterior_components}}}
\usage{
plot_posterior_components_sub1(fit, subsamples, time_is_xvar, marker, ...)
}
\arguments{
\item{fit}{An object of class \code{lgpfit}.}

\item{subsamples}{How many samples to plot. If this is \code{NULL},
average over all samples is plotted. If this is \code{"all"}, all
samples are plotted.}

\item{time_is_xvar}{is the time variable the x-axis variable
in all subplots?}

\item{marker}{point type}

\item{...}{additional arguments for \code{\link{plot_components}}}
}
\value{
an object returned by \code{ggpubr::ggarrange} or a list
}
\description{
Helper for \code{\link{plot_posterior_components}}
}
