% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-predict.R
\name{plot_posterior_predictions}
\alias{plot_posterior_predictions}
\title{Plot posterior of f or predictive distribution for y}
\usage{
plot_posterior_predictions(fit, mode, PRED = NULL,
  color_scheme = "red", color_scheme_t_effect = "gray", alpha = 0.5,
  alpha_line = 1, alpha2 = 0.5, plot_uncertainty = TRUE,
  title = NULL, ylim = NULL, plot_obs_onset = FALSE,
  plot_t_effect_samples = FALSE, ypos_dens = NULL, test_data = NULL,
  color_test = "deepskyblue2", pch_test = 21, size_test = 2,
  error_bar = FALSE, n_sds = 2, reference_times = NULL,
  post_t_effect_stat = "none", original_y_scale = TRUE,
  data_color = "black", data_marker = 21, ons_linetypes = c(1, 2, 3),
  ons_linecolors = c("black", "red", "gray50"))
}
\arguments{
\item{fit}{An object of class \code{lgpfit}.}

\item{mode}{Must be either "posterior" or "predictive".}

\item{PRED}{Predictions computed using \code{lgp_predict}.}

\item{color_scheme}{Name of bayesplot color scheme or a list with fieds 'dark' and 'light'.}

\item{color_scheme_t_effect}{color scheme name for effect time density plotting}

\item{alpha}{Ribbon fill opacity.}

\item{alpha_line}{Line opacity.}

\item{alpha2}{alpha of t_onset density}

\item{plot_uncertainty}{Should an uncertainty ribbon be plotted?}

\item{title}{optional prefix to plot title}

\item{ylim}{y axis limits}

\item{plot_obs_onset}{should the observed disease onset/initiation time be plotted
by a vertical line}

\item{plot_t_effect_samples}{should a distribution of sampled effect times be plotted}

\item{ypos_dens}{y-position of the density plot}

\item{test_data}{Test data frame}

\item{color_test}{test point color}

\item{pch_test}{test point marker}

\item{size_test}{test point size}

\item{error_bar}{should uncertainty be plotted using error bars instead of a ribbon}

\item{n_sds}{number of standard deviations for the uncertainty band width}

\item{reference_times}{reference onset times}

\item{post_t_effect_stat}{statistic computed from effect time samples (mean or median)}

\item{original_y_scale}{should the predictions be scaled back to original data scale}

\item{data_color}{data marker color}

\item{data_marker}{data marker type}

\item{ons_linetypes}{onset line types}

\item{ons_linecolors}{onset line colors}
}
\value{
a ggplot object
}
\description{
Plot posterior of f or predictive distribution for y
}
