% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-lgpfit.R
\name{plot_posterior_warp}
\alias{plot_posterior_warp}
\title{Visualize the input warping function for different parameter samples}
\usage{
plot_posterior_warp(fit, p = 300, R = 48, color_scheme = "brightblue")
}
\arguments{
\item{fit}{an object of class \linkS4class{lgpfit}}

\item{p}{number of plot points}

\item{R}{width of time window}

\item{color_scheme}{name of \code{bayesplot} color scheme}
}
\value{
a \code{ggplot} object or list of them
}
\description{
Visualize the input warping function for different parameter samples
}
