% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-predict.R
\name{plot_posterior_y}
\alias{plot_posterior_y}
\title{Plot posterior predictive distribution}
\usage{
plot_posterior_y(fit, PRED, uncertainty = "ribbon", test_data = NULL,
  data_marker = 16, n_sds = 2, ...)
}
\arguments{
\item{fit}{An object of class \code{lgpfit}.}

\item{PRED}{Predictions computed using \code{lgp_predict}.}

\item{uncertainty}{Either "none", "ribbon" or "errorbar".}

\item{test_data}{Test data set.}

\item{data_marker}{pch for data points}

\item{n_sds}{number of standard deviations for the uncertainty band width}

\item{...}{additional arguments to \code{\link{plot_posterior_predictions}}}
}
\value{
a ggplot object
}
\description{
This is a wrapper for \code{\link{plot_posterior_predictions}}.
}
