% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-plot_pred.R
\name{plot_pred.create_input}
\alias{plot_pred.create_input}
\title{Helper function for plot_pred and plot_f}
\usage{
plot_pred.create_input(pred, x, draws, reduce, group_by, t_name)
}
\arguments{
\item{pred}{An object of class \linkS4class{GaussianPrediction} or
\linkS4class{Prediction}. It is not recommended to set \code{pred=NULL},
as the \code{pred} needs to be then computed inside this function by
calling \code{\link{pred}}.}

\item{x}{Deprecated argument. This is now taken from the \code{pred}
object to ensure compatibility.}

\item{draws}{only has effect if \code{pred} is \code{NULL}}

\item{reduce}{only has effect if \code{pred} is \code{NULL}}

\item{group_by}{name of the grouping variable (use \code{group_by=NA}
to avoid grouping)}

\item{t_name}{name of the x-axis variable}
}
\value{
a list with names \code{pred} and \code{df_base}
}
\description{
Helper function for plot_pred and plot_f
}
