% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pred.R
\name{plot_pred}
\alias{plot_pred}
\alias{plot_f}
\title{Visualizing model predictions or inferred covariate effects}
\usage{
plot_pred(
  fit,
  x = NULL,
  pred = NULL,
  group_by = "id",
  t_name = "age",
  draws = NULL,
  reduce = mean,
  MULT_STD = 2,
  ...
)

plot_f(
  fit,
  x = NULL,
  pred = NULL,
  group_by = "id",
  t_name = "age",
  draws = NULL,
  reduce = mean,
  MULT_STD = 2,
  comp_idx = NULL,
  color_by = NA,
  ...
)
}
\arguments{
\item{fit}{An object of class \linkS4class{lgpfit}.}

\item{x}{A data frame of prediction points, containing at least the
variables used as covariates in the model. This must be specified if
\code{pred} is not \code{NULL}.}

\item{pred}{An object of class \linkS4class{GaussianPrediction} or
\linkS4class{Prediction}.}

\item{group_by}{name of the grouping variable}

\item{t_name}{name of the x-axis variable}

\item{draws}{only has effect if \code{pred} is \code{NULL}}

\item{reduce}{only has effect if \code{pred} is \code{NULL}}

\item{MULT_STD}{a multiplier for standard deviation}

\item{...}{additional arguments to \code{\link{plot_api_g}} or
\code{\link{plot_api_c}}}

\item{comp_idx}{Index of component to plot. The total sum is plotted
if this is \code{NULL}.}

\item{color_by}{name of coloring factor}
}
\value{
a \code{\link[ggplot2]{ggplot}} object
}
\description{
\itemize{
  \item Predictions at data points can be visualized using \code{plot_pred}.
    Out-of-sample predictions can be visualized by giving the \code{pred}
    and \code{x} arguments.
  \item The total signal \code{f} or any of its
  additive components can be plotted using \code{plot_f}.
}
}
\seealso{
Other main plot functions: 
\code{\link{plot_data}()},
\code{\link{plot_sim}()}
}
\concept{main plot functions}
