% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-spaghetti.R
\name{plot_predictions}
\alias{plot_predictions}
\title{Plot computed predictions componentwise}
\usage{
plot_predictions(fit, PRED = NULL, componentwise = FALSE,
  color_scheme = "red", alpha = 1 - 0.9 * as.numeric(componentwise),
  plot_ribbon = TRUE, title = NULL, theme = ggplot2::theme_gray(),
  original_y_scale = TRUE, plot_ribbon_edge = !componentwise,
  plot_pred = TRUE, alpha_line = alpha, ylim = NULL)
}
\arguments{
\item{fit}{An object of class \code{lgpfit}.}

\item{PRED}{Predictions computed using \code{lgp_predict}.}

\item{componentwise}{Should the predictions be plotted componentwise?}

\item{color_scheme}{Name of bayesplot color scheme or a list with fieds 'dark' and 'light'.}

\item{alpha}{Ribbon fill opacity.}

\item{plot_ribbon}{Should an uncertainty ribbon be plotted?}

\item{title}{optional prefix to plot title}

\item{theme}{ggplot theme}

\item{original_y_scale}{should the predictions be scaled back to the original data y scale}

\item{plot_ribbon_edge}{Should an uncertainty ribbon edge be plotted?}

\item{plot_pred}{should the (mean) predictions line be plotted?}

\item{alpha_line}{Line opacity}

\item{ylim}{y axis limits}
}
\value{
a ggplot object
}
\description{
Plot computed predictions componentwise
}
