% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-predict.R
\name{plot_predictions_add_onsets}
\alias{plot_predictions_add_onsets}
\title{Add disease onsets to predictions plot}
\usage{
plot_predictions_add_onsets(fit, h, plot_obs_onset, plot_onset_samples,
  idvar, timevar, ypos_dens, color_scheme_onset, alpha2 = 1)
}
\arguments{
\item{fit}{An object of class \code{lgpfit}.}

\item{h}{a ggplot object}

\item{plot_obs_onset}{a boolean value}

\item{plot_onset_samples}{a boolean value}

\item{idvar}{id variable name}

\item{timevar}{time variable name}

\item{ypos_dens}{y position of the estimated onset density}

\item{color_scheme_onset}{color scheme}

\item{alpha2}{alpha parameter}
}
\value{
a modified ggplot object
}
\description{
NOTE: currently assumes that diseased individuals come first.
}
