% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-predict.R
\name{plot_predictions_options}
\alias{plot_predictions_options}
\title{Do input checks and set options for plotting predictions}
\usage{
plot_predictions_options(fit, color_scheme, original_y_scale, PRED,
  test_data, color_scheme_t_effect, mode, n_sds)
}
\arguments{
\item{fit}{An object of class \code{lgpfit}.}

\item{color_scheme}{Name of bayesplot color scheme.}

\item{original_y_scale}{Boolean value.}

\item{PRED}{Predictions computed using \code{lgp_predict}.}

\item{test_data}{test data}

\item{color_scheme_t_effect}{Another color scheme.}

\item{mode}{mode}

\item{n_sds}{number of standard deviations for the uncertainty band width}
}
\value{
a list
}
\description{
Do input checks and set options for plotting predictions
}
