% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-samples.R
\name{plot_relevances}
\alias{plot_relevances}
\title{Barplot of covariate relevances}
\usage{
plot_relevances(object, color_scheme = "red")
}
\arguments{
\item{object}{an object of class \code{lgpfit}}

\item{color_scheme}{bayesplot color scheme name}
}
\value{
a ggplot object
}
\description{
Barplot of covariate relevances
}
