% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-lgpfit.R
\name{plot_samples}
\alias{plot_samples}
\title{Visualize posterior samples}
\usage{
plot_samples(
  fit,
  type = "intervals",
  regex_pars = c("alpha", "ell", "wrp", "sigma", "phi"),
  ...
)
}
\arguments{
\item{fit}{an object of class \linkS4class{lgpfit}}

\item{type}{plot type}

\item{regex_pars}{regex for parameter names to plot}

\item{...}{other arguments to bayesplot functions}
}
\value{
a \code{ggplot} object
}
\description{
Visualize posterior samples
}
