% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-simulate.R
\name{plot_sim}
\alias{plot_sim}
\alias{plot_sim.data}
\alias{plot_sim.component}
\alias{plot_sim.component.df}
\title{Visualize an lgpsim object (simulated data)}
\usage{
plot_sim(
  simdata,
  group_by = "id",
  x_name = "age",
  h_name = "h",
  y_name = "y",
  comp_idx = NULL,
  color_by = NA,
  verbose = TRUE,
  ...
)

plot_sim.data(simdata, group_by, x_name, h_name, y_name, verbose, ...)

plot_sim.component(simdata, comp_idx, group_by, x_name, color_by, verbose, ...)

plot_sim.component.df(simdata, x_name, group_by, color_by, comp_idx)
}
\arguments{
\item{simdata}{an object of class \linkS4class{lgpsim}}

\item{group_by}{grouping factor}

\item{x_name}{name of x-axis variable}

\item{h_name}{name of the signal in \code{simdata$components}}

\item{y_name}{name of response variable}

\item{comp_idx}{Possible index of a component to be shown.
If this is NULL, the data and total signal are shown.}

\item{color_by}{coloring factor}

\item{verbose}{should some information be printed?}

\item{...}{additional arguments to \code{\link{plot_api_g}} or
\code{\link{plot_api_c}}}
}
\value{
a \code{\link[ggplot2]{ggplot}} object
}
\description{
Visualize an lgpsim object (simulated data)
}
\seealso{
Other main plot functions: 
\code{\link{plot_components}()},
\code{\link{plot_data}()},
\code{\link{plot_pred}()}
}
\concept{main plot functions}
