% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-plot_sim.R
\name{plot_sim_data_title}
\alias{plot_sim_data_title}
\alias{plot_sim.data.title}
\alias{plot_sim.data.title_teff}
\title{Helpers for plot_sim.data}
\usage{
plot_sim.data.title(noise_type)

plot_sim.data.title_teff()
}
\arguments{
\item{noise_type}{noise type name}
}
\description{
Helpers for plot_sim.data
}
