% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-simdata.R
\name{plot_simdata_by_component}
\alias{plot_simdata_by_component}
\title{Plot each component of a simulated longitudinal data set separately}
\usage{
plot_simdata_by_component(simData, linecolor = "black", nrow = NULL,
  ncol = NULL, plot_point = TRUE)
}
\arguments{
\item{simData}{a list returned by \code{\link{simulate_data}}}

\item{linecolor}{line color}

\item{nrow}{an argument for \code{ggplot2::facet_wrap}}

\item{ncol}{an argument for \code{ggplot2::facet_wrap}}

\item{plot_point}{should points be plotted also}
}
\value{
a ggplot object
}
\description{
Plot each component of a simulated longitudinal data set separately
}
