% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-simdata.R
\name{plot_simdata_by_individual}
\alias{plot_simdata_by_individual}
\title{Plot a simulated longitudinal data set for each individual separately}
\usage{
plot_simdata_by_individual(simData, linecolor = "firebrick3",
  nrow = NULL, ncol = NULL, plot_point = FALSE,
  vlinecolors = rep("steelblue4", 2), vlinetypes = c(1, 2))
}
\arguments{
\item{simData}{a list returned by \code{\link{simulate_data}}}

\item{linecolor}{line color}

\item{nrow}{an argument for \code{ggplot2::facet_wrap}}

\item{ncol}{an argument for \code{ggplot2::facet_wrap}}

\item{plot_point}{should \code{geom_point} be used also for the underlying signal?}

\item{vlinecolors}{colors to plot the real and observed disease onsets if they exist}

\item{vlinetypes}{line types to use for the real and observed disease onsets if they exist}
}
\value{
a ggplot object
}
\description{
Plot a simulated longitudinal data set for each individual separately
}
