% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-lgpfit.R
\name{plot_warp}
\alias{plot_warp}
\title{Visualize the input warping function for different parameter samples}
\usage{
plot_warp(fit, p = 300, R = 48)
}
\arguments{
\item{fit}{an object of class \linkS4class{lgpfit}}

\item{p}{number of plot points}

\item{R}{width of time window}
}
\value{
a \code{ggplot} object or list of them
}
\description{
Visualize the input warping function for different parameter samples
}
