% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_warp_helper}
\alias{plot_warp_helper}
\title{Helper function}
\usage{
plot_warp_helper(par_summary, dis_age, color_scheme = "brightblue")
}
\arguments{
\item{par_summary}{summary of warping parameter}

\item{dis_age}{the x-axis values}

\item{color_scheme}{name of \code{bayesplot} color scheme}
}
\value{
a \code{ggplot} object
}
\description{
Helper function
}
