% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-post_post.R
\name{post_pred}
\alias{post_pred}
\title{Get posterior predictive distribution (draws) from function posterior (draws)}
\usage{
post_pred(post)
}
\arguments{
\item{post}{an object of class \linkS4class{FunctionPosterior}}
}
\value{
TODO
}
\description{
Get posterior predictive distribution (draws) from function posterior (draws)
}
