% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-posterior_f.R
\name{posterior_f}
\alias{posterior_f}
\title{Evaluate or extract function posterior distributions}
\usage{
posterior_f(
  fit,
  x = NULL,
  reduce = function(x) base::mean(x),
  draws = NULL,
  verbose = TRUE,
  STREAM = get_stream(),
  debug_km = FALSE,
  debug_dims = FALSE,
  force = FALSE
)
}
\arguments{
\item{fit}{An object of class \linkS4class{lgpfit}.}

\item{x}{A data frame of points where the predictions are computed.
The function \code{\link{new_x}} provides an easy way to create it.}

\item{reduce}{Reduction for parameters draws. Can be a function that
is applied to reduce all parameter draws into one parameter set, or
NULL (no reduction). Has no effect if \code{draws} is specified.}

\item{draws}{Indices of parameter draws to use, or \code{NULL} to use all
draws.}

\item{verbose}{Should more information and a progress bar be printed?}

\item{STREAM}{External pointer. By default obtained with
\code{\link[rstan]{get_stream}}.}

\item{debug_km}{Should this only return the required kernel matrices.
Can be used for debugging or testing \code{\link{kernels_fpost}}.}

\item{debug_dims}{Should this print dimensions of some variables.
Can be used for debugging \code{\link{kernels_fpost}}.}

\item{force}{This is by default \code{FALSE} to prevent unnecessarily
large computations that might crash R or take forever.}
}
\value{
An object of class \linkS4class{FunctionPosteriors} or
\linkS4class{FunctionDraws}.
}
\description{
Evaluates (or extracts) the posterior distribution of the
total signal \code{f} and its additive components (or draws from these
distributions). All these are computed for
each parameter draw (defined by \code{draws}), or other parameter set
(obtained by a reduction defined by \code{reduce}).
}
