% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior.R
\name{posterior_predict}
\alias{posterior_predict}
\title{Posterior predictive distribution at test points}
\usage{
posterior_predict(fit, X_pred, STREAM = get_stream())
}
\arguments{
\item{fit}{An object of class \linkS4class{lgpfit}.}

\item{X_pred}{A data frame of points where predictions are computed.
The function \code{\link{new_x}} can help in creating it.}

\item{STREAM}{external pointer}
}
\value{
A list.
}
\description{
Posterior predictive distribution at test points
}
\seealso{
Other GP posterior computation functions: 
\code{\link{compute_gp_posteriors}()},
\code{\link{compute_kernel_matrices}()},
\code{\link{format_X_pred}()},
\code{\link{get_draws_kernel_pars}()}
}
\concept{GP posterior computation functions}
