% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior.R
\name{posterior_predict_kernels}
\alias{posterior_predict_kernels}
\title{Compute all kernel matrices required for computing the function
posterior distributions at test points}
\usage{
posterior_predict_kernels(fit, X_pred, STREAM = get_stream())
}
\arguments{
\item{fit}{An object of class \linkS4class{lgpfit}.}

\item{X_pred}{A data frame of points where predictions are computed.
The function \code{\link{new_x}} can help in creating it.}

\item{STREAM}{External pointer. By default obtained with
\code{\link[rstan]{get_stream}}.}
}
\value{
A list.
}
\description{
Compute all kernel matrices required for computing the function
posterior distributions at test points
}
\seealso{
Other GP posterior computation functions: 
\code{\link{compute_gp_posteriors}()},
\code{\link{compute_kernel_matrices_const}()},
\code{\link{compute_kernel_matrices}()},
\code{\link{create_kernel_input}()},
\code{\link{get_draws_kernel_pars}()},
\code{\link{posterior_predict}()}
}
\concept{GP posterior computation functions}
