% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postproc-relevances.R
\name{postproc_relevances}
\alias{postproc_relevances}
\title{Compute component relevances and estimate amount of noise
(one MCMC sample)}
\usage{
postproc_relevances(
  fit,
  relevance_method = "f_mean",
  noise_method = "SSE",
  verbose = FALSE
)
}
\arguments{
\item{fit}{An (incomplete) object of class \code{lgpfit}.}

\item{relevance_method}{Component relevance determination method. 
Must be either \code{"f_mean"} or \code{"alpha"}.}

\item{noise_method}{Noise level determination method. 
Currently must be \code{"SSE"}.}

\item{verbose}{Should some output be printed?}
}
\value{
An updated object of class \code{lgpfit}.
}
\description{
Compute component relevances and estimate amount of noise
(one MCMC sample)
}
