% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_s4_generics.R
\name{ppc}
\alias{ppc}
\alias{prior_predict,}
\alias{posterior_predict}
\alias{prior_predict}
\title{Prior and posterior predictive checks}
\usage{
posterior_predict(object, ...)

prior_predict(object, ...)
}
\arguments{
\item{object}{an object of class \linkS4class{lgpmodel}}
}
\description{
Prior and posterior predictive checks
}
