% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_get.R
\name{pred.gaussian.f_to_y}
\alias{pred.gaussian.f_to_y}
\title{Tranform distribution of f to distribution of y}
\usage{
pred.gaussian.f_to_y(f_mean, f_std, sigma, y_norm_inv)
}
\arguments{
\item{f_mean}{an array of shape \code{num_draws} x \code{num_points}}

\item{f_std}{an array of shape \code{num_draws} x \code{num_points}}

\item{sigma}{a vector of length \code{num_draws}}

\item{y_norm_inv}{inverse normalization function for y}
}
\value{
a list with names \code{mean} and \code{std}, both of which are
arrays of shape \code{num_draws} x \code{num_points}
}
\description{
Tranform distribution of f to distribution of y
}
