% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred.R
\name{pred.kr_compute}
\alias{pred.kr_compute}
\title{Compute out-of-sample predictions using kernel regression on
sampled function values}
\usage{
pred.kr_compute(kernels, pred, delta, STREAM = get_stream())
}
\arguments{
\item{kernels}{A list returned by \code{\link{predict_kernels}}.}

\item{pred}{An object of class \linkS4class{Prediction}.}

\item{delta}{jitter to ensure positive definite matrices}
}
\value{
An object of class \linkS4class{Prediction}.
}
\description{
Compute out-of-sample predictions using kernel regression on
sampled function values
}
\seealso{
Other prediction functions: 
\code{\link{kernel_matrices}()},
\code{\link{pred.gaussian_compute}()},
\code{\link{pred_kernels}()},
\code{\link{pred}()}
}
\concept{prediction functions}
