% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-pred.R
\name{pred.marginal_compute}
\alias{pred.marginal_compute}
\title{Compute analytic out-of-sample predictions}
\usage{
pred.marginal_compute(kernels, y, sigma, delta, verbose, STREAM = get_stream())
}
\arguments{
\item{kernels}{A list returned by \code{\link{pred_kernels}}.}

\item{y}{response variable vector of length \code{num_obs}}

\item{sigma}{a vector with length equal to number of parameter sets}

\item{delta}{jitter to ensure positive definite matrices}

\item{verbose}{Should more information and a progress bar be printed?}

\item{STREAM}{External pointer. By default obtained with
\code{\link[rstan]{get_stream}}.}
}
\value{
A list.
}
\description{
Compute analytic out-of-sample predictions
}
\seealso{
Other prediction functions: 
\code{\link{kernel_matrices}()},
\code{\link{pred.latent_compute}()},
\code{\link{pred_kernels}()},
\code{\link{pred}()}
}
\concept{prediction functions}
