% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-pred.R
\name{pred}
\alias{pred}
\title{Compute posterior predictive distribution or get draws from it}
\usage{
pred(
  fit,
  x,
  c_hat_pred = NULL,
  reduce = function(x) base::mean(x),
  draws = NULL,
  verbose = TRUE,
  STREAM = get_stream(),
  ...
)
}
\arguments{
\item{fit}{An object of class \linkS4class{lgpfit}.}

\item{x}{A data frame of points where the predictions are computed.
The function \code{\link{new_x}} provides an easy way to create it.}

\item{c_hat_pred}{This is only used if the latent signal \code{f} was
sampled. This input contains the values added to the sum \code{f} before
passing through inverse link function. Must be a vector with length equal to
the number of prediction points. If original \code{c_hat} was constant,
then \code{c_hat_pred} can be ignored, in which case this will by default
use the same constant.}

\item{reduce}{Reduction for parameters draws. Can be a function that
is applied to reduce all parameter draws into one parameter set, or
NULL (no reduction). Has no effect if \code{draws} is specified.}

\item{draws}{Indices of parameter draws to use, or \code{NULL} to use all
draws.}

\item{verbose}{Should more information and a progress bar be printed?}

\item{STREAM}{External pointer. By default obtained with
\code{\link[rstan]{get_stream}}.}

\item{...}{Other optional arguments passed to \code{\link{posterior_f}}.}
}
\value{
An object of class \linkS4class{GaussianPrediction} or
\linkS4class{Prediction}.
}
\description{
Compute posterior predictive distribution or get draws from it
}
