% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-pred_input.R
\name{pred_input}
\alias{pred_input}
\alias{pred_input_x}
\alias{pred_input_draws}
\alias{pred_input_draws.marginal}
\alias{pred_input_draws.latent}
\alias{pred_input_draws.common}
\alias{get_draw_arr}
\alias{get_draw_arr_vec}
\title{Create input for Stan model that computes predictions}
\usage{
pred_input(fit, x, reduce, draws, refresh)

pred_input_x(fit, x)

pred_input_draws(fit, reduce, draws, refresh)

pred_input_draws.marginal(fit, reduce, draws)

pred_input_draws.latent(fit, reduce, draws)

pred_input_draws.common(fit, reduce, draws, refresh)

get_draw_arr(fit, draws, reduce, par_name, S, D)

get_draw_arr_vec(fit, draws, reduce, par_name, S, B, V)
}
\arguments{
\item{fit}{An object of class \linkS4class{lgpfit}.}

\item{x}{A data frame of points where the predictions are computed.
The function \code{\link{new_x}} can help in creating it.}

\item{reduce}{Reduction for parameters draws. Can be a function that
is applied to reduce all parameter draws into one parameter set, or
NULL (no reduction). Has no effect if \code{draws} is specified.}

\item{draws}{Indices of parameter draws to use, or \code{NULL} to use all
draws.}

\item{refresh}{How often to print progress? Has no effect if \code{verbose}
is \code{FALSE}.}
}
\value{
a list that can be used as data for \code{lgp_predict.stan}
}
\description{
Create input for Stan model that computes predictions
}
