% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred.R
\name{pred_kernels}
\alias{pred_kernels}
\title{Compute all kernel matrices required for computing predictions}
\usage{
pred_kernels(fit, x, reduce, draws, STREAM = get_stream())
}
\arguments{
\item{fit}{An object of class \linkS4class{lgpfit}.}

\item{x}{A data frame of points where the predictions are computed.
The function \code{\link{new_x}} provides an easy way to create it.}

\item{reduce}{reduction for parameters draws}

\item{draws}{indices of parameter draws}

\item{STREAM}{External pointer. By default obtained with
\code{\link[rstan]{get_stream}}.}
}
\value{
Returns a list with fields named \code{data_vs_data},
\code{pred_vs_data} and \code{pred_vs_pred}. Each of them is a list with
length equal to the number of parameter sets.
}
\description{
Compute all kernel matrices required for computing predictions
}
\seealso{
Other prediction functions: 
\code{\link{kernel_matrices}()},
\code{\link{pred.gaussian_compute}()},
\code{\link{pred.kr_compute}()},
\code{\link{pred}()}
}
\concept{prediction functions}
