% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-utils.R
\name{PRED_to_arrays}
\alias{PRED_to_arrays}
\title{PRED object to arrays}
\usage{
PRED_to_arrays(PRED)
}
\arguments{
\item{PRED}{an object returned by \code{\link{lgp_predict}}}
}
\value{
a list containing two arrays
}
\description{
PRED object to arrays
}
