% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_oos.R
\name{predict.gaussian_compute}
\alias{predict.gaussian_compute}
\title{Compute analytic out-of-sample predictions}
\usage{
predict.gaussian_compute(kernels, y, sigma, delta, STREAM = get_stream())
}
\arguments{
\item{kernels}{A list returned by \code{\link{predict_kernels}}.}

\item{y}{response variable vector of length \code{num_obs}}

\item{sigma}{a vector with length equal to number of parameter sets}

\item{delta}{jitter to ensure positive definite matrices}
}
\value{
A list.
}
\description{
Compute analytic out-of-sample predictions
}
\seealso{
Other prediction functions: 
\code{\link{kernel_matrices}()},
\code{\link{predict.kr_compute}()},
\code{\link{predict_kernels}()},
\code{\link{predict}()}
}
\concept{prediction functions}
