% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_oos.R
\name{predict.kr_compute}
\alias{predict.kr_compute}
\title{Compute out-of-sample predictions using kernel regression on
sampled function values}
\usage{
predict.kr_compute(kernels, f, delta, STREAM = get_stream())
}
\arguments{
\item{f}{A list returned by \code{\link{get_f}}.}

\item{inheritParams}{predict.gaussian_compute}
}
\value{
An object of class \linkS4class{Prediction}.
}
\description{
Compute out-of-sample predictions using kernel regression on
sampled function values
}
\seealso{
Other prediction functions: 
\code{\link{kernel_matrices}()},
\code{\link{predict.gaussian_compute}()},
\code{\link{predict_kernels}()},
\code{\link{predict}()}
}
\concept{prediction functions}
