% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgpr.R
\name{predict.lgpr}
\alias{predict.lgpr}
\title{Predict method for lgpr object}
\usage{
\method{predict}{lgpr}(object, ...)
}
\arguments{
\item{object}{an lgpr object}

\item{...}{....}
}
\value{
a 3-column matrix with prediction interval defined by the gaussian
uncertainty model.
}
\description{
Predict method for lgpr object
}
\author{
David Senhora Navega
}
