% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict_preproc}
\alias{predict_preproc}
\title{Preprocess some things before computing predictions}
\usage{
predict_preproc(fit, X_test, samples)
}
\arguments{
\item{fit}{An object of class \code{lgpfit}.}

\item{X_test}{The test points where the predictions should be computed.}

\item{samples}{The samples argument to \code{\link{lgp_predict}}}
}
\description{
This is a helper function for \code{\link{lgp_predict}}.
}
