% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-lgpr.R, R/methods-Prediction.R
\docType{class}
\name{Prediction-class}
\alias{Prediction-class}
\alias{Prediction}
\alias{show,Prediction-method}
\alias{plot,Prediction,missing-method}
\title{An S4 class to represent general additive model predictions}
\usage{
\S4method{show}{Prediction}(object)

\S4method{plot}{Prediction,missing}(x, y, ...)
}
\arguments{
\item{x}{a \linkS4class{Prediction} object to visualize}

\item{y}{unused argument}
}
\description{
An S4 class to represent general additive model predictions
}
\section{Methods (by generic)}{
\itemize{
\item \code{show}: Print a summary about the object.

\item \code{plot}: Visualization.
Optional arguments (\code{...}) are passed to
\code{\link{plot_pred}}.
}}

\section{Slots}{

\describe{
\item{\code{f_comp}}{component predictions}

\item{\code{f}}{signal prediction}

\item{\code{h}}{prediction (signal predictions transformed through inverse link
function)}
}}

\seealso{
\linkS4class{GaussianPrediction}
}
