% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print_list}
\alias{print_list}
\title{Printing a list in a more compact format}
\usage{
print_list(input)
}
\arguments{
\item{input}{a named list}
}
\description{
Printing a list in a more compact format
}
\seealso{
Other list utilities: 
\code{\link{dollar}()},
\code{\link{list_if_named}()},
\code{\link{list_to_matrix}()},
\code{\link{matrix_to_list}()},
\code{\link{zip_lists}()}
}
\concept{list utilities}
