% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-prior.R
\name{print_prior}
\alias{print_prior}
\title{Human-readable description of a specified prior}
\usage{
print_prior(object)
}
\arguments{
\item{object}{An object of class \code{lgpfit} or a valid prior argument for the `lgp` function.}
}
\value{
nothing
}
\description{
Print human-readable info about the prior specification that was used or will be used
}
