% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/character.R
\name{print_stan_input}
\alias{print_stan_input}
\title{Print the Stan input of an lgpmodel}
\usage{
print_stan_input(model)
}
\arguments{
\item{model}{an object of class \linkS4class{lgpmodel}}
}
\description{
Print the Stan input of an lgpmodel
}
