% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-prior.R
\name{prior_default}
\alias{prior_default}
\title{Create the default prior}
\usage{
prior_default(sigma_alpha = 1)
}
\arguments{
\item{sigma_alpha}{Sigma parameter of the student-t distribution for all alpha.}
}
\value{
A list defining a valid \code{prior} argument for the \code{lgp} function.
}
\description{
Create the default prior
}
