% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{prior_df}
\alias{prior_df}
\alias{prior_to_df}
\alias{prior_to_df_oneparam}
\alias{prior_to_df_beta}
\alias{prior_to_df_teff}
\title{Convert the Stan input encoding of a prior to a human-readable format}
\usage{
prior_to_df(stan_input, digits = 3)

prior_to_df_oneparam(stan_input, parname, digits)

prior_to_df_beta(stan_input)

prior_to_df_teff(stan_input, digits)
}
\arguments{
\item{stan_input}{a list of stan input fields}

\item{digits}{number of digits to show for floats}

\item{parname}{parameter name}
}
\value{
a data frame
}
\description{
Convert the Stan input encoding of a prior to a human-readable format
}
