% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{prior_get_default}
\alias{prior_get_default}
\title{Default prior for a named parameter}
\usage{
prior_get_default(name)
}
\arguments{
\item{name}{parameter name}
}
\value{
a named list
}
\description{
Default prior for a named parameter
}
