% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-prior.R
\name{prior_LonGP}
\alias{prior_LonGP}
\title{Create a similar default prior as in \code{LonGP} (Cheng et. al, 2019)}
\usage{
prior_LonGP()
}
\value{
A list defining a valid \code{prior} argument for the \code{lgp_model} function.
}
\description{
Not recommended, because a lengthscale close to 0 is possible.
}
