% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-prior_pred.R
\name{prior_pred}
\alias{prior_pred}
\alias{sample_param_prior}
\title{Prior (predictive) sampling}
\usage{
prior_pred(
  model,
  verbose = TRUE,
  quiet = FALSE,
  refresh = 0,
  STREAM = get_stream(),
  ...
)

sample_param_prior(model, verbose = TRUE, quiet = FALSE, ...)
}
\arguments{
\item{model}{An object of class \linkS4class{lgpmodel}.}

\item{verbose}{Should more information and a possible progress bar be
printed?}

\item{quiet}{Should all output be suppressed?}

\item{refresh}{Argument for \code{\link[rstan]{sampling}}.}

\item{STREAM}{External pointer. By default obtained with
\code{rstan::get_stream()}.}

\item{...}{Additional arguments for \code{\link[rstan]{sampling}}.}
}
\description{
\itemize{
  \item The \code{prior_pred} function takes an
  \linkS4class{lgpmodel} object, samples from the prior predictive
  distribution and returns an object of class \linkS4class{Prediction}.
  \item The \code{sample_param_prior} function takes an
  \linkS4class{lgpmodel} object, samples only its parameter prior
  using \code{\link[rstan]{sampling}} and returns
  an object of class \code{\link[rstan]{stanfit}}.
}
}
\seealso{
Other main functions: 
\code{\link{create_model}()},
\code{\link{get_draws}()},
\code{\link{lgp}()},
\code{\link{pred}()},
\code{\link{sample_model}()}
}
\concept{main functions}
