% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgpr.R
\name{prior_predict}
\alias{prior_predict}
\title{Prior predictive check}
\usage{
prior_predict(formula, data, ...)
}
\arguments{
\item{formula}{A formula specified using the common \code{formula}
syntax, such as \code{y ~ x1 + x2:z1 + } \code{x2:z2 + z2}.
\itemize{
  \item The formula must contain exatly one tilde (\code{~}), with response
  variable on the left-hand side and model terms on the right-hand side.
  \item Terms are be separated by a plus (\code{+}) sign.
  \item Terms can consits of a single variable name, such as \code{x}, or
  an interaction of two variables, such as \code{x:z}.
  \item In single-variable terms, the variable can be either continuous or
  categorical, whereas in interaction terms the variable
  on the left-hand side of the colon (\code{:}) has to be continuous and the
  one on the right-hand side has to be categorical (a factor).
  \item All variables appearing in \code{formula} must be
  found in \code{data}.
}}

\item{data}{A \code{data.frame} where each column corresponds to one
variable, and each row is one observation. Continuous covariates and the
response variable must have type \code{"numeric"} and categorical covariates
must have type \code{"factor"}. Missing values should be indicated with
\code{NaN} or \code{NA}. The response variable cannot contain missing
values.}

\item{...}{keyword arguments to \code{\link{lgp}}}
}
\value{
an array of shape \code{num_draws} x \code{num_obs}
}
\description{
Prior predictive check
}
