% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-prior.R
\name{prior_stan_to_readable}
\alias{prior_stan_to_readable}
\title{Human-readable information about the priors in the Stan data object}
\usage{
prior_stan_to_readable(stan_dat)
}
\arguments{
\item{stan_dat}{The list that is passed as data to \code{rstan::sampling}.}
}
\value{
Info as a string.
}
\description{
Human-readable information about the priors in the Stan data object
}
