% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-prior.R
\name{prior_statement}
\alias{prior_statement}
\title{Human-readable prior statement}
\usage{
prior_statement(parname, TYP, P, dist, row_change = TRUE)
}
\arguments{
\item{parname}{parameter name}

\item{TYP}{two integers}

\item{P}{three real numbers}

\item{dist}{list of distribution names}

\item{row_change}{should a newline be last character?}
}
\value{
Sampling statement as a string.
}
\description{
Human-readable prior statement
}
