% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_s4_generics.R, R/s4-summary.R
\name{prior_summary}
\alias{prior_summary}
\alias{prior_summary,lgpmodel-method}
\alias{prior_summary,lgpfit-method}
\title{Summary of model prior}
\usage{
prior_summary(object, ...)

\S4method{prior_summary}{lgpmodel}(object, digits = 3)

\S4method{prior_summary}{lgpfit}(object, digits = 3)
}
\arguments{
\item{object}{an object of class \linkS4class{lgpmodel} or
\linkS4class{lgpfit}}

\item{...}{other arguments}

\item{digits}{number of digits to show for floating point numbers}
}
\value{
a data frame
}
\description{
Summary of model prior
}
