% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-prior.R
\name{prior_to_df}
\alias{prior_to_df}
\title{Convert the given prior in a Stan input to a human-readable format}
\usage{
prior_to_df(stan_input, digits = 3)
}
\arguments{
\item{stan_input}{a list of stan input fields}

\item{digits}{number of digits to show for floats}
}
\value{
a data frame
}
\description{
Convert the given prior in a Stan input to a human-readable format
}
