% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{prior_to_df_oneparam}
\alias{prior_to_df_oneparam}
\title{Convert the given prior for one parameter type into a human-readable format}
\usage{
prior_to_df_oneparam(stan_input, parname, digits)
}
\arguments{
\item{stan_input}{a list of stan input fields}

\item{parname}{parameter name}

\item{digits}{number of digits to show for floats}
}
\value{
a data frame
}
\description{
Convert the given prior for one parameter type into a human-readable format
}
