% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preproc-parse_prior.R
\name{prior_to_stan}
\alias{prior_to_stan}
\title{Get priors as a format that can be input to Stan}
\usage{
prior_to_stan(D, prior, HMGNS, UNCRT, N_cases, T_observed, T_last)
}
\arguments{
\item{D}{an integer vector of length 6}

\item{prior}{The \code{prior} argument supplied to \code{lgp()}.}

\item{HMGNS}{Is diseaseAge assumed to have a homogenous effect (1) or not (0)?}

\item{UNCRT}{Boolean value, is uncertainty of disease onset modeled?}

\item{N_cases}{number of case individuals}

\item{T_observed}{observed disease onsets}

\item{T_last}{last time point for each diseased individual}
}
\value{
a list with all things related to priors that Stan needs
}
\description{
Get priors as a format that can be input to Stan
}
