% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{prior_to_str}
\alias{prior_to_str}
\title{Human-readable prior statement}
\usage{
prior_to_str(parname, prior, hyper, digits)
}
\arguments{
\item{parname}{parameter name}

\item{prior}{two integers}

\item{hyper}{three real numbers}

\item{digits}{number of digits to show for floats}
}
\value{
A list.
}
\description{
Human-readable prior statement
}
