% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-priors.R
\name{prior_type_names}
\alias{prior_type_names}
\title{Names of allowed  prior types}
\usage{
prior_type_names(idx = NULL)
}
\arguments{
\item{idx}{an integer or NULL}
}
\value{
a character vector
}
\description{
Names of allowed  prior types
}
