% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-utils.R
\name{progbar}
\alias{progbar}
\alias{progbar_header}
\alias{progbar_print}
\title{Progress bar for iterative functions}
\usage{
progbar_header(L)

progbar_print(idx, idx_print)
}
\arguments{
\item{L}{length of bar}

\item{idx}{current iteration index}

\item{idx_print}{indices of iterations to print a bar block}
}
\description{
\itemize{
  \item \code{progbar_header} creates header for a progress bar
  \item \code{progbar_print} prints part of the bar depending on iteration
  index
}
}
