% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{progbar_header}
\alias{progbar_header}
\title{Create header for a progress bar}
\usage{
progbar_header(L, width = 4)
}
\arguments{
\item{L}{length}

\item{width}{width of one block}
}
\value{
s string
}
\description{
Create header for a progress bar
}
