% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-inv_gamma.R
\name{qinvgamma_stanlike}
\alias{qinvgamma_stanlike}
\title{Quantiles of inverse gamma distribution pdf with the same parametrization
as in Stan}
\usage{
qinvgamma_stanlike(p, alpha, beta)
}
\arguments{
\item{p}{quantile (must be between 0 and 1)}

\item{alpha}{positive real number}

\item{beta}{positive real number}
}
\value{
a positive real number
}
\description{
See
\href{https://mc-stan.org/docs/2_21/functions-reference/inverse-gamma-distribution.html}
{here}.
}
