% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-datautil.R
\name{read_proteomics_data}
\alias{read_proteomics_data}
\title{Function for reading the built-in proteomics data}
\usage{
read_proteomics_data(parentDir = NULL, protein = NULL, verbose = TRUE)
}
\arguments{
\item{parentDir}{Path to local parent directory for the data.
If this is \code{NULL}, data is downloaded from
\url{https://github.com/jtimonen/lgpr-usage/tree/master/data/proteomics}.}

\item{protein}{Index or name of protein.}

\item{verbose}{Can this print some output?}
}
\value{
a \code{data.frame}
}
\description{
Function for reading the built-in proteomics data
}
\concept{built-in data}
