% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-components.R
\name{reduce_factors_gp}
\alias{reduce_factors_gp}
\title{Check for gp expressions in a term}
\usage{
reduce_factors_gp(factors)
}
\arguments{
\item{factors}{list of \linkS4class{lgpexpression} objects}
}
\value{
an updated list with no \code{gp*} expressions, and name of
the covariate in the original \code{gp*} expression
}
\description{
Check for gp expressions in a term
}
