% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-idx_expand.R
\name{reduce_rows}
\alias{reduce_rows}
\title{Check that each row of array is identical}
\usage{
reduce_rows(rows)
}
\arguments{
\item{rows}{rows}
}
\value{
the indentical row
}
\description{
Check that each row of array is identical
}
