% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relevances.R
\name{relevances}
\alias{relevances}
\alias{relevances.f_sampled}
\alias{relevances.f_marginalized}
\title{Assess component relevances}
\usage{
relevances(fit)

relevances.f_sampled(fit)

relevances.f_marginalized(fit)
}
\arguments{
\item{fit}{an object of class \code{lgpfit}}
}
\description{
\itemize{
  \item \code{relevances} assesses covariate relevances
  \item \code{relevances.f_marginalized} assesses covariate relevances for a
  model where the latent function \code{f} was not sampled
  \item \code{relevances.f_sampled} assesses covariate relevances for a
  model where the latent function \code{f} was sampled
}
}
