% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-relevances.R
\name{relevances_default}
\alias{relevances_default}
\alias{relevances.default.noise}
\alias{relevances.default.comp}
\title{Helper functions for relevances.default}
\usage{
relevances.default.noise(fit, pred)

relevances.default.comp(fit, pred)
}
\arguments{
\item{fit}{an object of class \code{lgpfit}}

\item{pred}{an object of class \linkS4class{FunctionDraws} or
\linkS4class{FunctionPosterior}}
}
\value{
an array or vector
}
\description{
\itemize{
  \item \code{relevances.default.noise} computes the noise proportion
  in each draws, and returns a vector with length \code{num_draws}
  \item \code{relevances.default.comp} divides the signal proportion for
  each component in each draw, and returns a data frame with
  \code{num_draws} rows and \code{num_comps} colums
}
}
