% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-misc.R
\name{rep_df}
\alias{rep_df}
\title{Repeat a data frame vertically}
\usage{
rep_df(df, times)
}
\arguments{
\item{df}{a data frame}

\item{times}{number of times to repeat}
}
\value{
a data frame
}
\description{
Repeat a data frame vertically
}
\seealso{
Other data utilities: 
\code{\link{add_factor_crossing}()},
\code{\link{data_types}()},
\code{\link{to_long_format}()}
}
\concept{data utilities}
