% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-utils.R
\name{repvec}
\alias{repvec}
\title{Repeat a vector as a rows of an array}
\usage{
repvec(v, n)
}
\arguments{
\item{v}{a vector of length \code{m}}

\item{n}{number of times to repeat}
}
\value{
returns an array of size \code{n} x \code{m}
}
\description{
Repeat a vector as a rows of an array
}
