% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula.R
\name{rhs_variables}
\alias{rhs_variables}
\title{Get names of all variables appearing on formula right-hand side}
\usage{
rhs_variables(rhs)
}
\arguments{
\item{rhs}{an object of class \linkS4class{lgprhs}}
}
\value{
a list of variable names
}
\description{
Get names of all variables appearing on formula right-hand side
}
\seealso{
Other model formula functions: 
\code{\link{formula_to_advanced}()},
\code{\link{is_advanced_formula}()},
\code{\link{parse_expr}()},
\code{\link{parse_formula}()},
\code{\link{parse_rhs}()},
\code{\link{parse_term}()},
\code{\link{parse_whitespace}()},
\code{\link{rhs_to_terms}()},
\code{\link{term_variables}()}
}
\concept{model formula functions}
