% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_array.R
\name{row_vars}
\alias{row_vars}
\title{Compute row variances for a 2-dimensional array}
\usage{
row_vars(x)
}
\arguments{
\item{x}{an array of size \code{n} x \code{m}}
}
\value{
returns a vector
}
\description{
Compute row variances for a 2-dimensional array
}
\seealso{
Other array utilities: 
\code{\link{add_to_columns}()},
\code{\link{apply_reduce}()},
\code{\link{arr3_select}()},
\code{\link{arr3_to_list}()},
\code{\link{array_to_arraylist}()},
\code{\link{ensure_2dim}()},
\code{\link{make_draw_df}()},
\code{\link{normalize_rows}()},
\code{\link{null_if_all_nan}()},
\code{\link{reduce_rows}()},
\code{\link{repvec}()},
\code{\link{select_row}()},
\code{\link{squeeze_second_dim}()}
}
\concept{array utilities}
