% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-fit_model.R
\name{sample_model}
\alias{sample_model}
\alias{optimize_model}
\title{Fitting a model}
\usage{
sample_model(model, ...)

optimize_model(model, ...)
}
\arguments{
\item{model}{An object of class \linkS4class{lgpmodel}.}

\item{...}{Optional arguments passed to
\code{\link[rstan]{sampling}} or \code{\link[rstan]{optimizing}}.}
}
\description{
\itemize{
  \item The \code{sample_model} function takes an \linkS4class{lgpmodel}
  object, fits it using  \code{\link[rstan]{sampling}}, and
  returns an object of class \linkS4class{lgpfit}.
  \item The \code{optimize_model} function takes an \linkS4class{lgpmodel}
  object and fits it using \code{\link[rstan]{optimizing}}.
}
}
\seealso{
Other main functions: 
\code{\link{create_model}()},
\code{\link{lgp}()}
}
\concept{main functions}
