% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{scale_color}
\alias{scale_color}
\alias{scale_fill}
\title{A color scales}
\usage{
scale_color(n)

scale_fill(n)
}
\arguments{
\item{n}{an integer from 1 to 5}
}
\value{
a ggplot object
}
\description{
A color scales
}
\seealso{
Other color utilities: 
\code{\link{color_palette}()},
\code{\link{colorset}()},
\code{\link{plot_color_palette}()}
}
\concept{color utilities}
