% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-lgpfit_helper.R
\name{scale_f_post_helper}
\alias{scale_f_post_helper}
\title{Helper function}
\usage{
scale_f_post_helper(fun, arr)
}
\arguments{
\item{fun}{a function}

\item{arr}{an array of shape \code{S} x \code{n}}
}
\value{
an array with same shape as \code{arr}
}
\description{
Helper function
}
