% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate-create_F.R
\name{scaleRelevances}
\alias{scaleRelevances}
\title{Scale the effect sizes}
\usage{
scaleRelevances(FFF, relevances, force_zeromean, i_dis)
}
\arguments{
\item{FFF}{matrix where one column corresponds to one additive data component}

\item{relevances}{the desired variance of each component (column)}

\item{force_zeromean}{Should each component (excluding the disease age component)
be forced to have a zero mean.}

\item{i_dis}{index of a component for which the zero-mean forcing is skipped}
}
\value{
a new matrix \code{FFF}
}
\description{
Scale the effect sizes
}
