% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{select.all_draws}
\alias{select.all_draws}
\title{Selection probabilities using a fixed threshold}
\usage{
select.all_draws(rel, threshold)
}
\arguments{
\item{rel}{An array of shape \code{num_draws} x \code{num_comps+1},
where rows sum to 1.}

\item{threshold}{Threshold for relevance sum.
Must be a value between 0 and 1.}
}
\value{
a binary array with same shape as \code{rel}
}
\description{
Selection probabilities using a fixed threshold
}
