% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{select.helper}
\alias{select.helper}
\alias{select.all_thresholds}
\alias{select_freq.all_thresholds}
\alias{select.compute_integral}
\alias{select.all_draws}
\alias{select.one_draw}
\alias{result_df}
\title{Helper functions for component selection}
\usage{
select.all_thresholds(fit, reduce, p, h, verbose, ...)

select_freq.all_thresholds(fit, p, h, verbose, ...)

select.compute_integral(arr, p)

select.all_draws(rel, threshold)

select.one_draw(rel, threshold)

result_df(val, val_name, comp_names)
}
\arguments{
\item{fit}{An object of class \code{lgpfit}.}

\item{reduce}{The \code{reduce} argument for \code{\link{relevances}}.}

\item{p}{A threshold density over interval [0,1].}

\item{h}{A discretization parameter for computing a quadrature.}

\item{verbose}{Should this show a progress bar?}

\item{...}{Additional arguments to \code{\link{relevances}}.}

\item{arr}{an array of frequencies or selections}

\item{rel}{an array of relevances}

\item{threshold}{Threshold for relevance sum.
Must be a value between 0 and 1.}

\item{val}{values}

\item{val_name}{name of variable that \code{val} presents}

\item{comp_names}{component names}
}
\description{
Helper functions for component selection
}
