% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{select.one_draw}
\alias{select.one_draw}
\title{Select relevant components}
\usage{
select.one_draw(rel, threshold)
}
\arguments{
\item{rel}{A vector of component relevances that sums to 1, noise being
the last element.}

\item{threshold}{Threshold for relevance sum.
Must be a value between 0 and 1.}
}
\value{
indices of selected components (including "noise" always)
}
\description{
Select relevant components
}
