% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_data.R
\name{select_factors_and}
\alias{select_factors_and}
\title{Select data columns which are factors or have a certain name}
\usage{
select_factors_and(data, valid)
}
\arguments{
\item{data}{a data frame}

\item{valid}{names of variables that do not need to be factors}
}
\value{
a data frame
}
\description{
Select data columns which are factors or have a certain name
}
\seealso{
Other data utilities: 
\code{\link{add_dis_age}()},
\code{\link{add_factor}()},
\code{\link{data_types}()},
\code{\link{get_teff_obs}()},
\code{\link{new_x}()},
\code{\link{pick_one_row_each}()},
\code{\link{split}()},
\code{\link{valid_data}}
}
\concept{data utilities}
