% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{select_integrate}
\alias{select_integrate}
\alias{select_integrate.all}
\alias{select_integrate.integrate}
\title{Compute expected selection frequencies by integrating the threshold
over [0,1]}
\usage{
select_integrate(
  fit,
  p = function(x) {     stats::dbeta(x, 100, 5) },
  h = 0.01,
  show_progbar = TRUE,
  ...
)

select_integrate.all(fit, p, h, show_progbar, ...)

select_integrate.integrate(freqs, p)
}
\arguments{
\item{p}{The threshold density over interval [0,1].}

\item{h}{A discretization parameter for computing a quadrature.}

\item{show_progbar}{Should this show a progress bar?}

\item{...}{additional arguments to \code{\link{relevances}}}

\item{freqs}{an object returned by \code{select_integrate.all}}

\item{object}{An object of class \code{lgpfit}.}
}
\value{
A named list
}
\description{
Compute expected selection frequencies by integrating the threshold
over [0,1]
}
