% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-utils.R
\name{select_row}
\alias{select_row}
\title{Select named row of an array}
\usage{
select_row(x, name)
}
\arguments{
\item{x}{an array of shape \code{n} x \code{m}}

\item{name}{name of the row}
}
\value{
a vector of length \code{m}
}
\description{
Select named row of an array
}
\seealso{
Other array utilities: 
\code{\link{add_to_columns}()},
\code{\link{apply_reduce}()},
\code{\link{arr3_select}()},
\code{\link{arr3_to_list}()},
\code{\link{array_to_arraylist}()},
\code{\link{ensure_2dim}()},
\code{\link{make_draw_df}()},
\code{\link{normalize_rows}()},
\code{\link{null_if_all_nan}()},
\code{\link{reduce_rows}()},
\code{\link{repvec}()},
\code{\link{row_vars}()},
\code{\link{squeeze_second_dim}()}
}
\concept{array utilities}
