% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postproc-selection.R
\name{selection}
\alias{selection}
\title{Selection of relevant components using a threshold
for the total proportion of explained variance}
\usage{
selection(object, threshold = 0.95)
}
\arguments{
\item{object}{An object of class \code{lgpfit}.}

\item{threshold}{A value between 0 and 1}
}
\value{
A named list
}
\description{
Selection of relevant components using a threshold
for the total proportion of explained variance
}
