% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postproc-selection.R
\name{selection_fixed_threshold}
\alias{selection_fixed_threshold}
\title{Select relevant components}
\usage{
selection_fixed_threshold(rel, threshold)
}
\arguments{
\item{rel}{a named vector of component relevances}

\item{threshold}{value between 0 and 1}
}
\value{
indices of selected components (including "noise" always)
}
\description{
Select relevant components
}
