% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postproc-selection.R
\name{selection_prob_fixed_threshold}
\alias{selection_prob_fixed_threshold}
\title{Selection probabilities using a fixed threshold}
\usage{
selection_prob_fixed_threshold(object, threshold = 0.95)
}
\arguments{
\item{object}{An object of class \code{lgpfit}.}

\item{threshold}{value between 0 and 1}
}
\value{
selection probabilities for each component
}
\description{
Selection probabilities using a fixed threshold
}
