% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postproc-selection.R
\name{selection_prob_plot}
\alias{selection_prob_plot}
\title{Plot of probabilistic selection of relevant components}
\usage{
selection_prob_plot(PROB, H, P)
}
\arguments{
\item{PROB}{computed probabilities at points H}

\item{H}{a grid on interval [0,1]}

\item{P}{threshold probability distribution evaluated at H}
}
\value{
a ggplot object
}
\description{
Plot of probabilistic selection of relevant components
}
