% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postproc-helper.R
\name{separate_effects}
\alias{separate_effects}
\title{Separate the covariate effects from an interaction components of a categorical 
covariate and age}
\usage{
separate_effects(f_post, t, D, ell, i_edit)
}
\arguments{
\item{f_post}{a matrix of size \code{n} x \code{sum(D)}}

\item{t}{vector of \code{n} time points corresponding to \code{f_post}}

\item{D}{a vector of length 6}

\item{ell}{kernel lengthscale}

\item{i_edit}{Indices of columns whose effect should be moved to shared age.}
}
\value{
a corrected \code{f_post}
}
\description{
Separate the covariate effects from an interaction components of a categorical 
covariate and age
}
