% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preproc-helper.R
\name{set_basis}
\alias{set_basis}
\title{Set basis functions}
\usage{
set_basis(n_basisfun, L_basisfun)
}
\arguments{
\item{n_basisfun}{number of basis functions}

\item{L_basisfun}{basis function domain size}
}
\value{
a list
}
\description{
Set basis functions
}
