% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preproc-helper.R
\name{set_C_hat}
\alias{set_C_hat}
\title{Set C_hat (Poisson and NB observation models)}
\usage{
set_C_hat(C_hat, response, LH)
}
\arguments{
\item{C_hat}{the C_hat argument given as input to \code{lgp}}

\item{response}{response variable}

\item{LH}{likelihood as int}
}
\value{
a real number
}
\description{
Set C_hat (Poisson and NB observation models)
}
