% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preproc-helper.R
\name{set_N_trials}
\alias{set_N_trials}
\title{Set N_trials (binomial and Bernoulli observation models)}
\usage{
set_N_trials(N_trials, response, LH)
}
\arguments{
\item{N_trials}{the N_trials argument given as input to \code{lgp}}

\item{response}{response variable}

\item{LH}{likelihood as int}
}
\value{
a numeric vector
}
\description{
Set N_trials (binomial and Bernoulli observation models)
}
