% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preproc-helper.R
\name{set_norm_factors}
\alias{set_norm_factors}
\title{Check and possibly edit the norm_factors input}
\usage{
set_norm_factors(norm_factors, response, LH)
}
\arguments{
\item{norm_factors}{the \code{norm_factors} argument given as input to 
\code{lgp_model}}

\item{response}{response variable}

\item{LH}{likelihood as int}
}
\value{
a numeric vector
}
\description{
Check and possibly edit the norm_factors input
}
