% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-likelihood.R
\name{set_num_cat}
\alias{set_num_cat}
\title{Count numbers of different categories for each categorical variable}
\usage{
set_num_cat(X, D)
}
\arguments{
\item{X}{the design matrix}

\item{D}{a vector of length 6}
}
\value{
a numeric vector
}
\description{
Count numbers of different categories for each categorical variable
}
