% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-likelihood.R
\name{set_num_trials}
\alias{set_num_trials}
\title{Set num_trials (binomial and Bernoulli observation models)}
\usage{
set_num_trials(num_trials, response, LH)
}
\arguments{
\item{num_trials}{the \code{num_trials} argument given as input to
\code{lgp_model}}

\item{response}{response variable}

\item{LH}{likelihood as int}
}
\value{
a numeric vector
}
\description{
Set num_trials (binomial and Bernoulli observation models)
}
