% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_functions.R
\name{sf_warp_input}
\alias{sf_warp_input}
\title{Warp input (Cpp)}
\usage{
sf_warp_input(t, a, b, c)
}
\arguments{
\item{t}{an input}

\item{a}{an input}

\item{b}{an input}

\item{c}{an input}
}
\description{
Warp input (Cpp)
}
