% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4-show.R
\name{show}
\alias{show}
\alias{show,lgpfit-method}
\alias{show,lgpmodel-method}
\alias{show,lgpsim-method}
\title{Printing S4 object info using the show generic}
\usage{
\S4method{show}{lgpformula}(object)

\S4method{show}{lgpfit}(object)

\S4method{show}{lgpmodel}(object)

\S4method{show}{lgpsim}(object)
}
\arguments{
\item{object}{an object of some S4 class}
}
\value{
the object invisibly
}
\description{
Printing S4 object info using the show generic
}
