% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4-show.R
\name{show_default}
\alias{show_default}
\title{Print the character representation of an S4 object}
\usage{
show_default(object)
}
\arguments{
\item{object}{an object of some S4 class}
}
\value{
the object invisibly
}
\description{
Print the character representation of an S4 object
}
