% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_s4_classes.R
\name{show,lgpfit-method}
\alias{show,lgpfit-method}
\title{Show a summary of results of the \code{lgp} function}
\usage{
\S4method{show}{lgpfit}(object)
}
\arguments{
\item{object}{an object of class \code{lgpfit}}
}
\value{
nothing
}
\description{
Show a summary of results of the \code{lgp} function
}
