% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_s4_classes.R
\name{show,lgpformula-method}
\alias{show,lgpformula-method}
\title{Show a summary of an \code{lgpformula}}
\usage{
\S4method{show}{lgpformula}(object)
}
\arguments{
\item{object}{an object of class \code{lgpformula}}
}
\description{
Show a summary of an \code{lgpformula}
}
