% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_s4_classes.R
\name{show,lgpmodel-method}
\alias{show,lgpmodel-method}
\title{Show a summary of an \code{lgpmodel}}
\usage{
\S4method{show}{lgpmodel}(object)
}
\arguments{
\item{object}{an object of class \code{lgpmodel}}
}
\value{
nothing
}
\description{
Show a summary of an \code{lgpmodel}
}
