% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_s4_classes.R
\name{show,lgpsum-method}
\alias{show,lgpsum-method}
\title{Show a summary of an \code{lgpsum}}
\usage{
\S4method{show}{lgpsum}(object)
}
\arguments{
\item{object}{an object of class \code{lgpsum}}
}
\description{
Show a summary of an \code{lgpsum}
}
