% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-utils.R
\name{show_relevances}
\alias{show_relevances}
\title{Print info about component and covariate relevances}
\usage{
show_relevances(fit)
}
\arguments{
\item{fit}{an object of class \code{lgpfit}}
}
\value{
nothing
}
\description{
Print info about component and covariate relevances
}
