% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethodsFit.R
\name{signal}
\alias{signal}
\alias{get_f_total}
\alias{get_f_components}
\alias{get_g}
\title{Extract modeled signal or its components}
\usage{
get_f_total(fit, field = "mean")

get_f_components(fit, field = "mean")

get_g(fit)
}
\arguments{
\item{fit}{an object of class \linkS4class{lgpfit}}

\item{field}{Should be \code{"mean"}, \code{"std"}, or \code{NULL}.}
}
\value{
an array of shape \code{num_samples} x \code{num_obs} (or a list of
them for \code{get_f_components})
}
\description{
\itemize{
  \item \code{get_f_total} gets total signal for each sample
  \item \code{get_f_components} gets signal components for each sample
  \item \code{get_g} gets total signal, mapped through
  the inverse link function, for each sample
}
}
