% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{sim.apply_onset_fun}
\alias{sim.apply_onset_fun}
\title{Check if trying to simulate onset happening after measurement interval}
\usage{
sim.apply_onset_fun(fun_obs, t0_real, rem)
}
\arguments{
\item{fun_obs}{function for generating the observed onset
from real onset}

\item{t0_real}{the real onset}

\item{rem}{measurement points after real onset}
}
\value{
a list with names \code{idx0} and \code{t0}
}
\description{
Check if trying to simulate onset happening after measurement interval
}
