% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-simulate.R
\name{sim.check_covariates}
\alias{sim.check_covariates}
\title{Input check for the covariates-related arguments of \code{simulate_data}}
\usage{
sim.check_covariates(covariates, relevances, names, n_cat)
}
\arguments{
\item{covariates}{argument to \code{simulate_data}}

\item{relevances}{argument to \code{simulate_data}}

\item{names}{argument to \code{simulate_data}}

\item{n_cat}{the \code{n_categs} argument to \code{simulate_data}}
}
\value{
the covariate names
}
\description{
Input check for the covariates-related arguments of \code{simulate_data}
}
