% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-simulate.R
\name{sim.check_too_far}
\alias{sim.check_too_far}
\title{Check if trying to simulate onset happening after measurement interval}
\usage{
sim.check_too_far(idx, rem)
}
\arguments{
\item{idx}{index of onset point relative to real onset}

\item{rem}{measurement points after real onset}
}
\description{
Check if trying to simulate onset happening after measurement interval
}
