% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{sim.create_x_D}
\alias{sim.create_x_D}
\title{Helper function for sim.create_x}
\usage{
sim.create_x_D(covariates)
}
\arguments{
\item{covariates}{Integer vector that defines the types of covariates
(other than id and age). If not given, only the id and age
covariates are created. Different integers correspond to the following
covariate types:
\itemize{
  \item 0 = disease-related age
  \item 1 = other continuous covariate
  \item 2 = a categorical covariate that interacts with age
  \item 3 = a categorical covariate that acts as a group offset
  \item 4 = a categorical covariate that that acts as a group offset AND
  is restricted to have value 0 for controls and 1 for cases
}}
}
\value{
an array of five integers
}
\description{
Helper function for sim.create_x
}
