% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-simulate.R
\name{sim.create_x_other}
\alias{sim.create_x_other}
\title{Helper function for sim.create_x}
\usage{
sim.create_x_other(X, k, N, D, n_categs, dis_age, continuous_info)
}
\arguments{
\item{X}{current covariate matrix}

\item{k}{number of time points per individual}

\item{N}{Number of individuals.}

\item{D}{covariate type array}

\item{n_categs}{An integer vector defining the number of categories
for each categorical covariate, so that \code{length(n_categs)} equals to
the number of 2's and 3's in the \code{covariates} vector.}

\item{dis_age}{a disease-age vector}

\item{continuous_info}{Info for generating continuous covariates. Must be a
list containing fields \code{lambda} and \code{mu}, which have length 3.
The continuous covariates are generated so that \code{x <- sin(a*t + b) + c},
where
\itemize{
  \item \code{t <- seq(0, 2*pi, length.out = k)}
  \item \code{a <- mu[1] + lambda[1]*stats::runif(1)}
  \item \code{b <- mu[2] + lambda[2]*stats::runif(1)}
  \item \code{c <- mu[3] + lambda[3]*stats::runif(1)}
}}
}
\value{
a list
}
\description{
Helper function for sim.create_x
}
