% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-simulate.R
\name{sim.data_to_observed}
\alias{sim.data_to_observed}
\title{Real generated disease ages to observed ones}
\usage{
sim.data_to_observed(dat, t_observed)
}
\arguments{
\item{dat}{data frame}

\item{t_observed}{Determines how the disease onset is observed. See
documentation of \code{\link{simulate_data}}.}
}
\value{
a new data frame and observed onsets
}
\description{
Real generated disease ages to observed ones
}
