% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-simulate.R
\name{sim.disease_effect}
\alias{sim.disease_effect}
\title{Draw disease component from a parameteric form}
\usage{
sim.disease_effect(X_id, X_disAge, dis_fun)
}
\arguments{
\item{X_id}{the id covariate}

\item{X_disAge}{the diseaseAge covariate}

\item{dis_fun}{the disease age effect function}
}
\value{
a vector
}
\description{
Draw disease component from a parameteric form
}
