% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-simulate.R
\name{sim.draw_categorical}
\alias{sim.draw_categorical}
\title{Independently draw categorical variables for each individual}
\usage{
sim.draw_categorical(N, k, v)
}
\arguments{
\item{N}{number of individuals}

\item{k}{number of timepoints}

\item{v}{vector of numbers of different categories}
}
\value{
a data frame with \code{N} rows and \code{D} = \code{length(v)}
columns
}
\description{
Independently draw categorical variables for each individual
}
