% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-simulate.R
\name{sim.draw_components}
\alias{sim.draw_components}
\title{Draw realizations of multivariate normals}
\usage{
sim.draw_components(KK)
}
\arguments{
\item{KK}{3D matrix where \code{KK[,,j]} is the \code{j}th kernel matrix}
}
\value{
a matrix \code{FFF}
}
\description{
Draw realizations of multivariate normals
}
