% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-simulate.R
\name{sim.draw_continuous}
\alias{sim.draw_continuous}
\title{Indepedently draw continuous variables for each individual}
\usage{
sim.draw_continuous(N, k, D, mu, lambda)
}
\arguments{
\item{N}{number of individuals}

\item{k}{number of timepoints}

\item{D}{number of variables}

\item{mu}{a vector of length 3}

\item{lambda}{a vector of length 3}
}
\value{
a matrix of size \code{N} x \code{D}
}
\description{
Indepedently draw continuous variables for each individual
}
