% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-simulate.R
\name{sim.draw_measurement_times}
\alias{sim.draw_measurement_times}
\title{Draw the age covariate}
\usage{
sim.draw_measurement_times(N, t_data, t_jitter)
}
\arguments{
\item{N}{number of individuals}

\item{t_data}{a vector of length \code{k}}

\item{t_jitter}{Standard deviation of the jitter added to the given
measurement times.}
}
\value{
a vector of length \code{N*k}
}
\description{
Draw the age covariate
}
