% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-simulate.R
\name{sim.generate_names}
\alias{sim.generate_names}
\title{Generate names for covariates}
\usage{
sim.generate_names(covariates)
}
\arguments{
\item{covariates}{vector of covariate types}
}
\value{
covariate names
}
\description{
Generate names for covariates
}
