% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{sim.kernel_ns}
\alias{sim.kernel_ns}
\title{Compute a nonstationary kernel matrix using input warping}
\usage{
sim.kernel_ns(x1, x2 = NULL, alpha = 1, ell, a, b, c)
}
\arguments{
\item{x1}{vector of length n}

\item{x2}{vector of length m}

\item{alpha}{marginal std (default = 1)}

\item{ell}{lengthscale in the warped space}

\item{a}{steepness of the warping function rise}

\item{b}{location of the effective time window}

\item{c}{maximum range}
}
\value{
A kernel matrix of size n x m
}
\description{
Compute a nonstationary kernel matrix using input warping
}
