% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-simulate.R
\name{sim.name_components}
\alias{sim.name_components}
\title{Create names for all components based on covariate names and types}
\usage{
sim.name_components(types, names)
}
\arguments{
\item{types}{vector of covariate types}

\item{names}{names of the covariates}
}
\value{
a vector of component names
}
\description{
Create names for all components based on covariate names and types
}
