% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{sim.onsets_to_dis_age}
\alias{sim.onsets_to_dis_age}
\title{Compute the disease-related ages}
\usage{
sim.onsets_to_dis_age(onsets, age, k)
}
\arguments{
\item{onsets}{true disease effect times, a vector of length \code{N}}

\item{age}{the age covariate, a vector of length \code{N*k}}

\item{k}{number of measurements per individual}
}
\value{
the diseaseAge covariate, a vector of length \code{N*k}
}
\description{
Compute the disease-related ages
}
