% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-simulate.R
\name{sim.rtgeom}
\alias{sim.rtgeom}
\title{Sample from the 'truncated geometric' distribution}
\usage{
sim.rtgeom(s, p, n = 1)
}
\arguments{
\item{s}{an integer}

\item{p}{a number between 0 and 1}

\item{n}{number of samples}
}
\value{
an integer from the interval 1...n
}
\description{
Sample from the 'truncated geometric' distribution
}
