% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-simulate.R
\name{sim.scale_relevances}
\alias{sim.scale_relevances}
\title{Scale the effect sizes}
\usage{
sim.scale_relevances(FFF, relevances, force_zeromean, i_skip)
}
\arguments{
\item{FFF}{matrix where one column corresponds to one additive component}

\item{relevances}{the desired variance of each component (column)}

\item{force_zeromean}{Should each component (excluding the disease age
component) be forced to have a zero mean.}

\item{i_skip}{induces of components for which the zero-mean forcing is
skipped}
}
\value{
a new matrix \code{FFF}
}
\description{
Scale the effect sizes
}
