% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-simulate.R
\name{sim.warp_input}
\alias{sim.warp_input}
\title{Input warping function}
\usage{
sim.warp_input(t, a, b, c)
}
\arguments{
\item{t}{a vector}

\item{a}{steepness of the rise}

\item{b}{location of the effective time window}

\item{c}{maximum range}
}
\value{
a vector of warped inputs \code{w(t)}
}
\description{
Input warping function
}
