% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-covariates.R
\name{sim_create_x_check}
\alias{sim_create_x_check}
\title{Input check helper function for sim_create_x}
\usage{
sim_create_x_check(n_categs, D, t_data, t_effect_range)
}
\arguments{
\item{n_categs}{An integer vector defining the number of categories
for each categorical covariate, so that \code{length(n_categs)} equals to
the number of 2's and 3's in the \code{covariates} vector.}

\item{D}{covariate type array}

\item{t_data}{Measurement times.}

\item{t_effect_range}{Time interval from which the disease effect times are
sampled uniformly. Alternatively, This can any function that returns the
(possibly randomly generated) real disease effect time for one individual.}
}
\value{
a list
}
\description{
Input check helper function for sim_create_x
}
