% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-covariates.R
\name{sim_create_x_dis_age}
\alias{sim_create_x_dis_age}
\title{Helper function for sim_create_x}
\usage{
sim_create_x_dis_age(X, k, N, D, age, t_effect_range)
}
\arguments{
\item{X}{current covariate matrix}

\item{k}{number of time points per individual}

\item{N}{Number of individuals.}

\item{D}{covariate type array}

\item{age}{the age covariate}

\item{t_effect_range}{Parsed version of \code{t_effect_range}}
}
\value{
list
}
\description{
Helper function for sim_create_x
}
