% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate-main.R
\name{sim_data_to_observed}
\alias{sim_data_to_observed}
\title{Real generated disease ages to observed ones}
\usage{
sim_data_to_observed(dat, p = 0)
}
\arguments{
\item{dat}{data frame}

\item{p}{decay factor}
}
\value{
a new data frame and observed onsets
}
\description{
Real generated disease ages to observed ones
}
