% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-x.R
\name{sim_draw_categorical}
\alias{sim_draw_categorical}
\title{Independently draw categorical variables for each individual}
\usage{
sim_draw_categorical(N, k, v)
}
\arguments{
\item{N}{number of individuals}

\item{k}{number of timepoints}

\item{v}{vector of numbers of different categories}
}
\value{
a matrix of size \code{N} x \code{D}, where \code{D <- length(v)}
}
\description{
Independently draw categorical variables for each individual
}
