% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-f.R
\name{sim_draw_components}
\alias{sim_draw_components}
\title{Draw realizations of multivariate normals}
\usage{
sim_draw_components(KK)
}
\arguments{
\item{KK}{3D matrix where \code{KK[,,j]} is the \code{j}th kernel matrix}
}
\value{
a matrix \code{FFF}
}
\description{
Draw realizations of multivariate normals
}
