% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{sim_generate_names}
\alias{sim_generate_names}
\title{Generate names for covariates}
\usage{
sim_generate_names(covariates)
}
\arguments{
\item{covariates}{vector of covariate types}
}
\value{
covariate names
}
\description{
Generate names for covariates
}
