% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-kernels.R
\name{sim_kernel_bin}
\alias{sim_kernel_bin}
\title{Compute a binary kernel matrix}
\usage{
sim_kernel_bin(x1, x2 = NULL, alpha = 1, pos_class = 1)
}
\arguments{
\item{x1}{(integer) vector of length n}

\item{x2}{(integer) vector of length m}

\item{alpha}{marginal std (default = 1)}

\item{pos_class}{the positive class label}
}
\value{
A kernel matrix of size n x m
}
\description{
Compute a binary kernel matrix
}
