% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{sim_kernel_var_mask}
\alias{sim_kernel_var_mask}
\title{Compute the variance mask kernel matrix}
\usage{
sim_kernel_var_mask(disAge1, disAge2, vm_params, stp, nan_replace = 0)
}
\arguments{
\item{disAge1}{disease-related age covariate vector of length \code{n1}}

\item{disAge2}{disease-related age covariate vector of length \code{n2}}

\item{vm_params}{vector of two mask function parameters}

\item{stp}{input warping steepness}

\item{nan_replace}{value to replace nans in disAge vectors}
}
\value{
a matrix of size \code{n1} x \code{n2}
}
\description{
Compute the variance mask kernel matrix
}
