% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-kernels.R
\name{sim_kernel_zerosum}
\alias{sim_kernel_zerosum}
\title{Compute a zero-sum kernel matrix}
\usage{
sim_kernel_zerosum(x1, x2, M, alpha = 1)
}
\arguments{
\item{x1}{(integer) vector of length n}

\item{x2}{(integer) vector of length m}

\item{M}{number of categories}

\item{alpha}{marginal std (default = 1)}
}
\value{
A (binary) kernel matrix of size n x m
}
\description{
Compute a zero-sum kernel matrix
}
