% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate-main.R
\name{sim_parse_t_obs}
\alias{sim_parse_t_obs}
\title{Parse the t_observed argument of \code{simulate_data}}
\usage{
sim_parse_t_obs(t_observed)
}
\arguments{
\item{t_observed}{a string}
}
\value{
a list with a name and number
}
\description{
Parse the t_observed argument of \code{simulate_data}
}
