% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-plot.R
\name{sim_plot}
\alias{sim_plot}
\title{Visualize a simulated longitudinal data set}
\usage{
sim_plot(
  simdata,
  signal_name = "signal",
  signal_column = "f",
  x_name = "age",
  y_name = "y",
  group_by = "id",
  ...
)
}
\arguments{
\item{simdata}{an object of class \linkS4class{lgpsim}}

\item{signal_name}{signal label to show in legend}

\item{signal_column}{name of the signal column in \code{simdata$components}}

\item{x_name}{name of x-axis variable}

\item{y_name}{name of y-axis variable}

\item{group_by}{grouping variable}

\item{...}{keyword arguments to \code{\link{plot_panel}}}
}
\value{
a \code{ggplot object}
}
\description{
Creates plots where each observation unit has a separate panel.
}
