% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-plot.R
\name{sim_plot_add_onsets}
\alias{sim_plot_add_onsets}
\title{Add real and observed disease onsets to plot}
\usage{
sim_plot_add_onsets(h, effect_times, N)
}
\arguments{
\item{h}{a \code{ggplot} object}

\item{effect_times}{a list}

\item{N}{number of individuals}
}
\value{
a list containing the updated \code{ggplot} object and info
}
\description{
A helper function for \code{sim_plot}.
}
