% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-plot.R
\name{sim_plot_add_signal_and_data}
\alias{sim_plot_add_signal_and_data}
\title{Add generated signal and data to plot}
\usage{
sim_plot_add_signal_and_data(h, i_test, linecolor, dotcolor, testcolor)
}
\arguments{
\item{h}{a \code{ggplot} object}

\item{i_test}{indices of possible test points}

\item{linecolor}{line color for signal}

\item{dotcolor}{dot color for data points}

\item{testcolor}{dot color for possible test points}
}
\value{
the updated \code{ggplot} object
}
\description{
A helper function for \code{sim_plot}.
}
