% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-lgpsim.R
\name{sim_plot_components}
\alias{sim_plot_components}
\title{Visualize the components of a simulated data set}
\usage{
sim_plot_components(simdata, time_is_xvar = TRUE, marker = 16, ...)
}
\arguments{
\item{simdata}{an object of class \linkS4class{lgpsim}}

\item{time_is_xvar}{is the time variable the x-axis variable
in all subplots?}

\item{marker}{point marker}

\item{...}{keyword arguments for \code{\link{plot_components}}}
}
\value{
an object returned by \code{ggpubr::ggarrange} or a list
}
\description{
Visualize the components of a simulated data set
}
