% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-plot.R
\name{sim_plot_create}
\alias{sim_plot_create}
\title{Initialize a simulated data plot}
\usage{
sim_plot_create(simdata, i_test, y_transform, signal_name)
}
\arguments{
\item{simdata}{an object of class \linkS4class{lgpsim}}

\item{i_test}{indices of possible test points}

\item{y_transform}{function to transform the data y}

\item{signal_name}{name of signal in the legend}
}
\value{
updated \code{ggplot} object
}
\description{
A helper function for \code{sim_plot}.
}
