% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{sim_var_mask}
\alias{sim_var_mask}
\title{Variance masking function}
\usage{
sim_var_mask(x, a)
}
\arguments{
\item{x}{vector of length \code{n}}

\item{a}{a positive real number}
}
\value{
a vector of length \code{n}
}
\description{
Variance masking function
}
