% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-helper.R
\name{simdata_colnames_pretty}
\alias{simdata_colnames_pretty}
\title{Simulated data column names in a prettier form}
\usage{
simdata_colnames_pretty(cn)
}
\arguments{
\item{cn}{column names}
}
\value{
names of model components
}
\description{
Simulated data column names in a prettier form
}
