% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-utils.R
\name{simplify_list}
\alias{simplify_list}
\title{Return first list element if the list has length is one}
\usage{
simplify_list(x)
}
\arguments{
\item{x}{a list}
}
\value{
the original list or just its first element if its length is one
}
\description{
Return first list element if the list has length is one
}
\seealso{
Other list utilities: 
\code{\link{dollar}()},
\code{\link{list_if_named}()},
\code{\link{list_to_matrix}()},
\code{\link{matrix_to_list}()}
}
\concept{list utilities}
