% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-utils.R
\name{simplify_str}
\alias{simplify_str}
\title{Remove quotes and whitespace from a string}
\usage{
simplify_str(s)
}
\arguments{
\item{s}{a string}
}
\value{
a string
}
\description{
Remove quotes and whitespace from a string
}
