% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate-create_F.R
\name{simulate_kernels}
\alias{simulate_kernels}
\title{Compute all kernel matrices when simulating data}
\usage{
simulate_kernels(X, types, lengthscales, X_affected, useBinKernel = TRUE)
}
\arguments{
\item{X}{covariates}

\item{types}{vector of covariate types, so that
\itemize{
  \item 1 = ID
  \item 2 = age
  \item 3 = diseaseAge
  \item 4 = other continuous covariate
  \item 5 = a categorical covariate that interacts with age
  \item 6 = a categorical covariate that acts as an offset
}}

\item{lengthscales}{vector of lengthscales}

\item{X_affected}{which individuals are affected by the disease}

\item{useBinKernel}{whether or not binary (mask) kernel should be used for
categorical covariates}
}
\value{
a 3D array
}
\description{
Compute all kernel matrices when simulating data
}
