% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-utils.R
\name{split_data_by_id}
\alias{split_data_by_id}
\title{Split data into training and test data according to given individuals}
\usage{
split_data_by_id(data, test_ids, id_variable = "id")
}
\arguments{
\item{data}{a data frame}

\item{test_ids}{test data individual identifiers}

\item{id_variable}{name of id variable}
}
\value{
a \code{list(train, test)}
}
\description{
Split data into training and test data according to given individuals
}
