% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-split.R
\name{split_data_by_timepoint}
\alias{split_data_by_timepoint}
\title{Split data into training and test data according to time point indices}
\usage{
split_data_by_timepoint(data, test_idx, id_variable = "id")
}
\arguments{
\item{data}{a data frame}

\item{test_idx}{indices of test time points}

\item{id_variable}{name of id variable}
}
\value{
a \code{list(train, test)}
}
\description{
Split data into training and test data according to time point indices
}
