% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-split.R
\name{split_data_random_each}
\alias{split_data_random_each}
\title{Split data into training and test data by selecting randomly k points from each individual}
\usage{
split_data_random_each(
  data,
  n_test = 1,
  id_variable = "id",
  time_variable = "age"
)
}
\arguments{
\item{data}{a data frame}

\item{n_test}{desired number of test data points per individual}

\item{id_variable}{name of id variable}

\item{time_variable}{name of time variable}
}
\value{
a \code{list(train, test)}
}
\description{
Split data into training and test data by selecting randomly k points from each individual
}
