% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula.R
\name{split_formula}
\alias{split_formula}
\title{Split formula into left and right-hand side}
\usage{
split_formula(formula)
}
\arguments{
\item{formula}{The model formula, where
\itemize{
  \item it must contain exatly one tilde (\code{~}), with response
  variable on the left-hand side and model terms on the right-hand side
  \item terms are be separated by a plus (\code{+}) sign
  \item all variables appearing in \code{formula} must be
  found in \code{data}
}
See the "Model formula syntax" section below (\code{\link{lgp}}) for
instructions on how to specify the model terms.}
}
\description{
Split formula into left and right-hand side
}
