% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-parse_covs_and_comps.R
\name{stan_data_covariates}
\alias{stan_data_covariates}
\title{Create covariate data for Stan input}
\usage{
stan_data_covariates(data, x_names, x_cont_scl)
}
\arguments{
\item{data}{a data frame}

\item{x_names}{unique covariate names}

\item{x_cont_scl}{possible existing continuous covariate scaling (list)}
}
\value{
a named list with fields
\itemize{
  \item \code{to_stan}: a list of stan data
  \item \code{x_cont_scaling}: normalization function and inverse for each
  continuous covariate
  \item \code{x_cat_levels}: names of the levels of each categorical
  covariate before conversion from factor to numeric
}
}
\description{
Creates the following Stan data input list fields:
\itemize{
  \item \code{num_cov_cat}
  \item \code{num_cov_cont}
  \item \code{x_cat}
  \item \code{x_cat_num_levels}
  \item \code{x_cont}
  \item \code{x_cont_unnorm}
  \item \code{x_cont_mask}
}
}
