% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-parse_covs_and_comps.R
\name{stan_data_expanding}
\alias{stan_data_expanding}
\title{Create mapping from observation index to index of beta or teff parameter}
\usage{
stan_data_expanding(covariates, components)
}
\arguments{
\item{covariates}{a covariates information list returned by
\code{\link{stan_data_covariates}}}

\item{components}{a components information list returned by
\code{\link{stan_data_components}}}
}
\value{
a list
}
\description{
Creates the following Stan data input list fields:
\itemize{
  \item \code{num_bt}
  \item \code{idx_expand} (\code{_PRED})
}
}
