% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preproc-helper.R
\name{stan_input_X_and_D}
\alias{stan_input_X_and_D}
\title{Predictor covariates and types to Stan input}
\usage{
stan_input_X_and_D(data, varInfo, types, formula, verbose)
}
\arguments{
\item{data}{a data frame containing the covariates}

\item{varInfo}{original variable type info}

\item{types}{types of the covariates}

\item{formula}{model formula}

\item{verbose}{can this print some info?}
}
\value{
X and needed types and updated varInfo
}
\description{
Reorders covariates and takes only those that are needed
}
