% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-utils.R
\name{stan_list_names}
\alias{stan_list_names}
\title{Names that the list given as data to Stan should contain}
\usage{
stan_list_names()
}
\value{
a character vector
}
\description{
Names that the list given as data to Stan should contain
}
