% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preproc-helper.R
\name{standardize_inputs}
\alias{standardize_inputs}
\title{Standardize continuous input variables in X}
\usage{
standardize_inputs(X, D)
}
\arguments{
\item{X}{the design matrix}

\item{D}{the covariate types, a vector of length 6}
}
\value{
updated X and info about scaling
}
\description{
Standardize continuous input variables in X
}
