% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4-as_character.R
\name{term_as_character}
\alias{term_as_character}
\title{Character format of an lgpterm}
\usage{
term_as_character(x, verbose = TRUE)
}
\arguments{
\item{x}{an object of class \linkS4class{lgpterm}}

\item{verbose}{should the format be more verbose}
}
\value{
a string
}
\description{
Character format of an lgpterm
}
