% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-data.R
\name{term_names}
\alias{term_names}
\title{Map a list of terms to their "names"}
\usage{
term_names(rhs)
}
\arguments{
\item{rhs}{an object of class \linkS4class{lgprhs}}
}
\value{
a character vector
}
\description{
Map a list of terms to their "names"
}
