% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-parse_covs_and_comps.R
\name{term_to_numeric}
\alias{term_to_numeric}
\title{An lgpterm to numeric representation for Stan}
\usage{
term_to_numeric(term, covariates)
}
\arguments{
\item{term}{an object of class \linkS4class{lgpterm}}

\item{covariates}{a list returned by \code{\link{stan_data_covariates}}}
}
\value{
a vector of 9 integers
}
\description{
An lgpterm to numeric representation for Stan
}
