% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula_advanced.R
\name{term_variables}
\alias{term_variables}
\title{Get names of all variables appearing in a term}
\usage{
term_variables(term)
}
\arguments{
\item{term}{an object of class \linkS4class{lgpterm}}
}
\value{
a list of variable names
}
\description{
Get names of all variables appearing in a term
}
\seealso{
Other advanced formula parsers: 
\code{\link{parse_expr}()},
\code{\link{parse_formula_advanced}()},
\code{\link{parse_rhs}()},
\code{\link{parse_term}()},
\code{\link{rhs_variables}()}
}
\concept{advanced formula parsers}
