% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-test_data.R
\name{test_data_x}
\alias{test_data_x}
\title{Create data for unit tests of Stan kernel array functions}
\usage{
test_data_x(N = 3)
}
\arguments{
\item{N}{number of individuals}
}
\value{
A list of data suitable for testing \code{STAN_kernel_all}
}
\description{
Create data for unit tests of Stan kernel array functions
}
