% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preproc-disease_info.R
\name{test_id_variable}
\alias{test_id_variable}
\title{Test that the id variable is correctly specified to allow
disease effect modeling}
\usage{
test_id_variable(X_id)
}
\arguments{
\item{X_id}{the id covariate in X}
}
\value{
a list
}
\description{
Test that the id variable is correctly specified to allow
disease effect modeling
}
