% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\docType{data}
\name{testdata_001}
\alias{testdata_001}
\title{A very small artificial test data, used mostly for unit tests}
\format{
A data frame with 24 rows and 6 variables:
\describe{
  \item{id}{individual id, a factor with levels: 1, 2, 3, 4}
  \item{age}{age}
  \item{dis_age}{disease-related age}
  \item{blood}{a continuous variable}
  \item{sex}{a factor with 2 levels: Male, Female}
  \item{y}{a continuous variable}
}
}
\usage{
testdata_001
}
\description{
A very small artificial test data, used mostly for unit tests
}
\seealso{
Other built-in datasets: 
\code{\link{testdata_002}}
}
\concept{built-in datasets}
\keyword{datasets}
