% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\docType{data}
\name{testdata_002}
\alias{testdata_002}
\title{Medium-size artificial test data, used mostly for tutorials}
\format{
A data frame with 96 rows and 6 variables:
\describe{
  \item{id}{individual id, a factor with levels: 01-12}
  \item{age}{age}
  \item{diseaseAge}{disease-related age}
  \item{sex}{a factor with 2 levels: Male, Female}
  \item{group}{a factor with 2 levels: Case, Control}
  \item{y}{a continuous variable}
}
}
\usage{
testdata_002
}
\description{
Medium-size artificial test data, used mostly for tutorials
}
\seealso{
\code{\link{read_proteomics_data}}

Other built-in datasets: 
\code{\link{testdata_001}}
}
\concept{built-in datasets}
\keyword{datasets}
