% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_data.R
\name{to_long_format}
\alias{to_long_format}
\title{Data frame and additional information to long format}
\usage{
to_long_format(df)
}
\arguments{
\item{df}{a data frame in wide format}
}
\value{
a data frame where first column is a factor that determines the
column variable name in the original data and second column contains
the actual values
}
\description{
Data frame and additional information to long format
}
\seealso{
Other data utilities: 
\code{\link{add_dis_age}()},
\code{\link{add_factor_crossing}()},
\code{\link{add_factor}()},
\code{\link{data_types}()},
\code{\link{get_teff_obs}()},
\code{\link{new_x}()},
\code{\link{pick_one_row_each}()},
\code{\link{select_factors_and}()},
\code{\link{split}()},
\code{\link{valid_data}}
}
\concept{data utilities}
