% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assist.R
\name{valid_data}
\alias{valid_data}
\alias{validate_data}
\alias{validate_factor}
\title{Validate data array}
\usage{
validate_data(data, id_name = "id")

validate_factor(x, id, name)

validate_data(data, id_name = "id")
}
\arguments{
\item{data}{a data frame}

\item{id_name}{name of the id variable}

\item{x}{a factor}

\item{id}{the id factor}

\item{name}{factor name}
}
\value{
a data frame
}
\description{
Validate data array
}
