% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{validate_dimension_list}
\alias{validate_dimension_list}
\title{Check that all listed dimensions are equal}
\usage{
validate_dimension_list(dims)
}
\arguments{
\item{dims}{a list of dimensions}
}
\value{
list of errors or NULL
}
\description{
Check that all listed dimensions are equal
}
