% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-all_classes.R
\name{validate_lengths}
\alias{validate_lengths}
\title{Check that arguments have equal lengths}
\usage{
validate_lengths(a, b)
}
\arguments{
\item{a}{first argument}

\item{b}{second argument}
}
\value{
list of errors or NULL
}
\description{
Check that arguments have equal lengths
}
