% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-prior.R
\name{validate_prior}
\alias{validate_prior}
\title{Validate prior by sampling the signal and noise from it}
\usage{
validate_prior(model, chains = 4, iter = 1000, parallel = FALSE)
}
\arguments{
\item{model}{An object of class \code{\link{lgpmodel}}.}

\item{chains}{how many chains are used to sample from the prior}

\item{iter}{for how many iterations are the chains run}

\item{parallel}{should the chains be run in parallel?}
}
\value{
An object of class \code{lgpfit} and random samples of
both `f` and `y`.
}
\description{
Validate prior by sampling the signal and noise from it
}
