% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan-functions.R
\name{var_mask}
\alias{var_mask}
\title{Variance masking function}
\usage{
var_mask(x, a, STREAM = get_stream())
}
\arguments{
\item{x}{a numeric vector}

\item{a}{steepness parameter}

\item{STREAM}{an external pointer}
}
\value{
a vector
}
\description{
This is a wrapper for \code{STAN_var_mask}
}
