% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postproc-main.R
\name{varsel}
\alias{varsel}
\title{Covariate selection}
\usage{
varsel(object, threshold = 0.95, verbose = TRUE)
}
\arguments{
\item{object}{An object of class \code{lgpfit}.}

\item{threshold}{A threshold for proportion of explained variance}

\item{verbose}{should this print some output}
}
\value{
the selected covariates
}
\description{
Covariate selection
}
