% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-utils.R
\name{warn_msg_default_prior}
\alias{warn_msg_default_prior}
\title{Warning message about using a default prior}
\usage{
warn_msg_default_prior(desc, name, model_desc)
}
\arguments{
\item{desc}{parameter description}

\item{name}{parameter name}

\item{model_desc}{model description}
}
\description{
Warning message about using a default prior
}
