% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stanfunctions.R
\name{warp_input}
\alias{warp_input}
\title{Input warping function}
\usage{
warp_input(x, a, STREAM = get_stream())
}
\arguments{
\item{x}{a numeric vector}

\item{a}{steepness parameter}

\item{STREAM}{an external pointer}
}
\value{
a vector
}
\description{
This is a wrapper for \code{STAN_warp_input}
}
