% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgp-formula.R
\name{zerosum}
\alias{zerosum}
\title{Create a zerosum term}
\usage{
zerosum(covariate)
}
\arguments{
\item{covariate}{name of a categorical covariate}
}
\value{
an object of class \code{\link{lgpterm}}
}
\description{
Create a zerosum term
}
