% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assist.R
\name{add_categorical_covariate}
\alias{add_categorical_covariate}
\title{Easily add a categorical covariate to a data frame}
\usage{
add_categorical_covariate(data, x, id_var = "id")
}
\arguments{
\item{data}{the original data frame}

\item{x}{A named vector containing the category for each individual.
The names should specify the individual id.}

\item{id_var}{name of the id variable in \code{data}}
}
\value{
A data frame with one column added. The new column will
have same name as the variable passed as input \code{x}.
}
\description{
Easily add a categorical covariate to a data frame
}
