% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_data.R
\name{add_factor}
\alias{add_factor}
\title{Easily add a categorical covariate to a data frame}
\usage{
add_factor(data, x, id_var = "id")
}
\arguments{
\item{data}{the original data frame}

\item{x}{A named vector containing the category for each individual.
The names should specify the individual id.}

\item{id_var}{name of the id variable in \code{data}}
}
\value{
A data frame with one column added. The new column will
have same name as the variable passed as input \code{x}.
}
\description{
Easily add a categorical covariate to a data frame
}
\seealso{
Other data utilities: 
\code{\link{add_dis_age}()},
\code{\link{add_factor_crossing}()},
\code{\link{data_types}()},
\code{\link{get_teff_obs}()},
\code{\link{new_x}()},
\code{\link{pick_one_row_each}()},
\code{\link{select_factors_and}()},
\code{\link{split}()},
\code{\link{to_long_format}()},
\code{\link{valid_data}}
}
\concept{data utilities}
