% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_data.R
\name{add_factor_crossing}
\alias{add_factor_crossing}
\title{Add a crossing of two factors to a data frame}
\usage{
add_factor_crossing(df, fac1, fac2, new_name)
}
\arguments{
\item{df}{a data frame}

\item{fac1}{name of first factor, must be found in \code{df}}

\item{fac2}{name of second factor, must be found in \code{df}}

\item{new_name}{name of the new factor}
}
\value{
a data frame
}
\description{
Add a crossing of two factors to a data frame
}
\seealso{
Other data utilities: 
\code{\link{add_dis_age}()},
\code{\link{add_factor}()},
\code{\link{data_types}()},
\code{\link{get_teff_obs}()},
\code{\link{new_x}()},
\code{\link{pick_one_row_each}()},
\code{\link{select_factors_and}()},
\code{\link{split}()},
\code{\link{to_long_format}()},
\code{\link{valid_data}}
}
\concept{data utilities}
