% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-datautil.R
\name{add_factor_crossing}
\alias{add_factor_crossing}
\title{Add a crossing of two factors to a data frame}
\usage{
add_factor_crossing(data, fac1, fac2, new_name)
}
\arguments{
\item{data}{a data frame}

\item{fac1}{name of first factor, must be found in \code{df}}

\item{fac2}{name of second factor, must be found in \code{df}}

\item{new_name}{name of the new factor}
}
\value{
a data frame
}
\description{
Add a crossing of two factors to a data frame
}
\seealso{
Other data frame handling functions: 
\code{\link{add_dis_age}()},
\code{\link{add_factor}()},
\code{\link{adjusted_c_hat}()},
\code{\link{new_x}()},
\code{\link{split}()}
}
\concept{data frame handling functions}
