% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_array.R
\name{add_sum_arraylist}
\alias{add_sum_arraylist}
\title{Add the sum of all arrays in a list to the list}
\usage{
add_sum_arraylist(x)
}
\arguments{
\item{x}{a list of arrays of shape \code{n} x \code{m}, with
length \code{L}}
}
\value{
a list of arrays of shape \code{n} x \code{m},
with length \code{L + 1}
}
\description{
Add the sum of all arrays in a list to the list
}
\seealso{
Other array utilities: 
\code{\link{array_to_arraylist}()},
\code{\link{ensure_2dim}()},
\code{\link{normalize_rows}()},
\code{\link{null_if_all_nan}()},
\code{\link{reduce_rows}()},
\code{\link{repvec}()},
\code{\link{row_vars}()},
\code{\link{select_row}()},
\code{\link{squeeze_second_dim}()}
}
\concept{array utilities}
