% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assist.R
\name{adjusted_c_hat}
\alias{adjusted_c_hat}
\title{Create the GP mean input for \code{lgp}, so that it accounts
for normalization between data points in the Poisson or NB
observation model}
\usage{
adjusted_c_hat(y, norm_factors)
}
\arguments{
\item{y}{response variable, vector of length \code{n}}

\item{norm_factors}{normalization factors, vector of length \code{n}}
}
\value{
a vector of length \code{n}, which can be used as
the \code{c_hat} input to the \code{lgp} function
}
\description{
Create the GP mean input for \code{lgp}, so that it accounts
for normalization between data points in the Poisson or NB
observation model
}
