% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-utils_array.R
\name{apply_reduce}
\alias{apply_reduce}
\title{Apply possible reduction to rows of an array}
\usage{
apply_reduce(x, reduce)
}
\arguments{
\item{x}{an array of shape \code{n} x \code{m}}

\item{reduce}{a function or \code{NULL}}
}
\value{
an array of shape \code{1} x \code{m} if \code{reduce} is not
\code{NULL}, otherwise the original array \code{x}
}
\description{
Apply possible reduction to rows of an array
}
\seealso{
Other array utilities: 
\code{\link{add_to_columns}()},
\code{\link{arr3_select}()},
\code{\link{arr3_to_list}()},
\code{\link{array_to_arraylist}()},
\code{\link{ensure_2dim}()},
\code{\link{make_draw_df}()},
\code{\link{normalize_rows}()},
\code{\link{null_if_all_nan}()},
\code{\link{reduce_rows}()},
\code{\link{repvec}()},
\code{\link{row_vars}()},
\code{\link{select_row}()},
\code{\link{squeeze_second_dim}()}
}
\concept{array utilities}
