% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-utils.R
\name{array_to_arraylist}
\alias{array_to_arraylist}
\title{Array reshaping}
\usage{
array_to_arraylist(x, L)
}
\arguments{
\item{x}{an array of shape \code{n} x \code{m}}

\item{L}{an integer}
}
\value{
a list of length \code{L}, where each element is an array of
shape \code{n} x (\code{m}/\code{L})
}
\description{
Array reshaping
}
