% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-misc.R
\name{array_to_arraylist}
\alias{array_to_arraylist}
\title{Array transformation utility}
\usage{
array_to_arraylist(x, L)
}
\arguments{
\item{x}{an array of shape \code{n} x \code{m}}

\item{L}{an integer}
}
\value{
a list of length \code{L}, where each element is an array of
shape \code{h} x (\code{m}/\code{L}), where \code{h = nrow(x)}
}
\description{
Array transformation utility
}
\seealso{
Other array utilities: 
\code{\link{apply_reduce}()},
\code{\link{ensure_2dim}()},
\code{\link{repvec}()},
\code{\link{squeeze_second_dim}()}
}
\concept{array utilities}
