% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-lgpformula.R
\name{as_character}
\alias{as_character}
\alias{as.character,lgpexpr-method}
\alias{as.character,lgpterm-method}
\alias{as.character,lgpformula-method}
\title{Character representations of different formula objects}
\usage{
\S4method{as.character}{lgpexpr}(x)

\S4method{as.character}{lgpterm}(x)

\S4method{as.character}{lgpformula}(x)
}
\arguments{
\item{x}{an object of some S4 class}
}
\value{
a character representation of the object
}
\description{
Character representations of different formula objects
}
