% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postproc-main.R
\name{assess_convergence}
\alias{assess_convergence}
\title{Assess convergence of the chains}
\usage{
assess_convergence(fit, verbose = FALSE, recompute = F)
}
\arguments{
\item{fit}{An (incomplete) object of class \code{lgpfit}.}

\item{verbose}{should convergence info be printed?}

\item{recompute}{Should the Rhat statistics be recomputed?}
}
\value{
Potential scale reduction factors (R_hat).
}
\description{
Assess convergence of the chains
}
