% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-misc.R
\name{call_fun}
\alias{call_fun}
\title{A linter-friendly way to call a function}
\usage{
call_fun(fun, arg)
}
\arguments{
\item{fun}{a function which takes one argument}

\item{arg}{the argument}
}
\value{
The value of \code{fun(arg)}
}
\description{
This exists just to allow calling a function which is a slot
of an S4 object, without some linter warnings.
}
\seealso{
Other function utilities: 
\code{\link{enclose_fun}()}
}
\concept{function utilities}
