% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preproc-helper.R
\name{check_data}
\alias{check_data}
\title{Check that given data is consistent with the given formula}
\usage{
check_data(formula, data)
}
\arguments{
\item{formula}{an object of class \code{\link{lgpformula}}}

\item{data}{an object of class \code{data.frame}}
}
\description{
Check that given data is consistent with the given formula
}
