% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-response.R
\name{check_response}
\alias{check_response}
\title{Check that the response is numeric and compatibile with observation model}
\usage{
check_response(y, obs_model)
}
\arguments{
\item{y}{the response variable measurements}

\item{obs_model}{observation model (integer encoding)}
}
\description{
Check that the response is numeric and compatibile with observation model
}
