% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argcheck.R
\name{checks}
\alias{checks}
\alias{check_type}
\alias{check_numeric}
\alias{check_positive}
\alias{check_non_negative}
\alias{check_interval}
\alias{check_length}
\alias{check_not_null}
\alias{check_null}
\alias{check_lengths}
\alias{check_false}
\alias{check_in_data}
\alias{check_all_leq}
\alias{check_not_named}
\alias{check_named}
\alias{check_dim}
\title{Argument validation}
\usage{
check_type(arg, allowed)

check_numeric(arg)

check_positive(arg)

check_non_negative(arg)

check_interval(arg, lower, upper)

check_length(arg, expected)

check_not_null(arg)

check_null(arg, reason = NULL)

check_lengths(a, b)

check_false(arg)

check_in_data(var_name, data)

check_all_leq(arg, maximums)

check_not_named(arg)

check_named(arg)

check_dim(arg, D)
}
\arguments{
\item{arg}{An argument to check.}

\item{allowed}{Allowed class names.}

\item{lower}{interval minimum}

\item{upper}{interval maximum}

\item{expected}{Expected length.}

\item{reason}{explanation why the argument should be NULL}

\item{a}{An argument to check.}

\item{b}{An argument to check.}

\item{var_name}{the variable to be searched for}

\item{data}{a data frame}

\item{maximums}{maximum allowed values for \code{arg}}

\item{D}{expected number of dimensions}
}
\value{
\code{TRUE} if the check passes.
}
\description{
\itemize{
  \item \code{check_type} checks if argument has correct class
  \item \code{check_numeric} checks if argument is numeric
  \item \code{check_positive} checks if argument is positive
  \item \code{check_non_negative} checks if argument is non-negative
  \item \code{check_interval} checks if argument is inside a given interval
  \item \code{check_not_null} checks if argument is not null
  \item \code{check_false} checks if argument is false or zero
  \item \code{check_length} checks if argument has correct length
  \item \code{check_lengths} checks if two argument have equal length
  \item \code{check_in_data} checks that variable exists in a data frame
  \item \code{check_all_leq} checks that argument has values less than equal
  to given maximums, elementwise
  \item \code{check_not_named} checks that \code{names(arg)} is NULL
  \item \code{check_named} checks that length of \code{names(arg)} is
  greater than zero
  \item \code{check_dim} checks that \code{arg} has expected number of
  dimensions
  \item \code{check_null} checks that \code{arg} is NULL
}
}
\seealso{
Other argument checks: 
\code{\link{check_allowed}()}
}
\concept{argument checks}
