% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{color_palette}
\alias{color_palette}
\title{A color palette function}
\usage{
color_palette(n)
}
\arguments{
\item{n}{an integer from 1 to 6}
}
\value{
an array of \code{n} hex values
}
\description{
A color palette function
}
