% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{colorset}
\alias{colorset}
\title{Plot colors to use}
\usage{
colorset(main, variant = "mid")
}
\arguments{
\item{main}{Color name. Must be a valid \code{bayesplot} color name.}

\item{variant}{Must be one of {"light", "light_highlight", "mid",
"mid_highlight", "dark", "dark_highlight"}.}
}
\value{
A hex value of the color.
}
\description{
Plot commands which specify the \code{col} argument should
mostly use colors that are obtained using this function.
}
\seealso{
Other color utilities: 
\code{\link{color_palette}()},
\code{\link{line_alpha_fun}()},
\code{\link{plot_color_palette}()}
}
\concept{color utilities}
