% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-plot.R
\name{colorset}
\alias{colorset}
\title{Plot colors}
\usage{
colorset(main, variant = "mid")
}
\arguments{
\item{main}{Color name. Must be a valid scheme name for
\code{\link[bayesplot]{color_scheme_get}}.}

\item{variant}{Must be one of {"light", "light_highlight", "mid",
"mid_highlight", "dark", "dark_highlight"}.}
}
\value{
A hex value of the color.
}
\description{
Plot colors
}
