% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postproc-predict.R
\name{compute_kernel_matrices}
\alias{compute_kernel_matrices}
\title{Evaluate kernel matrices for each component}
\usage{
compute_kernel_matrices(X1, X2, kernel_info)
}
\arguments{
\item{X1}{Covariate matrix of size \code{n1} x \code{sum(D)}.}

\item{X2}{Covariate matrix of size \code{n2} x \code{sum(D)}.}

\item{kernel_info}{A list of parameters and other kernel info.}
}
\value{
An array of size \code{n1} x \code{n2} x \code{sum(D)}.
}
\description{
Used by \code{\link{compute_predictions}}.
}
