% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict-helper.R
\name{compute_kernel_matrices}
\alias{compute_kernel_matrices}
\title{Evaluate kernel matrices for each component}
\usage{
compute_kernel_matrices(X1, X2, D, alpha, ell, stp, beta, info)
}
\arguments{
\item{X1}{Covariate matrix of size \code{n1} x \code{sum(D)}.}

\item{X2}{Covariate matrix of size \code{n2} x \code{sum(D)}.}

\item{D}{A vector of length 6.}

\item{alpha}{Marginal standard deviation parameters.}

\item{ell}{Lengthscale parameters}

\item{stp}{Input warp steepness parameter.}

\item{beta}{Individual-specific disease effect magnitudes}

\item{info}{Other model info}
}
\value{
An array of size \code{n1} x \code{n2} x \code{sum(D)}.
}
\description{
Used by \code{\link{compute_predictions}}.
}
