% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict-main.R
\name{compute_lppd}
\alias{compute_lppd}
\title{Compute log-posterior predictive density at test points}
\usage{
compute_lppd(PRED, y_test, reduction = "mean")
}
\arguments{
\item{PRED}{predictions}

\item{y_test}{values of the response variable at the test points}

\item{reduction}{must be either "mean", "sum" or "none"}
}
\value{
a vector
}
\description{
Compute log-posterior predictive density at test points
}
