% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postproc-predict.R
\name{compute_lppd}
\alias{compute_lppd}
\title{Compute log-posterior predictive density at test points}
\usage{
compute_lppd(PRED, y_test)
}
\arguments{
\item{PRED}{predictions}

\item{y_test}{values of the response variable at the test points}
}
\value{
a matrix with size \code{n_samples} x \code{n_data}
}
\description{
Compute log-posterior predictive density at test points
}
