% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postproc-predict.R
\name{compute_predictions}
\alias{compute_predictions}
\title{Compute component-wise predictions at test points}
\usage{
compute_predictions(X_data, y_data, X_test, params, D, info, cnames)
}
\arguments{
\item{X_data}{Covariate matrix (data points).}

\item{y_data}{Response variable (data points).}

\item{X_test}{Covariate matrix (test points).}

\item{params}{Kernel function and other hyperparameters.}

\item{D}{A vector of length 6.}

\item{info}{Other model info.}

\item{cnames}{Names of the model components.}
}
\value{
A list.
}
\description{
Used by \code{\link{lgp_predict}}.
}
