% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postproc-helper.R
\name{compute_relevances}
\alias{compute_relevances}
\title{Covariate and component relevance calculations}
\usage{
compute_relevances(FFF, y_data, info, D)
}
\arguments{
\item{FFF}{a data frame of size \code{n_data} x \code{n_components+2}}

\item{y_data}{(scaled) measurements of the response variable}

\item{info}{model info}

\item{D}{a vector of length 6}
}
\value{
a list
}
\description{
Covariate and component relevance calculations
}
