% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgp-model.R
\name{covariate_names}
\alias{covariate_names}
\title{Get covariate names for a model}
\usage{
covariate_names(model, type = "all")
}
\arguments{
\item{model}{an object of class \linkS4class{lgpmodel}}

\item{type}{what types of covariates to include. Must be either
"all", "categorical" or "continuous".}
}
\value{
A string where names are separated by a comma.
}
\description{
Get covariate names for a model
}
