% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stanfunctions_bfa.R
\name{cpp_spd_eq}
\alias{cpp_spd_eq}
\title{Spectral density function of the exponentiated quadratic kernel}
\usage{
cpp_spd_eq(w, alpha, ell, STREAM = get_stream())
}
\arguments{
\item{w}{the frequency, real number}

\item{alpha}{marginal std of the kernel}

\item{ell}{lengthscale of the kernel}

\item{STREAM}{an external pointer}
}
\value{
a real number
}
\description{
This is a wrapper for \code{STAN_spd_eq}, with some input
validation added.
}
\seealso{
Other basisfunction approximation functions: 
\code{\link{cpp_bfa_lambda}()},
\code{\link{cpp_bfa_multi_normal_lpdf}()},
\code{\link{cpp_bfa_phi}()}
}
\concept{basisfunction approximation functions}
