% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate-create_F.R
\name{create_F}
\alias{create_F}
\title{Simulate latent function components for longitudinal data analysis}
\usage{
create_F(X, covariates, relevances, lengthscales, X_affected, dis_fun,
  useBinKernel = TRUE)
}
\arguments{
\item{X}{input data matrix (generated by \code{\link{create_X}})}

\item{covariates}{Integer vector that defines the types of covariates
(other than id and age). Different integers correspond to the
following covariate types:
\itemize{
  \item 0 = disease-related age
  \item 1 = other continuous covariate
  \item 2 = a categorical covariate that interacts with age
  \item 3 = a categorical covariate that acts as a group offset
}}

\item{relevances}{Relative relevance of each component. Must have be a vector
so that \cr
 \code{length(relevances) =  2 + length(covariates)}. \cr
First two values define the relevance of the infividual-specific age and 
shared age component, respectively.}

\item{lengthscales}{A vector so that \cr \code{length(lengthscales) = }
\code{2 + sum(covariates \%in\% c(0,1,2))}.}

\item{X_affected}{which individuals are affected by the disease}

\item{dis_fun}{A function that defines the disease effect. If NULL, the effect
is is drawn from a nonstationary GP prior.}

\item{useBinKernel}{Should the binary kernel be used for categorical covariates? 
If this is \code{TRUE}, the effect will exist only for group 1.}
}
\value{
a data frame FFF where one column corresponds to one additive data component
}
\description{
Simulate latent function components for longitudinal data analysis
}
