% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-model_methods.R
\name{create_plot_df}
\alias{create_plot_df}
\title{Helper function for plots}
\usage{
create_plot_df(object, x = "age", group_by = "id")
}
\arguments{
\item{object}{an object of class \linkS4class{lgpmodel} or
\linkS4class{lgpfit}}

\item{x}{x-axis variable name}

\item{group_by}{grouping variable name (use \code{NULL} for no grouping)}
}
\value{
a data frame
}
\description{
Helper function for plots
}
