% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethodsModel.R
\name{create_plot_df}
\alias{create_plot_df}
\title{Helper function for plots}
\usage{
create_plot_df(object)
}
\arguments{
\item{object}{an object of class \linkS4class{lgpmodel} or
\linkS4class{lgpfit}}
}
\value{
a data frame
}
\description{
Helper function for plots
}
