% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-parse_covs_and_comps.R
\name{create_scaling}
\alias{create_scaling}
\title{Create the function that does a standardizing transform and its inverse}
\usage{
create_scaling(y, name)
}
\arguments{
\item{y}{variable measurements}

\item{name}{variable name}
}
\value{
an object of class \linkS4class{lgpscaling}
}
\description{
Create the function that does a standardizing transform and its inverse
}
