% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-create_model.R
\name{create_scaling}
\alias{create_scaling}
\title{Create a standardizing transform}
\usage{
create_scaling(x, name)
}
\arguments{
\item{x}{variable measurements (might contain \code{NA} or \code{NaN})}

\item{name}{variable name}
}
\value{
an object of class \linkS4class{lgpscaling}
}
\description{
Create a standardizing transform
}
\seealso{
Other variable scaling functions: 
\code{\link{apply_scaling}()}
}
\concept{variable scaling functions}
