% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preproc-main.R
\name{create_stan_input}
\alias{create_stan_input}
\title{Create input for Stan}
\usage{
create_stan_input(formula, data, prior, likelihood, varInfo, standardize,
  uncertain_effect_time, equal_effect, C_hat, DELTA, sample_F, t_test,
  verbose, variance_mask, cat_interact_kernel_type, N_trials)
}
\arguments{
\item{formula}{A formula of the form
\code{y ~ } \code{x1 + x2 + x3}
defining the response variable \code{y} and covariates \code{xi}. All variables that
appear in the formula must exist as columns of \code{data}.}

\item{data}{A data frame containing (at least) the variables given in \code{formula}.}

\item{prior}{Prior distribution. Can be created for example using the function 
\code{\link{prior_default}}.}

\item{likelihood}{Determines the observation model. Must be either \code{"Gaussian"} 
(default), \code{"Poisson"}, \code{"NB" (negative binomial) or "binomial"}.}

\item{varInfo}{Variable type info.}

\item{standardize}{Should the response variable be standardized?}

\item{uncertain_effect_time}{Do we wish to model uncertainty in the disease effect time?}

\item{equal_effect}{Is the disease effect assumed to be equally strong for all diseased individuals?}

\item{C_hat}{This can only be given if likelihood is Poisson or NB. The signal \code{f} 
will the be transformed so that \code{g = } \code{exp(C_hat + f)}. If \code{NULL}, it will be
set to \code{C_hat = } \code{log(mean(y))}, where \code{y} is the response variable.}

\item{DELTA}{the amount of added jitter to ensure positive definiteness of the kernel}

\item{sample_F}{Determines if the function values are be sampled (must be \code{TRUE} if
likelihood is not Gaussian).}

\item{t_test}{Optional test time points. Should only be used if \code{sample_F = TRUE}.
Otherwise use \code{\link{lgp_predict}} after fitting the model.}

\item{verbose}{Should more verbose output be printed?}

\item{variance_mask}{Should a variance mask be used to force disease component
variance to zero before disease onset?}

\item{cat_interact_kernel_type}{Kernel type for categorical variables (other than id). 
Possible options are \code{"categorical"} (default) and \code{"binary"} (mask kernel where
only category \code{"1"} will have an effect).}

\item{N_trials}{This argument (number of trials) is only needed when likelihood is binomial.
Must have length one or equal to number of data points. Setting \code{N_trials=1} corresponds to 
Bernoulli observation model.}
}
\value{
A list containing the data to be given to \code{rstan::sampling}, some info about
preprocessing and all the information about scaling the inputs and response, and updated
variable type info.
}
\description{
Parses the \code{formula} and \code{data} input to \code{\link{lgp_model}}. Also performs
many input checks.
}
