% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preproc-main.R
\name{create_stan_input}
\alias{create_stan_input}
\title{Create input for Stan}
\usage{
create_stan_input(
  formula,
  data,
  prior,
  likelihood,
  varInfo,
  standardize,
  uncertain_effect_time,
  equal_effect,
  C_hat,
  DELTA,
  sample_F,
  verbose,
  variance_mask,
  N_trials,
  skip_gen_quant
)
}
\arguments{
\item{formula}{A formula specified using the common \code{\link{formula}}
syntax, such as \code{y ~ x1 + x2:z1 + } \code{x2:z2 + z2}.
\itemize{
  \item The formula must contain exatly one tilde (\code{~}), with response
  variable on the left-hand side and model terms on the right-hand side.
  \item Terms are be separated by a plus (\code{+}) sign.
  \item Terms can consits of a single variable name, such as \code{x}, or
  an interaction of two variables, such as \code{x:z}.
  \item In single-variable terms, the variable can be either continuous or
  categorical, whereas in interaction terms the variable
  on the left-hand side of the colon (\code{:}) has to be continuous and the
  one on the right-hand side has to be categorical (a factor).
  \item All variables appearing in \code{formula} must be
  found in \code{data}.
}}

\item{data}{A data frame where each column corresponds to one variable, and
each row is one observation. Continuous covariates must have type
\code{"numeric"} and categorical ones must have type \code{"factor"}.
Note that you need to indicate missing values of \code{disAge_variable} with
\code{NaN}.}

\item{prior}{A named list, defining the prior distribution of model
(hyper)parameters. It is recommended to first create this using the function
\code{\link{prior_default}}, and then possibly modify it.}

\item{likelihood}{Determines the observation model. Must be either
\code{"Gaussian"} (default), \code{"Poisson"}, \code{"NB"} (negative
binomial) or \code{"binomial"}. To
use Bernoulli likelihood, use \code{likelihood="binomial"} and set
\code{N_trials} as a vector of ones.}

\item{varInfo}{Variable type info.}

\item{standardize}{Should the response variable be standardized?}

\item{uncertain_effect_time}{Do we wish to model uncertainty in the disease
effect time?}

\item{equal_effect}{Is the disease effect assumed to be equally strong for
all diseased individuals?}

\item{C_hat}{The GP mean. Must be a vector of length \code{dim(data)[1]}, or
a real number defining a constant GP mean. If \code{NULL}, this is set to
 \itemize{
   \item \code{C_hat = 0}, if \code{likelihood} is \code{"Gaussian"}, because
   with
   Gaussian likelihood the response variable is by default centered to have
   zero mean.
   \item \code{C_hat = } \code{log(mean(y))} if \code{likelihood} is
   \code{"Poisson"} or \code{"NB"},
   \item \code{C_hat = } \code{log(p/(1-p))}, where
   \code{p = mean(y/N_trials)} if \code{likelihood} is \code{"binomial"},
 }
where \code{y} denotes the response variable. You can modify this vector to
account for normalization between data points. With Gaussian likelihood
though, do not modify this argument, normalize the data beforehand instead.}

\item{DELTA}{the amount of added jitter to ensure positive definiteness of
the kernel}

\item{sample_F}{Determines if the function values are be sampled (must be
\code{TRUE} if likelihood is not \code{"Gaussian"}).}

\item{verbose}{Should more verbose output be printed?}

\item{variance_mask}{Should a variance mask be used to force disease
component variance to zero before disease onset?}

\item{N_trials}{This argument (number of trials) is only needed when
likelihood is binomial. Must have length one or equal to number of data
points. Setting \code{N_trials=1} corresponds to Bernoulli observation model.}

\item{skip_gen_quant}{If this is true, the generated quantities block of Stan
is skipped.}
}
\value{
A list containing the data to be given to \code{rstan::sampling},
some info about preprocessing and all the information about scaling the
inputs and response, and updated variable type info.
}
\description{
Parses the \code{formula} and \code{data} input to
\code{\link{lgp_model}}. Also performs many input checks.
}
