% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preproc-main.R
\name{create_stan_input}
\alias{create_stan_input}
\title{Create input for Stan}
\usage{
create_stan_input(formula, data, prior, likelihood, varInfo, standardize,
  uncertain_diagnosis, equal_effect, C_hat, DELTA, sample_F, t_test,
  verbose, variance_mask)
}
\arguments{
\item{formula}{A formula of the form
\code{y ~ } \code{x1 + x2 + x3}
defining the response variable \code{y} and covariates \code{xi}. All variables that
appear in the formula must exist as columns of \code{data}.}

\item{data}{A data frame containing (at least) the variables given in \code{formula}.}

\item{prior}{Prior distribution. Can be created for example using the function 
\code{\link{prior_default}}.}

\item{likelihood}{Either \code{"Gaussian"} (default), \code{"Poisson"} or \code{"NB"}.}

\item{varInfo}{Variable type info.}

\item{standardize}{Should the response variable be standardized?}

\item{uncertain_diagnosis}{Do we wish to model uncertainty in the disease onset?}

\item{equal_effect}{Is the disease effect assumed to be equally strong for all diseased individuals?}

\item{C_hat}{This can only be given if likelihood is not Gaussian. The signal \code{f} 
will the be transformed so that \code{g = } \code{exp(C_hat + f)}. If \code{NULL}, it will be
set to \code{C_hat = } \code{log(mean(y))}, where \code{y} is the response variable.}

\item{DELTA}{the amount of added jitter to ensure positive definiteness of the kernel}

\item{sample_F}{Determines if the function values are be sampled (must be \code{TRUE} if
likelihood is not Gaussian).}

\item{t_test}{Optional test time points. Should only be used if \code{sample_F = TRUE}.
Otherwise use \code{\link{lgp_predict}} after fitting the model.}

\item{verbose}{Can this print some info?}

\item{variance_mask}{Should a variance mask be used to force disease component
variance to zero before disease onset?}
}
\value{
A list containing the data to be given to \code{rstan::sampling}, some info about
preprocessing and all the information about scaling the inputs and response, and updated
variable type info.
}
\description{
Parses the \code{formula} and \code{data} input to \code{\link{lgp_model}}. Also performs
many input checks.
}
