% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_data.R
\name{data_types}
\alias{data_types}
\title{Get type of each data column}
\usage{
data_types(data)
}
\arguments{
\item{data}{a data frame}
}
\value{
a named list
}
\description{
Get type of each data column
}
\seealso{
Other data utilities: 
\code{\link{add_dis_age}()},
\code{\link{add_factor}()},
\code{\link{get_teff_obs}()},
\code{\link{new_x}()},
\code{\link{pick_one_row_each}()},
\code{\link{select_factors_and}()},
\code{\link{split}()},
\code{\link{valid_data}}
}
\concept{data utilities}
