% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-invgamma.R
\name{dinvgamma_stanlike}
\alias{dinvgamma_stanlike}
\alias{qinvgamma_stanlike}
\title{Density and quantile functions of the inverse gamma distribution}
\usage{
dinvgamma_stanlike(x, alpha, beta, log = FALSE)

qinvgamma_stanlike(p, alpha, beta)
}
\arguments{
\item{x}{point where to compute the density}

\item{alpha}{positive real number}

\item{beta}{positive real number}

\item{log}{is log-scale used?}

\item{p}{quantile (must be between 0 and 1)}
}
\value{
density/quantile value
}
\description{
Using the same parametrization as Stan. More info
\href{https://mc-stan.org/docs/2_24/functions-reference/inverse-gamma-distribution.html}{here}.
}
\seealso{
Other functions related to the inverse-gamma distribution: 
\code{\link{plot_invgamma}()},
\code{\link{priors}}
}
\concept{functions related to the inverse-gamma distribution}
