% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dollar}
\alias{dollar}
\title{Dollar operator but so that the variable must exist}
\usage{
dollar(object, var_name)
}
\arguments{
\item{object}{a list or data frame}

\item{var_name}{name of the variable to access}
}
\value{
equivalent to \code{object[[var_name]]} if variable exists
}
\description{
throws error if \code{var_name} is not in \code{names(object)}
}
\seealso{
Other list utilities: 
\code{\link{list_to_matrix}()},
\code{\link{matrix_to_list}()},
\code{\link{print_list}()},
\code{\link{zip_lists}()}
}
\concept{list utilities}
