% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-draw_pred.R
\name{draw_pred}
\alias{draw_pred}
\title{Draw pseudo-observations from a predictive distribution}
\usage{
draw_pred(fit, pred = NULL)
}
\arguments{
\item{fit}{An object of class \linkS4class{lgpfit} that can been created
with \code{sample_f=TRUE}.}

\item{pred}{An object of \linkS4class{Prediction}.}
}
\description{
This doesn't work for \linkS4class{GaussianPrediction} objects
because for memory reasons they don't store the
covariance matrices of the predictive distributions.
}
\seealso{
Other main functions: 
\code{\link{create_model}()},
\code{\link{get_draws}()},
\code{\link{lgp}()},
\code{\link{pred}()},
\code{\link{prior_pred}()},
\code{\link{sample_model}()}
}
\concept{main functions}
