% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-draw_pred.R
\name{draw_pred}
\alias{draw_pred}
\title{Draw pseudo-observations from a predictive distribution}
\usage{
draw_pred(fit, pred)
}
\arguments{
\item{fit}{An object of class \linkS4class{lgpfit}.}

\item{pred}{An object of \linkS4class{Prediction}.}
}
\description{
Draw pseudo-observations from a predictive distribution
}
\seealso{
Other main functions: 
\code{\link{create_model}()},
\code{\link{get_draws}()},
\code{\link{lgp}()},
\code{\link{pred}()},
\code{\link{prior_pred}()},
\code{\link{sample_model}()}
}
\concept{main functions}
