% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ensure_2dim}
\alias{ensure_2dim}
\title{Ensure that v is 2-dimensional}
\usage{
ensure_2dim(v)
}
\arguments{
\item{v}{a vector of length \code{m} or an array of size \code{n} x \code{m}}
}
\value{
returns a 2-dimensional array
}
\description{
Ensure that v is 2-dimensional
}
\seealso{
Other array utilities: 
\code{\link{array_to_arraylist}()},
\code{\link{normalize_rows}()},
\code{\link{null_if_all_nan}()},
\code{\link{reduce_rows}()},
\code{\link{repvec}()},
\code{\link{row_vars}()},
\code{\link{select_row}()},
\code{\link{squeeze_second_dim}()}
}
\concept{array utilities}
