% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-misc.R
\name{example_fit}
\alias{example_fit}
\alias{example_fit_adv}
\alias{example_adv_formula}
\alias{example_xpar_prior_info}
\title{Quick way to create an example lgpfit, useful for debugging}
\usage{
example_fit(
  formula = y ~ id + age + age | SEX + age | LOC,
  likelihood = "gaussian",
  chains = 1,
  iter = 30,
  num_indiv = 6,
  num_timepoints = 5,
  ...
)

example_fit_adv(
  formula = example_adv_formula(),
  likelihood = "nb",
  chains = 1,
  iter = 30,
  num_indiv = 6,
  num_timepoints = 5,
  ...
)

example_adv_formula()

example_xpar_prior_info()
}
\arguments{
\item{formula}{model formula}

\item{likelihood}{observation model}

\item{chains}{number of chains to run}

\item{iter}{number of iterations to run}

\item{num_indiv}{number of individuals (data simulation)}

\item{num_timepoints}{number of time points (data simulation)}

\item{...}{additional arguments to \code{\link{lgp}}}
}
\value{
An \linkS4class{lgpfit} object created by fitting
the example model.
}
\description{
Quick way to create an example lgpfit, useful for debugging
}
\section{Functions}{
\itemize{
\item \code{example_fit}: Simple model

\item \code{example_fit_adv}: Model with advanced features

\item \code{example_adv_formula}: Create an example formula for a model with
advanced features

\item \code{example_xpar_prior_info}: Create an example covariate uncertainty prior info
}}

