% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-misc.R
\name{example_fit}
\alias{example_fit}
\title{Quick way to create an example lgpfit, useful for debugging}
\usage{
example_fit(
  formula = y ~ id + age + age | SEX + age | LOC,
  likelihood = "gaussian",
  chains = 1,
  iter = 30,
  num_indiv = 6,
  num_timepoints = 5,
  ...
)
}
\arguments{
\item{formula}{model formula}

\item{likelihood}{observation model}

\item{chains}{number of chains to run}

\item{iter}{number of iterations to run}

\item{num_indiv}{number of individuals (data simulation)}

\item{num_timepoints}{number of time points (data simulation)}

\item{...}{additional arguments to \code{\link{lgp}}}
}
\value{
An \linkS4class{lgpfit} object created by fitting
the example model.
}
\description{
Quick way to create an example lgpfit, useful for debugging
}
