% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethodsFit.R
\name{fit_summary}
\alias{fit_summary}
\title{Posterior summary}
\usage{
fit_summary(fit, ignore_pars = c("f_post", "f_latent", "teff_raw", "lp__"))
}
\arguments{
\item{fit}{an object of class \linkS4class{lgpfit}}

\item{ignore_pars}{names of parameters and generated quantities to ingore}
}
\value{
a character representation
}
\description{
Posterior summary
}
\seealso{
Other fit postprocessing functions: 
\code{\link{get_draws}()},
\code{\link{get_f}()},
\code{\link{scale_f_total}()}
}
\concept{fit postprocessing functions}
