% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-lgpfit.R
\name{fit_summary}
\alias{fit_summary}
\title{Print a fit summary.}
\usage{
fit_summary(fit, ignore_pars = c("f_latent", "eta", "teff_raw", "lp__"))
}
\arguments{
\item{fit}{an object of class \linkS4class{lgpfit}}

\item{ignore_pars}{parameters and generated quantities to ignore from output}
}
\value{
\code{object} invisibly.
}
\description{
Print a fit summary.
}
