% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula.R
\name{formula_to_advanced}
\alias{formula_to_advanced}
\title{Parse a model formula that uses the common formula syntax}
\usage{
formula_to_advanced(rhs)
}
\arguments{
\item{rhs}{The formula right-hand side in text format.}

\item{description}{Translates formula to advanced syntax format.}
}
\value{
an object of class \linkS4class{lgpformula}
}
\description{
Parse a model formula that uses the common formula syntax
}
\seealso{
Other model formula functions: 
\code{\link{is_advanced_formula}()},
\code{\link{parse_expr}()},
\code{\link{parse_formula}()},
\code{\link{parse_rhs}()},
\code{\link{parse_term}()},
\code{\link{parse_whitespace}()},
\code{\link{rhs_to_terms}()},
\code{\link{rhs_variables}()},
\code{\link{term_variables}()}
}
\concept{model formula functions}
