% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-lgpr.R, R/methods-FunctionPosterior.R
\docType{class}
\name{FunctionDraws-class}
\alias{FunctionDraws-class}
\alias{FunctionDraws}
\alias{show,FunctionDraws-method}
\title{An S4 class to represent posterior or prior function (component)
draws of an additive latent GP model}
\usage{
\S4method{show}{FunctionDraws}(object)
}
\arguments{
\item{object}{\linkS4class{FunctionDraws} object for which to apply a class
method.}
}
\description{
An S4 class to represent posterior or prior function (component)
draws of an additive latent GP model
}
\section{Methods (by generic)}{
\itemize{
\item \code{show}: Print a summary about the object.
}}

\section{Slots}{

\describe{
\item{\code{components}}{Data frame representing the draws of each model component.}

\item{\code{x}}{The evaluation points (values of covariates) where the function
draws are evaluated.}

\item{\code{model}}{The \linkS4class{lgpmodel} for which these draws are
computed. Contains important information about how to transform the
total draws through an inverse link function.}
}}

\seealso{
\linkS4class{FunctionPosterior}
}
