% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-lgpr.R, R/methods-FunctionPosteriors.R
\docType{class}
\name{FunctionPosteriors-class}
\alias{FunctionPosteriors-class}
\alias{FunctionPosteriors}
\alias{show,FunctionPosteriors-method}
\title{An S4 class to represent conditional function posterior distributions of an
additive marginal GP.}
\usage{
\S4method{show}{FunctionPosteriors}(object)
}
\arguments{
\item{object}{\linkS4class{FunctionPosteriors} object for which to apply a
class method.}
}
\description{
An S4 class to represent conditional function posterior distributions of an
additive marginal GP.
}
\section{Methods (by generic)}{
\itemize{
\item \code{show}: Print a summary about the object.
}}

\section{Slots}{

\describe{
\item{\code{components}}{A list with length equal to \code{num_paramsets}.
Each list element is a data frame representing a multivariate Gaussian
distribution of each model component (on normalized scale) and the sum of
the components (on normalized scale).}

\item{\code{x}}{The evaluation points (values of covariates) where the posteriors
have been evaluated. Original scale.}

\item{\code{model}}{The \linkS4class{lgpmodel} for which these posteriors are
computed. Contains important information about how to scale the total
posterior from normalized scale to the original scale.}

\item{\code{num_paramsets}}{Number of parameter sets for which the posteriors have
been computed.}

\item{\code{sigma2}}{Vector of \eqn{\sigma^2} values (noise variance parameter),
with length equal to \code{num_paramsets}.}
}}

\seealso{
\linkS4class{FunctionDraws}
}
