% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-lgpr.R, R/methods-GaussianPrediction.R
\docType{class}
\name{GaussianPrediction-class}
\alias{GaussianPrediction-class}
\alias{GaussianPrediction}
\alias{show,GaussianPrediction-method}
\alias{plot,GaussianPrediction,missing-method}
\title{An S4 class to represent analytic predictive distributions
(multivariate Gaussians), conditional on model parameters}
\usage{
\S4method{show}{GaussianPrediction}(object)

\S4method{plot}{GaussianPrediction,missing}(x, y, ...)
}
\arguments{
\item{object}{\linkS4class{GaussianPrediction} object for which to apply a
class method.}

\item{x}{a \linkS4class{GaussianPrediction} object to visualize}

\item{y}{unused argument}

\item{...}{optional arguments passed to a subroutine}
}
\description{
An S4 class to represent analytic predictive distributions
(multivariate Gaussians), conditional on model parameters
}
\section{Methods (by generic)}{
\itemize{
\item \code{show}: Print a summary about the object.

\item \code{plot}: Visualization.
Optional arguments (\code{...}) are passed to
\code{\link{plot_GaussianPrediction}}.
}}

\section{Slots}{

\describe{
\item{\code{f_post}}{An object of class \linkS4class{FunctionPosterior}.}

\item{\code{pred}}{A data frame.}
}}

\seealso{
\linkS4class{Prediction}
}
