% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-lgpr.R, R/methods-GaussianPrediction.R
\docType{class}
\name{GaussianPrediction-class}
\alias{GaussianPrediction-class}
\alias{GaussianPrediction}
\alias{show,GaussianPrediction-method}
\alias{plot,GaussianPrediction,missing-method}
\title{An S4 class to represent analytically computed predictive distributions
(conditional on hyperparameters) of an additive GP model}
\usage{
\S4method{show}{GaussianPrediction}(object)

\S4method{plot}{GaussianPrediction,missing}(x, y, ...)
}
\arguments{
\item{x}{a \linkS4class{GaussianPrediction} object to visualize}

\item{y}{unused argument}
}
\description{
An S4 class to represent analytically computed predictive distributions
(conditional on hyperparameters) of an additive GP model
}
\section{Methods (by generic)}{
\itemize{
\item \code{show}: Print a summary about the object.

\item \code{plot}: Visualization.
Optional arguments (\code{...}) are passed to
\code{\link{plot_pred}}.
}}

\section{Slots}{

\describe{
\item{\code{f_comp_mean}}{component means}

\item{\code{f_comp_std}}{component standard deviations}

\item{\code{f_mean}}{signal mean (on normalized scale)}

\item{\code{f_std}}{signal standard deviation (on normalized scale)}

\item{\code{y_mean}}{predictive mean (on original data scale)}

\item{\code{y_std}}{predictive standard deviation (on original data scale)}
}}

\seealso{
\linkS4class{Prediction}
}
