% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preproc-disease_info.R
\name{get_case_row_mappings}
\alias{get_case_row_mappings}
\title{Create case ID to rows and back mappings}
\usage{
get_case_row_mappings(X_notnan, X_id, only_R2C = FALSE)
}
\arguments{
\item{X_notnan}{binary vector indicating if diseaseAge is available for that measurement}

\item{X_id}{the id covariate in X}

\item{only_R2C}{should this return only the rows-to-caseID mapping}
}
\value{
a list
}
\description{
Create mappings
\itemize{
  \item from case ID to data rows (caseID_to_rows, caseID_nrows)
  \item  from row number to case ID (row_to_caseID)
}
}
