% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-fit_info.R
\name{get_draws}
\alias{get_draws}
\alias{get_draws.catch}
\alias{get_num_draws}
\title{Extract posterior draws}
\usage{
get_draws(fit, draws = NULL, reduce = NULL, ...)

get_draws.catch(fit, draws = NULL, reduce = NULL, ...)

get_num_draws(fit)
}
\arguments{
\item{fit}{an object of class \linkS4class{lgpfit}}

\item{draws}{Indices of parameter draws to return use. All post-warmup
draws are returned if this is \code{NULL}.}

\item{reduce}{Function used to reduce all parameter draws into
one set of parameters. Ignored if \code{NULL}, or if \code{draws} is not
\code{NULL}.}

\item{...}{additional keyword arguments to \code{rstan::extract}}
}
\value{
an array of shape \code{num_param_sets} x \code{num_params}
}
\description{
These functions use \code{\link[rstan]{extract}} with
\code{permuted = FALSE} and \code{inc_warmup = FALSE}. Chains are stacked
so that the return value is always a 2-dimensional array.
\itemize{
 \item \code{get_draws} extracts posterior draws after warmup
  \item \code{get_draws.catch} function calls \code{get_draws} but catches
errors and returns NULL if an error occurs.
  \item \code{get_num_draws} returns totals number of post-warmup draws
}
}
\seealso{
Other fit postprocessing functions: 
\code{\link{fit_summary}()},
\code{\link{get_y_rng}()}
}
\concept{fit postprocessing functions}
