% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethodsFit.R
\name{get_draws}
\alias{get_draws}
\title{Extract posterior draws}
\usage{
get_draws(fit, ...)
}
\arguments{
\item{fit}{an object of class \linkS4class{lgpfit}}

\item{...}{other keyword arguments to \code{rstan::extract}}
}
\value{
a named list
}
\description{
Uses \code{rstan::extract} with \code{permuted = FALSE} and
\code{inc_warmup = FALSE}.
}
\seealso{
Other fit postprocessing functions: 
\code{\link{fit_summary}()},
\code{\link{get_f}()},
\code{\link{scale_f_total}()}
}
\concept{fit postprocessing functions}
