% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-lgpmodel.R
\name{get_integer_encoding}
\alias{get_integer_encoding}
\title{Get integer encoding of model components}
\usage{
get_integer_encoding(object)
}
\arguments{
\item{object}{an object of class \linkS4class{lgpmodel} or
\linkS4class{lgpfit}}
}
\value{
A data frame.
}
\description{
Get integer encoding of model components
}
