% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred.R
\name{get_kernel_pars}
\alias{get_kernel_pars}
\title{Extract kernel parameter draws}
\usage{
get_kernel_pars(fit, reduce = NULL, draws = NULL)
}
\arguments{
\item{fit}{An object of class \linkS4class{lgpfit}.}

\item{reduce}{reduction for parameters draws}

\item{draws}{indices of parameter draws}
}
\value{
a list with names "alpha", "ell", "wrp", "beta" and "teff",
where each element is an array with number of rows equal to number of
parameter sets
}
\description{
Extract kernel parameter draws
}
