% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preproc-helper.R
\name{get_model_dims}
\alias{get_model_dims}
\title{Get some dimension variables that the Stan model needs as input}
\usage{
get_model_dims(X, D)
}
\arguments{
\item{X}{the design matrix}

\item{D}{a vector of length 6}
}
\value{
a list
}
\description{
Get some dimension variables that the Stan model needs as input
}
