% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-lgpfit.R
\name{get_posterior_f}
\alias{get_posterior_f}
\title{Extract posterior of the function f and its components}
\usage{
get_posterior_f(fit)
}
\arguments{
\item{fit}{an object of class \linkS4class{lgpfit}}
}
\value{
a named list where each element is a named list
(length \code{num_components + 1}) of arrays of size
\code{num_draws} x \code{num_chains}
}
\description{
Extract posterior of the function f and its components
}
