% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-pred.R
\name{get_pred}
\alias{get_pred}
\alias{get_pred.gaussian}
\alias{get_pred.sampled}
\title{Extract model predictions}
\usage{
get_pred(fit, draws = NULL, reduce = NULL)

get_pred.gaussian(fit, draws, reduce)

get_pred.sampled(fit, draws, reduce)
}
\arguments{
\item{fit}{an object of class \linkS4class{lgpfit}}

\item{draws}{Indices of parameter draws to return use. All post-warmup
draws are returned if this is \code{NULL}.}

\item{reduce}{Function used to reduce all parameter draws into
one set of parameters. Ignored if \code{NULL}, or if \code{draws} is not
\code{NULL}.}
}
\value{
an object of class \linkS4class{GaussianPrediction} or
\linkS4class{Prediction}
}
\description{
Extract model predictions
}
\seealso{
Other prediction extraction functions: 
\code{\link{get_pred_gaussian}},
\code{\link{get_pred_sampled}}
}
\concept{prediction extraction functions}
