% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_stan_model}
\alias{get_stan_model}
\title{Get a stan model of the package}
\usage{
get_stan_model(name = "lgp")
}
\arguments{
\item{name}{name of the model}
}
\value{
an object of class stanmodel
}
\description{
Get a stan model of the package
}
