% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethodsModel.R
\name{get_teff_obs}
\alias{get_teff_obs}
\title{Get observed disease effect times}
\usage{
get_teff_obs(object, dra = "diseaseAge")
}
\arguments{
\item{object}{an object of class \linkS4class{lgpmodel} or
\linkS4class{lgpfit}}

\item{dra}{name of the disease-related age variable}
}
\value{
a named vector, where the names are individual IDs
}
\description{
Get observed disease effect times
}
