% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_data.R
\name{get_teff_obs}
\alias{get_teff_obs}
\title{Get observed effect times from a data frame}
\usage{
get_teff_obs(data, group_by = "id", x = "age", x_ns = "diseaseAge")
}
\arguments{
\item{data}{a data frame}

\item{group_by}{name of the grouping varible, must be a factor
in \code{data}}

\item{x}{of the variable along which to extend,
must be a numeric in \code{data}}

\item{x_ns}{of a nonstationary variable derived from \code{x},
must be a numeric in \code{data}}
}
\value{
a named vector, where the names are the levels of \code{group_by}
}
\description{
Get observed effect times from a data frame
}
\seealso{
Other data utilities: 
\code{\link{add_dis_age}()},
\code{\link{add_factor_crossing}()},
\code{\link{add_factor}()},
\code{\link{data_types}()},
\code{\link{new_x}()},
\code{\link{pick_one_row_each}()},
\code{\link{select_factors_and}()},
\code{\link{split}()},
\code{\link{to_long_format}()},
\code{\link{valid_data}}
}
\concept{data utilities}
