% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula.R
\name{is_advanced_formula}
\alias{is_advanced_formula}
\title{Check if formula is in advanced format.}
\usage{
is_advanced_formula(rhs)
}
\arguments{
\item{rhs}{The formula right-hand side in text format.}
}
\value{
TRUE or FALSE
}
\description{
Check if formula is in advanced format.
}
\seealso{
Other model formula functions: 
\code{\link{formula_to_advanced}()},
\code{\link{parse_expr}()},
\code{\link{parse_formula}()},
\code{\link{parse_rhs}()},
\code{\link{parse_term}()},
\code{\link{parse_whitespace}()},
\code{\link{rhs_to_terms}()},
\code{\link{rhs_variables}()},
\code{\link{term_variables}()}
}
\concept{model formula functions}
