% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-formula.R
\name{is_advanced_formula}
\alias{is_advanced_formula}
\title{Check if formula is in advanced format.}
\usage{
is_advanced_formula(formula)
}
\arguments{
\item{formula}{The model formula, where
\itemize{
  \item it must contain exatly one tilde (\code{~}), with response
  variable on the left-hand side and model terms on the right-hand side
  \item terms are be separated by a plus (\code{+}) sign
  \item all variables appearing in \code{formula} must be
  found in \code{data}
}
See the "Model formula syntax" section below (\code{\link{lgp}}) for
instructions on how to specify the model terms.}
}
\value{
TRUE or FALSE
}
\description{
Checks if right-hand side contains opening parentheses.
}
\seealso{
Other model formula functions: 
\code{\link{parse_formula}()},
\code{\link{rhs_to_terms}()}
}
\concept{model formula functions}
