% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-kernel.R
\name{kernel}
\alias{kernel}
\alias{kernel_se}
\alias{kernel_ns}
\alias{kernel_zerosum}
\alias{kernel_bin}
\alias{kernel_var_mask}
\alias{kernel_beta}
\title{Compute a kernel matrix (covariance matrix)}
\usage{
kernel_se(x1, x2, alpha = 1, ell)

kernel_ns(x1, x2, alpha = 1, ell, a)

kernel_zerosum(x1, x2, M, alpha = 1)

kernel_bin(x1, x2 = NULL, alpha = 1, pos_class = 1)

kernel_var_mask(x1, x2, vm_params, a, nan_replace = 0)

kernel_beta(beta, row_to_caseID_1, row_to_caseID_2)
}
\arguments{
\item{x1}{vector of length \eqn{n}}

\item{x2}{vector of length \eqn{m}}

\item{alpha}{marginal std (default = 1)}

\item{ell}{lengthscale}

\item{a}{steepness of the warping function rise}

\item{M}{number of categories}

\item{pos_class}{the positive class label (integer)}

\item{vm_params}{vector of two mask function parameters}

\item{nan_replace}{value to replace \code{NaN}s in \code{x1} and \code{x2}}

\item{beta}{a parameter vector (row vector) of length \code{N_cases}}

\item{row_to_caseID_1}{mapinf from row index to case ID (integer vector of
length \eqn{n})}

\item{row_to_caseID_2}{mapping from row index to case ID (integer vector of
length \eqn{m})}
}
\value{
A matrix of size \eqn{n} x \eqn{m}.
}
\description{
Compute a kernel matrix (covariance matrix)
}
\section{Functions}{
\itemize{
\item \code{kernel_se}: Use the squared exponential kernel.

\item \code{kernel_ns}: Use the nonstationary kernel (input warping + squared
exponential).

\item \code{kernel_zerosum}: Uses the zero-sum kernel. Here, \code{x1} and 
\code{x2} must be integer vectors (integers denoting different categories).
Returns a binary matrix.

\item \code{kernel_bin}: Uses the binary (mask) kernel. Here, \code{x1} and 
\code{x2} must be integer vectors (integers denoting different categories).
Returns a binary matrix.

\item \code{kernel_var_mask}: Computes variance mask multiplier matrix.

\item \code{kernel_beta}: Computes the heterogeneity multiplier matrix.
}}

