% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior.R
\name{kernel_matrices}
\alias{kernel_matrices}
\alias{kernel_matrices_const}
\alias{kernel_matrices_create_input}
\alias{get_draws_kernel_pars}
\title{Kernel matrix computation using the C++ functions of the package namespace}
\usage{
kernel_matrices(model, x1, x2, theta, STREAM = get_stream())

kernel_matrices_const(model, x1, x2, STREAM = get_stream())

kernel_matrices_create_input(model, x)

get_draws_kernel_pars(fit)
}
\arguments{
\item{model}{An object of class \linkS4class{lgpmodel}.}

\item{x1}{A list defining the first input argument. Can be
created by \code{\link{kernels_matrices_create_input}}.}

\item{x2}{A list defining the second input argument. Can be
created by \code{\link{kernels_matrices_create_input}}.}

\item{theta}{A list of kernel parameter draws with length \code{num_draws}.}

\item{STREAM}{External pointer. By default obtained with
\code{\link[rstan]{get_stream}}.}

\item{x}{A data frame of points where the posteriors are computed.
The function \code{\link{new_x}} provides an easy way to create it.}

\item{fit}{An object of class \linkS4class{lgpfit}.}
}
\description{
\itemize{
  \item Kernel matrices using parameter values defined by each posterior
  draw can be computed using \code{kernel_matrices}. It returns a
  4-dimensional array of shape \code{num_draws} x \code{num_comps} x
  \code{n1} x \code{n2}.
  \item The constant kernel matrices that do not depend on parameters
  can be computed using \code{kernel_matrices_const}. It returns a list
  of kernel matrices with length equal to number of model components.
  \item Input information can be translated from a data frame to the
  needed format using \code{kernels_matrices_create_input}.
  \item Kernel parameter draws can be extracted using
  \code{get_draws_kernel_params}. It returns a list with names
  "alpha", "ell", "wrp", "beta" and "teff", where each element is an
  array with number of rows equal to number of posterior draws.
}
}
