% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R, R/methods-KernelComputer.R
\docType{class}
\name{KernelComputer-class}
\alias{KernelComputer-class}
\alias{KernelComputer}
\alias{show,KernelComputer-method}
\title{An S4 class to represent input for kernel matrix computations}
\usage{
\S4method{show}{KernelComputer}(object)
}
\arguments{
\item{object}{The object for which to call a class method.}
}
\description{
An S4 class to represent input for kernel matrix computations
}
\section{Methods (by generic)}{
\itemize{
\item \code{show}: Print a summary about the object.
}}

\section{Slots}{

\describe{
\item{\code{input}}{a list}

\item{\code{init}}{a list}

\item{\code{param_draws}}{formatted parameter draws}

\item{\code{STREAM}}{external pointer (for calling Stan functions)}
}}

