% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-kernel.R
\name{kernels_fpost}
\alias{kernels_fpost}
\title{Kernel matrices for function posterior computation}
\usage{
kernels_fpost(
  fit,
  x,
  reduce,
  draws,
  verbose,
  debug_dims,
  STREAM = get_stream()
)
}
\arguments{
\item{fit}{An object of class \linkS4class{lgpfit}.}

\item{x}{A data frame of points where function posterior distributions
and predictions should be computed or sampled.
The function \code{\link{new_x}} provides an easy way to create it.
If this is NULL, the data points are used.}

\item{reduce}{Reduction for parameters draws. Can be a function that
is applied to reduce all parameter draws into one parameter set, or
NULL (no reduction). Has no effect if \code{draws} is specified.}

\item{draws}{Indices of parameter draws to use, or \code{NULL} to use all
draws.}

\item{verbose}{Should more information and a possible progress bar be
printed?}

\item{debug_dims}{Should this print dimensions of some variables.
Can be used for debugging \code{\link{kernels_fpost}}.}

\item{STREAM}{External pointer. By default obtained with
\code{\link[rstan]{get_stream}}.}
}
\description{
Kernel matrices for function posterior computation
}
