% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgp-main.R
\name{lgp}
\alias{lgp}
\title{The main function of the `lgpr` package}
\usage{
lgp(formula, data, likelihood = "Gaussian", prior = prior_default(),
  uncertain_effect_time = FALSE, equal_effect = TRUE,
  id_variable = "id", time_variable = "age", disAge_variable = NULL,
  continuous_vars = NULL, categorical_vars = NULL,
  offset_vars = NULL, C_hat = NULL, DELTA = 1e-12,
  sample_F = (likelihood != "Gaussian"), parallel = FALSE,
  skip_postproc = FALSE, t_test = NULL, threshold = 0.95,
  variance_mask = TRUE, ell_smooth = "ell_shared",
  ell_smooth_multip = 1, cat_interact_kernel_type = "categorical",
  N_trials = NULL, ...)
}
\arguments{
\item{formula}{A formula of the form
\code{y ~ } \code{x1 + x2 + x3}
defining the response variable \code{y} and covariates \code{xi}. All variables that
appear in the formula must exist as columns of \code{data}.}

\item{data}{A data frame containing (at least) the variables given in \code{formula}.}

\item{likelihood}{Determines the observation model. Must be either \code{"Gaussian"} 
(default), \code{"Poisson"}, \code{"NB" (negative binomial) or "binomial"}.}

\item{prior}{Prior distribution. Can be created for example using the function 
\code{\link{prior_default}}.}

\item{uncertain_effect_time}{Do we wish to model uncertainty in the disease effect time?}

\item{equal_effect}{Is the disease effect assumed to be equally strong for all diseased individuals?}

\item{id_variable}{Name of the unique subject identifier variable.}

\item{time_variable}{Name of the time variable.}

\item{disAge_variable}{Name of the disease-related age variable. If {NULL}, this
will be chosen to be "diseaseAge", if such covariate is found in the data.}

\item{continuous_vars}{Names of other continuous covariates. If \code{NULL}, the
remaining covariates that have floating point values are interpreted as continuous.}

\item{categorical_vars}{Names of categorical covariates that interact with the time variable.
If \code{NULL}, the remaining covariates that have integer values are interpreted 
as categorical.}

\item{offset_vars}{Names of the categorical covariates that are treated as 
time-independent group offsets. If \code{NULL}, no variables are interpreted as such
covariates.}

\item{C_hat}{This can only be given if likelihood is Poisson or NB. The signal \code{f} 
will the be transformed so that \code{g = } \code{exp(C_hat + f)}. If \code{NULL}, it will be
set to \code{C_hat = } \code{log(mean(y))}, where \code{y} is the response variable.}

\item{DELTA}{the amount of added jitter to ensure positive definiteness of the kernel}

\item{sample_F}{Determines if the function values are be sampled (must be \code{TRUE} if
likelihood is not Gaussian).}

\item{parallel}{Determines if the chain will be run in parallel (\code{default = FALSE}).
If \code{TRUE}, then Stan is run by first defining 
\code{options(mc.cores = } \code{parallel::detectCores())}.}

\item{skip_postproc}{In this mode the postprocessing after running Stan is skipped.}

\item{t_test}{Optional test time points. Should only be used if \code{sample_F = TRUE}.
Otherwise use \code{\link{lgp_predict}} after fitting the model.}

\item{threshold}{Covariate selection threshold.}

\item{variance_mask}{Should a variance mask be used to force disease component
variance to zero before disease onset?}

\item{ell_smooth}{Defines how to determine smoothing lengthscale for corrected shared age 
effect inference. Possible options are
\enumerate{
   \item \code{"ell_shared"} (default) - the sampled lengthscale of the shared age 
   component is used as \code{ell_smooth}
   \item \code{"none"} - no correction will be performed
   \item A numeric argument that directly defines \code{ell_smooth}
}}

\item{ell_smooth_multip}{a multiplier for ell_smooth}

\item{cat_interact_kernel_type}{Kernel type for categorical variables (other than id). 
Possible options are \code{"categorical"} (default) and \code{"binary"} (mask kernel where
only category \code{"1"} will have an effect).}

\item{N_trials}{This argument (number of trials) is only needed when likelihood is binomial.
Must have length one or equal to number of data points. Setting \code{N_trials=1} corresponds to 
Bernoulli observation model.}

\item{...}{Optional arguments passed to \code{rstan::sampling}, for example \code{iter},
\code{chains} or \code{control}. See \code{\link[rstan]{sampling}} for the possible arguments.}
}
\value{
An object of class \code{lgpfit}.
}
\description{
This is a wrapper for both \code{\link{lgp_model}} and \code{\link{lgp_fit}}.
It first creates an \code{lgpmodel} object and then fits the model, finally returning an
\code{lgpfit} object.
}
