% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgpr.R
\name{lgp}
\alias{lgp}
\title{Main function of the lgpr package}
\usage{
lgp(
  formula,
  data,
  likelihood = "gaussian",
  prior = NULL,
  c_hat = NULL,
  num_trials = NULL,
  options = NULL,
  prior_only = FALSE,
  verbose = FALSE,
  sample_f = !(likelihood == "gaussian"),
  ...
)
}
\arguments{
\item{formula}{The model formula, where
\itemize{
  \item it must contain exatly one tilde (\code{~}), with response
  variable on the left-hand side and model terms on the right-hand side
  \item terms are be separated by a plus (\code{+}) sign
  \item all variables appearing in \code{formula} must be
  found in \code{data}
}
See the "Model formula syntax" section below (\code{\link{lgp}}) for
instructions on how to specify the model terms.}

\item{data}{A \code{data.frame} where each column corresponds to one
variable, and each row is one observation. Continuous covariates and the
response variable must have type \code{"numeric"} and categorical covariates
must have type \code{"factor"}. Missing values should be indicated with
\code{NaN} or \code{NA}. The response variable cannot contain missing
values.}

\item{likelihood}{Determines the observation model. Must be either
\code{"gaussian"} (default), \code{"poisson"}, \code{"nb"} (negative
binomial), \code{"binomial"} or \code{"bb"} (beta binomial).}

\item{prior}{A named list, defining the prior distribution of model
(hyper)parameters. See the "Defining priors" section below
(\code{\link{lgp}}).}

\item{c_hat}{The GP mean. Must be a vector of length \code{dim(data)[1]}, or
a real number defining a constant GP mean. If \code{NULL}, this is set to
 \itemize{
   \item \code{c_hat = 0}, if \code{likelihood} is \code{"gaussian"}, because
   with Gaussian likelihood the response variable is by default centered to
   have zero mean.
   \item \code{c_hat = } \code{log(mean(y))} if \code{likelihood} is
   \code{"poisson"} or \code{"nb"},
   \item \code{c_hat = } \code{log(p/(1-p))}, where
   \code{p = mean(y/num_trials)} if \code{likelihood} is \code{"binomial"},
 }
where \code{y} denotes the response variable. You can modify this vector to
account for normalization between data points. With Gaussian likelihood
though, do not modify this argument, normalize the data beforehand instead.}

\item{num_trials}{This argument (number of trials) is only needed when
likelihood is binomial or beta binomial. Must have length one or equal
to number of data points. Setting \code{num_trials=1} and
\code{likelihood="binomial"} corresponds to Bernoulli observation model.}

\item{options}{A named list with the following possible fields:
\itemize{
  \item \code{delta} Amount of added jitter to ensure positive definite
  covariance matrices.
}}

\item{prior_only}{Should this run in prior sampling mode,
where likelihood is ignored?}

\item{verbose}{Should more verbose output be printed?}

\item{sample_f}{Determines if the latent function values are be sampled
(must be \code{TRUE} if likelihood is not \code{"gaussian"}).}

\item{...}{Optional arguments passed to
\code{\link[rstan]{sampling}} or \code{\link[rstan]{optimizing}}.}
}
\value{
Returns an object of the S4 class \linkS4class{lgpfit}.
}
\description{
Creates a model by calling \code{\link{create_model}} and fits by calling
\code{\link{sample_model}}.
}
\section{Model formula syntax}{

There are two ways to define the model formula:
\enumerate{
  \item Using a common \code{\link[stats]{formula}}-like syntax, like in
  \code{y ~ age +} \code{age|id} \code{ + sex}. Terms can consits of a
  single variable, such as \code{age}, or an interaction of two variables,
  such as \code{age|id}. In single-variable terms, the variable can be either
  continuous (numeric) or categorical (factor), whereas in interaction terms
  the variable on the left-hand side of the vertical bar (\code{|}) has to
  be continuous and the one on the right-hand side has to be categorical.
  Formulae specified using this syntax are translated to the advanced format
  so that
  \itemize{
    \item single-variable terms become \code{gp(x)} if
    variable \code{x} is numeric and \code{zs(x)} if \code{x} is a factor
    \item interaction terms \code{x|z} become \code{gp(x)*zs(z)}
  }
  \item Using the advanced syntax, like in \code{y ~ gp(age) +}
  \code{gp(age)*zs(id) +} \code{het(id)*gp_vm(disAge)}.
  This creates \linkS4class{lgprhs} objects, which consist of
 \linkS4class{lgpterm}s, which consist of \linkS4class{lgpexpr}s.
 This approach must be used if creating nonstationary, heterogeneous or
 temporally uncertain components.
}
Either one of the approaches should be used and they should not be mixed.
}

\section{Defining priors}{

The \code{priors} argument must be a named list.
\enumerate{
  \item Ok.
}
}

\seealso{
Other main functions: 
\code{\link{create_model}()},
\code{\link{sample_model}()}
}
\concept{main functions}
