% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgpr.R
\name{lgp}
\alias{lgp}
\alias{sample_model}
\alias{optimize_model}
\title{Main functions of the `lgpr` package}
\usage{
lgp(
  formula,
  data,
  likelihood = "gaussian",
  prior = NULL,
  c_hat = NULL,
  num_trials = NULL,
  options = NULL,
  prior_only = FALSE,
  verbose = FALSE,
  sample_f = !(likelihood == "gaussian"),
  ...
)

sample_model(model, ...)

optimize_model(model, ...)
}
\arguments{
\item{formula}{A formula specified using the common \code{formula}
syntax, such as \code{y ~ x1 + x2:z1 + } \code{x2:z2 + z2}.
\itemize{
  \item The formula must contain exatly one tilde (\code{~}), with response
  variable on the left-hand side and model terms on the right-hand side.
  \item Terms are be separated by a plus (\code{+}) sign.
  \item Terms can consits of a single variable name, such as \code{x}, or
  an interaction of two variables, such as \code{x:z}.
  \item In single-variable terms, the variable can be either continuous or
  categorical, whereas in interaction terms the variable
  on the left-hand side of the colon (\code{:}) has to be continuous and the
  one on the right-hand side has to be categorical (a factor).
  \item All variables appearing in \code{formula} must be
  found in \code{data}.
}}

\item{data}{A \code{data.frame} where each column corresponds to one
variable, and each row is one observation. Continuous covariates and the
response variable must have type \code{"numeric"} and categorical covariates
must have type \code{"factor"}. Missing values should be indicated with
\code{NaN} or \code{NA}. The response variable cannot contain missing
values.}

\item{likelihood}{Determines the observation model. Must be either
\code{"gaussian"} (default), \code{"poisson"}, \code{"nb"} (negative
binomial) or \code{"binomial"}. To use Bernoulli likelihood, use
\code{likelihood="binomial"} and set \code{num_trials} as a vector of ones.}

\item{prior}{A named list, defining the prior distribution of model
(hyper)parameters.}

\item{c_hat}{The GP mean. Must be a vector of length \code{dim(data)[1]}, or
a real number defining a constant GP mean. If \code{NULL}, this is set to
 \itemize{
   \item \code{c_hat = 0}, if \code{likelihood} is \code{"gaussian"}, because
   with Gaussian likelihood the response variable is by default centered to
   have zero mean.
   \item \code{c_hat = } \code{log(mean(y))} if \code{likelihood} is
   \code{"poisson"} or \code{"nb"},
   \item \code{c_hat = } \code{log(p/(1-p))}, where
   \code{p = mean(y/num_trials)} if \code{likelihood} is \code{"binomial"},
 }
where \code{y} denotes the response variable. You can modify this vector to
account for normalization between data points. With Gaussian likelihood
though, do not modify this argument, normalize the data beforehand instead.}

\item{num_trials}{This argument (number of trials) is only needed when
likelihood is binomial. Must have length one or equal to number of data
points. Setting \code{num_trials=1} corresponds to Bernoulli observation
model.}

\item{options}{A named list with the following possible fields:
\itemize{
  \item \code{skip_generated} If this is true, the generated quantities
  block of Stan is skipped.
  \item \code{delta} Amount of added jitter to ensure positive definite
  covariance matrices.
}}

\item{prior_only}{Should this run in prior sampling mode,
where likelihood is ignored?}

\item{verbose}{Should more verbose output be printed?}

\item{sample_f}{Determines if the latent function values are be sampled
(must be \code{TRUE} if likelihood is not \code{"gaussian"}).}

\item{...}{Optional arguments passed to \code{rstan::optimizing}, such as
\code{algorithm}. See \code{\link[rstan]{optimizing}} for the possible
arguments.}

\item{model}{An object of class \linkS4class{lgpmodel}.}
}
\description{
\itemize{
  \item The \code{lgp} function is a wrapper for both \code{create_model}
  and \code{sample_model}, and its returns an object of class
  \linkS4class{lgpfit}.
  \item The \code{\link{create_model}} function creates an object of class
  \linkS4class{lgpmodel}.
  \item The \code{sample_model} function takes an \linkS4class{lgpmodel}
  object, fits the model using \code{rstan::sampling}, and returns an object
  of class \linkS4class{lgpfit}.
  \item The \code{optimize_model} function takes an \linkS4class{lgpmodel}
  object and fits the model using \code{rstan::optimizing}.
}
}
