% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgp-main.R
\name{lgp}
\alias{lgp}
\title{The main function of the `lgpr` package}
\usage{
lgp(
  formula,
  data,
  likelihood = "Gaussian",
  prior = prior_default(),
  uncertain_effect_time = FALSE,
  equal_effect = TRUE,
  id_variable = "id",
  time_variable = "age",
  disAge_variable = NULL,
  continuous_vars = NULL,
  categorical_vars = NULL,
  offset_vars = NULL,
  C_hat = NULL,
  DELTA = 1e-08,
  sample_F = NULL,
  parallel = FALSE,
  skip_postproc = FALSE,
  threshold = 0.95,
  variance_mask = TRUE,
  N_trials = NULL,
  relevance_method = "f_mean",
  verbose = FALSE,
  ...
)
}
\arguments{
\item{formula}{A formula specified using the common \code{\link{formula}}
syntax, such as \code{y ~ x1 + x2:z1 + } \code{x2:z2 + z2}.
\itemize{
  \item The formula must contain exatly one tilde (\code{~}), with response
  variable on the left-hand side and model terms on the right-hand side.
  \item Terms are be separated by a plus (\code{+}) sign.
  \item Terms can consits of a single variable name, such as \code{x}, or
  an interaction of two variables, such as \code{x:z}.
  \item In single-variable terms, the variable can be either continuous or
  categorical, whereas in interaction terms the variable
  on the left-hand side of the colon (\code{:}) has to be continuous and the
  one on the right-hand side has to be categorical (a factor).
  \item All variables appearing in \code{formula} must be
  found in \code{data}.
}}

\item{data}{A \code{data.frame} where each column corresponds to one
variable, and each row is one observation. Continuous covariates must have
type \code{"numeric"} and categorical ones must have type \code{"factor"}.
Missing values should be indicated with \code{NaN} or \code{NA}.}

\item{likelihood}{Determines the observation model. Must be either
\code{"gaussian"} (default), \code{"poisson"}, \code{"nb"} (negative
binomial) or \code{"binomial"}. To use Bernoulli likelihood, use
\code{likelihood="binomial"} and set \code{num_trials} as a vector of ones.}

\item{prior}{A named list, defining the prior distribution of model
(hyper)parameters. It is recommended to first create this using the function
\code{\link{prior_default}}, and then possibly modify it.}

\item{id_variable}{Name of the unique subject identifier variable
(default = \code{"id"}).}

\item{parallel}{This argument has been deprecated and throws and error if
\code{TRUE}. Use the \code{cores} argument of \code{rstan::sampling} instead.}

\item{skip_postproc}{In this mode the postprocessing after running Stan is
skipped.}

\item{threshold}{Component selection threshold for relevance sum.}

\item{relevance_method}{Component relevance determination method.
Must be either \code{"f_mean"} or \code{"alpha"}.}

\item{verbose}{should some output be printed?}

\item{...}{Additional arguments that are passed to \code{rstan::sampling}.}
}
\value{
An object of class \code{lgpfit}.
}
\description{
This is a wrapper for both \code{\link{lgp_model}}
 and \code{\link{lgp_fit}}. It first creates an
 \code{lgpmodel} object and then fits the model,
 finally returning an \code{lgpfit} object. Note that the
 covariate types are automatically inferred from the given
 \code{data}. If you wish to change these, see the
 arguments
 \itemize{
    \item \code{id_variable}
    \item \code{time_variable}
    \item \code{disAge_variable}
    \item \code{continuous_vars} and
    \item \code{categorical_vars}.
 }
}
