% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgp-fit.R
\name{lgp_fit}
\alias{lgp_fit}
\title{Fit an lgp model}
\usage{
lgp_fit(
  model,
  threshold = 0.95,
  parallel = FALSE,
  skip_postproc = FALSE,
  relevance_method = "f_mean",
  verbose = FALSE,
  ...
)
}
\arguments{
\item{model}{An object of class \code{lgpmodel}.}

\item{threshold}{Component selection threshold for relevance sum.}

\item{parallel}{This argument has been deprecated and throws and error if
\code{TRUE}. Use the \code{cores} argument of \code{rstan::sampling} instead.}

\item{skip_postproc}{In this mode the postprocessing after running Stan is
skipped.}

\item{relevance_method}{Component relevance determination method.
Must be either \code{"f_mean"} or \code{"alpha"}.}

\item{verbose}{should some output be printed?}

\item{...}{Additional arguments that are passed to \code{rstan::sampling}.}
}
\value{
An object of class \code{lgpfit}.
}
\description{
Samples the posterior of an additive Gaussian process regression
model using \code{\link{rstan}}.
}
\seealso{
For all possible additional arguments (\code{...}), see
\code{?rstan::sampling}. For creating the \code{lgpmodel} input, see
\code{\link{lgp_model}}.
}
