% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgp-fit.R
\name{lgp_fit}
\alias{lgp_fit}
\title{Fit a model}
\usage{
lgp_fit(model, ...)
}
\arguments{
\item{model}{An object of class \linkS4class{lgpmodel}.}

\item{...}{Optional arguments passed to \code{rstan::sampling}, for example
\code{iter}, \code{chains} or \code{control}. See
\code{\link[rstan]{sampling}} for the possible arguments.}
}
\description{
Fit a model
}
