% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgp-main.R
\name{lgp_fit}
\alias{lgp_fit}
\title{Fit an lgp model}
\usage{
lgp_fit(model, threshold, parallel = FALSE, skip_postproc = FALSE,
  ell_smooth = "ell_shared", ell_smooth_multip = 1, ...)
}
\arguments{
\item{model}{An object of class \code{lgpmodel}.}

\item{threshold}{Covariate selection threshold.}

\item{parallel}{Determines if the chain will be run in parallel (\code{default = FALSE}).
If \code{TRUE}, then Stan is run by first defining 
\code{options(mc.cores = } \code{parallel::detectCores())}.}

\item{skip_postproc}{In this mode the postprocessing after running Stan is skipped.}

\item{ell_smooth}{Defines how to determine smoothing lengthscale for corrected shared age 
effect inference. Possible options are
\enumerate{
   \item \code{"ell_shared"} (default) - the sampled lengthscale of the shared age 
   component is used as \code{ell_smooth}
   \item \code{"none"} - no correction will be performed
   \item A numeric argument that directly defines \code{ell_smooth}
}}

\item{ell_smooth_multip}{a multiplier for ell_smooth}

\item{...}{Optional arguments passed to \code{rstan::sampling}, for example \code{iter},
\code{chains} or \code{control}. See \code{\link[rstan]{sampling}} for the possible arguments.}
}
\value{
An object of class \code{lgpfit}.
}
\description{
Samples the posterior of an additive Gaussian process regression model
using \code{\link{rstan}}.
}
\seealso{
For the possible additional arguments, see \code{\link[rstan]{sampling}}.
For creating the \code{lgpmodel} input, see \code{\link{lgp_model}}.
}
