% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgp-main.R
\name{lgp_test}
\alias{lgp_test}
\title{Compute predictions and log-posterior predictive density at test points}
\usage{
lgp_test(fit, test_data, plot = FALSE, reduction = "mean",
  samples = "mean")
}
\arguments{
\item{fit}{an object of class \code{lgpfit}}

\item{test_data}{a test data matrix}

\item{plot}{should this plot the data and predictions?}

\item{reduction}{must be either "mean", "sum" or "none"}

\item{samples}{The predictions can be computed either by using only the posterior mean
\cr (\code{samples="mean"}) or median (\code{samples="median"}) parameters, or by averaging over
all parameter samples (\code{samples="all"}). This can also be a set of indices, for example 
\code{samples=c(1:10)} gives the averaged predictions over the parameter samples 1...10.}
}
\value{
a ggplot object or lppd
}
\description{
This is a convenience function that wraps \code{\link{lgp_predict}},
\code{\link{compute_lppd}} and \code{\link{plot_predictions}}.
}
