% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\docType{class}
\name{lgpexpr-class}
\alias{lgpexpr-class}
\alias{lgpexpr}
\title{An S4 class to represent an lgp expression}
\description{
An S4 class to represent an lgp expression
}
\section{Slots}{

\describe{
\item{\code{covariate}}{name of a covariate}

\item{\code{fun}}{function name}
}}

\seealso{
See \code{\link{operations}} for performing arithmetics
on \linkS4class{lgprhs}, \linkS4class{lgpterm} and \linkS4class{lgpexpr}
objects.
}
