% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-classes.R
\docType{class}
\name{lgpfit-class}
\alias{lgpfit-class}
\alias{lgpfit}
\title{An S4 class to represent the output of the \code{lgp_fit} function}
\description{
An S4 class to represent the output of the \code{lgp_fit} function
}
\section{Slots}{

\describe{
\item{\code{stan_fit}}{The \code{stanfit} object returned by \code{rstan::sampling}.}

\item{\code{model}}{The \code{lgpmodel} object returned by \code{lgp_model}.}

\item{\code{relevances}}{Inferred component relevances.}

\item{\code{selection}}{Component selection info.}

\item{\code{pkg_version}}{Package version number.}

\item{\code{diagnostics}}{A data frame with columns \code{c("Rhat", "Bulk_ESS", "Tail_ESS")}.}
}}

