% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{lgpfit-class}
\alias{lgpfit-class}
\alias{lgpfit}
\title{An S4 class to represent the output of the \code{lgp} function}
\description{
An S4 class to represent the output of the \code{lgp} function
}
\section{Slots}{

\describe{
\item{\code{stan_fit}}{An object of class \code{stanfit}.}

\item{\code{model}}{An object of class \code{lgpmodel}.}
}}

\section{Related methods and functions}{

}

\seealso{
All methods that work on \linkS4class{lgpmodel}
objects work also on \linkS4class{lgpfit} objects.

For complete info on
accessing the properties of the \code{stan_fit} slot, see
\href{https://cran.r-project.org/web/packages/rstan/vignettes/stanfit-objects.html}{here}.

For plotting, see \code{\link{plot_fit}}.
}
\concept{model fit vizualization functions}
