% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-all_classes.R, R/S4-lgpfit.R
\docType{class}
\name{lgpfit-class}
\alias{lgpfit-class}
\alias{lgpfit}
\alias{show,lgpfit-method}
\alias{get_draws,lgpfit-method}
\title{An S4 class to represent the output of the \code{lgp} function}
\usage{
\S4method{show}{lgpfit}(object)

\S4method{get_draws}{lgpfit}(object, draws = NULL, reduce = NULL, ...)
}
\arguments{
\item{object}{\linkS4class{lgpfit} object for which to apply a class method.}

\item{draws}{Indices of the parameter draws. \code{NULL} corresponds to
all post-warmup draws.}

\item{reduce}{Function used to reduce all parameter draws into
one set of parameters. Ignored if \code{NULL}, or if \code{draws} is not
\code{NULL}.}

\item{...}{additional keyword arguments to \code{\link[rstan]{extract}}}
}
\description{
An S4 class to represent the output of the \code{lgp} function
}
\section{Methods (by generic)}{
\itemize{
\item \code{show}: Print information and summary about the fit object.

\item \code{get_draws}: Extract parameter draws. Uses \code{\link[rstan]{extract}} with
\code{permuted = FALSE} and \code{inc_warmup = FALSE}, so that the return
value is always a 2-dimensional array of shape
\code{num_param_sets} x \code{num_params}.
}}

\section{Slots}{

\describe{
\item{\code{stan_fit}}{An object of class \code{stanfit}.}

\item{\code{model}}{An object of class \code{lgpmodel}.}

\item{\code{num_draws}}{Total number of parameter draws.}
}}

\seealso{
For complete info on accessing the properties of the
\code{stan_fit} slot, see
\href{https://cran.r-project.org/web/packages/rstan/vignettes/stanfit-objects.html}{here}.
}
