% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_s4_classes.R
\docType{class}
\name{lgpfit-class}
\alias{lgpfit-class}
\alias{lgpfit}
\title{An S4 class to represent the output of the \code{lgp} function}
\description{
An S4 class to represent the output of the \code{lgp} function
}
\section{Slots}{

\describe{
\item{\code{stan_fit}}{An object of class \code{stanfit}.}

\item{\code{model}}{An object of class \code{lgpmodel}.}
}}

\section{Related methods and functions}{

\itemize{
    \item \code{\link{plot_posterior}}
    \item \code{\link{plot_posterior_warp}}
    \item In addition, all methods that work on \linkS4class{lgpmodel}
    objects work also on \linkS4class{lgpfit} objects.
}
}

\seealso{
The main function \code{\link{lgp}}.
}
