% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-classes.R
\docType{class}
\name{lgpfit-class}
\alias{lgpfit-class}
\alias{lgpfit}
\title{An S4 class to represent the output of the \code{lgp_fit} function}
\description{
All slots that are lists contain fields 'samples' and 'average'.
}
\section{Slots}{

\describe{
\item{\code{stan_fit}}{The \code{stanfit} object returned by \code{rstan::sampling}.}

\item{\code{model}}{The \code{lgpmodel} object returned by \code{lgp_model}.}

\item{\code{components}}{Inferred components.}

\item{\code{components_corrected}}{Covariate-effect corrected components.}

\item{\code{component_relevances}}{Inferred component relevances.}

\item{\code{covariate_relevances}}{Inferred covariate relevances.}

\item{\code{covariate_selection}}{Covariate selection info.}

\item{\code{signal_variance}}{Signal variance.}

\item{\code{residual_variance}}{Residual variance.}

\item{\code{postproc_info}}{Postprocessing information.}

\item{\code{pkg_version}}{Package version number.}

\item{\code{Rhat}}{Split Rhat statistics.}
}}

