% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_s4_classes.R
\docType{class}
\name{lgpformula-class}
\alias{lgpformula-class}
\alias{lgpformula}
\title{An S4 class to represent an lgp formula}
\description{
An S4 class to represent an lgp formula
}
\section{Slots}{

\describe{
\item{\code{terms}}{an object of class \code{\link{lgpsum}}}

\item{\code{response}}{name of the response variable}

\item{\code{call}}{original formula call}
}}

