% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\docType{class}
\name{lgpformula-class}
\alias{lgpformula-class}
\alias{lgpformula}
\title{An S4 class to represent an lgp formula}
\description{
An S4 class to represent an lgp formula
}
\section{Slots}{

\describe{
\item{\code{terms}}{an object of class \linkS4class{lgprhs}}

\item{\code{y_name}}{name of the response variable}

\item{\code{call}}{original formula call}
}}

\seealso{
See \code{\link{operations}} for performing arithmetics
on \linkS4class{lgprhs}, \linkS4class{lgpterm} and \linkS4class{lgpexpr}
objects.
}
