% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_s4_classes.R
\docType{class}
\name{lgpmodel-class}
\alias{lgpmodel-class}
\alias{lgpmodel}
\title{An S4 class to represent an lgp model}
\description{
An S4 class to represent an lgp model
}
\section{Slots}{

\describe{
\item{\code{formula}}{An object of class \linkS4class{lgpformula}}

\item{\code{stan_dat}}{The data to be given as input to \code{rstan::sampling}}

\item{\code{x_scaling}}{Continous covariate normalization functions and their
inverse operations. Must be a named list with each element is an
\linkS4class{lgpscaling} object.}

\item{\code{y_scaling}}{Response variable normalization function and its
inverse operation. Must be an \linkS4class{lgpscaling} object.}

\item{\code{info}}{Model info.}
}}

