% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_s4_classes.R
\docType{class}
\name{lgpmodel-class}
\alias{lgpmodel-class}
\alias{lgpmodel}
\title{An S4 class to represent an lgp model}
\description{
An S4 class to represent an lgp model
}
\section{Slots}{

\describe{
\item{\code{formula}}{An object of class \linkS4class{lgpformula}}

\item{\code{stan_dat}}{The data to be given as input to \code{rstan::sampling}}

\item{\code{var_names}}{List of variable names grouped by type.}

\item{\code{var_scalings}}{A named list with fields
\itemize{
  \item \code{y} - Response variable normalization function and its
  inverse operation. Must be an \linkS4class{lgpscaling} object.
  \item \code{x_cont} - Continous covariate normalization functions and
  their inverse operations. Must be a named list with each element is an
  \linkS4class{lgpscaling} object.
}}

\item{\code{var_info}}{A named list with fields
\itemize{
  \item \code{x_cat_levels} - Names of the levels of categorical covariates
  before converting from factor to numeric.
}}

\item{\code{info}}{Other info in text format.}

\item{\code{stan_model_name}}{Name of Stan model.}
}}

\section{Related methods and functions}{

\itemize{
    \item \code{\link{model_summary}}
    \item \code{\link{prior_summary}} and \code{\link{param_summary}}
    \item \code{\link{model_getters}}
}
}

