% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgpr-package.R
\docType{package}
\name{lgpr-package}
\alias{lgpr-package}
\alias{lgpr}
\title{The 'lgpr' package.}
\description{
Longitudinal Gaussian Process regression. 
The package features
\itemize{
  \item Additive Gaussian process modeling of longitudinal data
  \item Posterior inference of the model (hyper)parameters using Stan
  \item Computation of covariate relevances, i.e. how much each covariate explains the target variable
  \item Specialized modeling of a non-stationary disease effect
  \item Functions for visualizing longitudinal data, posterior samples and model predictions
  \item Gaussian, Poisson or Negative Binomial observation models
}
}
\section{Basic usage}{

\itemize{
\item See the main function \code{\link{lgp}} for creating and fitting additive 
 longitudinal GP models.
\item Predictions outside the data can be computed using the function \code{\link[lgpr]{lgp_predict}}.
\item See documentation of the function \code{\link[lgpr]{simulate_data}} for generating
 artificial data.
\item For visualizing the data and results, see for example the functions 
  \itemize{
    \item \code{\link[lgpr]{plot_data}}
    \item \code{\link[lgpr]{plot_samples}}
    \item \code{\link[lgpr]{plot_components}}
    \item \code{\link[lgpr]{plot_posterior_y}}
    \item \code{\link[lgpr]{plot_simdata}}
  }
}
}

\references{
\enumerate{
  \item Carpenter, B. et al. (2017). \emph{Stan: A probabilistic programming language}. Journal of Statistical Software 76(1).
  \item Jonah Gabry, Ben Goodrich and Martin Lysy (2019). \emph{rstantools: Tools for Developing R Packages
  Interfacing with 'Stan'}. R package version 2.0.0.
  \item Gabry, J. and Mahr, T. (2019). \emph{bayesplot: Plotting for Bayesian Models}. R package version 1.7.0, http://mc-stan.org/bayesplot.
  \item Stan Development Team (2019). \emph{RStan: the R interface to Stan.} R package version 2.19.2. http://mc-stan.org/.
}
}
\author{
Juho Timonen (first.last at aalto.fi)
}
\keyword{Gaussian}
\keyword{Stan,}
\keyword{covariate}
\keyword{data,}
\keyword{interpretable}
\keyword{longitudinal}
\keyword{models}
\keyword{processes,}
\keyword{selection,}
