% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgpr-package.R
\docType{package}
\name{lgpr-package}
\alias{lgpr-package}
\alias{lgpr}
\title{The 'lgpr' package.}
\description{
Longitudinal Gaussian Process regression.
The package features
\itemize{
  \item Additive Gaussian process modeling of longitudinal data
  \item Posterior inference of the model (hyper)parameters using Stan
  \item Computation of covariate relevances
  \item Specialized modeling of a non-stationary disease effect
  \item Functions for visualizing longitudinal data, posterior samples and
  model predictions
  \item Gaussian, Poisson, binomial or negative binomial observation models
}
}
\section{Basic usage}{

\itemize{
\item See the main function documentation: \code{\link{lgp}}.
 \item See tutorials at
 \url{https://jtimonen.github.io/lgpr-usage/index.html}
}
}

\section{Citation}{


\emph{An interpretable probabilistic machine learning method for
heterogeneous longitudinal studies}. Juho Timonen, Henrik Mannerstrom,
Aki Vehtari and Harri Lahdesmaki, 2019.
\url{https://arxiv.org/abs/1912.03549}
}

\references{
\enumerate{
  \item Carpenter, B. et al. (2017).
  \emph{Stan: A probabilistic programming language}. Journal of Statistical
   Software 76(1).
  \item Jonah Gabry, Ben Goodrich and Martin Lysy (2019).
  \emph{rstantools: Tools for Developing R Packages Interfacing with 'Stan'}.
  R package version 2.0.0.
  \item Gabry, J. and Mahr, T. (2019).
  \emph{bayesplot: Plotting for Bayesian Models}. R package version 1.7.0,
   http://mc-stan.org/bayesplot.
  \item Stan Development Team (2019). \emph{RStan: the R interface to Stan.}
  R package version 2.19.2. http://mc-stan.org/.
}
}
\author{
Juho Timonen (first.last at iki.fi)
}
\keyword{Gaussian}
\keyword{Stan,}
\keyword{covariate}
\keyword{data,}
\keyword{interpretable}
\keyword{longitudinal}
\keyword{models}
\keyword{processes,}
\keyword{selection,}
