% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\docType{class}
\name{lgpscaling-class}
\alias{lgpscaling-class}
\alias{lgpscaling}
\title{An S4 class to represent variable scaling}
\description{
An S4 class to represent variable scaling
}
\section{Slots}{

\describe{
\item{\code{loc}}{original location (mean)}

\item{\code{scale}}{original scale (standard deviation)}

\item{\code{var_name}}{variable name}
}}

