% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_s4_classes.R
\docType{class}
\name{lgpterm-class}
\alias{lgpterm-class}
\alias{lgpterm}
\title{An S4 class to represent one formula term}
\description{
An S4 class to represent one formula term
}
\section{Slots}{

\describe{
\item{\code{factors}}{a list of at most two \code{\link{lgpexpr}}s}
}}

