% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\docType{class}
\name{lgpterm-class}
\alias{lgpterm-class}
\alias{lgpterm}
\title{An S4 class to represent one formula term}
\description{
An S4 class to represent one formula term
}
\section{Slots}{

\describe{
\item{\code{factors}}{a list of at most two \linkS4class{lgpexpr}s}
}}

\seealso{
See \code{\link{operations}} for performing arithmetics
on \linkS4class{lgprhs}, \linkS4class{lgpterm} and \linkS4class{lgpexpr}
objects.
}
