% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-utils.R
\name{list_to_matrix}
\alias{list_to_matrix}
\title{List elements to matrix rows}
\usage{
list_to_matrix(x, n)
}
\arguments{
\item{x}{a listof length \code{m} where each element is a vector of
length code \code{n}}

\item{n}{length of each vector}
}
\value{
a matrix with shape \code{m} x \code{n}
}
\description{
List elements to matrix rows
}
