% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\docType{class}
\name{MarginalGPModel-class}
\alias{MarginalGPModel-class}
\alias{MarginalGPModel}
\alias{LatentGPModel}
\title{An S4 class to represent a marginal GP model}
\arguments{
\item{object}{The object for which to apply a class method.}
}
\description{
An S4 class to represent a marginal GP model

An S4 class to represent a latent GP model
}
\section{Slots}{

\describe{
\item{\code{stan_input}}{Model specific Stan input list.}

\item{\code{stan_input}}{Model specific Stan input list.}
}}

