% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R, R/methods-lgpmodel.R
\docType{class}
\name{MarginalGPModel-class}
\alias{MarginalGPModel-class}
\alias{MarginalGPModel}
\alias{get_stanmodel,MarginalGPModel-method}
\title{An S4 class to represent a marginal GP model}
\usage{
\S4method{get_stanmodel}{MarginalGPModel}(object)
}
\arguments{
\item{object}{The object for which to apply a class method.}
}
\description{
An S4 class to represent a marginal GP model
}
\section{Methods (by generic)}{
\itemize{
\item \code{get_stanmodel}: Get name of corresponding 'Stan' model.
}}

\seealso{
This class inherits from \linkS4class{lgpmodel}, so also the
methods of that class work for \code{MarginalGPModel}s.
}
