% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\docType{class}
\name{MarginalGPModel-class}
\alias{MarginalGPModel-class}
\alias{MarginalGPModel}
\title{An lgpmodel where the signal f is marginalized}
\description{
An lgpmodel where the signal f is marginalized
}
