% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-model_info.R
\name{model_summary}
\alias{model_summary}
\alias{model_summary.brief}
\title{Print a model summary}
\usage{
model_summary(object)

model_summary.brief(object)
}
\arguments{
\item{object}{an object of class \linkS4class{lgpmodel} or
\linkS4class{lgpfit}}
}
\value{
\itemize{
  \item \code{\link{model_summary.brief}} returns a string
  \item \code{\link{model_summary}} prints the summary and
returns \code{object} invisibly
  and returns \code{object} invisibly
}
}
\description{
Print a model summary
}
