% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-lgpmodel.R
\name{model_summary}
\alias{model_summary}
\alias{param_summary}
\title{Print a model summary.}
\usage{
model_summary(object, digits = 3)

param_summary(object, digits = 3)
}
\arguments{
\item{object}{a model or fit}

\item{digits}{number of digits to round floats to}
}
\value{
\code{object} invisibly.
}
\description{
Print a model summary.
}
