% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_array.R
\name{null_if_all_nan}
\alias{null_if_all_nan}
\title{Return NULL if vector contains only NaN values}
\usage{
null_if_all_nan(x)
}
\arguments{
\item{x}{a vector}
}
\value{
\code{x} unchanged or \code{NULL}
}
\description{
Return NULL if vector contains only NaN values
}
\seealso{
Other array utilities: 
\code{\link{add_to_columns}()},
\code{\link{apply_reduce}()},
\code{\link{arr3_select}()},
\code{\link{arr3_to_list}()},
\code{\link{array_to_arraylist}()},
\code{\link{ensure_2dim}()},
\code{\link{make_draw_df}()},
\code{\link{normalize_rows}()},
\code{\link{reduce_rows}()},
\code{\link{repvec}()},
\code{\link{row_vars}()},
\code{\link{select_row}()},
\code{\link{squeeze_second_dim}()}
}
\concept{array utilities}
