% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-data.R
\name{parse_data}
\alias{parse_data}
\title{Parse the covariates and model components from given data and formula}
\usage{
parse_data(data, model_formula, id_variable)
}
\arguments{
\item{data}{A \code{data.frame} where each column corresponds to one
variable, and each row is one observation. Continuous covariates and the
response variable must have type \code{"numeric"} and categorical covariates
must have type \code{"factor"}. Missing values should be indicated with
\code{NaN} or \code{NA}. The response variable cannot contain missing
values.}

\item{model_formula}{An object of class \code{lgpformula}.}

\item{id_variable}{Name of the unique subject identifier variable
(default = \code{"id"}).}
}
\value{
parsed input to stan and covariate scaling
}
\description{
Parse the covariates and model components from given data and formula
}
