% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-formula_advanced.R
\name{parse_expr}
\alias{parse_expr}
\title{Parse a string representation of one formula expression}
\usage{
parse_expr(expr)
}
\arguments{
\item{expr}{the expression as a string, e.g. \code{"gp(x)"}}
}
\value{
an object of class \linkS4class{lgpexpr}
}
\description{
Parse a string representation of one formula expression
}
\seealso{
Other advanced formula parsers: 
\code{\link{parse_formula_advanced}()},
\code{\link{parse_rhs}()},
\code{\link{parse_term}()},
\code{\link{rhs_variables}()},
\code{\link{term_variables}()}
}
\concept{advanced formula parsers}
