% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula.R
\name{parse_formula}
\alias{parse_formula}
\title{Create a model formula}
\usage{
parse_formula(formula)
}
\arguments{
\item{formula}{The model formula, where
\itemize{
  \item it must contain exatly one tilde (\code{~}), with response
  variable on the left-hand side and model terms on the right-hand side
  \item terms are be separated by a plus (\code{+}) sign
  \item all variables appearing in \code{formula} must be
  found in \code{data}
}
See the ´Syntax for formula terms´ section in the documentation of
\code{\link{lgp}} for instructions on how to specify the model terms.}
}
\value{
an object of class \linkS4class{lgpformula}
}
\description{
Create a model formula
}
\seealso{
Other model formula functions: 
\code{\link{formula_to_advanced}()},
\code{\link{is_advanced_formula}()},
\code{\link{parse_expr}()},
\code{\link{parse_rhs}()},
\code{\link{parse_term}()},
\code{\link{parse_whitespace}()},
\code{\link{rhs_to_terms}()},
\code{\link{rhs_variables}()},
\code{\link{term_variables}()}
}
\concept{model formula functions}
