% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula.R
\name{parse_formula}
\alias{parse_formula}
\title{Create a model formula}
\usage{
parse_formula(formula, data)
}
\arguments{
\item{formula}{The model formula, where
\itemize{
  \item it must contain exatly one tilde (\code{~}), with response
  variable on the left-hand side and model terms on the right-hand side
  \item terms are be separated by a plus (\code{+}) sign
  \item all variables appearing in \code{formula} must be
  found in \code{data}
}
See the "Model formula syntax" section below (\code{\link{lgp}}) for
instructions on how to specify the model terms.}

\item{data}{A \code{data.frame} where each column corresponds to one
variable, and each row is one observation. Continuous covariates and the
response variable must have type \code{"numeric"} and categorical covariates
must have type \code{"factor"}. Missing values should be indicated with
\code{NaN} or \code{NA}. The response variable cannot contain missing
values.}
}
\value{
an object of class \linkS4class{lgpformula}
}
\description{
Checks if formula is in advanced format and translates if not.
}
\seealso{
Other model formula functions: 
\code{\link{is_advanced_formula}()},
\code{\link{rhs_to_terms}()}
}
\concept{model formula functions}
