% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{parse_likelihood}
\alias{parse_likelihood}
\title{Parse the given observation model}
\usage{
parse_likelihood(likelihood, c_hat, num_trials, list_y, sample_f)
}
\arguments{
\item{likelihood}{Determines the observation model. Must be either
\code{"gaussian"} (default), \code{"poisson"}, \code{"nb"} (negative
binomial) or \code{"binomial"}. To use Bernoulli likelihood, use
\code{likelihood="binomial"} and set \code{num_trials} as a vector of ones.}

\item{c_hat}{The GP mean. Must be a vector of length \code{dim(data)[1]}, or
a real number defining a constant GP mean. If \code{NULL}, this is set to
 \itemize{
   \item \code{c_hat = 0}, if \code{likelihood} is \code{"gaussian"}, because
   with Gaussian likelihood the response variable is by default centered to
   have zero mean.
   \item \code{c_hat = } \code{log(mean(y))} if \code{likelihood} is
   \code{"poisson"} or \code{"nb"},
   \item \code{c_hat = } \code{log(p/(1-p))}, where
   \code{p = mean(y/num_trials)} if \code{likelihood} is \code{"binomial"},
 }
where \code{y} denotes the response variable. You can modify this vector to
account for normalization between data points. With Gaussian likelihood
though, do not modify this argument, normalize the data beforehand instead.}

\item{num_trials}{This argument (number of trials) is only needed when
likelihood is binomial. Must have length one or equal to number of data
points. Setting \code{num_trials=1} corresponds to Bernoulli observation
model.}

\item{list_y}{a list field returned by \code{\link{parse_response}}}

\item{sample_f}{Determines if the latent function values are be sampled
(must be \code{TRUE} if likelihood is not \code{"gaussian"}).}
}
\value{
a list of parsed options
}
\description{
Parse the given observation model
}
