% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{parse_options}
\alias{parse_options}
\title{Parse the given modeling options}
\usage{
parse_options(options = NULL)
}
\arguments{
\item{options}{A named list with the following possible fields:
\itemize{
  \item \code{skip_generated} If this is true, the generated quantities
  block of Stan is skipped.
  \item \code{delta} Amount of added jitter to ensure positive definite
  covariance matrices.
}}
}
\value{
a named list of parsed options
}
\description{
Parse the given modeling options
}
