% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-misc.R
\name{parse_options}
\alias{parse_options}
\title{Parse the given modeling options}
\usage{
parse_options(options = NULL)
}
\arguments{
\item{options}{A named list with the following possible fields:
\itemize{
  \item \code{sample_f} Determines if the function values are be sampled
  (must be \code{TRUE} if likelihood is not \code{"gaussian"}).
  \item \code{skip_generated} If this is true, the generated quantities
  block of Stan is skipped.
  \item \code{delta} Amount of added jitter to ensure positive definite
  covariance matrices.
  \item \code{verbose} Should more verbose output be printed?
}}
}
\value{
a named list of parsed options
}
\description{
Parse the given modeling options
}
