% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-prior.R
\name{parse_prior}
\alias{parse_prior}
\title{Parse the given prior}
\usage{
parse_prior(prior)
}
\arguments{
\item{prior}{A named list, defining the prior distribution of model
(hyper)parameters. It is recommended to first create this using the function
\code{\link{prior_default}}, and then possibly modify it.}
}
\value{
a named list of parsed options
}
\description{
Parse the given prior
}
