% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors-parse.R
\name{parse_prior_full}
\alias{parse_prior_full}
\title{Parse given fully defined prior}
\usage{
parse_prior_full(prior, stan_input, obs_model)
}
\arguments{
\item{prior}{A named list, defining the prior distribution of model
(hyper)parameters.}

\item{stan_input}{a list of stan input fields}

\item{obs_model}{observation model as integer}
}
\value{
a named list of parsed options
}
\description{
Parse given fully defined prior
}
