% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{parse_prior_teff}
\alias{parse_prior_teff}
\title{Create the teff_ inputs to Stan}
\usage{
parse_prior_teff(desc, info, num_uncrt, num_bt)
}
\arguments{
\item{desc}{Prior description as a named list, containing fields
\itemize{
  \item \code{dist} - Distribution name. Must be one of
  {'uniform', 'normal', 'student-t', 'gamma', 'inv-gamma', 'log-normal'}
  (case-insensitive)
  \item \code{square} - Is the prior for a square-transformed parameter.
}
Other list fields are interpreted as hyperparameters.}

\item{info}{effect time info}

\item{num_uncrt}{the \code{num_uncrt} input created for Stan}

\item{num_bt}{the \code{num_bt} input created for Stan}
}
\value{
a named list
}
\description{
Create the teff_ inputs to Stan
}
