% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula_advanced.R
\name{parse_rhs}
\alias{parse_rhs}
\title{Parse a string representation of the right-hand side of a formula}
\usage{
parse_rhs(rhs)
}
\arguments{
\item{rhs}{the formula right-hand side in text format}
}
\value{
an object of class \linkS4class{lgprhs}
}
\description{
Parse a string representation of the right-hand side of a formula
}
\seealso{
Other advanced formula parsers: 
\code{\link{parse_expr}()},
\code{\link{parse_formula_advanced}()},
\code{\link{parse_term}()},
\code{\link{rhs_variables}()},
\code{\link{term_variables}()}
}
\concept{advanced formula parsers}
