% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-formula.R
\name{parse_rhs}
\alias{parse_rhs}
\title{Parse a string representation of the right-hand side of a formula}
\usage{
parse_rhs(rhs)
}
\arguments{
\item{rhs}{the formula right-hand side, e.g. \code{a + b*c}}
}
\value{
an object of class \linkS4class{lgprhs}
}
\description{
Parse a string representation of the right-hand side of a formula
}
