% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-formula_advanced.R
\name{parse_term}
\alias{parse_term}
\title{Parse a string representation of one formula term}
\usage{
parse_term(term)
}
\arguments{
\item{term}{a term without any plus signs, e.g. \code{b*c}}
}
\value{
an object of class \linkS4class{lgpterm}
}
\description{
Parse a string representation of one formula term
}
\seealso{
Other advanced formula parsers: 
\code{\link{parse_expr}()},
\code{\link{parse_formula_advanced}()},
\code{\link{parse_rhs}()},
\code{\link{rhs_variables}()},
\code{\link{term_variables}()}
}
\concept{advanced formula parsers}
