% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgp-formula.R
\name{parse_term}
\alias{parse_term}
\title{Parse a string representation of one formula term}
\usage{
parse_term(term)
}
\arguments{
\item{term}{a term without any plus signs, e.g. \code{b*c}}
}
\value{
an object of class \code{\link{lgpterm}}
}
\description{
Parse a string representation of one formula term
}
