% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-formula.R
\name{parse_whitespace}
\alias{parse_whitespace}
\title{Remove quotes and whitespace from a formula}
\usage{
parse_whitespace(rhs)
}
\arguments{
\item{rhs}{the formula right-hand side, e.g. \code{a + b*c}}
}
\value{
a character string
}
\description{
Remove quotes and whitespace from a formula
}
