% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-lgpfit.R
\name{plot,lgpfit,missing-method}
\alias{plot,lgpfit,missing-method}
\title{Visualize a model posterior}
\usage{
\S4method{plot}{lgpfit,missing}(x, y, ...)
}
\arguments{
\item{x}{an object of class \linkS4class{lgpfit}}

\item{y}{not used}

\item{...}{keyword arguments passed to \code{\link{plot_posterior}}}
}
\value{
a \code{ggplot} object
}
\description{
Visualize a model posterior
}
