% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-lgpfit.R
\name{plot,lgpfit,missing-method}
\alias{plot,lgpfit,missing-method}
\title{Visualize parameter draws or model fit.}
\usage{
\S4method{plot}{lgpfit,missing}(x, y, plotfun = "plot_draws", ...)
}
\arguments{
\item{x}{an \linkS4class{lgpfit} object to visualize}

\item{y}{not used}

\item{plotfun}{name of plotting function}

\item{...}{optional arguments to \code{plotfun}}
}
\description{
Visualize parameter draws or model fit.
}
