% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4-plot.R
\name{plot}
\alias{plot}
\alias{plot,lgpsim,missing-method}
\title{Visualize S4 class objects using the plot generic}
\usage{
\S4method{plot}{lgpfit,missing}(x, y, ...)

\S4method{plot}{lgpsim,missing}(x, y, ...)
}
\arguments{
\item{x}{object to be plotted}

\item{y}{unused}

\item{...}{keyword arguments to be passed to the signature-specific plotting
function}
}
\value{
All methods return a \code{ggplot} object.
}
\description{
Depending on the signature of the first argument, one
of the following signature-specific plotting functions is called
\itemize{
    \item \code{\link{plot_relevances}} for objects with signature
      \linkS4class{lgpfit}
    \item \code{\link{sim_plot}} for objects with signature
      \linkS4class{lgpsim}
}
}
