% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-components.R
\name{plot_component}
\alias{plot_component}
\title{Helper function for plotting one component}
\usage{
plot_component(MMM, SSS, model, idx, time_is_xvar, linealpha, linetype)
}
\arguments{
\item{MMM}{a n array of size n_samples x n_data x n_components}

\item{SSS}{a n array of size n_samples x n_data x n_components}

\item{model}{an object of class 'lgpmodel'}

\item{idx}{Index of component to be plotted.}

\item{time_is_xvar}{boolean value}

\item{linealpha}{line alpha}

\item{linetype}{line type}
}
\value{
a ggplot object
}
\description{
Helper function for plotting one component
}
