% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-components.R
\name{plot_component}
\alias{plot_component}
\title{Helper function for plotting one component}
\usage{
plot_component(MMM, SSS, model, idx, time_is_xvar, linealpha, linetype,
  fill_alpha, X_test, marker, sum_highlight, viridis_option)
}
\arguments{
\item{MMM}{a n array of size n_samples x n_data x n_components}

\item{SSS}{a n array of size n_samples x n_data x n_components}

\item{model}{an object of class 'lgpmodel'}

\item{idx}{Index of component to be plotted.}

\item{time_is_xvar}{is the time variable the x-axis variable}

\item{linealpha}{line alpha}

\item{linetype}{line type}

\item{fill_alpha}{fill alpha for geom_ribbons}

\item{X_test}{optional matrix of test points}

\item{marker}{point type}

\item{sum_highlight}{name of a categorical covariate to be highlighted}

\item{viridis_option}{the option argument of \code{ggplot2::scale_colour_viridis_c}
by colour in the sum plot}
}
\value{
a ggplot object
}
\description{
Helper function for plotting one component
}
