% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-components.R
\name{plot_component}
\alias{plot_component}
\title{Visualize one (average) inferred component evaluated at data points}
\usage{
plot_component(fit, idx, linealpha = 0.2, linetype = 1,
  subsamples = 100)
}
\arguments{
\item{fit}{An object of class \code{lgpfit}.}

\item{idx}{Index of component to be plotted.}

\item{linealpha}{line alpha}

\item{linetype}{line type}

\item{subsamples}{how many samples to plot}
}
\value{
a ggplot object
}
\description{
Visualize one (average) inferred component evaluated at data points
}
