% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-components.R
\name{plot_components}
\alias{plot_components}
\title{Visualize the (average) inferred components evaluated at data points}
\usage{
plot_components(fit, corrected = TRUE, title = NULL,
  sample_idx = NULL, linealpha = 1, linetype = 3,
  return_list = FALSE, ncol = NULL, nrow = NULL, legend = NULL,
  labels = "auto", ylim = NULL)
}
\arguments{
\item{fit}{An object of class \code{lgpfit}.}

\item{corrected}{Should this plot the covariate-effect corrected components?}

\item{title}{optional prefix to plot title}

\item{sample_idx}{If given, only one sample will be plotted, else the average
components over all samples.}

\item{linealpha}{line alpha}

\item{linetype}{line type}

\item{return_list}{should this return a list of ggplot objects?}

\item{ncol}{number of plot columns}

\item{nrow}{number of plot rows}

\item{legend}{legend argument for ggarrange, use "none" to remove legends}

\item{labels}{labels argument for ggarrange}

\item{ylim}{y axis limits}
}
\value{
alist of ggplot objects or one combined plot
}
\description{
Visualize the (average) inferred components evaluated at data points
}
