% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-spaghetti.R
\name{plot_components}
\alias{plot_components}
\title{Visualize the (average) inferred components evaluated at data points}
\usage{
plot_components(fit, corrected = TRUE, title = NULL,
  sample_idx = NULL, linealpha = 0.6)
}
\arguments{
\item{fit}{An object of class \code{lgpfit}.}

\item{corrected}{Should this plot the covariate-effect corrected components?}

\item{title}{optional prefix to plot title}

\item{sample_idx}{If given, only one sample will be plotted, else the average
components over all samples.}

\item{linealpha}{line alpha}
}
\value{
a ggplot object
}
\description{
Visualize the (average) inferred components evaluated at data points
}
