% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-post_plot.R
\name{plot_components}
\alias{plot_components}
\title{Visualize all model components}
\usage{
plot_components(
  fit,
  x = NULL,
  pred = NULL,
  group_by = "id",
  t_name = "age",
  draws = NULL,
  reduce = mean,
  MULT_STD = 2,
  color_by = NA,
  no_err = FALSE,
  ylim = NULL,
  draw = TRUE,
  nrow = NULL,
  ncol = NULL,
  gg_add = NULL,
  ...
)
}
\arguments{
\item{fit}{An object of class \linkS4class{lgpfit}.}

\item{x}{A data frame of prediction points, containing at least the
variables used as covariates in the model. This must be specified if
\code{pred} is not \code{NULL}.}

\item{pred}{An object of class \linkS4class{FunctionPosterior} or
\linkS4class{FunctionDraws}.}

\item{group_by}{name of the grouping variable (use \code{group_by=NA}
to avoid grouping)}

\item{t_name}{name of the x-axis variable}

\item{draws}{only has effect if \code{pred} is \code{NULL}}

\item{reduce}{only has effect if \code{pred} is \code{NULL}}

\item{MULT_STD}{a multiplier for standard deviation}

\item{color_by}{Names of coloring factors. Can have length 1 or equal to
the number of components. See the \code{color_by} argument of
\code{\link{plot_f}}.}

\item{no_err}{Should the error ribbons be skipped even though they
otherwise would be shown? Can have length 1 or equal to number of
components + 1. See the \code{no_err} argument of \code{\link{plot_api_c}}.}

\item{ylim}{a vector of length 2 (upper and lower y-axis limits), or NULL}

\item{draw}{if this is TRUE, the plot grid is drawn using
\code{\link[gridExtra]{arrangeGrob}}}

\item{nrow}{number of grid rows}

\item{ncol}{number of grid columns}

\item{gg_add}{additional ggplot obejct to add to each plot}

\item{...}{additional arguments to \code{\link{plot_api_c}}}
}
\value{
a list of ggplot objects invisibly
}
\description{
This calls \code{\link{plot_f}} for all model components.
}
\seealso{
Other main plot functions: 
\code{\link{plot_data}()},
\code{\link{plot_pred}()},
\code{\link{plot_sim}()}
}
\concept{main plot functions}
