% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-components.R
\name{plot_components}
\alias{plot_components}
\title{Helper function for plotting components}
\usage{
plot_components(MMM, SSS, model, time_is_xvar, linealpha = 1,
  linetype = 1, ncol = NULL, nrow = NULL, legend = NULL,
  labels = NULL, ylim = NULL, font_size = 9,
  theme = ggplot2::theme_linedraw(), legend_dir = "horizontal",
  xlabel = NULL, ylabel = " ")
}
\arguments{
\item{MMM}{a n array of size n_samples x n_data x n_components}

\item{SSS}{a n array of size n_samples x n_data x n_components}

\item{model}{an object of class 'lgpmodel'}

\item{time_is_xvar}{boolean value}

\item{linealpha}{line alpha}

\item{linetype}{line type}

\item{ncol}{number of plot columns}

\item{nrow}{number of plot rows}

\item{legend}{legend argument for ggarrange, use "none" to remove legends}

\item{labels}{labels argument for ggarrange}

\item{ylim}{y axis limits}

\item{font_size}{font size for plots}

\item{theme}{ggplot theme}

\item{legend_dir}{direction of legend}

\item{xlabel}{x-axis label}

\item{ylabel}{y-axis label}
}
\value{
an object returned by ggpubr::ggarrange
}
\description{
Helper function for plotting components
}
