% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-components.R
\name{plot_components_posterior}
\alias{plot_components_posterior}
\title{Visualize average inferred components or component samples,
evaluated at data points}
\usage{
plot_components_posterior(fit, subsamples = NULL, time_is_xvar = TRUE,
  ...)
}
\arguments{
\item{fit}{An object of class \code{lgpfit}.}

\item{subsamples}{How many samples to plot. If this is NULL,
average over all samples is plotted. If this is "all", all
samples are plotted.}

\item{time_is_xvar}{is age the x-axis variable?}

\item{...}{additional arguments for \code{\link{plot_components}}}
}
\value{
an object returned by ggpubr::ggarrange
}
\description{
Visualize average inferred components or component samples,
evaluated at data points
}
