% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-spaghetti.R
\name{plot_data}
\alias{plot_data}
\title{A spaghetti plot of longitudinal data.}
\usage{
plot_data(data, highlight = NULL, response = "y", id_variable = "id",
  time_variable = "age", psize = 2, lwd = 0.5, title = NULL)
}
\arguments{
\item{data}{A data frame.}

\item{highlight}{Name of a covariate to be highlighted with color, or id of a subject
to be highlighted.}

\item{response}{Name of the response variable.}

\item{id_variable}{Name of id variable.}

\item{time_variable}{Name of time variable.}

\item{psize}{point size}

\item{lwd}{line width}

\item{title}{additional string added to title}
}
\value{
a ggplot object
}
\description{
A spaghetti plot of longitudinal data.
}
