% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-datautil.R
\name{plot_data}
\alias{plot_data}
\title{Vizualizing longitudinal data}
\usage{
plot_data(
  data,
  x_name = "age",
  y_name = "y",
  group_by = "id",
  facet_by = NULL,
  color_by = NULL,
  highlight = NULL,
  main = NULL,
  sub = NULL
)
}
\arguments{
\item{data}{A data frame.}

\item{x_name}{Name of x-axis variable.}

\item{y_name}{Name of the y-axis variable.}

\item{group_by}{Name of grouping variable (must be a factor).}

\item{facet_by}{Name of the faceting variable (must be a factor).}

\item{color_by}{Name of coloring variable (must be a factor).}

\item{highlight}{Value of category of the \code{group_by} variable
that is highlighted. Can only be used if \code{color_by} is \code{NULL}.}

\item{main}{main plot title}

\item{sub}{plot subtitle}
}
\value{
a \code{\link[ggplot2]{ggplot}} object
}
\description{
Vizualizing longitudinal data
}
\concept{data frame handling}
