% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-fit_info.R
\name{plot_draws}
\alias{plot_draws}
\alias{plot_warp}
\alias{plot_beta}
\alias{plot_effect_times}
\title{Visualize the distribution of the obtained parameter draws}
\usage{
plot_draws(
  fit,
  type = "intervals",
  regex_pars = c("alpha", "ell", "wrp", "sigma", "phi", "gamma"),
  ...
)

plot_warp(
  fit,
  num_points = 300,
  window_size = 48,
  color = colorset("red", "dark"),
  alpha = 0.5,
  color_scheme = "brightblue"
)

plot_beta(fit, type = "dens", ...)

plot_effect_times(fit, type = "areas", ...)
}
\arguments{
\item{fit}{an object of class \linkS4class{lgpfit}}

\item{type}{plot type, allowed options are "intervals", "dens",
"areas", and "trace"}

\item{regex_pars}{regex for parameter names to plot}

\item{...}{additional arguments for the \code{bayesplot} function
\code{\link[bayesplot]{mcmc_intervals}}, \code{\link[bayesplot]{mcmc_dens}},
\code{\link[bayesplot]{mcmc_areas}} or \code{\link[bayesplot]{mcmc_trace}}}

\item{num_points}{number of plot points}

\item{window_size}{width of time window}

\item{color}{line color}

\item{alpha}{line alpha}

\item{color_scheme}{deprecated argument, has no effect}
}
\value{
a \code{ggplot} object or list of them

a ggplot object
}
\description{
\itemize{
  \item \code{plot_draws} visualizes the distribution of any set of
  model parameters (defaults to kernel hyperparameters and possible
  observation model parameters)
  \item \code{plot_beta} visualizes the distribution of the
  individual-specific disease effect magnitude parameter draws
  \item \code{plot_effect_times} visualizes the input warping function for
  different parameter draws
  \item \code{plot_warp} visualizes the input warping function for
  different draws of the warping steepness parameter
}
}
\concept{model fit visualization functions}
