% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pred.R
\name{plot_f.get_color_fac}
\alias{plot_f.get_color_fac}
\title{Helper function for plot_f}
\usage{
plot_f.get_color_fac(fit, pred, x, color_by)
}
\arguments{
\item{fit}{An object of class \linkS4class{lgpfit}.}

\item{pred}{An object of class \linkS4class{GaussianPrediction} or
\linkS4class{Prediction}.}

\item{x}{A data frame of prediction points. Must be specified if
\code{pred} is not \code{NULL}.}

\item{color_by}{name of coloring factor}
}
\value{
a factor
}
\description{
Helper function for plot_f
}
