% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fit.R
\name{plot_fit}
\alias{plot_fit}
\alias{plot_fit_component}
\title{Visualize a model fit}
\usage{
plot_fit(fit, group_by = "id", x_name = "age", draws = NULL, ...)

plot_fit_component(
  fit,
  component_idx,
  color_by = NA,
  x_name = "age",
  group_by = "id",
  draws = NULL,
  ...
)
}
\arguments{
\item{fit}{an object of class \linkS4class{lgpfit}}

\item{group_by}{grouping variable name or a factor to be used for grouping
the plot into panels}

\item{x_name}{name of x-axis variable}

\item{draws}{see the \code{draws} argument of \code{\link{get_f}}}

\item{...}{additional keyword arguments to \code{rstan::extract}}

\item{component_idx}{index of the component to be shown}

\item{color_by}{coloring factor}
}
\value{
a \code{\link[ggplot2]{ggplot}} object
}
\description{
\itemize{
  \item In \code{plot_fit}, the data and model fit are plotted for
  each individual separately using \code{\link{plot_api_g}}
  \item In \code{plot_fit_component}, one component of the model fit
  is plotted using \code{\link{plot_api_c}}
}
}
\seealso{
Other main plot functions: 
\code{\link{plot_data}()},
\code{\link{plot_sim}()}
}
\concept{main plot functions}
