% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethodsFit.R
\name{plot_fit}
\alias{plot_fit}
\title{Visualize a model fit against longitudinal data set}
\usage{
plot_fit(fit, x = "age", group_by = "id", draws = NULL, ...)
}
\arguments{
\item{fit}{an object of class \linkS4class{lgpfit}}

\item{x}{x-axis variable name or a vector to be used as x-axis values}

\item{group_by}{grouping variable name or a factor to be used for grouping
the plot into panels}

\item{draws}{see the \code{draws} argument of \code{\link{get_f}}}

\item{...}{keyword arguments to \code{\link{plot_panel}}}
}
\value{
a \code{ggplot object}
}
\description{
Creates plots where each observation unit has a separate panel.
}
\seealso{
Other model fit visualization functions: 
\code{\link{plot,lgpfit,missing-method}},
\code{\link{plot_draws}()}
}
\concept{model fit visualization functions}
