% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethodsFit.R
\name{plot_fit}
\alias{plot_fit}
\title{Visualize a model fit against longitudinal data set}
\usage{
plot_fit(
  fit,
  data,
  x_name = "age",
  y_name = "y",
  group_by = "id",
  draws = NULL,
  ...
)
}
\arguments{
\item{fit}{an object of class \linkS4class{lgpfit}}

\item{data}{a data frame}

\item{x_name}{name of x-axis variable}

\item{y_name}{name of y-axis variable}

\item{group_by}{grouping variable}

\item{draws}{see the \code{draws} argument of \code{\link{get_posterior_f}}}

\item{...}{keyword arguments to \code{\link{plot_panel}}}
}
\value{
a \code{ggplot object}
}
\description{
Creates plots where each observation unit has a separate panel.
}
\seealso{
Other model fit visualization functions: 
\code{\link{plot,lgpfit,missing-method}},
\code{\link{plot_beta}()},
\code{\link{plot_effect_times}()},
\code{\link{plot_posterior_warp}()},
\code{\link{plot_posterior}()}
}
\concept{model fit visualization functions}
