% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_fit_create_df}
\alias{plot_fit_create_df}
\title{Helper function}
\usage{
plot_fit_create_df(df, f)
}
\arguments{
\item{df}{a data frame with group_by factor and x-variable}

\item{f}{an array of size \code{num_draws} x \code{num_obs}}
}
\value{
a data frame
}
\description{
Helper function
}
