% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-plot.R
\name{plot_inputwarp}
\alias{plot_inputwarp}
\title{Visualize input warping function with several steepness parameter values}
\usage{
plot_inputwarp(wrp, x, color = colorset("red", "dark"), alpha = 0.5)
}
\arguments{
\item{wrp}{a vector of values of the warping steepness parameter}

\item{x}{a vector of input values}

\item{color}{line color}

\item{alpha}{line alpha}
}
\value{
a \code{ggplot} object
}
\description{
Visualize input warping function with several steepness parameter values
}
