% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-inv_gamma.R
\name{plot_invgamma}
\alias{plot_invgamma}
\title{Plot pdf of inverse gamma distribution}
\usage{
plot_invgamma(
  alpha,
  beta,
  by = 0.01,
  max = 1.5,
  log = FALSE,
  IQR = 0.95,
  return_quantiles = FALSE
)
}
\arguments{
\item{alpha}{positive real number}

\item{beta}{positive real number}

\item{by}{grid size}

\item{max}{maximum warping steepness to be visualized}

\item{log}{if prior pdf should be on log scale}

\item{IQR}{inter-quantile range width}

\item{return_quantiles}{should this return a list}
}
\value{
a ggplot object or a list
}
\description{
Uses the same parametrization as described
\href{https://mc-stan.org/docs/2_21/functions-reference/inverse-gamma-distribution.html}
{here}.
}
