% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-panel.R
\name{plot_panel}
\alias{plot_panel}
\title{Visuzalize longitudinal data in panels}
\usage{
plot_panel(
  df_data,
  df_signal = NULL,
  df_fit = NULL,
  i_test = NULL,
  linecolors = color_palette(2),
  vlinecolor = colorset("gray", "mid_highlight"),
  nrow = NULL,
  ncol = NULL,
  y_transform = function(x) x,
  true_teff = NULL,
  signal_teff = NULL,
  fit_teff = list(),
  fit_alpha = 0.1
)
}
\arguments{
\item{df_data}{a data frame}

\item{df_signal}{a data frame or \code{NULL}}

\item{df_fit}{a data frame or \code{NULL}}

\item{i_test}{indices of test points}

\item{linecolors}{two line colors (signal and fit)}

\item{vlinecolor}{color of vertical lines (true and obs. effect time)}

\item{nrow}{number of rows, an argument for \code{ggplot2::facet_wrap}}

\item{ncol}{number of columns, an argument for \code{ggplot2::facet_wrap}}

\item{y_transform}{a function}

\item{true_teff}{a named vector}

\item{signal_teff}{a named vector}

\item{fit_teff}{a list of named vectors}

\item{fit_alpha}{line alpha for fit}
}
\value{
a \code{ggplot} object
}
\description{
Creates plots where each observation unit has a separate panel
}
