% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-panel.R
\name{plot_panel}
\alias{plot_panel}
\title{Visuzalize longitudinal data in panels}
\usage{
plot_panel(
  data,
  signal = NULL,
  fit = NULL,
  i_test = NULL,
  linecolors = color_palette(2),
  nrow = NULL,
  ncol = NULL,
  y_transform = function(x) x,
  true_teff = NULL,
  signal_teff = NULL,
  fit_teff = list()
)
}
\arguments{
\item{data}{a data frame}

\item{signal}{a data frame}

\item{fit}{a list of data frames}

\item{i_test}{indices of test points}

\item{linecolors}{three line colors}

\item{nrow}{number of rows, an argument for \code{ggplot2::facet_wrap}}

\item{ncol}{number of columns, an argument for \code{ggplot2::facet_wrap}}

\item{y_transform}{a function}

\item{true_teff}{a named vector}

\item{signal_teff}{a named vector}

\item{fit_teff}{a list of named vectors}
}
\value{
a \code{ggplot} object
}
\description{
Creates plots where each observation unit has a separate panel
}
