% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-posterior-main.R
\name{plot_posterior_components}
\alias{plot_posterior_components}
\title{Visualize inferred components}
\usage{
plot_posterior_components(
  fit,
  subsamples = NULL,
  time_is_xvar = TRUE,
  PRED = NULL,
  marker = NULL,
  sample_idx = 1,
  n_sd = 2,
  ...
)
}
\arguments{
\item{fit}{An object of class \code{lgpfit}.}

\item{subsamples}{How many samples to plot. If this is \code{NULL},
average over all samples is plotted. If this is \code{"all"}, all
samples are plotted.}

\item{time_is_xvar}{is the time variable the x-axis variable
in all subplots?}

\item{PRED}{object returned by \code{\link{lgp_predict}}}

\item{marker}{point type}

\item{sample_idx}{Which sample to plot.}

\item{n_sd}{number of standard deviations (ribbon width)}

\item{...}{additional arguments for \code{\link{plot_components}}}
}
\value{
an object returned by \code{ggpubr::ggarrange} or a
list of ggplot2 objects
}
\description{
Visualize inferred components
}
