% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-predict.R
\name{plot_posterior_f}
\alias{plot_posterior_f}
\title{Plot posterior of f}
\usage{
plot_posterior_f(fit, PRED = NULL, plot_uncertainty = TRUE,
  data_marker = 16, n_sds = 2)
}
\arguments{
\item{fit}{An object of class \code{lgpfit}.}

\item{PRED}{Predictions computed using \code{lgp_predict}.}

\item{plot_uncertainty}{Should an uncertainty ribbon be plotted?}

\item{data_marker}{pch for data points}

\item{n_sds}{number of standard deviations for the uncertainty band width}
}
\value{
a ggplot object
}
\description{
This is a wrapper for \code{\link{plot_posterior_predictions}}.
}
