% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_predict.R
\name{plot_posterior_f}
\alias{plot_posterior_f}
\title{Plot posterior of f}
\usage{
plot_posterior_f(fit, componentwise = FALSE, PRED = NULL)
}
\arguments{
\item{fit}{An object of class \code{lgpfit}.}

\item{componentwise}{A boolean value.}

\item{PRED}{Predictions computed using \code{lgp_predict}.}
}
\value{
a ggplot object
}
\description{
Plot posterior of f
}
