% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-predict.R
\name{plot_posterior_f}
\alias{plot_posterior_f}
\title{Plot posterior of f}
\usage{
plot_posterior_f(fit, PRED = NULL, componentwise = FALSE,
  plot_uncertainty = TRUE, n_sds = n_sds)
}
\arguments{
\item{fit}{An object of class \code{lgpfit}.}

\item{PRED}{Predictions computed using \code{lgp_predict}.}

\item{componentwise}{A boolean value.}

\item{plot_uncertainty}{Should an uncertainty ribbon be plotted?}

\item{n_sds}{number of standard deviations for the uncertainty band width}
}
\value{
a ggplot object
}
\description{
This is a wrapper for \code{\link{plot_posterior_components}}.
and \code{\link{plot_posterior_predictions}}.
}
