% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-predict.R
\name{plot_posterior_predictions}
\alias{plot_posterior_predictions}
\title{Plot posterior of f or predictive distribution for y}
\usage{
plot_posterior_predictions(fit, mode, PRED = NULL,
  color_scheme = "red", color_scheme_onset = "gray", alpha = 0.5,
  alpha_line = 1, alpha2 = 0.5, plot_uncertainty = TRUE,
  title = NULL, original_y_scale = TRUE, ylim = NULL,
  plot_obs_onset = FALSE, plot_onset_samples = FALSE,
  ypos_dens = NULL, test_data = NULL, color_test = "deepskyblue2",
  pch_test = 21, size_test = 2, error_bar = FALSE)
}
\arguments{
\item{fit}{An object of class \code{lgpfit}.}

\item{mode}{Must be either "posterior" or "predictive".}

\item{PRED}{Predictions computed using \code{lgp_predict}.}

\item{color_scheme}{Name of bayesplot color scheme or a list with fieds 'dark' and 'light'.}

\item{color_scheme_onset}{color scheme name for onset density plotting}

\item{alpha}{Ribbon fill opacity.}

\item{alpha_line}{Line opacity.}

\item{alpha2}{alpha of t_onset density}

\item{plot_uncertainty}{Should an uncertainty ribbon be plotted?}

\item{title}{optional prefix to plot title}

\item{original_y_scale}{should the predictions be scaled back to the original data y scale}

\item{ylim}{y axis limits}

\item{plot_obs_onset}{should the observed disease onset be plotted by a vertical line}

\item{plot_onset_samples}{should a distribution of sampled onsets be plotted}

\item{ypos_dens}{y-position of the density plot}

\item{test_data}{Test data frame}

\item{color_test}{test point color}

\item{pch_test}{test point marker}

\item{size_test}{test point size}

\item{error_bar}{should uncertainty be plotted using error bars instead of a ribbon}
}
\value{
a ggplot object
}
\description{
NOTE: currently plotting the onsets assumes that the case individuals come first!
}
