% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_predict.R
\name{plot_posterior_y}
\alias{plot_posterior_y}
\title{Plot posterior predictive distribution}
\usage{
plot_posterior_y(fit, PRED, uncertainty = "ribbon", test_data = NULL)
}
\arguments{
\item{fit}{An object of class \code{lgpfit}.}

\item{PRED}{Predictions computed using \code{lgp_predict}.}

\item{uncertainty}{Either "none", "ribbon" or "errorbar".}

\item{test_data}{Test data set.}
}
\value{
a ggplot object
}
\description{
Plot posterior predictive distribution
}
