% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-plot_pred.R
\name{plot_pred.create}
\alias{plot_pred.create}
\alias{plot_pred.create.df_line}
\alias{plot_pred.create.df_ribbon}
\title{Helper functions for plot_pred}
\usage{
plot_pred.create.df_line(fit, pred, df_base)

plot_pred.create.df_ribbon(fit, pred, df_base, MULT_STD)
}
\arguments{
\item{fit}{An object of class \linkS4class{lgpfit}.}

\item{pred}{An object of class \linkS4class{GaussianPrediction} or
\linkS4class{Prediction}.}

\item{df_base}{a data frame with two columns}

\item{MULT_STD}{a multiplier for standard deviation}
}
\value{
a data frame
}
\description{
Helper functions for plot_pred
}
