% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-posterior_plot.R
\name{plot_pred.create_input}
\alias{plot_pred.create_input}
\title{Helper function for plot_pred and plot_f}
\usage{
plot_pred.create_input(fit, pred, x, draws, reduce, group_by, t_name)
}
\arguments{
\item{fit}{An object of class \linkS4class{lgpfit}.}

\item{pred}{An object of class \linkS4class{FunctionPosterior} or
\linkS4class{FunctionDraws}.}

\item{x}{A data frame of prediction points, containing at least the
variables used as covariates in the model. This must be specified if
\code{pred} is not \code{NULL}.}

\item{draws}{only has effect if \code{pred} is \code{NULL}}

\item{reduce}{only has effect if \code{pred} is \code{NULL}}

\item{group_by}{name of the grouping variable (use \code{group_by=NA}
to avoid grouping)}

\item{t_name}{name of the x-axis variable}
}
\value{
a list with names \code{pred} and \code{df_base}
}
\description{
Helper function for plot_pred and plot_f
}
