% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-predict.R
\name{plot_predictions}
\alias{plot_predictions}
\title{Plot posterior of f or predictive distribution for y}
\usage{
plot_predictions(fit, mode, PRED = NULL, componentwise = FALSE,
  color_scheme = "red", alpha = 0.5 - 0.4 * as.numeric(componentwise),
  plot_uncertainty = TRUE, title = NULL,
  theme = ggplot2::theme_gray(), original_y_scale = TRUE,
  alpha_line = 1, ylim = NULL, plot_obs_onset = FALSE,
  alpha2 = 0.5, color_scheme_onset = "gray",
  plot_onset_samples = FALSE, ypos_dens = NULL, test_data = NULL,
  color_test = "deepskyblue2", pch_test = 21, size_test = 2,
  error_bar = FALSE)
}
\arguments{
\item{fit}{An object of class \code{lgpfit}.}

\item{mode}{Must be either "posterior" or "predictive".}

\item{PRED}{Predictions computed using \code{lgp_predict}.}

\item{componentwise}{Should the predictions be plotted componentwise?}

\item{color_scheme}{Name of bayesplot color scheme or a list with fieds 'dark' and 'light'.}

\item{alpha}{Ribbon fill opacity.}

\item{plot_uncertainty}{Should an uncertainty ribbon be plotted?}

\item{title}{optional prefix to plot title}

\item{theme}{ggplot theme}

\item{original_y_scale}{should the predictions be scaled back to the original data y scale}

\item{alpha_line}{Line opacity}

\item{ylim}{y axis limits}

\item{plot_obs_onset}{should the observed disease onset be plotted by a vertical line}

\item{alpha2}{alpha of t_onset density}

\item{color_scheme_onset}{color scheme name for onset density plotting}

\item{plot_onset_samples}{should a distribution of sampled onsets be plotted}

\item{ypos_dens}{y-position of the density plot}

\item{test_data}{Test data frame}

\item{color_test}{test point color}

\item{pch_test}{test point marker}

\item{size_test}{test point size}

\item{error_bar}{should uncertainty be plotted using error bars instead of a ribbon}
}
\value{
a ggplot object
}
\description{
NOTE: currently plotting the onsets assumes that the case individuals come first!
}
