% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-predict.R
\name{plot_predictions_add_onsets}
\alias{plot_predictions_add_onsets}
\title{Add disease onset / effect times to predictions plot}
\usage{
plot_predictions_add_onsets(
  fit,
  h,
  plot_obs_onset,
  plot_t_effect_samples,
  idvar,
  timevar,
  ypos_dens,
  color_scheme_t_effect,
  reference_times,
  post_t_effect_stat,
  linetypes = c(1, 2, 3),
  linecolors = c("black", "red", "gray50"),
  alpha2 = 1
)
}
\arguments{
\item{fit}{An object of class \code{lgpfit}.}

\item{h}{a ggplot object}

\item{plot_obs_onset}{a boolean value}

\item{plot_t_effect_samples}{a boolean value}

\item{idvar}{id variable name}

\item{timevar}{time variable name}

\item{ypos_dens}{y position of the estimated onset density}

\item{color_scheme_t_effect}{color scheme}

\item{reference_times}{reference onset times}

\item{post_t_effect_stat}{statistic computed from effect time samples}

\item{linetypes}{onset line types}

\item{linecolors}{onset line colors}

\item{alpha2}{alpha parameter}
}
\value{
a modified ggplot object
}
\description{
NOTE: currently assumes that diseased individuals come first.
}
