% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_predict.R
\name{plot_predictions_options}
\alias{plot_predictions_options}
\title{Do input checks and set options for plotting predictions}
\usage{
plot_predictions_options(fit, color_scheme, componentwise,
  original_y_scale, PRED, test_data, color_scheme_onset, mode)
}
\arguments{
\item{fit}{An object of class \code{lgpfit}.}

\item{color_scheme}{Name of bayesplot color scheme.}

\item{componentwise}{Should the predictions be plotted componentwise?}

\item{original_y_scale}{Boolean value.}

\item{PRED}{Predictions computed using \code{lgp_predict}.}

\item{test_data}{test data}

\item{color_scheme_onset}{Another color scheme.}

\item{mode}{mode}
}
\value{
a list
}
\description{
Do input checks and set options for plotting predictions
}
