% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-samples.R
\name{plot_samples}
\alias{plot_samples}
\title{Visualize the distribution of the model parameter samples}
\usage{
plot_samples(
  object,
  pars = character(),
  regex_pars = character(),
  type = "intervals",
  prob = 0.5,
  prob_outer = 0.9,
  color_scheme = "red",
  point_est = "median",
  binwidth = NULL,
  transformations = list(),
  off_diag_args = list(size = 1),
  facet_args = list()
)
}
\arguments{
\item{object}{An object of class \code{lgpfit}.}

\item{pars}{parameter names}

\item{regex_pars}{regex for parameter names}

\item{type}{Visualization type. Must be either \code{"dens"}, \code{"areas"},
\code{"intervals"}(default) or \code{"hist"}.}

\item{prob}{inner interval}

\item{prob_outer}{outer interval}

\item{color_scheme}{See different color schemes in the
\code{bayesplot} package.}

\item{point_est}{the point estimate type}

\item{binwidth}{width of histogram bins if \code{type = "hist"}}

\item{transformations}{the parameter transformations}

\item{off_diag_args}{Additional argument list for the pairs plot.}

\item{facet_args}{additional faceting arguments}
}
\value{
a ggplot object
}
\description{
This is a wrapper for functions in the \code{bayesplot} package.
}
