% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sim.R
\name{plot_sim}
\alias{plot_sim}
\alias{plot_sim_component}
\alias{plot_sim_component_df}
\title{Visualize an lgpsim object (simulated data)}
\usage{
plot_sim(
  simdata,
  group_by = "id",
  x_name = "age",
  f_name = "f",
  y_name = "y",
  ...
)

plot_sim_component(
  simdata,
  component_idx,
  color_by = NA,
  x_name = "age",
  group_by = "id",
  f_name = "f",
  y_name = "y",
  ...
)

plot_sim_component_df(simdata, x_name, group_by, color_by, component_idx)
}
\arguments{
\item{simdata}{an object of class \linkS4class{lgpsim}}

\item{group_by}{grouping factor}

\item{x_name}{name of x-axis variable}

\item{f_name}{name of the signal in \code{simdata$components}}

\item{y_name}{name of response variable}

\item{...}{additional arguments to \code{\link{plot_api_g}} or
\code{\link{plot_api_c}}}

\item{component_idx}{index of the component to be shown}

\item{color_by}{coloring factor}
}
\value{
a \code{\link[ggplot2]{ggplot}} object
}
\description{
\itemize{
  \item In \code{plot_sim}, the data and generating signal are plotted for
  each individual separately using \code{\link{plot_api_g}}
  \item In \code{plot_sim_component}, one component of the generating signal
  is plotted using \code{\link{plot_api_c}}
}
}
\seealso{
Other main plot functions: 
\code{\link{plot_data}()},
\code{\link{plot_pred}()}
}
\concept{main plot functions}
