% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-lgpsim.R
\name{plot_simdata}
\alias{plot_simdata}
\alias{plot,lgpsim,missing-method}
\title{Visualize an lgpsim object}
\usage{
plot_simdata(simdata, ...)

\S4method{plot}{lgpsim,missing}(x, y, ...)
}
\arguments{
\item{simdata}{an object of class \linkS4class{lgpsim}}

\item{...}{other arguments to \code{\link{sim_plot}}}

\item{x}{an object of class \linkS4class{lgpsim}}

\item{y}{not used}
}
\value{
a \code{ggplot} object

a \code{ggplot} object
}
\description{
Visualize an lgpsim object
}
