% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-simdata.R
\name{plot_simdata}
\alias{plot_simdata}
\title{Visualize simulated data}
\usage{
plot_simdata(simData, componentwise = FALSE, nrow = NULL,
  ncol = NULL, i_test = NULL, color_test = "steelblue2",
  y_transform = function(x) {     x })
}
\arguments{
\item{simData}{a list returned by \code{\link{simulate_data}}}

\item{componentwise}{should each component be plotted separately?}

\item{nrow}{an argument for \code{ggplot2::facet_wrap}}

\item{ncol}{an argument for \code{ggplot2::facet_wrap}}

\item{i_test}{test point indices}

\item{color_test}{test point color}

\item{y_transform}{function to transform y}
}
\value{
a ggplot object
}
\description{
This is a wrapper for \code{plot_simdata_by_individual} and
\code{plot_simdata_by_component}
}
