% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-simdata.R
\name{plot_simdata}
\alias{plot_simdata}
\title{Plot a simulated longitudinal data set for each individual separately}
\usage{
plot_simdata(simData, linecolor = "gray70", nrow = NULL, ncol = NULL,
  i_test = NULL, color_test = "steelblue2", y_transform = function(x)
  {     x })
}
\arguments{
\item{simData}{a list returned by \code{\link{simulate_data}}}

\item{linecolor}{line color}

\item{nrow}{an argument for \code{ggplot2::facet_wrap}}

\item{ncol}{an argument for \code{ggplot2::facet_wrap}}

\item{i_test}{test point indices}

\item{color_test}{test point color}

\item{y_transform}{function to transform y}
}
\value{
a ggplot object
}
\description{
Plot a simulated longitudinal data set for each individual separately
}
\seealso{
For plotting each component separately,
 see \code{\link{plot_components_simdata}}
}
