% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-posterior_f.R
\name{posterior_f}
\alias{posterior_f}
\title{Function posterior distributions}
\usage{
posterior_f(
  fit,
  x = NULL,
  reduce = function(x) base::mean(x),
  draws = NULL,
  verbose = TRUE,
  debug_km = FALSE,
  STREAM = get_stream(),
  force = FALSE
)
}
\arguments{
\item{fit}{An object of class \linkS4class{lgpfit}.}

\item{x}{A data frame of points where function posterior distributions
and predictions should be computed or sampled.
The function \code{\link{new_x}} provides an easy way to create it.
If this is NULL, the data points are used.}

\item{reduce}{Reduction for parameters draws. Can be a function that
is applied to reduce all parameter draws into one parameter set, or
NULL (no reduction). Has no effect if \code{draws} is specified.}

\item{draws}{Indices of parameter draws to use, or \code{NULL} to use all
draws.}

\item{verbose}{Should more information and a possible progress bar be
printed?}

\item{debug_km}{Debug kernel matrix computation initialization.}

\item{STREAM}{External pointer. By default obtained with
\code{\link[rstan]{get_stream}}.}

\item{force}{This is by default \code{FALSE} to prevent unintended
large computations that might crash R or take forever.}
}
\value{
A named list.
}
\description{
\itemize{
  \item If \code{fit} is for a model that marginalizes the latent
  signal \code{f} (i.e. \code{is_f_sampled(fit)} is \code{FALSE}), this
  computes the analytic conditional posterior
  distributions of each model component, and their sum.

  \item If \code{fit} is for a model that samples the latent
  signal \code{f} (i.e. \code{is_f_sampled(fit)} is \code{TRUE}), this will
  extract these function samples and compute their sum. If \code{x} is not
  \code{NULL}, the function draws are extrapolated to the points \code{x}
  using kernel regression.
}
}
