% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-posterior_f.R
\name{posterior_f}
\alias{posterior_f}
\title{Evaluate or extract function posterior distributions}
\usage{
posterior_f(
  fit,
  x = NULL,
  c_hat_pred = NULL,
  reduce = function(x) base::mean(x),
  draws = NULL,
  verbose = TRUE,
  debug_km = FALSE,
  debug_dims = FALSE,
  STREAM = get_stream()
)
}
\arguments{
\item{fit}{An object of class \linkS4class{lgpfit}.}

\item{x}{A data frame of points where the predictions are computed.
The function \code{\link{new_x}} can help in creating it.}

\item{c_hat_pred}{This is only used if the latent signal \code{f} was
sampled. This input contains the values added to the sum \code{f} before
passing through inverse link function. Must be a vector with length equal to
the number of prediction points. If original \code{c_hat} was constant,
then \code{c_hat_pred} can be ignored, in which case this will by default
use the same constant.}

\item{reduce}{Reduction for parameters draws. Can be a function that
is applied to reduce all parameter draws into one parameter set, or
\code{NULL} (no reduction). Has no effect if \code{draws} is specified.}

\item{draws}{Indices of parameter draws to use, or \code{NULL} to use all
draws.}

\item{verbose}{Should more some informational messages be printed?}

\item{debug_km}{Should this only return the required kernel matrices.
Can be used for debugging or testing \code{\link{kernels_fpost}}.}

\item{debug_dims}{Should this print dimensions of some variables.
Can be used for debugging \code{\link{kernels_fpost}}.}

\item{STREAM}{an external pointer}
}
\value{
An object of class \linkS4class{FunctionPosterior} or
\linkS4class{FunctionDraws}.
}
\description{
Evaluates (or extracts) the posterior distribution of the
total signal \code{f} and its additive components (or draws from these
distributions). All these are computed for
each parameter draw (defined by \code{draws}), or other parameter set
(obtained by a reduction defined by \code{reduce}).
}
