% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postproc-main.R
\name{postproc}
\alias{postproc}
\title{Finalize the lgpfit object after sampling}
\usage{
postproc(fit, threshold = 0.95, verbose = FALSE)
}
\arguments{
\item{fit}{An (incomplete) object of class \code{lgpfit}.}

\item{threshold}{Covariate selection threshold.}

\item{verbose}{Should some output be printed?}
}
\value{
An updated object of class \code{lgpfit}.
}
\description{
Finalize the lgpfit object after sampling
}
