% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postproc-main.R
\name{postproc}
\alias{postproc}
\title{Finalize the lgpfit object after sampling}
\usage{
postproc(fit, threshold = 0.95, relevance_method = "f_mean", verbose = FALSE)
}
\arguments{
\item{fit}{An (incomplete) object of class \code{lgpfit}.}

\item{threshold}{Threshold for relevance sum.
Must be a value between 0 and 1.}

\item{relevance_method}{Component relevance determination method.
Must be either \code{"f_mean"} or \code{"alpha"}.}

\item{verbose}{Should some output be printed?}
}
\value{
An updated object of class \code{lgpfit}.
}
\description{
Finalize the lgpfit object after sampling
}
