% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethodsFit.R
\name{ppc}
\alias{ppc}
\title{Posterior or prior predictive checks for lgpfit objects}
\usage{
ppc(fit, data, fun = bayesplot::ppc_dens_overlay, ...)
}
\arguments{
\item{fit}{an object of class \linkS4class{lgpfit}}

\item{data}{the original data frame}

\item{fun}{\code{bayesplot} function name}

\item{...}{arguments passed to the default \code{pp_check} method in
\code{bayesplot}}
}
\value{
a \code{ggplot} object
}
\description{
Posterior or prior predictive checks for lgpfit objects
}
