% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethodsFit.R
\name{ppc}
\alias{ppc}
\title{Graphical posterior or prior predictive checks}
\usage{
ppc(fit, data, fun = default_ppc_fun(fit), ...)
}
\arguments{
\item{fit}{an object of class \linkS4class{lgpfit}}

\item{data}{the original data frame}

\item{fun}{\code{bayesplot} function name}

\item{...}{additional arguments passed to the default
\code{\link[bayesplot]{pp_check}} method in
\code{bayesplot}}
}
\value{
a \code{ggplot} object
}
\description{
Graphical posterior or prior predictive checks
}
\seealso{
Introduction to graphical posterior predictive checks:
\href{here}{https://cran.r-project.org/web/packages/bayesplot/vignettes/graphical-ppcs.html}
}
