% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-ppc.R
\name{ppc}
\alias{ppc}
\title{Graphical posterior predictive checks}
\usage{
ppc(fit, data = NULL, fun = default_ppc_fun(fit), verbose = TRUE, ...)
}
\arguments{
\item{fit}{An object of class \linkS4class{lgpfit} that can been created
with \code{sample_f=TRUE}.}

\item{data}{the original data frame (deprecated argument with no
effect, now obtained from fit object)}

\item{fun}{\code{bayesplot} function name}

\item{verbose}{Can this print any messages?}

\item{...}{additional arguments passed to the default
\code{\link[bayesplot]{pp_check}} method in
\code{bayesplot}}
}
\value{
a \code{ggplot} object
}
\description{
Graphical posterior predictive checks
}
\seealso{
Introduction to graphical posterior predictive checks:
\href{https://CRAN.R-project.org/package=bayesplot/vignettes/graphical-ppcs.html}{here}.
Prior predictive check can be done by calling
\code{\link{prior_pred}} and then \code{bayesplot::pp_check()}.
}
