% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-pred.R
\name{pred}
\alias{pred}
\alias{pred.marginal}
\alias{pred.marginal_old}
\alias{pred.latent}
\title{Computing out-of-sample predictions}
\usage{
pred(
  fit,
  x = NULL,
  c_hat_pred = NULL,
  reduce = function(x) base::mean(x),
  draws = NULL,
  verbose = TRUE
)

pred.marginal(fit, x, reduce, draws, verbose)

pred.marginal_old(fit, x, reduce, draws, verbose)

pred.latent(fit, x, c_hat_pred, reduce, draws, verbose, STREAM = get_stream())
}
\arguments{
\item{fit}{An object of class \linkS4class{lgpfit}.}

\item{x}{A data frame of points where the predictions are computed.
The function \code{\link{new_x}} can help in creating it.}

\item{c_hat_pred}{This is only used if the latent signal \code{f} was
sampled. This input contains the values added to the sum \code{f} before
passing through inverse link function. Must be a vector with length equal to
the number of prediction points. If original \code{c_hat} was constant,
then \code{c_hat_pred} can be ignored, in which case this will by default
use the same constant.}

\item{reduce}{Reduction for parameters draws. Can be a function that
is applied to reduce all parameter draws into one parameter set, or
NULL (no reduction). Has no effect if \code{draws} is specified.}

\item{draws}{Indices of parameter draws to use, or \code{NULL} to use all
draws.}

\item{verbose}{Should more information and a progress bar be printed?}

\item{STREAM}{External pointer. By default obtained with
\code{\link[rstan]{get_stream}}.}
}
\value{
An object of class \linkS4class{GaussianPrediction} or
\linkS4class{Prediction}.
}
\description{
\itemize{
  \item \code{pred} calls either \code{pred.marginal} or
  \code{pred.latent} depending on whether \code{fit} is for a model
  that marginalized or sampled the latent signal \code{f}
  \item \code{pred.marginal} computes analytic posterior or prior
  predictive distribution and returns an object of class
  \linkS4class{GaussianPrediction}
  \item \code{pred.latent} computes predictions using kernel regression for
  the samples of \code{f} and returns an object of class
  \linkS4class{Prediction}
}
These also compute the predicted (distribution of the) total signal
\code{f} and its additive components. All these are computed for
each parameter draw (defined by \code{draws}), or other parameter set
(obtained by a reduction defined by \code{reduce}).
}
\seealso{
Other prediction functions: 
\code{\link{kernel_matrices}()},
\code{\link{pred.latent_compute}()},
\code{\link{pred.marginal_compute}()},
\code{\link{pred_kernels}()}
}
\concept{prediction functions}
