% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred.R
\name{pred}
\alias{pred}
\alias{pred.gaussian}
\alias{pred.kr}
\title{Computing out-of-sample predictions}
\usage{
pred(fit, x, reduce = mean, draws = NULL, STREAM = get_stream())

pred.gaussian(fit, x, reduce, draws, STREAM = get_stream())

pred.kr(fit, x, reduce, draws, STREAM = get_stream())
}
\arguments{
\item{fit}{An object of class \linkS4class{lgpfit}.}

\item{x}{A data frame of points where the predictions are computed.
The function \code{\link{new_x}} provides an easy way to create it.}

\item{reduce}{reduction for parameters draws}

\item{draws}{indices of parameter draws}

\item{STREAM}{External pointer. By default obtained with
\code{\link[rstan]{get_stream}}.}
}
\value{
An object of class \linkS4class{GaussianPrediction} or
\linkS4class{Prediction}.
}
\description{
\itemize{
  \item \code{predict} calls either \code{predict.gaussian} or
  \code{predict.kr} depending on whether \code{fit} is for a model
  that marginalized or sampled the latent signal \code{f}
  \item \code{predict.gaussian} computes analytic posterior or prior
  predictive distribution and returns an object of class
  \linkS4class{GaussianPrediction}
  \item \code{predict.kr} computes predictions using kernel regression for
  the samples of \code{f} and returns an object of class
  \linkS4class{Prediction}
}
These also compute the predicted (distribution of the) total signal
\code{f} and its additive components. All these are computed for
each parameter draw (defined by \code{draws}), or other parameter set
(obtained by a reduction defined by \code{reduce}).
}
\seealso{
Other prediction functions: 
\code{\link{kernel_matrices}()},
\code{\link{pred.gaussian_compute}()},
\code{\link{pred.kr_compute}()},
\code{\link{pred_kernels}()}
}
\concept{prediction functions}
