% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred.R
\name{pred}
\alias{pred}
\alias{pred.gaussian}
\alias{pred.kr}
\title{Computing out-of-sample predictions}
\usage{
pred(
  fit,
  x,
  reduce = mean,
  draws = NULL,
  verbose = TRUE,
  STREAM = get_stream()
)

pred.gaussian(fit, x, reduce, draws, verbose, STREAM = get_stream())

pred.kr(fit, x, reduce, draws, verbose, STREAM = get_stream())
}
\arguments{
\item{fit}{An object of class \linkS4class{lgpfit}.}

\item{x}{A data frame of points where the predictions are computed.
The function \code{\link{new_x}} provides an easy way to create it.}

\item{reduce}{Reduction for parameters draws. Can be a function that
is applied to reduce all parameter draws into one parameter set, or
NULL (no reduction). Has no effect if \code{draws} is specified.}

\item{draws}{Indices of parameter draws to use, or \code{NULL} to use all
draws.}

\item{verbose}{Should more information and a progress bar be printed?}

\item{STREAM}{External pointer. By default obtained with
\code{\link[rstan]{get_stream}}.}
}
\value{
An object of class \linkS4class{GaussianPrediction} or
\linkS4class{Prediction}.
}
\description{
\itemize{
  \item \code{pred} calls either \code{pred.gaussian} or
  \code{pred.kr} depending on whether \code{fit} is for a model
  that marginalized or sampled the latent signal \code{f}
  \item \code{pred.gaussian} computes analytic posterior or prior
  predictive distribution and returns an object of class
  \linkS4class{GaussianPrediction}
  \item \code{pred.kr} computes predictions using kernel regression for
  the samples of \code{f} and returns an object of class
  \linkS4class{Prediction}
}
These also compute the predicted (distribution of the) total signal
\code{f} and its additive components. All these are computed for
each parameter draw (defined by \code{draws}), or other parameter set
(obtained by a reduction defined by \code{reduce}).
}
\seealso{
Other prediction functions: 
\code{\link{kernel_matrices}()},
\code{\link{pred.gaussian_compute}()},
\code{\link{pred.kr_compute}()},
\code{\link{pred_kernels}()}
}
\concept{prediction functions}
