% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-pred.R
\name{pred}
\alias{pred}
\alias{pred_marginal}
\alias{pred_latent}
\alias{pred_marginal.create}
\title{Computing model predictions}
\usage{
pred(
  fit,
  x = NULL,
  c_hat_pred = NULL,
  reduce = function(x) base::mean(x),
  draws = NULL,
  verbose = TRUE,
  STREAM = NULL,
  refresh = NULL
)

pred_marginal(fit, x, reduce, draws, refresh)

pred_latent(fit, x, c_hat_pred, reduce, draws, refresh)

pred_marginal.create(fit, x, reduce, draws)
}
\arguments{
\item{fit}{An object of class \linkS4class{lgpfit}.}

\item{x}{A data frame of points where the predictions are computed.
The function \code{\link{new_x}} can help in creating it.}

\item{c_hat_pred}{This is only used if the latent signal \code{f} was
sampled. This input contains the values added to the sum \code{f} before
passing through inverse link function. Must be a vector with length equal to
the number of prediction points. If original \code{c_hat} was constant,
then \code{c_hat_pred} can be ignored, in which case this will by default
use the same constant.}

\item{reduce}{Reduction for parameters draws. Can be a function that
is applied to reduce all parameter draws into one parameter set, or
NULL (no reduction). Has no effect if \code{draws} is specified.}

\item{draws}{Indices of parameter draws to use, or \code{NULL} to use all
draws.}

\item{verbose}{Should more information be printed?}

\item{STREAM}{Deprecated argument, has no effect.}

\item{refresh}{How often to print progress? Has no effect if \code{verbose}
is \code{FALSE}.}
}
\value{
An object of class \linkS4class{GaussianPrediction} or
\linkS4class{Prediction}.
}
\description{
Computes the predicted (distribution of the) total signal
\code{f} and its additive components. All these are computed for
each parameter draw (defined by \code{draws}), or other parameter set
(obtained by a reduction defined by \code{reduce}).
}
\concept{prediction functions}
