% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-pred.R
\name{pred_kernels}
\alias{pred_kernels}
\title{Compute all kernel matrices required for computing predictions}
\usage{
pred_kernels(fit, x, reduce, draws, verbose, STREAM = get_stream())
}
\arguments{
\item{fit}{An object of class \linkS4class{lgpfit}.}

\item{x}{A data frame of points where the predictions are computed.
The function \code{\link{new_x}} can help in creating it.}

\item{reduce}{Reduction for parameters draws. Can be a function that
is applied to reduce all parameter draws into one parameter set, or
NULL (no reduction). Has no effect if \code{draws} is specified.}

\item{draws}{Indices of parameter draws to use, or \code{NULL} to use all
draws.}

\item{verbose}{Should more information and a progress bar be printed?}

\item{STREAM}{External pointer. By default obtained with
\code{\link[rstan]{get_stream}}.}
}
\value{
Returns a list with fields named \code{data_vs_data},
\code{pred_vs_data} and \code{pred_vs_pred}. Each of them is a list with
length equal to the number of parameter sets.
}
\description{
Compute all kernel matrices required for computing predictions
}
\seealso{
Other prediction functions: 
\code{\link{kernel_matrices}()},
\code{\link{pred.latent_compute}()},
\code{\link{pred.marginal_compute}()},
\code{\link{pred}()}
}
\concept{prediction functions}
