% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-lgpr.R, R/methods-Prediction.R
\docType{class}
\name{Prediction-class}
\alias{Prediction-class}
\alias{Prediction}
\alias{show,Prediction-method}
\alias{plot,Prediction,missing-method}
\title{An S4 class to represent draws from a general
predictive distribution}
\usage{
\S4method{show}{Prediction}(object)

\S4method{plot}{Prediction,missing}(x, y, ...)
}
\arguments{
\item{object}{\linkS4class{Prediction} object for which to apply a
class method.}

\item{x}{a \linkS4class{Prediction} object to visualize}

\item{y}{unused argument}

\item{...}{optional arguments passed to a subroutine}
}
\description{
An S4 class to represent draws from a general
predictive distribution
}
\section{Methods (by generic)}{
\itemize{
\item \code{show}: Print a summary about the object.

\item \code{plot}: Visualization.
Optional arguments (\code{...}) are passed to
\code{\link{plot_Prediction}}.
}}

\section{Slots}{

\describe{
\item{\code{f_draws}}{An object of class \linkS4class{FunctionDraws}.}

\item{\code{pred}}{A data frame.}
}}

\seealso{
\linkS4class{GaussianPrediction}
}
