% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{Prediction-class}
\alias{Prediction-class}
\alias{Prediction}
\title{An S4 class to represent general additive model predictions}
\description{
An S4 class to represent general additive model predictions
}
\section{Slots}{

\describe{
\item{\code{f_comp}}{component predictions}

\item{\code{f}}{signal prediction}

\item{\code{h}}{prediction (signal prediction transformed through inverse link
function)}
}}

\seealso{
\linkS4class{GaussianPrediction}
}
