% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-lgpmodel.R
\name{prior_summary}
\alias{prior_summary}
\title{Prior summary}
\usage{
prior_summary(object, digits = 3)
}
\arguments{
\item{object}{an object of class \linkS4class{lgpmodel} or
\linkS4class{lgpfit}}

\item{digits}{number of digits to show for floating point numbers}
}
\value{
a data frame
}
\description{
Prior summary
}
