% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-create_model-prior.R
\name{prior_to_num}
\alias{prior_to_num}
\title{Convert given prior to numeric format}
\usage{
prior_to_num(desc)
}
\arguments{
\item{desc}{Prior description as a named list, containing fields
\itemize{
  \item \code{dist} - Distribution name. Must be one of
  {'uniform', 'normal', 'student-t', 'gamma', 'inv-gamma', 'log-normal'}
  (case-insensitive)
  \item \code{square} - Is the prior for a square-transformed parameter.
}
Other list fields are interpreted as hyperparameters.}
}
\value{
a named list of parsed options
}
\description{
Convert given prior to numeric format
}
