% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors-main.R
\name{priors}
\alias{priors}
\alias{normal,}
\alias{log_normal,}
\alias{gamma,}
\alias{inv_gamma,}
\alias{uniform,}
\alias{student_t}
\alias{uniform}
\alias{normal}
\alias{gamma}
\alias{inv_gamma}
\alias{log_normal}
\alias{beta}
\title{Prior definitions}
\usage{
uniform(square = FALSE)

normal(mu, sigma, square = FALSE)

student_t(nu, square = FALSE)

gamma(shape, inv_scale, square = FALSE)

inv_gamma(shape, scale, square = FALSE)

log_normal(mu, sigma, square = FALSE)

beta(a, b)
}
\arguments{
\item{square}{is prior for a square-transformed parameter?}

\item{mu}{mean}

\item{sigma}{standard deviation}

\item{nu}{degrees of freedom}

\item{shape}{shape parameter (alpha)}

\item{inv_scale}{inverse scale parameter (beta)}

\item{scale}{scale parameter (beta)}

\item{a}{shape parameter}

\item{b}{shape parameter}
}
\value{
a named list
}
\description{
These use the same parametrizations as defined in the Stan
documentation. See the docs for
\href{https://mc-stan.org/docs/2_24/functions-reference/
gamma-distribution.html}{gamma} and
\href{https://mc-stan.org/docs/2_24/functions-reference/
inverse-gamma-distribution.html}{inverse gamma} distributions.
}
\examples{
# Log-normal prior
log_normal(mu = 1, sigma = 1)

# Cauchy prior
student_t(nu = 1)

# Exponential prior with rate = 0.1
gamma(shape = 1, inv_scale = 0.1)
}
