% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-utils_array.R
\name{reduce_rows}
\alias{reduce_rows}
\title{Check that each row of array is identical}
\usage{
reduce_rows(rows)
}
\arguments{
\item{rows}{rows}
}
\value{
the indentical row
}
\description{
Check that each row of array is identical
}
\seealso{
Other array utilities: 
\code{\link{add_to_columns}()},
\code{\link{apply_reduce}()},
\code{\link{arr3_select}()},
\code{\link{arr3_to_list}()},
\code{\link{array_to_arraylist}()},
\code{\link{ensure_2dim}()},
\code{\link{make_draw_df}()},
\code{\link{normalize_rows}()},
\code{\link{null_if_all_nan}()},
\code{\link{repvec}()},
\code{\link{row_vars}()},
\code{\link{select_row}()},
\code{\link{squeeze_second_dim}()}
}
\concept{array utilities}
