% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-relevances.R
\name{relevances}
\alias{relevances}
\alias{relevances.default}
\alias{relevances.default.all}
\title{Assess component relevances}
\usage{
relevances(fit, reduce = function(x) base::mean(x), ...)

relevances.default(fit, reduce, ...)

relevances.default.all(fit)
}
\arguments{
\item{fit}{an object of class \code{lgpfit}}

\item{reduce}{a function to apply to reduce the relevances in each
draw into one value}

\item{...}{currently has no effect}
}
\description{
\itemize{
  \item \code{relevances} returns a named vector with length equal to
  \code{num_comps + 1}
  \item \code{relevances.default} is the default method
  \item \code{relevances.default_all} returns a \code{data.frame} of size
  \code{num_draws} x \code{num_comps + 1}
}
}
