% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relevances.R
\name{relevances}
\alias{relevances}
\alias{relevances.default}
\title{Assess component relevances}
\usage{
relevances(fit)

relevances.default(fit)
}
\arguments{
\item{fit}{an object of class \code{lgpfit}}
}
\description{
\itemize{
  \item \code{relevances} assesses covariate relevances
  \item \code{relevances.default} is the only currently implemented method
}
}
