% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-relevances.R
\name{relevances}
\alias{relevances}
\title{Assess component relevances}
\usage{
relevances(fit, reduce = function(x) base::mean(x), verbose = TRUE, ...)
}
\arguments{
\item{fit}{an object of class \code{lgpfit}}

\item{reduce}{a function to apply to reduce the relevances given each
parameter draw into one value}

\item{verbose}{Can this print any messages?}

\item{...}{currently has no effect}
}
\value{
a named vector with length equal to \code{num_comps + 1}
}
\description{
Assess component relevances
}
