% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-utils.R
\name{rep_df}
\alias{rep_df}
\title{Repeat a data frame vertically}
\usage{
rep_df(df, times)
}
\arguments{
\item{df}{a data frame}

\item{times}{number of times to repeat}
}
\value{
a data frame
}
\description{
Repeat a data frame vertically
}
\seealso{
Other data utilities: 
\code{\link{add_dis_age}()},
\code{\link{add_factor_crossing}()},
\code{\link{add_factor}()},
\code{\link{data_types}()},
\code{\link{get_teff_obs}()},
\code{\link{new_x}()},
\code{\link{pick_one_row_each}()},
\code{\link{select_factors_and}()},
\code{\link{split}()},
\code{\link{to_long_format}()}
}
\concept{data utilities}
