% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-misc.R
\name{repvec}
\alias{repvec}
\title{Repeat a vector as a rows of an array}
\usage{
repvec(v, n)
}
\arguments{
\item{v}{a vector of length \code{m}}

\item{n}{number of times to repeat}
}
\value{
returns an array of size \code{n} x \code{m}
}
\description{
Throws an error if \code{v} is \code{NULL}.
}
\seealso{
Other array utilities: 
\code{\link{apply_reduce}()},
\code{\link{array_to_arraylist}()},
\code{\link{ensure_2dim}()},
\code{\link{squeeze_second_dim}()}
}
\concept{array utilities}
