% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula.R
\name{rhs_to_terms}
\alias{rhs_to_terms}
\title{Split a formula to terms}
\usage{
rhs_to_terms(rhs)
}
\arguments{
\item{rhs}{the formula right-hand side in text format}
}
\value{
a character vector
}
\description{
Split a formula to terms
}
\seealso{
Other model formula functions: 
\code{\link{formula_to_advanced}()},
\code{\link{is_advanced_formula}()},
\code{\link{parse_formula}()}
}
\concept{model formula functions}
