% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgp-formula.R
\name{rhs_variables}
\alias{rhs_variables}
\title{Get names of all variables appearing on formula right-hand side}
\usage{
rhs_variables(rhs)
}
\arguments{
\item{rhs}{an object of class \code{lgprhs}}
}
\value{
a list of variable names
}
\description{
Get names of all variables appearing on formula right-hand side
}
