% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula_advanced.R
\name{rhs_variables}
\alias{rhs_variables}
\title{Get names of all variables appearing on formula right-hand side}
\usage{
rhs_variables(rhs)
}
\arguments{
\item{rhs}{an object of class \linkS4class{lgprhs}}
}
\value{
a list of variable names
}
\description{
Get names of all variables appearing on formula right-hand side
}
\seealso{
Other advanced formula parsers: 
\code{\link{parse_expr}()},
\code{\link{parse_formula_advanced}()},
\code{\link{parse_rhs}()},
\code{\link{parse_term}()},
\code{\link{term_variables}()}
}
\concept{advanced formula parsers}
