% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-plot.R
\name{scale_color}
\alias{scale_color}
\alias{scale_fill}
\title{A color scale.}
\usage{
scale_color(n)

scale_fill(n)
}
\arguments{
\item{n}{an integer from 1 to 5}
}
\value{
a \code{ggplot} object (if \code{n <= 5}) or NULL (if \code{n > 5})
}
\description{
A color scale.
}
\seealso{
Other color utilities: 
\code{\link{color_palette}()},
\code{\link{colorset}()},
\code{\link{plot_color_palette}()}
}
\concept{color utilities}
