% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethodsFit.R
\name{scale_f_total}
\alias{scale_f_total}
\title{Scale the function f posterior to original unnormalized scale}
\usage{
scale_f_total(fit, f_total)
}
\arguments{
\item{fit}{an object of class \linkS4class{lgpfit}}

\item{f_total}{a list with fields \code{mean} and \code{std}}
}
\value{
a similar object as \code{f_total}
}
\description{
Can only be used with Gaussian observation model.
}
\seealso{
Other fit postprocessing functions: 
\code{\link{fit_summary}()},
\code{\link{get_draws}()},
\code{\link{get_f}()}
}
\concept{fit postprocessing functions}
