% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{select}
\alias{select}
\title{Select relevant components}
\usage{
select(fit, reduce = mean, threshold = 0.95, ...)
}
\arguments{
\item{fit}{An object of class \code{lgpfit}.}

\item{reduce}{The \code{reduce} argument for \code{\link{relevances}}.}

\item{threshold}{Threshold for relevance sum.
Must be a value between 0 and 1.}

\item{...}{Additional arguments to \code{\link{relevances}}.}
}
\value{
A binary array of shape \code{S} x \code{num_comps+1},
where 1 denotes selected components and 0 not selected. \code{S} is
equal to \code{num_draws} if \code{reduce} is NULL and otherwise one.
}
\description{
Select relevant components
}
