% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postproc-main.R
\name{selection_helper}
\alias{selection_helper}
\title{Select relevant components (one MCMC sample)}
\usage{
selection_helper(rel, threshold)
}
\arguments{
\item{rel}{a named vector of component relevances}

\item{threshold}{A threshold for proportion of explained variance}
}
\value{
names of selected components (including "noise" always)
}
\description{
Select relevant components (one MCMC sample)
}
