% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postproc-main.R
\name{selection_prob}
\alias{selection_prob}
\title{Probabilistic selection of relevant components}
\usage{
selection_prob(object, threshold = 0.95)
}
\arguments{
\item{object}{An object of class \code{lgpfit}.}

\item{threshold}{A threshold for proportion of explained variance}
}
\value{
the selected covariates
}
\description{
Probabilistic selection of relevant components
}
