% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postproc-selection.R
\name{selection_prob_fixed_threshold}
\alias{selection_prob_fixed_threshold}
\title{Selection probabilities using a fixed threshold}
\usage{
selection_prob_fixed_threshold(relevances, threshold)
}
\arguments{
\item{relevances}{The \code{relevances$samples} slot of an
\code{lgpfit} object.}

\item{threshold}{value between 0 and 1}
}
\value{
proportion of times each component was selected
}
\description{
Selection probabilities using a fixed threshold
}
