% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-likelihood.R
\name{set_c_hat}
\alias{set_c_hat}
\title{Set c_hat (non-gaussian observation models)}
\usage{
set_c_hat(c_hat, response, LH, num_trials)
}
\arguments{
\item{c_hat}{the \code{c_hat} argument given as input to
\code{\link{lgp_model}}}

\item{response}{response variable}

\item{LH}{likelihood as int}

\item{num_trials}{the num_trials data (binomial likelihood)}
}
\value{
a real number
}
\description{
Set c_hat (non-gaussian observation models)
}
