% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-parse_y.R
\name{set_num_trials}
\alias{set_num_trials}
\title{Set num_trials (binomial and beta-binomial observation models)}
\usage{
set_num_trials(num_trials, y, LH)
}
\arguments{
\item{num_trials}{the \code{num_trials} argument}

\item{y}{response variable observations}

\item{LH}{likelihood as integer encoding}
}
\value{
a numeric vector
}
\description{
Set num_trials (binomial and beta-binomial observation models)
}
