% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{set_num_trials}
\alias{set_num_trials}
\title{Set num_trials (binomial and Bernoulli observation models)}
\usage{
set_num_trials(num_trials, num_obs, LH)
}
\arguments{
\item{num_trials}{the \code{num_trials} argument}

\item{num_obs}{number of observations}

\item{LH}{likelihood as integer encoding}
}
\value{
a numeric vector
}
\description{
Set num_trials (binomial and Bernoulli observation models)
}
