% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-misc.R
\name{show}
\alias{show}
\alias{show,lgpformula-method}
\alias{show,lgprhs-method}
\alias{show,lgpterm-method}
\alias{show,lgpsim-method}
\alias{show,FunctionPosterior-method}
\alias{show,FunctionDraws-method}
\title{Printing S4 object info using the show generic}
\usage{
\S4method{show}{lgpformula}(object)

\S4method{show}{lgprhs}(object)

\S4method{show}{lgpterm}(object)

\S4method{show}{lgpsim}(object)

\S4method{show}{FunctionPosterior}(object)

\S4method{show}{FunctionDraws}(object)
}
\arguments{
\item{object}{an object of some S4 class}
}
\value{
the object invisibly
}
\description{
Printing S4 object info using the show generic
}
