% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-lgpformula.R
\name{show}
\alias{show}
\alias{show,lgpformula-method}
\alias{show,lgprhs-method}
\alias{show,lgpterm-method}
\title{Printing formula object info using the show generic}
\usage{
\S4method{show}{lgpformula}(object)

\S4method{show}{lgprhs}(object)

\S4method{show}{lgpterm}(object)
}
\arguments{
\item{object}{an object of some S4 class}
}
\value{
the object invisibly
}
\description{
Printing formula object info using the show generic
}
