% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate-main.R
\name{sim_data_to_observed}
\alias{sim_data_to_observed}
\title{Real generated disease ages to observed ones}
\usage{
sim_data_to_observed(dat, t_observed)
}
\arguments{
\item{dat}{data frame}

\item{t_observed}{Determines how the disease onset is observed. Must have the form \code{"after_n"}
or \code{"random_p"}}
}
\value{
a new data frame and observed onsets
}
\description{
Real generated disease ages to observed ones
}
