% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-plot.R
\name{sim_plot}
\alias{sim_plot}
\title{Plot a simulated longitudinal data set for each individual separately}
\usage{
sim_plot(
  simdata,
  nrow = NULL,
  ncol = NULL,
  i_test = NULL,
  linecolor = colorset("gray"),
  dotcolor = "black",
  testcolor = colorset("brightblue", "dark"),
  signal_name = "signal",
  y_transform = function(x) {     x }
)
}
\arguments{
\item{simdata}{an object of class \linkS4class{lgpsim}}

\item{nrow}{number of rows, an argument for \code{ggplot2::facet_wrap}}

\item{ncol}{number of columns, an argument for \code{ggplot2::facet_wrap}}

\item{i_test}{indices of possible test points}

\item{linecolor}{line color for signal}

\item{dotcolor}{dot color for data points}

\item{testcolor}{dot color for possible test points}

\item{signal_name}{name of signal in the legend}

\item{y_transform}{function to transform the data y}
}
\value{
a ggplot object
}
\description{
Plot a simulated longitudinal data set for each individual separately
}
\seealso{
For plotting each component separately,
 see \code{\link{sim_plot_components}}
}
