% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split.R
\name{split_data}
\alias{split_data}
\title{Split data into training and test data according to given row indices}
\usage{
split_data(data, i_test, sort_ids = TRUE)
}
\arguments{
\item{data}{a data frame}

\item{i_test}{test data row indices}

\item{sort_ids}{should the test indices be sorted into increasing order}
}
\value{
a \code{list(train, test)}
}
\description{
Split data into training and test data according to given row indices
}
