% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-split.R
\name{split_data_random}
\alias{split_data_random}
\title{Split data into training and test data randomly}
\usage{
split_data_random(data, p_test = 0.1, n_test = NULL)
}
\arguments{
\item{data}{a data frame}

\item{p_test}{desired proportion of test data}

\item{n_test}{desired number of test data points (if NULL, p_test is used
to compute this)}
}
\value{
a \code{list(train, test)}
}
\description{
Split data into training and test data randomly
}
