% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{squeeze_second_dim}
\alias{squeeze_second_dim}
\title{Squeeze the second dimension of an array}
\usage{
squeeze_second_dim(x)
}
\arguments{
\item{x}{an array of shape \code{n1} x \code{n2} x \code{n3}}
}
\value{
an array of shape \code{n1*n2} x \code{n3}
}
\description{
Squeeze the second dimension of an array
}
\seealso{
Other array utilities: 
\code{\link{array_to_arraylist}()},
\code{\link{ensure_2dim}()},
\code{\link{null_if_all_nan}()},
\code{\link{reduce_rows}()},
\code{\link{repvec}()}
}
\concept{array utilities}
