% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-data.R
\name{stan_data_components}
\alias{stan_data_components}
\title{Create model component data for Stan input}
\usage{
stan_data_components(model_formula, covariates)
}
\arguments{
\item{model_formula}{an object of class \linkS4class{lgpformula}}

\item{covariates}{a list returned by \code{\link{stan_data_covariates}}}
}
\value{
a named list with fields
\itemize{
  \item \code{to_stan}: a list of stan data
  \item TODO: something?
}
}
\description{
Create model component data for Stan input
}
