% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula.R
\name{term_variables}
\alias{term_variables}
\title{Get names of all variables appearing in a term}
\usage{
term_variables(term)
}
\arguments{
\item{term}{an object of class \linkS4class{lgpterm}}
}
\value{
a list of variable names
}
\description{
Get names of all variables appearing in a term
}
