% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_data.R
\name{valid_data}
\alias{valid_data}
\alias{validate_data}
\alias{validate_factor}
\title{Validate data array}
\usage{
validate_data(data, id_name = "id")

validate_factor(x, id, name)
}
\arguments{
\item{data}{a data frame}

\item{id_name}{name of the id variable}

\item{x}{a factor}

\item{id}{the id factor}

\item{name}{factor name}
}
\value{
a data frame
}
\description{
Validate data array
}
\seealso{
Other data utilities: 
\code{\link{add_dis_age}()},
\code{\link{add_factor}()},
\code{\link{get_teff_obs}()},
\code{\link{new_x}()},
\code{\link{pick_one_row_each}()},
\code{\link{select_factors_and}()},
\code{\link{split}()}
}
\concept{data utilities}
