% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-all_classes.R
\name{validate}
\alias{validate}
\alias{check_lgpexpr}
\alias{check_lgpformula}
\alias{check_lgpscaling}
\alias{check_GaussianPrediction}
\alias{check_Prediction}
\title{Validate S4 class objects}
\usage{
check_lgpexpr(object)

check_lgpformula(object)

check_lgpscaling(object)

check_GaussianPrediction(object)

check_Prediction(object)
}
\arguments{
\item{object}{an object to validate}
}
\value{
\code{TRUE} if valid, otherwise reasons for invalidity
}
\description{
Validate S4 class objects
}
