% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{validate}
\alias{validate}
\alias{validate_lgpexpr}
\alias{validate_lgpformula}
\alias{validate_lgpscaling}
\alias{validate_lgpfit}
\alias{validate_GaussianPrediction}
\alias{validate_Prediction}
\title{Validate S4 class objects}
\usage{
validate_lgpexpr(object)

validate_lgpformula(object)

validate_lgpscaling(object)

validate_lgpfit(object)

validate_GaussianPrediction(object)

validate_Prediction(object)
}
\arguments{
\item{object}{an object to validate}
}
\value{
\code{TRUE} if valid, otherwise reasons for invalidity
}
\description{
Validate S4 class objects
}
