% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-kernels.R
\name{var_mask}
\alias{var_mask}
\title{Variance masking function}
\usage{
var_mask(x, a)
}
\arguments{
\item{x}{vector of length \code{n}}

\item{a}{a positive real number}
}
\value{
a vector of length \code{n}
}
\description{
Variance masking function
}
