% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-kernel_reference.R
\name{var_mask}
\alias{var_mask}
\title{Variance masking function}
\usage{
var_mask(x, stp)
}
\arguments{
\item{x}{a vector of length \eqn{n}}

\item{stp}{a positive real number (steepness of mask function)}
}
\value{
a vector of length \eqn{n}
}
\description{
Variance masking function
}
\seealso{
Other kernel utility functions: 
\code{\link{warp_input}()}
}
\concept{kernel utility functions}
