% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-kernels.R
\name{warp_input}
\alias{warp_input}
\title{Input warping function}
\usage{
warp_input(t, a, b, c)
}
\arguments{
\item{t}{a vector}

\item{a}{steepness of the rise}

\item{b}{location of the effective time window}

\item{c}{maximum range}
}
\value{
a vector of warped inputs \code{w(t)}
}
\description{
Input warping function
}
