% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-kernel_reference.R
\name{warp_input}
\alias{warp_input}
\title{Input warping function}
\usage{
warp_input(x, a)
}
\arguments{
\item{x}{a vector of length \eqn{n}}

\item{a}{steepness of the warping function rise}
}
\value{
a vector of warped inputs \eqn{w(x)}, length \eqn{n}
}
\description{
Input warping function
}
\seealso{
Other kernel utility functions: 
\code{\link{var_mask}()}
}
\concept{kernel utility functions}
