% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{zip_lists}
\alias{zip_lists}
\title{Zip two lists into a list of lists of length 2}
\usage{
zip_lists(a, b)
}
\arguments{
\item{a}{a list of length \code{L}}

\item{b}{a list of length \code{L}}
}
\value{
a list of length \code{L}
}
\description{
Zip two lists into a list of lists of length 2
}
\seealso{
Other list utilities: 
\code{\link{dollar}()},
\code{\link{list_to_matrix}()},
\code{\link{matrix_to_list}()},
\code{\link{print_list}()}
}
\concept{list utilities}
